% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AdmissionregistrationApi.R
\name{AdmissionregistrationApi}
\alias{AdmissionregistrationApi}
\title{AdmissionregistrationApi}
\description{
AdmissionregistrationApi

AdmissionregistrationApi
}
\details{
Typed Kubernetes API client for the `Admissionregistration` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdmissionregistrationApi-new}{\code{AdmissionregistrationApi$new()}}
\item \href{#method-AdmissionregistrationApi-get_api_group}{\code{AdmissionregistrationApi$get_api_group()}}
\item \href{#method-AdmissionregistrationApi-clone}{\code{AdmissionregistrationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAdmissionregistrationAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
