% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AdmissionregistrationV1alpha1Api.R
\name{AdmissionregistrationV1alpha1Api}
\alias{AdmissionregistrationV1alpha1Api}
\title{AdmissionregistrationV1alpha1Api}
\description{
AdmissionregistrationV1alpha1Api

AdmissionregistrationV1alpha1Api
}
\details{
Typed Kubernetes API client for the `AdmissionregistrationV1alpha1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdmissionregistrationV1alpha1Api-new}{\code{AdmissionregistrationV1alpha1Api$new()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-get_api_resources}{\code{AdmissionregistrationV1alpha1Api$get_api_resources()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-list_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$list_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-create_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$create_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-delete_collection_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$delete_collection_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$read_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$replace_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$patch_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-delete_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$delete_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy_status}{\code{AdmissionregistrationV1alpha1Api$read_validating_admission_policy_status()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy_status}{\code{AdmissionregistrationV1alpha1Api$replace_validating_admission_policy_status()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy_status}{\code{AdmissionregistrationV1alpha1Api$patch_validating_admission_policy_status()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-list_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$list_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-create_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$create_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-delete_collection_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$delete_collection_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$read_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$replace_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$patch_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-delete_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$delete_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_list}{\code{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy_list()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy}{\code{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_binding_list}{\code{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy_binding_list()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_binding}{\code{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy_binding()}}
\item \href{#method-AdmissionregistrationV1alpha1Api-clone}{\code{AdmissionregistrationV1alpha1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAdmissionregistrationV1alpha1APIResources
get available resources
listAdmissionregistrationV1alpha1ValidatingAdmissionPolicy
list or watch objects of kind ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-list_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-list_validating_admission_policy}{}}}
\subsection{Method \code{list_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$list_validating_admission_policy(
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAdmissionregistrationV1alpha1ValidatingAdmissionPolicy
create a ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-create_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-create_validating_admission_policy}{}}}
\subsection{Method \code{create_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$create_validating_admission_policy(
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAdmissionregistrationV1alpha1CollectionValidatingAdmissionPolicy
delete collection of ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-delete_collection_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-delete_collection_validating_admission_policy}{}}}
\subsection{Method \code{delete_collection_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$delete_collection_validating_admission_policy(
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAdmissionregistrationV1alpha1ValidatingAdmissionPolicy
read the specified ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy}{}}}
\subsection{Method \code{read_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$read_validating_admission_policy(
  name,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAdmissionregistrationV1alpha1ValidatingAdmissionPolicy
replace the specified ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy}{}}}
\subsection{Method \code{replace_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$replace_validating_admission_policy(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAdmissionregistrationV1alpha1ValidatingAdmissionPolicy
partially update the specified ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy}{}}}
\subsection{Method \code{patch_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$patch_validating_admission_policy(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAdmissionregistrationV1alpha1ValidatingAdmissionPolicy
delete a ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-delete_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-delete_validating_admission_policy}{}}}
\subsection{Method \code{delete_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$delete_validating_admission_policy(
  name,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readAdmissionregistrationV1alpha1ValidatingAdmissionPolicyStatus
read status of the specified ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy_status"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy_status}{}}}
\subsection{Method \code{read_validating_admission_policy_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$read_validating_admission_policy_status(
  name,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAdmissionregistrationV1alpha1ValidatingAdmissionPolicyStatus
replace status of the specified ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy_status"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy_status}{}}}
\subsection{Method \code{replace_validating_admission_policy_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$replace_validating_admission_policy_status(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAdmissionregistrationV1alpha1ValidatingAdmissionPolicyStatus
partially update status of the specified ValidatingAdmissionPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy_status"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy_status}{}}}
\subsection{Method \code{patch_validating_admission_policy_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$patch_validating_admission_policy_status(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBinding
list or watch objects of kind ValidatingAdmissionPolicyBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-list_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-list_validating_admission_policy_binding}{}}}
\subsection{Method \code{list_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$list_validating_admission_policy_binding(
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBinding
create a ValidatingAdmissionPolicyBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-create_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-create_validating_admission_policy_binding}{}}}
\subsection{Method \code{create_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$create_validating_admission_policy_binding(
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAdmissionregistrationV1alpha1CollectionValidatingAdmissionPolicyBinding
delete collection of ValidatingAdmissionPolicyBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-delete_collection_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-delete_collection_validating_admission_policy_binding}{}}}
\subsection{Method \code{delete_collection_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$delete_collection_validating_admission_policy_binding(
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBinding
read the specified ValidatingAdmissionPolicyBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-read_validating_admission_policy_binding}{}}}
\subsection{Method \code{read_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$read_validating_admission_policy_binding(
  name,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBinding
replace the specified ValidatingAdmissionPolicyBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-replace_validating_admission_policy_binding}{}}}
\subsection{Method \code{replace_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$replace_validating_admission_policy_binding(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBinding
partially update the specified ValidatingAdmissionPolicyBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-patch_validating_admission_policy_binding}{}}}
\subsection{Method \code{patch_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$patch_validating_admission_policy_binding(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBinding
delete a ValidatingAdmissionPolicyBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-delete_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-delete_validating_admission_policy_binding}{}}}
\subsection{Method \code{delete_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$delete_validating_admission_policy_binding(
  name,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
watchAdmissionregistrationV1alpha1ValidatingAdmissionPolicyList
watch individual changes to a list of ValidatingAdmissionPolicy. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_list"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_list}{}}}
\subsection{Method \code{watch_validating_admission_policy_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy_list(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAdmissionregistrationV1alpha1ValidatingAdmissionPolicy
watch changes to an object of kind ValidatingAdmissionPolicy. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy}{}}}
\subsection{Method \code{watch_validating_admission_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy(
  name,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBindingList
watch individual changes to a list of ValidatingAdmissionPolicyBinding. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_binding_list"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_binding_list}{}}}
\subsection{Method \code{watch_validating_admission_policy_binding_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy_binding_list(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAdmissionregistrationV1alpha1ValidatingAdmissionPolicyBinding
watch changes to an object of kind ValidatingAdmissionPolicyBinding. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_binding"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-watch_validating_admission_policy_binding}{}}}
\subsection{Method \code{watch_validating_admission_policy_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$watch_validating_admission_policy_binding(
  name,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1alpha1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1alpha1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1alpha1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
