% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_AdmissionregistrationV1WebhookClientConfig.R
\name{AdmissionregistrationV1WebhookClientConfig}
\alias{AdmissionregistrationV1WebhookClientConfig}
\title{AdmissionregistrationV1WebhookClientConfig}
\description{
AdmissionregistrationV1WebhookClientConfig

AdmissionregistrationV1WebhookClientConfig
}
\details{
WebhookClientConfig contains the information to make a TLS connection with the webhook
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdmissionregistrationV1WebhookClientConfig-new}{\code{AdmissionregistrationV1WebhookClientConfig$new()}}
\item \href{#method-AdmissionregistrationV1WebhookClientConfig-to_list}{\code{AdmissionregistrationV1WebhookClientConfig$to_list()}}
\item \href{#method-AdmissionregistrationV1WebhookClientConfig-clone}{\code{AdmissionregistrationV1WebhookClientConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1WebhookClientConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1WebhookClientConfig-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1WebhookClientConfig$new(
  ca_bundle = NULL,
  service = NULL,
  url = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1WebhookClientConfig-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1WebhookClientConfig-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1WebhookClientConfig$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdmissionregistrationV1WebhookClientConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdmissionregistrationV1WebhookClientConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdmissionregistrationV1WebhookClientConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
