% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_ApiextensionsApi.R
\name{ApiextensionsApi}
\alias{ApiextensionsApi}
\title{ApiextensionsApi}
\description{
ApiextensionsApi

ApiextensionsApi
}
\details{
Typed Kubernetes API client for the `Apiextensions` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiextensionsApi-new}{\code{ApiextensionsApi$new()}}
\item \href{#method-ApiextensionsApi-get_api_group}{\code{ApiextensionsApi$get_api_group()}}
\item \href{#method-ApiextensionsApi-clone}{\code{ApiextensionsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getApiextensionsAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
