% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_ApiextensionsV1Api.R
\name{ApiextensionsV1Api}
\alias{ApiextensionsV1Api}
\title{ApiextensionsV1Api}
\description{
ApiextensionsV1Api

ApiextensionsV1Api
}
\details{
Typed Kubernetes API client for the `ApiextensionsV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiextensionsV1Api-new}{\code{ApiextensionsV1Api$new()}}
\item \href{#method-ApiextensionsV1Api-get_api_resources}{\code{ApiextensionsV1Api$get_api_resources()}}
\item \href{#method-ApiextensionsV1Api-list_custom_resource_definition}{\code{ApiextensionsV1Api$list_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-create_custom_resource_definition}{\code{ApiextensionsV1Api$create_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-delete_collection_custom_resource_definition}{\code{ApiextensionsV1Api$delete_collection_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-read_custom_resource_definition}{\code{ApiextensionsV1Api$read_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-replace_custom_resource_definition}{\code{ApiextensionsV1Api$replace_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-patch_custom_resource_definition}{\code{ApiextensionsV1Api$patch_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-delete_custom_resource_definition}{\code{ApiextensionsV1Api$delete_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-read_custom_resource_definition_status}{\code{ApiextensionsV1Api$read_custom_resource_definition_status()}}
\item \href{#method-ApiextensionsV1Api-replace_custom_resource_definition_status}{\code{ApiextensionsV1Api$replace_custom_resource_definition_status()}}
\item \href{#method-ApiextensionsV1Api-patch_custom_resource_definition_status}{\code{ApiextensionsV1Api$patch_custom_resource_definition_status()}}
\item \href{#method-ApiextensionsV1Api-watch_custom_resource_definition_list}{\code{ApiextensionsV1Api$watch_custom_resource_definition_list()}}
\item \href{#method-ApiextensionsV1Api-watch_custom_resource_definition}{\code{ApiextensionsV1Api$watch_custom_resource_definition()}}
\item \href{#method-ApiextensionsV1Api-clone}{\code{ApiextensionsV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getApiextensionsV1APIResources
get available resources
listApiextensionsV1CustomResourceDefinition
list or watch objects of kind CustomResourceDefinition
createApiextensionsV1CustomResourceDefinition
create a CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-list_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-list_custom_resource_definition}{}}}
\subsection{Method \code{list_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$list_custom_resource_definition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-create_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-create_custom_resource_definition}{}}}
\subsection{Method \code{create_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$create_custom_resource_definition(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteApiextensionsV1CollectionCustomResourceDefinition
delete collection of CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-delete_collection_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-delete_collection_custom_resource_definition}{}}}
\subsection{Method \code{delete_collection_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$delete_collection_custom_resource_definition(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readApiextensionsV1CustomResourceDefinition
read the specified CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-read_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-read_custom_resource_definition}{}}}
\subsection{Method \code{read_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$read_custom_resource_definition(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceApiextensionsV1CustomResourceDefinition
replace the specified CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-replace_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-replace_custom_resource_definition}{}}}
\subsection{Method \code{replace_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$replace_custom_resource_definition(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchApiextensionsV1CustomResourceDefinition
partially update the specified CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-patch_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-patch_custom_resource_definition}{}}}
\subsection{Method \code{patch_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$patch_custom_resource_definition(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteApiextensionsV1CustomResourceDefinition
delete a CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-delete_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-delete_custom_resource_definition}{}}}
\subsection{Method \code{delete_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$delete_custom_resource_definition(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readApiextensionsV1CustomResourceDefinitionStatus
read status of the specified CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-read_custom_resource_definition_status"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-read_custom_resource_definition_status}{}}}
\subsection{Method \code{read_custom_resource_definition_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$read_custom_resource_definition_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceApiextensionsV1CustomResourceDefinitionStatus
replace status of the specified CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-replace_custom_resource_definition_status"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-replace_custom_resource_definition_status}{}}}
\subsection{Method \code{replace_custom_resource_definition_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$replace_custom_resource_definition_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchApiextensionsV1CustomResourceDefinitionStatus
partially update status of the specified CustomResourceDefinition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-patch_custom_resource_definition_status"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-patch_custom_resource_definition_status}{}}}
\subsection{Method \code{patch_custom_resource_definition_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$patch_custom_resource_definition_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchApiextensionsV1CustomResourceDefinitionList
watch individual changes to a list of CustomResourceDefinition. deprecated: use the 'watch' parameter with a list operation instead.
watchApiextensionsV1CustomResourceDefinition
watch changes to an object of kind CustomResourceDefinition. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-watch_custom_resource_definition_list"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-watch_custom_resource_definition_list}{}}}
\subsection{Method \code{watch_custom_resource_definition_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$watch_custom_resource_definition_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-watch_custom_resource_definition"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-watch_custom_resource_definition}{}}}
\subsection{Method \code{watch_custom_resource_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$watch_custom_resource_definition(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
