% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_ApiextensionsV1ServiceReference.R
\name{ApiextensionsV1ServiceReference}
\alias{ApiextensionsV1ServiceReference}
\title{ApiextensionsV1ServiceReference}
\description{
ApiextensionsV1ServiceReference

ApiextensionsV1ServiceReference
}
\details{
ServiceReference holds a reference to Service.legacy.k8s.io
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiextensionsV1ServiceReference-new}{\code{ApiextensionsV1ServiceReference$new()}}
\item \href{#method-ApiextensionsV1ServiceReference-to_list}{\code{ApiextensionsV1ServiceReference$to_list()}}
\item \href{#method-ApiextensionsV1ServiceReference-clone}{\code{ApiextensionsV1ServiceReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1ServiceReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1ServiceReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1ServiceReference$new(
  name = NULL,
  namespace = NULL,
  path = NULL,
  port = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1ServiceReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1ServiceReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1ServiceReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiextensionsV1ServiceReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiextensionsV1ServiceReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiextensionsV1ServiceReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
