% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_ApiregistrationApi.R
\name{ApiregistrationApi}
\alias{ApiregistrationApi}
\title{ApiregistrationApi}
\description{
ApiregistrationApi

ApiregistrationApi
}
\details{
Typed Kubernetes API client for the `Apiregistration` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiregistrationApi-new}{\code{ApiregistrationApi$new()}}
\item \href{#method-ApiregistrationApi-get_api_group}{\code{ApiregistrationApi$get_api_group()}}
\item \href{#method-ApiregistrationApi-clone}{\code{ApiregistrationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getApiregistrationAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
