% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_ApiregistrationV1Api.R
\name{ApiregistrationV1Api}
\alias{ApiregistrationV1Api}
\title{ApiregistrationV1Api}
\description{
ApiregistrationV1Api

ApiregistrationV1Api
}
\details{
Typed Kubernetes API client for the `ApiregistrationV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiregistrationV1Api-new}{\code{ApiregistrationV1Api$new()}}
\item \href{#method-ApiregistrationV1Api-get_api_resources}{\code{ApiregistrationV1Api$get_api_resources()}}
\item \href{#method-ApiregistrationV1Api-list_api_service}{\code{ApiregistrationV1Api$list_api_service()}}
\item \href{#method-ApiregistrationV1Api-create_api_service}{\code{ApiregistrationV1Api$create_api_service()}}
\item \href{#method-ApiregistrationV1Api-delete_collection_api_service}{\code{ApiregistrationV1Api$delete_collection_api_service()}}
\item \href{#method-ApiregistrationV1Api-read_api_service}{\code{ApiregistrationV1Api$read_api_service()}}
\item \href{#method-ApiregistrationV1Api-replace_api_service}{\code{ApiregistrationV1Api$replace_api_service()}}
\item \href{#method-ApiregistrationV1Api-patch_api_service}{\code{ApiregistrationV1Api$patch_api_service()}}
\item \href{#method-ApiregistrationV1Api-delete_api_service}{\code{ApiregistrationV1Api$delete_api_service()}}
\item \href{#method-ApiregistrationV1Api-read_api_service_status}{\code{ApiregistrationV1Api$read_api_service_status()}}
\item \href{#method-ApiregistrationV1Api-replace_api_service_status}{\code{ApiregistrationV1Api$replace_api_service_status()}}
\item \href{#method-ApiregistrationV1Api-patch_api_service_status}{\code{ApiregistrationV1Api$patch_api_service_status()}}
\item \href{#method-ApiregistrationV1Api-watch_api_service_list}{\code{ApiregistrationV1Api$watch_api_service_list()}}
\item \href{#method-ApiregistrationV1Api-watch_api_service}{\code{ApiregistrationV1Api$watch_api_service()}}
\item \href{#method-ApiregistrationV1Api-clone}{\code{ApiregistrationV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getApiregistrationV1APIResources
get available resources
listApiregistrationV1APIService
list or watch objects of kind APIService
createApiregistrationV1APIService
create an APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-list_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-list_api_service}{}}}
\subsection{Method \code{list_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$list_api_service()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-create_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-create_api_service}{}}}
\subsection{Method \code{create_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$create_api_service(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteApiregistrationV1CollectionAPIService
delete collection of APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-delete_collection_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-delete_collection_api_service}{}}}
\subsection{Method \code{delete_collection_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$delete_collection_api_service(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readApiregistrationV1APIService
read the specified APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-read_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-read_api_service}{}}}
\subsection{Method \code{read_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$read_api_service(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceApiregistrationV1APIService
replace the specified APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-replace_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-replace_api_service}{}}}
\subsection{Method \code{replace_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$replace_api_service(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchApiregistrationV1APIService
partially update the specified APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-patch_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-patch_api_service}{}}}
\subsection{Method \code{patch_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$patch_api_service(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteApiregistrationV1APIService
delete an APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-delete_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-delete_api_service}{}}}
\subsection{Method \code{delete_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$delete_api_service(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readApiregistrationV1APIServiceStatus
read status of the specified APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-read_api_service_status"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-read_api_service_status}{}}}
\subsection{Method \code{read_api_service_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$read_api_service_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceApiregistrationV1APIServiceStatus
replace status of the specified APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-replace_api_service_status"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-replace_api_service_status}{}}}
\subsection{Method \code{replace_api_service_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$replace_api_service_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchApiregistrationV1APIServiceStatus
partially update status of the specified APIService}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-patch_api_service_status"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-patch_api_service_status}{}}}
\subsection{Method \code{patch_api_service_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$patch_api_service_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchApiregistrationV1APIServiceList
watch individual changes to a list of APIService. deprecated: use the 'watch' parameter with a list operation instead.
watchApiregistrationV1APIService
watch changes to an object of kind APIService. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-watch_api_service_list"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-watch_api_service_list}{}}}
\subsection{Method \code{watch_api_service_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$watch_api_service_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-watch_api_service"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-watch_api_service}{}}}
\subsection{Method \code{watch_api_service()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$watch_api_service(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiregistrationV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiregistrationV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiregistrationV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
