% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_ApisApi.R
\name{ApisApi}
\alias{ApisApi}
\title{ApisApi}
\description{
ApisApi

ApisApi
}
\details{
Typed Kubernetes API client for the `Apis` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApisApi-new}{\code{ApisApi$new()}}
\item \href{#method-ApisApi-get_api_versions}{\code{ApisApi$get_api_versions()}}
\item \href{#method-ApisApi-clone}{\code{ApisApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApisApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApisApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApisApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAPIVersions
get available API versions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApisApi-get_api_versions"></a>}}
\if{latex}{\out{\hypertarget{method-ApisApi-get_api_versions}{}}}
\subsection{Method \code{get_api_versions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApisApi$get_api_versions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApisApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApisApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApisApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
