% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AppsV1Api.R
\name{AppsV1Api}
\alias{AppsV1Api}
\title{AppsV1Api}
\description{
AppsV1Api

AppsV1Api
}
\details{
Typed Kubernetes API client for the `AppsV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AppsV1Api-new}{\code{AppsV1Api$new()}}
\item \href{#method-AppsV1Api-get_api_resources}{\code{AppsV1Api$get_api_resources()}}
\item \href{#method-AppsV1Api-list_controller_revision_for_all_namespaces}{\code{AppsV1Api$list_controller_revision_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_daemon_set_for_all_namespaces}{\code{AppsV1Api$list_daemon_set_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_deployment_for_all_namespaces}{\code{AppsV1Api$list_deployment_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_namespaced_controller_revision}{\code{AppsV1Api$list_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-create_namespaced_controller_revision}{\code{AppsV1Api$create_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_controller_revision}{\code{AppsV1Api$delete_collection_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-read_namespaced_controller_revision}{\code{AppsV1Api$read_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-replace_namespaced_controller_revision}{\code{AppsV1Api$replace_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-patch_namespaced_controller_revision}{\code{AppsV1Api$patch_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-delete_namespaced_controller_revision}{\code{AppsV1Api$delete_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-list_namespaced_daemon_set}{\code{AppsV1Api$list_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-create_namespaced_daemon_set}{\code{AppsV1Api$create_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_daemon_set}{\code{AppsV1Api$delete_collection_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-read_namespaced_daemon_set}{\code{AppsV1Api$read_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-replace_namespaced_daemon_set}{\code{AppsV1Api$replace_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-patch_namespaced_daemon_set}{\code{AppsV1Api$patch_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-delete_namespaced_daemon_set}{\code{AppsV1Api$delete_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-read_namespaced_daemon_set_status}{\code{AppsV1Api$read_namespaced_daemon_set_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_daemon_set_status}{\code{AppsV1Api$replace_namespaced_daemon_set_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_daemon_set_status}{\code{AppsV1Api$patch_namespaced_daemon_set_status()}}
\item \href{#method-AppsV1Api-list_namespaced_deployment}{\code{AppsV1Api$list_namespaced_deployment()}}
\item \href{#method-AppsV1Api-create_namespaced_deployment}{\code{AppsV1Api$create_namespaced_deployment()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_deployment}{\code{AppsV1Api$delete_collection_namespaced_deployment()}}
\item \href{#method-AppsV1Api-read_namespaced_deployment}{\code{AppsV1Api$read_namespaced_deployment()}}
\item \href{#method-AppsV1Api-replace_namespaced_deployment}{\code{AppsV1Api$replace_namespaced_deployment()}}
\item \href{#method-AppsV1Api-patch_namespaced_deployment}{\code{AppsV1Api$patch_namespaced_deployment()}}
\item \href{#method-AppsV1Api-delete_namespaced_deployment}{\code{AppsV1Api$delete_namespaced_deployment()}}
\item \href{#method-AppsV1Api-read_namespaced_deployment_scale}{\code{AppsV1Api$read_namespaced_deployment_scale()}}
\item \href{#method-AppsV1Api-replace_namespaced_deployment_scale}{\code{AppsV1Api$replace_namespaced_deployment_scale()}}
\item \href{#method-AppsV1Api-patch_namespaced_deployment_scale}{\code{AppsV1Api$patch_namespaced_deployment_scale()}}
\item \href{#method-AppsV1Api-read_namespaced_deployment_status}{\code{AppsV1Api$read_namespaced_deployment_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_deployment_status}{\code{AppsV1Api$replace_namespaced_deployment_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_deployment_status}{\code{AppsV1Api$patch_namespaced_deployment_status()}}
\item \href{#method-AppsV1Api-list_namespaced_replica_set}{\code{AppsV1Api$list_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-create_namespaced_replica_set}{\code{AppsV1Api$create_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_replica_set}{\code{AppsV1Api$delete_collection_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-read_namespaced_replica_set}{\code{AppsV1Api$read_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-replace_namespaced_replica_set}{\code{AppsV1Api$replace_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-patch_namespaced_replica_set}{\code{AppsV1Api$patch_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-delete_namespaced_replica_set}{\code{AppsV1Api$delete_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-read_namespaced_replica_set_scale}{\code{AppsV1Api$read_namespaced_replica_set_scale()}}
\item \href{#method-AppsV1Api-replace_namespaced_replica_set_scale}{\code{AppsV1Api$replace_namespaced_replica_set_scale()}}
\item \href{#method-AppsV1Api-patch_namespaced_replica_set_scale}{\code{AppsV1Api$patch_namespaced_replica_set_scale()}}
\item \href{#method-AppsV1Api-read_namespaced_replica_set_status}{\code{AppsV1Api$read_namespaced_replica_set_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_replica_set_status}{\code{AppsV1Api$replace_namespaced_replica_set_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_replica_set_status}{\code{AppsV1Api$patch_namespaced_replica_set_status()}}
\item \href{#method-AppsV1Api-list_namespaced_stateful_set}{\code{AppsV1Api$list_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-create_namespaced_stateful_set}{\code{AppsV1Api$create_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_stateful_set}{\code{AppsV1Api$delete_collection_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-read_namespaced_stateful_set}{\code{AppsV1Api$read_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-replace_namespaced_stateful_set}{\code{AppsV1Api$replace_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-patch_namespaced_stateful_set}{\code{AppsV1Api$patch_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-delete_namespaced_stateful_set}{\code{AppsV1Api$delete_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-read_namespaced_stateful_set_scale}{\code{AppsV1Api$read_namespaced_stateful_set_scale()}}
\item \href{#method-AppsV1Api-replace_namespaced_stateful_set_scale}{\code{AppsV1Api$replace_namespaced_stateful_set_scale()}}
\item \href{#method-AppsV1Api-patch_namespaced_stateful_set_scale}{\code{AppsV1Api$patch_namespaced_stateful_set_scale()}}
\item \href{#method-AppsV1Api-read_namespaced_stateful_set_status}{\code{AppsV1Api$read_namespaced_stateful_set_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_stateful_set_status}{\code{AppsV1Api$replace_namespaced_stateful_set_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_stateful_set_status}{\code{AppsV1Api$patch_namespaced_stateful_set_status()}}
\item \href{#method-AppsV1Api-list_replica_set_for_all_namespaces}{\code{AppsV1Api$list_replica_set_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_stateful_set_for_all_namespaces}{\code{AppsV1Api$list_stateful_set_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_controller_revision_list_for_all_namespaces}{\code{AppsV1Api$watch_controller_revision_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_daemon_set_list_for_all_namespaces}{\code{AppsV1Api$watch_daemon_set_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_deployment_list_for_all_namespaces}{\code{AppsV1Api$watch_deployment_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_namespaced_controller_revision_list}{\code{AppsV1Api$watch_namespaced_controller_revision_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_controller_revision}{\code{AppsV1Api$watch_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-watch_namespaced_daemon_set_list}{\code{AppsV1Api$watch_namespaced_daemon_set_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_daemon_set}{\code{AppsV1Api$watch_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-watch_namespaced_deployment_list}{\code{AppsV1Api$watch_namespaced_deployment_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_deployment}{\code{AppsV1Api$watch_namespaced_deployment()}}
\item \href{#method-AppsV1Api-watch_namespaced_replica_set_list}{\code{AppsV1Api$watch_namespaced_replica_set_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_replica_set}{\code{AppsV1Api$watch_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-watch_namespaced_stateful_set_list}{\code{AppsV1Api$watch_namespaced_stateful_set_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_stateful_set}{\code{AppsV1Api$watch_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-watch_replica_set_list_for_all_namespaces}{\code{AppsV1Api$watch_replica_set_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_stateful_set_list_for_all_namespaces}{\code{AppsV1Api$watch_stateful_set_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-clone}{\code{AppsV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAppsV1APIResources
get available resources
listAppsV1ControllerRevisionForAllNamespaces
list or watch objects of kind ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_controller_revision_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_controller_revision_for_all_namespaces}{}}}
\subsection{Method \code{list_controller_revision_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_controller_revision_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listAppsV1DaemonSetForAllNamespaces
list or watch objects of kind DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_daemon_set_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_daemon_set_for_all_namespaces}{}}}
\subsection{Method \code{list_daemon_set_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_daemon_set_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listAppsV1DeploymentForAllNamespaces
list or watch objects of kind Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_deployment_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_deployment_for_all_namespaces}{}}}
\subsection{Method \code{list_deployment_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_deployment_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listAppsV1NamespacedControllerRevision
list or watch objects of kind ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_controller_revision}{}}}
\subsection{Method \code{list_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_controller_revision(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAppsV1NamespacedControllerRevision
create a ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_controller_revision}{}}}
\subsection{Method \code{create_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_controller_revision(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedControllerRevision
delete collection of ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_controller_revision}{}}}
\subsection{Method \code{delete_collection_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_controller_revision(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedControllerRevision
read the specified ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_controller_revision}{}}}
\subsection{Method \code{read_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_controller_revision(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedControllerRevision
replace the specified ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_controller_revision}{}}}
\subsection{Method \code{replace_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_controller_revision(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedControllerRevision
partially update the specified ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_controller_revision}{}}}
\subsection{Method \code{patch_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_controller_revision(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAppsV1NamespacedControllerRevision
delete a ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_controller_revision}{}}}
\subsection{Method \code{delete_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_controller_revision(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
listAppsV1NamespacedDaemonSet
list or watch objects of kind DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_daemon_set}{}}}
\subsection{Method \code{list_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_daemon_set(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAppsV1NamespacedDaemonSet
create a DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_daemon_set}{}}}
\subsection{Method \code{create_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_daemon_set(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedDaemonSet
delete collection of DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_daemon_set}{}}}
\subsection{Method \code{delete_collection_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_daemon_set(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedDaemonSet
read the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_daemon_set}{}}}
\subsection{Method \code{read_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_daemon_set(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedDaemonSet
replace the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_daemon_set}{}}}
\subsection{Method \code{replace_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_daemon_set(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDaemonSet
partially update the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_daemon_set}{}}}
\subsection{Method \code{patch_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_daemon_set(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAppsV1NamespacedDaemonSet
delete a DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_daemon_set}{}}}
\subsection{Method \code{delete_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_daemon_set(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedDaemonSetStatus
read status of the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_daemon_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_daemon_set_status}{}}}
\subsection{Method \code{read_namespaced_daemon_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_daemon_set_status(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedDaemonSetStatus
replace status of the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_daemon_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_daemon_set_status}{}}}
\subsection{Method \code{replace_namespaced_daemon_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_daemon_set_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDaemonSetStatus
partially update status of the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_daemon_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_daemon_set_status}{}}}
\subsection{Method \code{patch_namespaced_daemon_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_daemon_set_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listAppsV1NamespacedDeployment
list or watch objects of kind Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_deployment}{}}}
\subsection{Method \code{list_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_deployment(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAppsV1NamespacedDeployment
create a Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_deployment}{}}}
\subsection{Method \code{create_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_deployment(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedDeployment
delete collection of Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_deployment}{}}}
\subsection{Method \code{delete_collection_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_deployment(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedDeployment
read the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_deployment}{}}}
\subsection{Method \code{read_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_deployment(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedDeployment
replace the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_deployment}{}}}
\subsection{Method \code{replace_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_deployment(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDeployment
partially update the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_deployment}{}}}
\subsection{Method \code{patch_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_deployment(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAppsV1NamespacedDeployment
delete a Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_deployment}{}}}
\subsection{Method \code{delete_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_deployment(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedDeploymentScale
read scale of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_deployment_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_deployment_scale}{}}}
\subsection{Method \code{read_namespaced_deployment_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_deployment_scale(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedDeploymentScale
replace scale of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_deployment_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_deployment_scale}{}}}
\subsection{Method \code{replace_namespaced_deployment_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_deployment_scale(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDeploymentScale
partially update scale of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_deployment_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_deployment_scale}{}}}
\subsection{Method \code{patch_namespaced_deployment_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_deployment_scale(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
readAppsV1NamespacedDeploymentStatus
read status of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_deployment_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_deployment_status}{}}}
\subsection{Method \code{read_namespaced_deployment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_deployment_status(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedDeploymentStatus
replace status of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_deployment_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_deployment_status}{}}}
\subsection{Method \code{replace_namespaced_deployment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_deployment_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDeploymentStatus
partially update status of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_deployment_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_deployment_status}{}}}
\subsection{Method \code{patch_namespaced_deployment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_deployment_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listAppsV1NamespacedReplicaSet
list or watch objects of kind ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_replica_set}{}}}
\subsection{Method \code{list_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_replica_set(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAppsV1NamespacedReplicaSet
create a ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_replica_set}{}}}
\subsection{Method \code{create_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_replica_set(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedReplicaSet
delete collection of ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_replica_set}{}}}
\subsection{Method \code{delete_collection_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_replica_set(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedReplicaSet
read the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_replica_set}{}}}
\subsection{Method \code{read_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_replica_set(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedReplicaSet
replace the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_replica_set}{}}}
\subsection{Method \code{replace_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_replica_set(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedReplicaSet
partially update the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_replica_set}{}}}
\subsection{Method \code{patch_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_replica_set(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAppsV1NamespacedReplicaSet
delete a ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_replica_set}{}}}
\subsection{Method \code{delete_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_replica_set(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedReplicaSetScale
read scale of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_replica_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_replica_set_scale}{}}}
\subsection{Method \code{read_namespaced_replica_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_replica_set_scale(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedReplicaSetScale
replace scale of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_replica_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_replica_set_scale}{}}}
\subsection{Method \code{replace_namespaced_replica_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_replica_set_scale(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedReplicaSetScale
partially update scale of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_replica_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_replica_set_scale}{}}}
\subsection{Method \code{patch_namespaced_replica_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_replica_set_scale(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
readAppsV1NamespacedReplicaSetStatus
read status of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_replica_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_replica_set_status}{}}}
\subsection{Method \code{read_namespaced_replica_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_replica_set_status(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedReplicaSetStatus
replace status of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_replica_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_replica_set_status}{}}}
\subsection{Method \code{replace_namespaced_replica_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_replica_set_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedReplicaSetStatus
partially update status of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_replica_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_replica_set_status}{}}}
\subsection{Method \code{patch_namespaced_replica_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_replica_set_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listAppsV1NamespacedStatefulSet
list or watch objects of kind StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_stateful_set}{}}}
\subsection{Method \code{list_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_stateful_set(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAppsV1NamespacedStatefulSet
create a StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_stateful_set}{}}}
\subsection{Method \code{create_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_stateful_set(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedStatefulSet
delete collection of StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_stateful_set}{}}}
\subsection{Method \code{delete_collection_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_stateful_set(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedStatefulSet
read the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_stateful_set}{}}}
\subsection{Method \code{read_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_stateful_set(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedStatefulSet
replace the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_stateful_set}{}}}
\subsection{Method \code{replace_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_stateful_set(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedStatefulSet
partially update the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_stateful_set}{}}}
\subsection{Method \code{patch_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_stateful_set(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAppsV1NamespacedStatefulSet
delete a StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_stateful_set}{}}}
\subsection{Method \code{delete_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_stateful_set(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readAppsV1NamespacedStatefulSetScale
read scale of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_stateful_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_stateful_set_scale}{}}}
\subsection{Method \code{read_namespaced_stateful_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_stateful_set_scale(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedStatefulSetScale
replace scale of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_stateful_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_stateful_set_scale}{}}}
\subsection{Method \code{replace_namespaced_stateful_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_stateful_set_scale(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedStatefulSetScale
partially update scale of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_stateful_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_stateful_set_scale}{}}}
\subsection{Method \code{patch_namespaced_stateful_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_stateful_set_scale(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
readAppsV1NamespacedStatefulSetStatus
read status of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_stateful_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_stateful_set_status}{}}}
\subsection{Method \code{read_namespaced_stateful_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_stateful_set_status(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAppsV1NamespacedStatefulSetStatus
replace status of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_stateful_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_stateful_set_status}{}}}
\subsection{Method \code{replace_namespaced_stateful_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_stateful_set_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedStatefulSetStatus
partially update status of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_stateful_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_stateful_set_status}{}}}
\subsection{Method \code{patch_namespaced_stateful_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_stateful_set_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listAppsV1ReplicaSetForAllNamespaces
list or watch objects of kind ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_replica_set_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_replica_set_for_all_namespaces}{}}}
\subsection{Method \code{list_replica_set_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_replica_set_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listAppsV1StatefulSetForAllNamespaces
list or watch objects of kind StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_stateful_set_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_stateful_set_for_all_namespaces}{}}}
\subsection{Method \code{list_stateful_set_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_stateful_set_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1ControllerRevisionListForAllNamespaces
watch individual changes to a list of ControllerRevision. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_controller_revision_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_controller_revision_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_controller_revision_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_controller_revision_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1DaemonSetListForAllNamespaces
watch individual changes to a list of DaemonSet. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_daemon_set_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_daemon_set_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_daemon_set_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_daemon_set_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1DeploymentListForAllNamespaces
watch individual changes to a list of Deployment. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_deployment_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_deployment_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_deployment_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_deployment_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedControllerRevisionList
watch individual changes to a list of ControllerRevision. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_controller_revision_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_controller_revision_list}{}}}
\subsection{Method \code{watch_namespaced_controller_revision_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_controller_revision_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedControllerRevision
watch changes to an object of kind ControllerRevision. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_controller_revision}{}}}
\subsection{Method \code{watch_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_controller_revision(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedDaemonSetList
watch individual changes to a list of DaemonSet. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_daemon_set_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_daemon_set_list}{}}}
\subsection{Method \code{watch_namespaced_daemon_set_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_daemon_set_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedDaemonSet
watch changes to an object of kind DaemonSet. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_daemon_set}{}}}
\subsection{Method \code{watch_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_daemon_set(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedDeploymentList
watch individual changes to a list of Deployment. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_deployment_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_deployment_list}{}}}
\subsection{Method \code{watch_namespaced_deployment_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_deployment_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedDeployment
watch changes to an object of kind Deployment. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_deployment}{}}}
\subsection{Method \code{watch_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_deployment(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedReplicaSetList
watch individual changes to a list of ReplicaSet. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_replica_set_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_replica_set_list}{}}}
\subsection{Method \code{watch_namespaced_replica_set_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_replica_set_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedReplicaSet
watch changes to an object of kind ReplicaSet. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_replica_set}{}}}
\subsection{Method \code{watch_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_replica_set(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedStatefulSetList
watch individual changes to a list of StatefulSet. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_stateful_set_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_stateful_set_list}{}}}
\subsection{Method \code{watch_namespaced_stateful_set_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_stateful_set_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1NamespacedStatefulSet
watch changes to an object of kind StatefulSet. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_stateful_set}{}}}
\subsection{Method \code{watch_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_stateful_set(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1ReplicaSetListForAllNamespaces
watch individual changes to a list of ReplicaSet. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_replica_set_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_replica_set_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_replica_set_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_replica_set_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAppsV1StatefulSetListForAllNamespaces
watch individual changes to a list of StatefulSet. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_stateful_set_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_stateful_set_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_stateful_set_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_stateful_set_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
