% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AuthenticationApi.R
\name{AuthenticationApi}
\alias{AuthenticationApi}
\title{AuthenticationApi}
\description{
AuthenticationApi

AuthenticationApi
}
\details{
Typed Kubernetes API client for the `Authentication` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthenticationApi-new}{\code{AuthenticationApi$new()}}
\item \href{#method-AuthenticationApi-get_api_group}{\code{AuthenticationApi$get_api_group()}}
\item \href{#method-AuthenticationApi-clone}{\code{AuthenticationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAuthenticationAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
