% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AuthenticationV1alpha1Api.R
\name{AuthenticationV1alpha1Api}
\alias{AuthenticationV1alpha1Api}
\title{AuthenticationV1alpha1Api}
\description{
AuthenticationV1alpha1Api

AuthenticationV1alpha1Api
}
\details{
Typed Kubernetes API client for the `AuthenticationV1alpha1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthenticationV1alpha1Api-new}{\code{AuthenticationV1alpha1Api$new()}}
\item \href{#method-AuthenticationV1alpha1Api-get_api_resources}{\code{AuthenticationV1alpha1Api$get_api_resources()}}
\item \href{#method-AuthenticationV1alpha1Api-create_self_subject_review}{\code{AuthenticationV1alpha1Api$create_self_subject_review()}}
\item \href{#method-AuthenticationV1alpha1Api-clone}{\code{AuthenticationV1alpha1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationV1alpha1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationV1alpha1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationV1alpha1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAuthenticationV1alpha1APIResources
get available resources
createAuthenticationV1alpha1SelfSubjectReview
create a SelfSubjectReview}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationV1alpha1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationV1alpha1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationV1alpha1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationV1alpha1Api-create_self_subject_review"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationV1alpha1Api-create_self_subject_review}{}}}
\subsection{Method \code{create_self_subject_review()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationV1alpha1Api$create_self_subject_review(
  body,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{pretty}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationV1alpha1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationV1alpha1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationV1alpha1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
