% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_AuthenticationV1TokenRequest.R
\name{AuthenticationV1TokenRequest}
\alias{AuthenticationV1TokenRequest}
\title{AuthenticationV1TokenRequest}
\description{
AuthenticationV1TokenRequest

AuthenticationV1TokenRequest
}
\details{
TokenRequest requests a token for a given service account.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthenticationV1TokenRequest-new}{\code{AuthenticationV1TokenRequest$new()}}
\item \href{#method-AuthenticationV1TokenRequest-to_list}{\code{AuthenticationV1TokenRequest$to_list()}}
\item \href{#method-AuthenticationV1TokenRequest-clone}{\code{AuthenticationV1TokenRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationV1TokenRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationV1TokenRequest-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationV1TokenRequest$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationV1TokenRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationV1TokenRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationV1TokenRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthenticationV1TokenRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthenticationV1TokenRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthenticationV1TokenRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
