% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AuthorizationApi.R
\name{AuthorizationApi}
\alias{AuthorizationApi}
\title{AuthorizationApi}
\description{
AuthorizationApi

AuthorizationApi
}
\details{
Typed Kubernetes API client for the `Authorization` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthorizationApi-new}{\code{AuthorizationApi$new()}}
\item \href{#method-AuthorizationApi-get_api_group}{\code{AuthorizationApi$get_api_group()}}
\item \href{#method-AuthorizationApi-clone}{\code{AuthorizationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAuthorizationAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
