% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AuthorizationV1Api.R
\name{AuthorizationV1Api}
\alias{AuthorizationV1Api}
\title{AuthorizationV1Api}
\description{
AuthorizationV1Api

AuthorizationV1Api
}
\details{
Typed Kubernetes API client for the `AuthorizationV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthorizationV1Api-new}{\code{AuthorizationV1Api$new()}}
\item \href{#method-AuthorizationV1Api-get_api_resources}{\code{AuthorizationV1Api$get_api_resources()}}
\item \href{#method-AuthorizationV1Api-create_namespaced_local_subject_access_review}{\code{AuthorizationV1Api$create_namespaced_local_subject_access_review()}}
\item \href{#method-AuthorizationV1Api-create_self_subject_access_review}{\code{AuthorizationV1Api$create_self_subject_access_review()}}
\item \href{#method-AuthorizationV1Api-create_self_subject_rules_review}{\code{AuthorizationV1Api$create_self_subject_rules_review()}}
\item \href{#method-AuthorizationV1Api-create_subject_access_review}{\code{AuthorizationV1Api$create_subject_access_review()}}
\item \href{#method-AuthorizationV1Api-clone}{\code{AuthorizationV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAuthorizationV1APIResources
get available resources
createAuthorizationV1NamespacedLocalSubjectAccessReview
create a LocalSubjectAccessReview}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationV1Api-create_namespaced_local_subject_access_review"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationV1Api-create_namespaced_local_subject_access_review}{}}}
\subsection{Method \code{create_namespaced_local_subject_access_review()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationV1Api$create_namespaced_local_subject_access_review(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
createAuthorizationV1SelfSubjectAccessReview
create a SelfSubjectAccessReview}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationV1Api-create_self_subject_access_review"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationV1Api-create_self_subject_access_review}{}}}
\subsection{Method \code{create_self_subject_access_review()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationV1Api$create_self_subject_access_review(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
createAuthorizationV1SelfSubjectRulesReview
create a SelfSubjectRulesReview}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationV1Api-create_self_subject_rules_review"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationV1Api-create_self_subject_rules_review}{}}}
\subsection{Method \code{create_self_subject_rules_review()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationV1Api$create_self_subject_rules_review(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
createAuthorizationV1SubjectAccessReview
create a SubjectAccessReview}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationV1Api-create_subject_access_review"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationV1Api-create_subject_access_review}{}}}
\subsection{Method \code{create_subject_access_review()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationV1Api$create_subject_access_review(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthorizationV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthorizationV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthorizationV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
