% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AutoscalingApi.R
\name{AutoscalingApi}
\alias{AutoscalingApi}
\title{AutoscalingApi}
\description{
AutoscalingApi

AutoscalingApi
}
\details{
Typed Kubernetes API client for the `Autoscaling` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutoscalingApi-new}{\code{AutoscalingApi$new()}}
\item \href{#method-AutoscalingApi-get_api_group}{\code{AutoscalingApi$get_api_group()}}
\item \href{#method-AutoscalingApi-clone}{\code{AutoscalingApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAutoscalingAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
