% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AutoscalingV1Api.R
\name{AutoscalingV1Api}
\alias{AutoscalingV1Api}
\title{AutoscalingV1Api}
\description{
AutoscalingV1Api

AutoscalingV1Api
}
\details{
Typed Kubernetes API client for the `AutoscalingV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutoscalingV1Api-new}{\code{AutoscalingV1Api$new()}}
\item \href{#method-AutoscalingV1Api-get_api_resources}{\code{AutoscalingV1Api$get_api_resources()}}
\item \href{#method-AutoscalingV1Api-list_horizontal_pod_autoscaler_for_all_namespaces}{\code{AutoscalingV1Api$list_horizontal_pod_autoscaler_for_all_namespaces()}}
\item \href{#method-AutoscalingV1Api-list_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$list_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-create_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$create_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-delete_collection_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$delete_collection_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-read_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$read_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-replace_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$replace_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-patch_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$patch_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-delete_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$delete_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-read_namespaced_horizontal_pod_autoscaler_status}{\code{AutoscalingV1Api$read_namespaced_horizontal_pod_autoscaler_status()}}
\item \href{#method-AutoscalingV1Api-replace_namespaced_horizontal_pod_autoscaler_status}{\code{AutoscalingV1Api$replace_namespaced_horizontal_pod_autoscaler_status()}}
\item \href{#method-AutoscalingV1Api-patch_namespaced_horizontal_pod_autoscaler_status}{\code{AutoscalingV1Api$patch_namespaced_horizontal_pod_autoscaler_status()}}
\item \href{#method-AutoscalingV1Api-watch_horizontal_pod_autoscaler_list_for_all_namespaces}{\code{AutoscalingV1Api$watch_horizontal_pod_autoscaler_list_for_all_namespaces()}}
\item \href{#method-AutoscalingV1Api-watch_namespaced_horizontal_pod_autoscaler_list}{\code{AutoscalingV1Api$watch_namespaced_horizontal_pod_autoscaler_list()}}
\item \href{#method-AutoscalingV1Api-watch_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV1Api$watch_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV1Api-clone}{\code{AutoscalingV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAutoscalingV1APIResources
get available resources
listAutoscalingV1HorizontalPodAutoscalerForAllNamespaces
list or watch objects of kind HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-list_horizontal_pod_autoscaler_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-list_horizontal_pod_autoscaler_for_all_namespaces}{}}}
\subsection{Method \code{list_horizontal_pod_autoscaler_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$list_horizontal_pod_autoscaler_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listAutoscalingV1NamespacedHorizontalPodAutoscaler
list or watch objects of kind HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-list_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-list_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{list_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$list_namespaced_horizontal_pod_autoscaler(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createAutoscalingV1NamespacedHorizontalPodAutoscaler
create a HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-create_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-create_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{create_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$create_namespaced_horizontal_pod_autoscaler(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAutoscalingV1CollectionNamespacedHorizontalPodAutoscaler
delete collection of HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-delete_collection_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-delete_collection_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{delete_collection_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$delete_collection_namespaced_horizontal_pod_autoscaler(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readAutoscalingV1NamespacedHorizontalPodAutoscaler
read the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-read_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-read_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{read_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$read_namespaced_horizontal_pod_autoscaler(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAutoscalingV1NamespacedHorizontalPodAutoscaler
replace the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-replace_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-replace_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{replace_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$replace_namespaced_horizontal_pod_autoscaler(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAutoscalingV1NamespacedHorizontalPodAutoscaler
partially update the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-patch_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-patch_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{patch_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$patch_namespaced_horizontal_pod_autoscaler(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteAutoscalingV1NamespacedHorizontalPodAutoscaler
delete a HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-delete_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-delete_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{delete_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$delete_namespaced_horizontal_pod_autoscaler(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readAutoscalingV1NamespacedHorizontalPodAutoscalerStatus
read status of the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-read_namespaced_horizontal_pod_autoscaler_status"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-read_namespaced_horizontal_pod_autoscaler_status}{}}}
\subsection{Method \code{read_namespaced_horizontal_pod_autoscaler_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$read_namespaced_horizontal_pod_autoscaler_status(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceAutoscalingV1NamespacedHorizontalPodAutoscalerStatus
replace status of the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-replace_namespaced_horizontal_pod_autoscaler_status"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-replace_namespaced_horizontal_pod_autoscaler_status}{}}}
\subsection{Method \code{replace_namespaced_horizontal_pod_autoscaler_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$replace_namespaced_horizontal_pod_autoscaler_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAutoscalingV1NamespacedHorizontalPodAutoscalerStatus
partially update status of the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-patch_namespaced_horizontal_pod_autoscaler_status"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-patch_namespaced_horizontal_pod_autoscaler_status}{}}}
\subsection{Method \code{patch_namespaced_horizontal_pod_autoscaler_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$patch_namespaced_horizontal_pod_autoscaler_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
watchAutoscalingV1HorizontalPodAutoscalerListForAllNamespaces
watch individual changes to a list of HorizontalPodAutoscaler. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-watch_horizontal_pod_autoscaler_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-watch_horizontal_pod_autoscaler_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_horizontal_pod_autoscaler_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$watch_horizontal_pod_autoscaler_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAutoscalingV1NamespacedHorizontalPodAutoscalerList
watch individual changes to a list of HorizontalPodAutoscaler. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-watch_namespaced_horizontal_pod_autoscaler_list"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-watch_namespaced_horizontal_pod_autoscaler_list}{}}}
\subsection{Method \code{watch_namespaced_horizontal_pod_autoscaler_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$watch_namespaced_horizontal_pod_autoscaler_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchAutoscalingV1NamespacedHorizontalPodAutoscaler
watch changes to an object of kind HorizontalPodAutoscaler. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-watch_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-watch_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{watch_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$watch_namespaced_horizontal_pod_autoscaler(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
