% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AutoscalingV2Api.R
\name{AutoscalingV2Api}
\alias{AutoscalingV2Api}
\title{AutoscalingV2Api}
\description{
AutoscalingV2Api

AutoscalingV2Api
}
\details{
Typed Kubernetes API client for the `AutoscalingV2` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutoscalingV2Api-new}{\code{AutoscalingV2Api$new()}}
\item \href{#method-AutoscalingV2Api-get_api_resources}{\code{AutoscalingV2Api$get_api_resources()}}
\item \href{#method-AutoscalingV2Api-list_horizontal_pod_autoscaler_for_all_namespaces}{\code{AutoscalingV2Api$list_horizontal_pod_autoscaler_for_all_namespaces()}}
\item \href{#method-AutoscalingV2Api-list_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$list_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-create_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$create_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-delete_collection_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$delete_collection_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-read_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$read_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-replace_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$replace_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-patch_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$patch_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-delete_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$delete_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-read_namespaced_horizontal_pod_autoscaler_status}{\code{AutoscalingV2Api$read_namespaced_horizontal_pod_autoscaler_status()}}
\item \href{#method-AutoscalingV2Api-replace_namespaced_horizontal_pod_autoscaler_status}{\code{AutoscalingV2Api$replace_namespaced_horizontal_pod_autoscaler_status()}}
\item \href{#method-AutoscalingV2Api-patch_namespaced_horizontal_pod_autoscaler_status}{\code{AutoscalingV2Api$patch_namespaced_horizontal_pod_autoscaler_status()}}
\item \href{#method-AutoscalingV2Api-watch_horizontal_pod_autoscaler_list_for_all_namespaces}{\code{AutoscalingV2Api$watch_horizontal_pod_autoscaler_list_for_all_namespaces()}}
\item \href{#method-AutoscalingV2Api-watch_namespaced_horizontal_pod_autoscaler_list}{\code{AutoscalingV2Api$watch_namespaced_horizontal_pod_autoscaler_list()}}
\item \href{#method-AutoscalingV2Api-watch_namespaced_horizontal_pod_autoscaler}{\code{AutoscalingV2Api$watch_namespaced_horizontal_pod_autoscaler()}}
\item \href{#method-AutoscalingV2Api-clone}{\code{AutoscalingV2Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAutoscalingV2APIResources
get available resources
listAutoscalingV2HorizontalPodAutoscalerForAllNamespaces
list or watch objects of kind HorizontalPodAutoscaler
listAutoscalingV2NamespacedHorizontalPodAutoscaler
list or watch objects of kind HorizontalPodAutoscaler
createAutoscalingV2NamespacedHorizontalPodAutoscaler
create a HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-list_horizontal_pod_autoscaler_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-list_horizontal_pod_autoscaler_for_all_namespaces}{}}}
\subsection{Method \code{list_horizontal_pod_autoscaler_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$list_horizontal_pod_autoscaler_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-list_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-list_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{list_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$list_namespaced_horizontal_pod_autoscaler()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-create_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-create_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{create_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$create_namespaced_horizontal_pod_autoscaler(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAutoscalingV2CollectionNamespacedHorizontalPodAutoscaler
delete collection of HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-delete_collection_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-delete_collection_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{delete_collection_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$delete_collection_namespaced_horizontal_pod_autoscaler(
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAutoscalingV2NamespacedHorizontalPodAutoscaler
read the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-read_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-read_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{read_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$read_namespaced_horizontal_pod_autoscaler(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAutoscalingV2NamespacedHorizontalPodAutoscaler
replace the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-replace_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-replace_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{replace_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$replace_namespaced_horizontal_pod_autoscaler(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAutoscalingV2NamespacedHorizontalPodAutoscaler
partially update the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-patch_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-patch_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{patch_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$patch_namespaced_horizontal_pod_autoscaler(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAutoscalingV2NamespacedHorizontalPodAutoscaler
delete a HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-delete_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-delete_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{delete_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$delete_namespaced_horizontal_pod_autoscaler(
  name,
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAutoscalingV2NamespacedHorizontalPodAutoscalerStatus
read status of the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-read_namespaced_horizontal_pod_autoscaler_status"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-read_namespaced_horizontal_pod_autoscaler_status}{}}}
\subsection{Method \code{read_namespaced_horizontal_pod_autoscaler_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$read_namespaced_horizontal_pod_autoscaler_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAutoscalingV2NamespacedHorizontalPodAutoscalerStatus
replace status of the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-replace_namespaced_horizontal_pod_autoscaler_status"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-replace_namespaced_horizontal_pod_autoscaler_status}{}}}
\subsection{Method \code{replace_namespaced_horizontal_pod_autoscaler_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$replace_namespaced_horizontal_pod_autoscaler_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAutoscalingV2NamespacedHorizontalPodAutoscalerStatus
partially update status of the specified HorizontalPodAutoscaler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-patch_namespaced_horizontal_pod_autoscaler_status"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-patch_namespaced_horizontal_pod_autoscaler_status}{}}}
\subsection{Method \code{patch_namespaced_horizontal_pod_autoscaler_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$patch_namespaced_horizontal_pod_autoscaler_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchAutoscalingV2HorizontalPodAutoscalerListForAllNamespaces
watch individual changes to a list of HorizontalPodAutoscaler. deprecated: use the 'watch' parameter with a list operation instead.
watchAutoscalingV2NamespacedHorizontalPodAutoscalerList
watch individual changes to a list of HorizontalPodAutoscaler. deprecated: use the 'watch' parameter with a list operation instead.
watchAutoscalingV2NamespacedHorizontalPodAutoscaler
watch changes to an object of kind HorizontalPodAutoscaler. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-watch_horizontal_pod_autoscaler_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-watch_horizontal_pod_autoscaler_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_horizontal_pod_autoscaler_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$watch_horizontal_pod_autoscaler_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-watch_namespaced_horizontal_pod_autoscaler_list"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-watch_namespaced_horizontal_pod_autoscaler_list}{}}}
\subsection{Method \code{watch_namespaced_horizontal_pod_autoscaler_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$watch_namespaced_horizontal_pod_autoscaler_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-watch_namespaced_horizontal_pod_autoscaler"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-watch_namespaced_horizontal_pod_autoscaler}{}}}
\subsection{Method \code{watch_namespaced_horizontal_pod_autoscaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$watch_namespaced_horizontal_pod_autoscaler(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoscalingV2Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutoscalingV2Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoscalingV2Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
