% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_BatchApi.R
\name{BatchApi}
\alias{BatchApi}
\title{BatchApi}
\description{
BatchApi

BatchApi
}
\details{
Typed Kubernetes API client for the `Batch` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchApi-new}{\code{BatchApi$new()}}
\item \href{#method-BatchApi-get_api_group}{\code{BatchApi$get_api_group()}}
\item \href{#method-BatchApi-clone}{\code{BatchApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getBatchAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-BatchApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
