% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_BatchV1Api.R
\name{BatchV1Api}
\alias{BatchV1Api}
\title{BatchV1Api}
\description{
BatchV1Api

BatchV1Api
}
\details{
Typed Kubernetes API client for the `BatchV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchV1Api-new}{\code{BatchV1Api$new()}}
\item \href{#method-BatchV1Api-get_api_resources}{\code{BatchV1Api$get_api_resources()}}
\item \href{#method-BatchV1Api-list_cron_job_for_all_namespaces}{\code{BatchV1Api$list_cron_job_for_all_namespaces()}}
\item \href{#method-BatchV1Api-list_job_for_all_namespaces}{\code{BatchV1Api$list_job_for_all_namespaces()}}
\item \href{#method-BatchV1Api-list_namespaced_cron_job}{\code{BatchV1Api$list_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-create_namespaced_cron_job}{\code{BatchV1Api$create_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-delete_collection_namespaced_cron_job}{\code{BatchV1Api$delete_collection_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-read_namespaced_cron_job}{\code{BatchV1Api$read_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-replace_namespaced_cron_job}{\code{BatchV1Api$replace_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-patch_namespaced_cron_job}{\code{BatchV1Api$patch_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-delete_namespaced_cron_job}{\code{BatchV1Api$delete_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-read_namespaced_cron_job_status}{\code{BatchV1Api$read_namespaced_cron_job_status()}}
\item \href{#method-BatchV1Api-replace_namespaced_cron_job_status}{\code{BatchV1Api$replace_namespaced_cron_job_status()}}
\item \href{#method-BatchV1Api-patch_namespaced_cron_job_status}{\code{BatchV1Api$patch_namespaced_cron_job_status()}}
\item \href{#method-BatchV1Api-list_namespaced_job}{\code{BatchV1Api$list_namespaced_job()}}
\item \href{#method-BatchV1Api-create_namespaced_job}{\code{BatchV1Api$create_namespaced_job()}}
\item \href{#method-BatchV1Api-delete_collection_namespaced_job}{\code{BatchV1Api$delete_collection_namespaced_job()}}
\item \href{#method-BatchV1Api-read_namespaced_job}{\code{BatchV1Api$read_namespaced_job()}}
\item \href{#method-BatchV1Api-replace_namespaced_job}{\code{BatchV1Api$replace_namespaced_job()}}
\item \href{#method-BatchV1Api-patch_namespaced_job}{\code{BatchV1Api$patch_namespaced_job()}}
\item \href{#method-BatchV1Api-delete_namespaced_job}{\code{BatchV1Api$delete_namespaced_job()}}
\item \href{#method-BatchV1Api-read_namespaced_job_status}{\code{BatchV1Api$read_namespaced_job_status()}}
\item \href{#method-BatchV1Api-replace_namespaced_job_status}{\code{BatchV1Api$replace_namespaced_job_status()}}
\item \href{#method-BatchV1Api-patch_namespaced_job_status}{\code{BatchV1Api$patch_namespaced_job_status()}}
\item \href{#method-BatchV1Api-watch_cron_job_list_for_all_namespaces}{\code{BatchV1Api$watch_cron_job_list_for_all_namespaces()}}
\item \href{#method-BatchV1Api-watch_job_list_for_all_namespaces}{\code{BatchV1Api$watch_job_list_for_all_namespaces()}}
\item \href{#method-BatchV1Api-watch_namespaced_cron_job_list}{\code{BatchV1Api$watch_namespaced_cron_job_list()}}
\item \href{#method-BatchV1Api-watch_namespaced_cron_job}{\code{BatchV1Api$watch_namespaced_cron_job()}}
\item \href{#method-BatchV1Api-watch_namespaced_job_list}{\code{BatchV1Api$watch_namespaced_job_list()}}
\item \href{#method-BatchV1Api-watch_namespaced_job}{\code{BatchV1Api$watch_namespaced_job()}}
\item \href{#method-BatchV1Api-clone}{\code{BatchV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getBatchV1APIResources
get available resources
listBatchV1CronJobForAllNamespaces
list or watch objects of kind CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-list_cron_job_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-list_cron_job_for_all_namespaces}{}}}
\subsection{Method \code{list_cron_job_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$list_cron_job_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listBatchV1JobForAllNamespaces
list or watch objects of kind Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-list_job_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-list_job_for_all_namespaces}{}}}
\subsection{Method \code{list_job_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$list_job_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listBatchV1NamespacedCronJob
list or watch objects of kind CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-list_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-list_namespaced_cron_job}{}}}
\subsection{Method \code{list_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$list_namespaced_cron_job(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createBatchV1NamespacedCronJob
create a CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-create_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-create_namespaced_cron_job}{}}}
\subsection{Method \code{create_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$create_namespaced_cron_job(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteBatchV1CollectionNamespacedCronJob
delete collection of CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-delete_collection_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-delete_collection_namespaced_cron_job}{}}}
\subsection{Method \code{delete_collection_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$delete_collection_namespaced_cron_job(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readBatchV1NamespacedCronJob
read the specified CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-read_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-read_namespaced_cron_job}{}}}
\subsection{Method \code{read_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$read_namespaced_cron_job(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceBatchV1NamespacedCronJob
replace the specified CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-replace_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-replace_namespaced_cron_job}{}}}
\subsection{Method \code{replace_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$replace_namespaced_cron_job(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchBatchV1NamespacedCronJob
partially update the specified CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-patch_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-patch_namespaced_cron_job}{}}}
\subsection{Method \code{patch_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$patch_namespaced_cron_job(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteBatchV1NamespacedCronJob
delete a CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-delete_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-delete_namespaced_cron_job}{}}}
\subsection{Method \code{delete_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$delete_namespaced_cron_job(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readBatchV1NamespacedCronJobStatus
read status of the specified CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-read_namespaced_cron_job_status"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-read_namespaced_cron_job_status}{}}}
\subsection{Method \code{read_namespaced_cron_job_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$read_namespaced_cron_job_status(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceBatchV1NamespacedCronJobStatus
replace status of the specified CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-replace_namespaced_cron_job_status"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-replace_namespaced_cron_job_status}{}}}
\subsection{Method \code{replace_namespaced_cron_job_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$replace_namespaced_cron_job_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchBatchV1NamespacedCronJobStatus
partially update status of the specified CronJob}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-patch_namespaced_cron_job_status"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-patch_namespaced_cron_job_status}{}}}
\subsection{Method \code{patch_namespaced_cron_job_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$patch_namespaced_cron_job_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listBatchV1NamespacedJob
list or watch objects of kind Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-list_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-list_namespaced_job}{}}}
\subsection{Method \code{list_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$list_namespaced_job(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createBatchV1NamespacedJob
create a Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-create_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-create_namespaced_job}{}}}
\subsection{Method \code{create_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$create_namespaced_job(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteBatchV1CollectionNamespacedJob
delete collection of Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-delete_collection_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-delete_collection_namespaced_job}{}}}
\subsection{Method \code{delete_collection_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$delete_collection_namespaced_job(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readBatchV1NamespacedJob
read the specified Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-read_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-read_namespaced_job}{}}}
\subsection{Method \code{read_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$read_namespaced_job(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceBatchV1NamespacedJob
replace the specified Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-replace_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-replace_namespaced_job}{}}}
\subsection{Method \code{replace_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$replace_namespaced_job(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchBatchV1NamespacedJob
partially update the specified Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-patch_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-patch_namespaced_job}{}}}
\subsection{Method \code{patch_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$patch_namespaced_job(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteBatchV1NamespacedJob
delete a Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-delete_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-delete_namespaced_job}{}}}
\subsection{Method \code{delete_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$delete_namespaced_job(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readBatchV1NamespacedJobStatus
read status of the specified Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-read_namespaced_job_status"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-read_namespaced_job_status}{}}}
\subsection{Method \code{read_namespaced_job_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$read_namespaced_job_status(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceBatchV1NamespacedJobStatus
replace status of the specified Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-replace_namespaced_job_status"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-replace_namespaced_job_status}{}}}
\subsection{Method \code{replace_namespaced_job_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$replace_namespaced_job_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchBatchV1NamespacedJobStatus
partially update status of the specified Job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-patch_namespaced_job_status"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-patch_namespaced_job_status}{}}}
\subsection{Method \code{patch_namespaced_job_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$patch_namespaced_job_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
watchBatchV1CronJobListForAllNamespaces
watch individual changes to a list of CronJob. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-watch_cron_job_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-watch_cron_job_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_cron_job_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$watch_cron_job_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchBatchV1JobListForAllNamespaces
watch individual changes to a list of Job. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-watch_job_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-watch_job_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_job_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$watch_job_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchBatchV1NamespacedCronJobList
watch individual changes to a list of CronJob. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-watch_namespaced_cron_job_list"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-watch_namespaced_cron_job_list}{}}}
\subsection{Method \code{watch_namespaced_cron_job_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$watch_namespaced_cron_job_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchBatchV1NamespacedCronJob
watch changes to an object of kind CronJob. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-watch_namespaced_cron_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-watch_namespaced_cron_job}{}}}
\subsection{Method \code{watch_namespaced_cron_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$watch_namespaced_cron_job(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchBatchV1NamespacedJobList
watch individual changes to a list of Job. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-watch_namespaced_job_list"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-watch_namespaced_job_list}{}}}
\subsection{Method \code{watch_namespaced_job_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$watch_namespaced_job_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchBatchV1NamespacedJob
watch changes to an object of kind Job. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-watch_namespaced_job"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-watch_namespaced_job}{}}}
\subsection{Method \code{watch_namespaced_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$watch_namespaced_job(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
