% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CertificatesApi.R
\name{CertificatesApi}
\alias{CertificatesApi}
\title{CertificatesApi}
\description{
CertificatesApi

CertificatesApi
}
\details{
Typed Kubernetes API client for the `Certificates` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CertificatesApi-new}{\code{CertificatesApi$new()}}
\item \href{#method-CertificatesApi-get_api_group}{\code{CertificatesApi$get_api_group()}}
\item \href{#method-CertificatesApi-clone}{\code{CertificatesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCertificatesAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
