% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CoordinationApi.R
\name{CoordinationApi}
\alias{CoordinationApi}
\title{CoordinationApi}
\description{
CoordinationApi

CoordinationApi
}
\details{
Typed Kubernetes API client for the `Coordination` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordinationApi-new}{\code{CoordinationApi$new()}}
\item \href{#method-CoordinationApi-get_api_group}{\code{CoordinationApi$get_api_group()}}
\item \href{#method-CoordinationApi-clone}{\code{CoordinationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCoordinationAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
