% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CoordinationV1Api.R
\name{CoordinationV1Api}
\alias{CoordinationV1Api}
\title{CoordinationV1Api}
\description{
CoordinationV1Api

CoordinationV1Api
}
\details{
Typed Kubernetes API client for the `CoordinationV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordinationV1Api-new}{\code{CoordinationV1Api$new()}}
\item \href{#method-CoordinationV1Api-get_api_resources}{\code{CoordinationV1Api$get_api_resources()}}
\item \href{#method-CoordinationV1Api-list_lease_for_all_namespaces}{\code{CoordinationV1Api$list_lease_for_all_namespaces()}}
\item \href{#method-CoordinationV1Api-list_namespaced_lease}{\code{CoordinationV1Api$list_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-create_namespaced_lease}{\code{CoordinationV1Api$create_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-delete_collection_namespaced_lease}{\code{CoordinationV1Api$delete_collection_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-read_namespaced_lease}{\code{CoordinationV1Api$read_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-replace_namespaced_lease}{\code{CoordinationV1Api$replace_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-patch_namespaced_lease}{\code{CoordinationV1Api$patch_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-delete_namespaced_lease}{\code{CoordinationV1Api$delete_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-watch_lease_list_for_all_namespaces}{\code{CoordinationV1Api$watch_lease_list_for_all_namespaces()}}
\item \href{#method-CoordinationV1Api-watch_namespaced_lease_list}{\code{CoordinationV1Api$watch_namespaced_lease_list()}}
\item \href{#method-CoordinationV1Api-watch_namespaced_lease}{\code{CoordinationV1Api$watch_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-clone}{\code{CoordinationV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCoordinationV1APIResources
get available resources
listCoordinationV1LeaseForAllNamespaces
list or watch objects of kind Lease
listCoordinationV1NamespacedLease
list or watch objects of kind Lease
createCoordinationV1NamespacedLease
create a Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-list_lease_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-list_lease_for_all_namespaces}{}}}
\subsection{Method \code{list_lease_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$list_lease_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-list_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-list_namespaced_lease}{}}}
\subsection{Method \code{list_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$list_namespaced_lease()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-create_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-create_namespaced_lease}{}}}
\subsection{Method \code{create_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$create_namespaced_lease(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCoordinationV1CollectionNamespacedLease
delete collection of Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-delete_collection_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-delete_collection_namespaced_lease}{}}}
\subsection{Method \code{delete_collection_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$delete_collection_namespaced_lease(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readCoordinationV1NamespacedLease
read the specified Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-read_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-read_namespaced_lease}{}}}
\subsection{Method \code{read_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$read_namespaced_lease(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceCoordinationV1NamespacedLease
replace the specified Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-replace_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-replace_namespaced_lease}{}}}
\subsection{Method \code{replace_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$replace_namespaced_lease(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchCoordinationV1NamespacedLease
partially update the specified Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-patch_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-patch_namespaced_lease}{}}}
\subsection{Method \code{patch_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$patch_namespaced_lease(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCoordinationV1NamespacedLease
delete a Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-delete_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-delete_namespaced_lease}{}}}
\subsection{Method \code{delete_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$delete_namespaced_lease(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchCoordinationV1LeaseListForAllNamespaces
watch individual changes to a list of Lease. deprecated: use the 'watch' parameter with a list operation instead.
watchCoordinationV1NamespacedLeaseList
watch individual changes to a list of Lease. deprecated: use the 'watch' parameter with a list operation instead.
watchCoordinationV1NamespacedLease
watch changes to an object of kind Lease. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-watch_lease_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-watch_lease_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_lease_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$watch_lease_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-watch_namespaced_lease_list"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-watch_namespaced_lease_list}{}}}
\subsection{Method \code{watch_namespaced_lease_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$watch_namespaced_lease_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-watch_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-watch_namespaced_lease}{}}}
\subsection{Method \code{watch_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$watch_namespaced_lease(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
