% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CoreApi.R
\name{CoreApi}
\alias{CoreApi}
\title{CoreApi}
\description{
CoreApi

CoreApi
}
\details{
Typed Kubernetes API client for the `Core` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoreApi-new}{\code{CoreApi$new()}}
\item \href{#method-CoreApi-get_api_versions}{\code{CoreApi$get_api_versions()}}
\item \href{#method-CoreApi-clone}{\code{CoreApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoreApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCoreAPIVersions
get available API versions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreApi-get_api_versions"></a>}}
\if{latex}{\out{\hypertarget{method-CoreApi-get_api_versions}{}}}
\subsection{Method \code{get_api_versions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreApi$get_api_versions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoreApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
