% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_CoreV1EndpointPort.R
\name{CoreV1EndpointPort}
\alias{CoreV1EndpointPort}
\title{CoreV1EndpointPort}
\description{
CoreV1EndpointPort

CoreV1EndpointPort
}
\details{
EndpointPort is a tuple that describes a single port.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoreV1EndpointPort-new}{\code{CoreV1EndpointPort$new()}}
\item \href{#method-CoreV1EndpointPort-to_list}{\code{CoreV1EndpointPort$to_list()}}
\item \href{#method-CoreV1EndpointPort-clone}{\code{CoreV1EndpointPort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1EndpointPort-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1EndpointPort-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1EndpointPort$new(
  app_protocol = NULL,
  name = NULL,
  port = NULL,
  protocol = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1EndpointPort-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1EndpointPort-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1EndpointPort$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1EndpointPort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1EndpointPort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1EndpointPort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
