% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_CoreV1Event.R
\name{CoreV1Event}
\alias{CoreV1Event}
\title{CoreV1Event}
\description{
CoreV1Event

CoreV1Event
}
\details{
Event is a report of an event somewhere in the cluster.  Events have a limited retention time and triggers and messages may evolve with time.  Event consumers should not rely on the timing of an event with a given Reason reflecting a consistent underlying trigger, or the continued existence of events with that Reason.  Events should be treated as informative, best-effort, supplemental data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoreV1Event-new}{\code{CoreV1Event$new()}}
\item \href{#method-CoreV1Event-to_list}{\code{CoreV1Event$to_list()}}
\item \href{#method-CoreV1Event-clone}{\code{CoreV1Event$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1Event-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1Event-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1Event$new(
  action = NULL,
  api_version = NULL,
  count = NULL,
  event_time = NULL,
  first_timestamp = NULL,
  involved_object = NULL,
  kind = NULL,
  last_timestamp = NULL,
  message = NULL,
  metadata = NULL,
  reason = NULL,
  related = NULL,
  reporting_component = NULL,
  reporting_instance = NULL,
  series = NULL,
  source = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1Event-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1Event-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1Event$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1Event-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1Event-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1Event$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
