% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_CoreV1EventSeries.R
\name{CoreV1EventSeries}
\alias{CoreV1EventSeries}
\title{CoreV1EventSeries}
\description{
CoreV1EventSeries

CoreV1EventSeries
}
\details{
EventSeries contain information on series of events, i.e. thing that was/is happening continuously for some time.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoreV1EventSeries-new}{\code{CoreV1EventSeries$new()}}
\item \href{#method-CoreV1EventSeries-to_list}{\code{CoreV1EventSeries$to_list()}}
\item \href{#method-CoreV1EventSeries-clone}{\code{CoreV1EventSeries$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1EventSeries-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1EventSeries-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1EventSeries$new(count = NULL, last_observed_time = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1EventSeries-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1EventSeries-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1EventSeries$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreV1EventSeries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoreV1EventSeries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreV1EventSeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
