% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_from_yaml}
\alias{create_from_yaml}
\title{Create one or more Kubernetes resources from a YAML manifest}
\usage{
create_from_yaml(
  client,
  yaml_file = NULL,
  yaml_string = NULL,
  namespace = "default"
)
}
\arguments{
\item{client}{An \link{ApiClient}.}

\item{yaml_file}{Path to a YAML file; mutually exclusive with \code{yaml_string}.}

\item{yaml_string}{A character string of YAML content.}

\item{namespace}{Default namespace for namespaced resources that don't
specify one. Defaults to "default".}
}
\value{
A list of server responses, one per document.
}
\description{
Accepts a path or a character string containing one or more YAML documents
(separated by \verb{---}). Each document is routed to the correct endpoint via
the \link{DynamicClient}. Mirrors \code{kubernetes.utils.create_from_yaml}.
}
\examples{
\dontrun{
client <- new_client_from_config()
create_from_yaml(client, yaml_file = "manifest.yaml")
}
}
