% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_DiscoveryApi.R
\name{DiscoveryApi}
\alias{DiscoveryApi}
\title{DiscoveryApi}
\description{
DiscoveryApi

DiscoveryApi
}
\details{
Typed Kubernetes API client for the `Discovery` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiscoveryApi-new}{\code{DiscoveryApi$new()}}
\item \href{#method-DiscoveryApi-get_api_group}{\code{DiscoveryApi$get_api_group()}}
\item \href{#method-DiscoveryApi-clone}{\code{DiscoveryApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getDiscoveryAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
