% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_DiscoveryV1Api.R
\name{DiscoveryV1Api}
\alias{DiscoveryV1Api}
\title{DiscoveryV1Api}
\description{
DiscoveryV1Api

DiscoveryV1Api
}
\details{
Typed Kubernetes API client for the `DiscoveryV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiscoveryV1Api-new}{\code{DiscoveryV1Api$new()}}
\item \href{#method-DiscoveryV1Api-get_api_resources}{\code{DiscoveryV1Api$get_api_resources()}}
\item \href{#method-DiscoveryV1Api-list_endpoint_slice_for_all_namespaces}{\code{DiscoveryV1Api$list_endpoint_slice_for_all_namespaces()}}
\item \href{#method-DiscoveryV1Api-list_namespaced_endpoint_slice}{\code{DiscoveryV1Api$list_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-create_namespaced_endpoint_slice}{\code{DiscoveryV1Api$create_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-delete_collection_namespaced_endpoint_slice}{\code{DiscoveryV1Api$delete_collection_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-read_namespaced_endpoint_slice}{\code{DiscoveryV1Api$read_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-replace_namespaced_endpoint_slice}{\code{DiscoveryV1Api$replace_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-patch_namespaced_endpoint_slice}{\code{DiscoveryV1Api$patch_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-delete_namespaced_endpoint_slice}{\code{DiscoveryV1Api$delete_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-watch_endpoint_slice_list_for_all_namespaces}{\code{DiscoveryV1Api$watch_endpoint_slice_list_for_all_namespaces()}}
\item \href{#method-DiscoveryV1Api-watch_namespaced_endpoint_slice_list}{\code{DiscoveryV1Api$watch_namespaced_endpoint_slice_list()}}
\item \href{#method-DiscoveryV1Api-watch_namespaced_endpoint_slice}{\code{DiscoveryV1Api$watch_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-clone}{\code{DiscoveryV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getDiscoveryV1APIResources
get available resources
listDiscoveryV1EndpointSliceForAllNamespaces
list or watch objects of kind EndpointSlice
listDiscoveryV1NamespacedEndpointSlice
list or watch objects of kind EndpointSlice
createDiscoveryV1NamespacedEndpointSlice
create an EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-list_endpoint_slice_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-list_endpoint_slice_for_all_namespaces}{}}}
\subsection{Method \code{list_endpoint_slice_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$list_endpoint_slice_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-list_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-list_namespaced_endpoint_slice}{}}}
\subsection{Method \code{list_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$list_namespaced_endpoint_slice()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-create_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-create_namespaced_endpoint_slice}{}}}
\subsection{Method \code{create_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$create_namespaced_endpoint_slice(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteDiscoveryV1CollectionNamespacedEndpointSlice
delete collection of EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-delete_collection_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-delete_collection_namespaced_endpoint_slice}{}}}
\subsection{Method \code{delete_collection_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$delete_collection_namespaced_endpoint_slice(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readDiscoveryV1NamespacedEndpointSlice
read the specified EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-read_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-read_namespaced_endpoint_slice}{}}}
\subsection{Method \code{read_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$read_namespaced_endpoint_slice(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceDiscoveryV1NamespacedEndpointSlice
replace the specified EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-replace_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-replace_namespaced_endpoint_slice}{}}}
\subsection{Method \code{replace_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$replace_namespaced_endpoint_slice(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchDiscoveryV1NamespacedEndpointSlice
partially update the specified EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-patch_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-patch_namespaced_endpoint_slice}{}}}
\subsection{Method \code{patch_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$patch_namespaced_endpoint_slice(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteDiscoveryV1NamespacedEndpointSlice
delete an EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-delete_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-delete_namespaced_endpoint_slice}{}}}
\subsection{Method \code{delete_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$delete_namespaced_endpoint_slice(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchDiscoveryV1EndpointSliceListForAllNamespaces
watch individual changes to a list of EndpointSlice. deprecated: use the 'watch' parameter with a list operation instead.
watchDiscoveryV1NamespacedEndpointSliceList
watch individual changes to a list of EndpointSlice. deprecated: use the 'watch' parameter with a list operation instead.
watchDiscoveryV1NamespacedEndpointSlice
watch changes to an object of kind EndpointSlice. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-watch_endpoint_slice_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-watch_endpoint_slice_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_endpoint_slice_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$watch_endpoint_slice_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-watch_namespaced_endpoint_slice_list"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-watch_namespaced_endpoint_slice_list}{}}}
\subsection{Method \code{watch_namespaced_endpoint_slice_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$watch_namespaced_endpoint_slice_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-watch_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-watch_namespaced_endpoint_slice}{}}}
\subsection{Method \code{watch_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$watch_namespaced_endpoint_slice(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
