% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_EventsV1Api.R
\name{EventsV1Api}
\alias{EventsV1Api}
\title{EventsV1Api}
\description{
EventsV1Api

EventsV1Api
}
\details{
Typed Kubernetes API client for the `EventsV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventsV1Api-new}{\code{EventsV1Api$new()}}
\item \href{#method-EventsV1Api-get_api_resources}{\code{EventsV1Api$get_api_resources()}}
\item \href{#method-EventsV1Api-list_event_for_all_namespaces}{\code{EventsV1Api$list_event_for_all_namespaces()}}
\item \href{#method-EventsV1Api-list_namespaced_event}{\code{EventsV1Api$list_namespaced_event()}}
\item \href{#method-EventsV1Api-create_namespaced_event}{\code{EventsV1Api$create_namespaced_event()}}
\item \href{#method-EventsV1Api-delete_collection_namespaced_event}{\code{EventsV1Api$delete_collection_namespaced_event()}}
\item \href{#method-EventsV1Api-read_namespaced_event}{\code{EventsV1Api$read_namespaced_event()}}
\item \href{#method-EventsV1Api-replace_namespaced_event}{\code{EventsV1Api$replace_namespaced_event()}}
\item \href{#method-EventsV1Api-patch_namespaced_event}{\code{EventsV1Api$patch_namespaced_event()}}
\item \href{#method-EventsV1Api-delete_namespaced_event}{\code{EventsV1Api$delete_namespaced_event()}}
\item \href{#method-EventsV1Api-watch_event_list_for_all_namespaces}{\code{EventsV1Api$watch_event_list_for_all_namespaces()}}
\item \href{#method-EventsV1Api-watch_namespaced_event_list}{\code{EventsV1Api$watch_namespaced_event_list()}}
\item \href{#method-EventsV1Api-watch_namespaced_event}{\code{EventsV1Api$watch_namespaced_event()}}
\item \href{#method-EventsV1Api-clone}{\code{EventsV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getEventsV1APIResources
get available resources
listEventsV1EventForAllNamespaces
list or watch objects of kind Event
listEventsV1NamespacedEvent
list or watch objects of kind Event
createEventsV1NamespacedEvent
create an Event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-list_event_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-list_event_for_all_namespaces}{}}}
\subsection{Method \code{list_event_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$list_event_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-list_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-list_namespaced_event}{}}}
\subsection{Method \code{list_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$list_namespaced_event()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-create_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-create_namespaced_event}{}}}
\subsection{Method \code{create_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$create_namespaced_event(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteEventsV1CollectionNamespacedEvent
delete collection of Event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-delete_collection_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-delete_collection_namespaced_event}{}}}
\subsection{Method \code{delete_collection_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$delete_collection_namespaced_event(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readEventsV1NamespacedEvent
read the specified Event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-read_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-read_namespaced_event}{}}}
\subsection{Method \code{read_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$read_namespaced_event(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceEventsV1NamespacedEvent
replace the specified Event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-replace_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-replace_namespaced_event}{}}}
\subsection{Method \code{replace_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$replace_namespaced_event(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchEventsV1NamespacedEvent
partially update the specified Event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-patch_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-patch_namespaced_event}{}}}
\subsection{Method \code{patch_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$patch_namespaced_event(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteEventsV1NamespacedEvent
delete an Event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-delete_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-delete_namespaced_event}{}}}
\subsection{Method \code{delete_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$delete_namespaced_event(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchEventsV1EventListForAllNamespaces
watch individual changes to a list of Event. deprecated: use the 'watch' parameter with a list operation instead.
watchEventsV1NamespacedEventList
watch individual changes to a list of Event. deprecated: use the 'watch' parameter with a list operation instead.
watchEventsV1NamespacedEvent
watch changes to an object of kind Event. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-watch_event_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-watch_event_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_event_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$watch_event_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-watch_namespaced_event_list"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-watch_namespaced_event_list}{}}}
\subsection{Method \code{watch_namespaced_event_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$watch_namespaced_event_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-watch_namespaced_event"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-watch_namespaced_event}{}}}
\subsection{Method \code{watch_namespaced_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$watch_namespaced_event(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
