% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_EventsV1Event.R
\name{EventsV1Event}
\alias{EventsV1Event}
\title{EventsV1Event}
\description{
EventsV1Event

EventsV1Event
}
\details{
Event is a report of an event somewhere in the cluster. It generally denotes some state change in the system. Events have a limited retention time and triggers and messages may evolve with time.  Event consumers should not rely on the timing of an event with a given Reason reflecting a consistent underlying trigger, or the continued existence of events with that Reason.  Events should be treated as informative, best-effort, supplemental data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventsV1Event-new}{\code{EventsV1Event$new()}}
\item \href{#method-EventsV1Event-to_list}{\code{EventsV1Event$to_list()}}
\item \href{#method-EventsV1Event-clone}{\code{EventsV1Event$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Event-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Event-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Event$new(
  action = NULL,
  api_version = NULL,
  deprecated_count = NULL,
  deprecated_first_timestamp = NULL,
  deprecated_last_timestamp = NULL,
  deprecated_source = NULL,
  event_time = NULL,
  kind = NULL,
  metadata = NULL,
  note = NULL,
  reason = NULL,
  regarding = NULL,
  related = NULL,
  reporting_controller = NULL,
  reporting_instance = NULL,
  series = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Event-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Event-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Event$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1Event-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1Event-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1Event$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
