% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_EventsV1EventList.R
\name{EventsV1EventList}
\alias{EventsV1EventList}
\title{EventsV1EventList}
\description{
EventsV1EventList

EventsV1EventList
}
\details{
EventList is a list of Event objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventsV1EventList-new}{\code{EventsV1EventList$new()}}
\item \href{#method-EventsV1EventList-to_list}{\code{EventsV1EventList$to_list()}}
\item \href{#method-EventsV1EventList-clone}{\code{EventsV1EventList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1EventList-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1EventList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1EventList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1EventList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1EventList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1EventList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1EventList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1EventList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1EventList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
