% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_EventsV1EventSeries.R
\name{EventsV1EventSeries}
\alias{EventsV1EventSeries}
\title{EventsV1EventSeries}
\description{
EventsV1EventSeries

EventsV1EventSeries
}
\details{
EventSeries contain information on series of events, i.e. thing that was/is happening continuously for some time. How often to update the EventSeries is up to the event reporters. The default event reporter in "k8s.io/client-go/tools/events/event_broadcaster.go" shows how this struct is updated on heartbeats and can guide customized reporter implementations.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventsV1EventSeries-new}{\code{EventsV1EventSeries$new()}}
\item \href{#method-EventsV1EventSeries-to_list}{\code{EventsV1EventSeries$to_list()}}
\item \href{#method-EventsV1EventSeries-clone}{\code{EventsV1EventSeries$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1EventSeries-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1EventSeries-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1EventSeries$new(count = NULL, last_observed_time = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1EventSeries-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1EventSeries-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1EventSeries$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventsV1EventSeries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventsV1EventSeries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsV1EventSeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
