% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_FlowcontrolApiserverApi.R
\name{FlowcontrolApiserverApi}
\alias{FlowcontrolApiserverApi}
\title{FlowcontrolApiserverApi}
\description{
FlowcontrolApiserverApi

FlowcontrolApiserverApi
}
\details{
Typed Kubernetes API client for the `FlowcontrolApiserver` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlowcontrolApiserverApi-new}{\code{FlowcontrolApiserverApi$new()}}
\item \href{#method-FlowcontrolApiserverApi-get_api_group}{\code{FlowcontrolApiserverApi$get_api_group()}}
\item \href{#method-FlowcontrolApiserverApi-clone}{\code{FlowcontrolApiserverApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getFlowcontrolApiserverAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
