% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_FlowcontrolApiserverV1Api.R
\name{FlowcontrolApiserverV1Api}
\alias{FlowcontrolApiserverV1Api}
\title{FlowcontrolApiserverV1Api}
\description{
FlowcontrolApiserverV1Api

FlowcontrolApiserverV1Api
}
\details{
Typed Kubernetes API client for the `FlowcontrolApiserverV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlowcontrolApiserverV1Api-new}{\code{FlowcontrolApiserverV1Api$new()}}
\item \href{#method-FlowcontrolApiserverV1Api-get_api_resources}{\code{FlowcontrolApiserverV1Api$get_api_resources()}}
\item \href{#method-FlowcontrolApiserverV1Api-list_flow_schema}{\code{FlowcontrolApiserverV1Api$list_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-create_flow_schema}{\code{FlowcontrolApiserverV1Api$create_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-delete_collection_flow_schema}{\code{FlowcontrolApiserverV1Api$delete_collection_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-read_flow_schema}{\code{FlowcontrolApiserverV1Api$read_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-replace_flow_schema}{\code{FlowcontrolApiserverV1Api$replace_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-patch_flow_schema}{\code{FlowcontrolApiserverV1Api$patch_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-delete_flow_schema}{\code{FlowcontrolApiserverV1Api$delete_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-read_flow_schema_status}{\code{FlowcontrolApiserverV1Api$read_flow_schema_status()}}
\item \href{#method-FlowcontrolApiserverV1Api-replace_flow_schema_status}{\code{FlowcontrolApiserverV1Api$replace_flow_schema_status()}}
\item \href{#method-FlowcontrolApiserverV1Api-patch_flow_schema_status}{\code{FlowcontrolApiserverV1Api$patch_flow_schema_status()}}
\item \href{#method-FlowcontrolApiserverV1Api-list_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$list_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-create_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$create_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-delete_collection_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$delete_collection_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-read_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$read_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-replace_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$replace_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-patch_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$patch_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-delete_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$delete_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-read_priority_level_configuration_status}{\code{FlowcontrolApiserverV1Api$read_priority_level_configuration_status()}}
\item \href{#method-FlowcontrolApiserverV1Api-replace_priority_level_configuration_status}{\code{FlowcontrolApiserverV1Api$replace_priority_level_configuration_status()}}
\item \href{#method-FlowcontrolApiserverV1Api-patch_priority_level_configuration_status}{\code{FlowcontrolApiserverV1Api$patch_priority_level_configuration_status()}}
\item \href{#method-FlowcontrolApiserverV1Api-watch_flow_schema_list}{\code{FlowcontrolApiserverV1Api$watch_flow_schema_list()}}
\item \href{#method-FlowcontrolApiserverV1Api-watch_flow_schema}{\code{FlowcontrolApiserverV1Api$watch_flow_schema()}}
\item \href{#method-FlowcontrolApiserverV1Api-watch_priority_level_configuration_list}{\code{FlowcontrolApiserverV1Api$watch_priority_level_configuration_list()}}
\item \href{#method-FlowcontrolApiserverV1Api-watch_priority_level_configuration}{\code{FlowcontrolApiserverV1Api$watch_priority_level_configuration()}}
\item \href{#method-FlowcontrolApiserverV1Api-clone}{\code{FlowcontrolApiserverV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getFlowcontrolApiserverV1APIResources
get available resources
listFlowcontrolApiserverV1FlowSchema
list or watch objects of kind FlowSchema
createFlowcontrolApiserverV1FlowSchema
create a FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-list_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-list_flow_schema}{}}}
\subsection{Method \code{list_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$list_flow_schema()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-create_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-create_flow_schema}{}}}
\subsection{Method \code{create_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$create_flow_schema(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteFlowcontrolApiserverV1CollectionFlowSchema
delete collection of FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-delete_collection_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-delete_collection_flow_schema}{}}}
\subsection{Method \code{delete_collection_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$delete_collection_flow_schema(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readFlowcontrolApiserverV1FlowSchema
read the specified FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-read_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-read_flow_schema}{}}}
\subsection{Method \code{read_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$read_flow_schema(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceFlowcontrolApiserverV1FlowSchema
replace the specified FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-replace_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-replace_flow_schema}{}}}
\subsection{Method \code{replace_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$replace_flow_schema(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchFlowcontrolApiserverV1FlowSchema
partially update the specified FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-patch_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-patch_flow_schema}{}}}
\subsection{Method \code{patch_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$patch_flow_schema(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteFlowcontrolApiserverV1FlowSchema
delete a FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-delete_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-delete_flow_schema}{}}}
\subsection{Method \code{delete_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$delete_flow_schema(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readFlowcontrolApiserverV1FlowSchemaStatus
read status of the specified FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-read_flow_schema_status"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-read_flow_schema_status}{}}}
\subsection{Method \code{read_flow_schema_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$read_flow_schema_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceFlowcontrolApiserverV1FlowSchemaStatus
replace status of the specified FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-replace_flow_schema_status"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-replace_flow_schema_status}{}}}
\subsection{Method \code{replace_flow_schema_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$replace_flow_schema_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchFlowcontrolApiserverV1FlowSchemaStatus
partially update status of the specified FlowSchema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-patch_flow_schema_status"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-patch_flow_schema_status}{}}}
\subsection{Method \code{patch_flow_schema_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$patch_flow_schema_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
listFlowcontrolApiserverV1PriorityLevelConfiguration
list or watch objects of kind PriorityLevelConfiguration
createFlowcontrolApiserverV1PriorityLevelConfiguration
create a PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-list_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-list_priority_level_configuration}{}}}
\subsection{Method \code{list_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$list_priority_level_configuration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-create_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-create_priority_level_configuration}{}}}
\subsection{Method \code{create_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$create_priority_level_configuration(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteFlowcontrolApiserverV1CollectionPriorityLevelConfiguration
delete collection of PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-delete_collection_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-delete_collection_priority_level_configuration}{}}}
\subsection{Method \code{delete_collection_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$delete_collection_priority_level_configuration(
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readFlowcontrolApiserverV1PriorityLevelConfiguration
read the specified PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-read_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-read_priority_level_configuration}{}}}
\subsection{Method \code{read_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$read_priority_level_configuration(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceFlowcontrolApiserverV1PriorityLevelConfiguration
replace the specified PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-replace_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-replace_priority_level_configuration}{}}}
\subsection{Method \code{replace_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$replace_priority_level_configuration(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchFlowcontrolApiserverV1PriorityLevelConfiguration
partially update the specified PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-patch_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-patch_priority_level_configuration}{}}}
\subsection{Method \code{patch_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$patch_priority_level_configuration(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteFlowcontrolApiserverV1PriorityLevelConfiguration
delete a PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-delete_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-delete_priority_level_configuration}{}}}
\subsection{Method \code{delete_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$delete_priority_level_configuration(
  name,
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readFlowcontrolApiserverV1PriorityLevelConfigurationStatus
read status of the specified PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-read_priority_level_configuration_status"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-read_priority_level_configuration_status}{}}}
\subsection{Method \code{read_priority_level_configuration_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$read_priority_level_configuration_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceFlowcontrolApiserverV1PriorityLevelConfigurationStatus
replace status of the specified PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-replace_priority_level_configuration_status"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-replace_priority_level_configuration_status}{}}}
\subsection{Method \code{replace_priority_level_configuration_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$replace_priority_level_configuration_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchFlowcontrolApiserverV1PriorityLevelConfigurationStatus
partially update status of the specified PriorityLevelConfiguration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-patch_priority_level_configuration_status"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-patch_priority_level_configuration_status}{}}}
\subsection{Method \code{patch_priority_level_configuration_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$patch_priority_level_configuration_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchFlowcontrolApiserverV1FlowSchemaList
watch individual changes to a list of FlowSchema. deprecated: use the 'watch' parameter with a list operation instead.
watchFlowcontrolApiserverV1FlowSchema
watch changes to an object of kind FlowSchema. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-watch_flow_schema_list"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-watch_flow_schema_list}{}}}
\subsection{Method \code{watch_flow_schema_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$watch_flow_schema_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-watch_flow_schema"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-watch_flow_schema}{}}}
\subsection{Method \code{watch_flow_schema()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$watch_flow_schema(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchFlowcontrolApiserverV1PriorityLevelConfigurationList
watch individual changes to a list of PriorityLevelConfiguration. deprecated: use the 'watch' parameter with a list operation instead.
watchFlowcontrolApiserverV1PriorityLevelConfiguration
watch changes to an object of kind PriorityLevelConfiguration. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-watch_priority_level_configuration_list"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-watch_priority_level_configuration_list}{}}}
\subsection{Method \code{watch_priority_level_configuration_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$watch_priority_level_configuration_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-watch_priority_level_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-watch_priority_level_configuration}{}}}
\subsection{Method \code{watch_priority_level_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$watch_priority_level_configuration(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolApiserverV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolApiserverV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolApiserverV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
