% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_FlowcontrolV1Subject.R
\name{FlowcontrolV1Subject}
\alias{FlowcontrolV1Subject}
\title{FlowcontrolV1Subject}
\description{
FlowcontrolV1Subject

FlowcontrolV1Subject
}
\details{
Subject matches the originator of a request, as identified by the request authentication system. There are three ways of matching an originator; by user, group, or service account.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlowcontrolV1Subject-new}{\code{FlowcontrolV1Subject$new()}}
\item \href{#method-FlowcontrolV1Subject-to_list}{\code{FlowcontrolV1Subject$to_list()}}
\item \href{#method-FlowcontrolV1Subject-clone}{\code{FlowcontrolV1Subject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolV1Subject-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolV1Subject-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolV1Subject$new(
  group = NULL,
  kind = NULL,
  service_account = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolV1Subject-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolV1Subject-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolV1Subject$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowcontrolV1Subject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlowcontrolV1Subject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowcontrolV1Subject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
