% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_Info.R
\name{Info}
\alias{Info}
\title{Info}
\description{
Info

Info
}
\details{
Info contains versioning information. how we'll want to distribute that information.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Info-new}{\code{Info$new()}}
\item \href{#method-Info-to_list}{\code{Info$to_list()}}
\item \href{#method-Info-clone}{\code{Info$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-new"></a>}}
\if{latex}{\out{\hypertarget{method-Info-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$new(
  build_date = NULL,
  compiler = NULL,
  git_commit = NULL,
  git_tree_state = NULL,
  git_version = NULL,
  go_version = NULL,
  major = NULL,
  minor = NULL,
  platform = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Info-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Info-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Info-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Info$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
