% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_InternalApiserverApi.R
\name{InternalApiserverApi}
\alias{InternalApiserverApi}
\title{InternalApiserverApi}
\description{
InternalApiserverApi

InternalApiserverApi
}
\details{
Typed Kubernetes API client for the `InternalApiserver` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InternalApiserverApi-new}{\code{InternalApiserverApi$new()}}
\item \href{#method-InternalApiserverApi-get_api_group}{\code{InternalApiserverApi$get_api_group()}}
\item \href{#method-InternalApiserverApi-clone}{\code{InternalApiserverApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getInternalApiserverAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
