% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_InternalApiserverV1alpha1Api.R
\name{InternalApiserverV1alpha1Api}
\alias{InternalApiserverV1alpha1Api}
\title{InternalApiserverV1alpha1Api}
\description{
InternalApiserverV1alpha1Api

InternalApiserverV1alpha1Api
}
\details{
Typed Kubernetes API client for the `InternalApiserverV1alpha1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InternalApiserverV1alpha1Api-new}{\code{InternalApiserverV1alpha1Api$new()}}
\item \href{#method-InternalApiserverV1alpha1Api-get_api_resources}{\code{InternalApiserverV1alpha1Api$get_api_resources()}}
\item \href{#method-InternalApiserverV1alpha1Api-list_storage_version}{\code{InternalApiserverV1alpha1Api$list_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-create_storage_version}{\code{InternalApiserverV1alpha1Api$create_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-delete_collection_storage_version}{\code{InternalApiserverV1alpha1Api$delete_collection_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-read_storage_version}{\code{InternalApiserverV1alpha1Api$read_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-replace_storage_version}{\code{InternalApiserverV1alpha1Api$replace_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-patch_storage_version}{\code{InternalApiserverV1alpha1Api$patch_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-delete_storage_version}{\code{InternalApiserverV1alpha1Api$delete_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-read_storage_version_status}{\code{InternalApiserverV1alpha1Api$read_storage_version_status()}}
\item \href{#method-InternalApiserverV1alpha1Api-replace_storage_version_status}{\code{InternalApiserverV1alpha1Api$replace_storage_version_status()}}
\item \href{#method-InternalApiserverV1alpha1Api-patch_storage_version_status}{\code{InternalApiserverV1alpha1Api$patch_storage_version_status()}}
\item \href{#method-InternalApiserverV1alpha1Api-watch_storage_version_list}{\code{InternalApiserverV1alpha1Api$watch_storage_version_list()}}
\item \href{#method-InternalApiserverV1alpha1Api-watch_storage_version}{\code{InternalApiserverV1alpha1Api$watch_storage_version()}}
\item \href{#method-InternalApiserverV1alpha1Api-clone}{\code{InternalApiserverV1alpha1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getInternalApiserverV1alpha1APIResources
get available resources
listInternalApiserverV1alpha1StorageVersion
list or watch objects of kind StorageVersion
createInternalApiserverV1alpha1StorageVersion
create a StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-list_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-list_storage_version}{}}}
\subsection{Method \code{list_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$list_storage_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-create_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-create_storage_version}{}}}
\subsection{Method \code{create_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$create_storage_version(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteInternalApiserverV1alpha1CollectionStorageVersion
delete collection of StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-delete_collection_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-delete_collection_storage_version}{}}}
\subsection{Method \code{delete_collection_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$delete_collection_storage_version(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readInternalApiserverV1alpha1StorageVersion
read the specified StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-read_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-read_storage_version}{}}}
\subsection{Method \code{read_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$read_storage_version(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceInternalApiserverV1alpha1StorageVersion
replace the specified StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-replace_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-replace_storage_version}{}}}
\subsection{Method \code{replace_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$replace_storage_version(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchInternalApiserverV1alpha1StorageVersion
partially update the specified StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-patch_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-patch_storage_version}{}}}
\subsection{Method \code{patch_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$patch_storage_version(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteInternalApiserverV1alpha1StorageVersion
delete a StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-delete_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-delete_storage_version}{}}}
\subsection{Method \code{delete_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$delete_storage_version(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readInternalApiserverV1alpha1StorageVersionStatus
read status of the specified StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-read_storage_version_status"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-read_storage_version_status}{}}}
\subsection{Method \code{read_storage_version_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$read_storage_version_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceInternalApiserverV1alpha1StorageVersionStatus
replace status of the specified StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-replace_storage_version_status"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-replace_storage_version_status}{}}}
\subsection{Method \code{replace_storage_version_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$replace_storage_version_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchInternalApiserverV1alpha1StorageVersionStatus
partially update status of the specified StorageVersion}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-patch_storage_version_status"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-patch_storage_version_status}{}}}
\subsection{Method \code{patch_storage_version_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$patch_storage_version_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchInternalApiserverV1alpha1StorageVersionList
watch individual changes to a list of StorageVersion. deprecated: use the 'watch' parameter with a list operation instead.
watchInternalApiserverV1alpha1StorageVersion
watch changes to an object of kind StorageVersion. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-watch_storage_version_list"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-watch_storage_version_list}{}}}
\subsection{Method \code{watch_storage_version_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$watch_storage_version_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-watch_storage_version"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-watch_storage_version}{}}}
\subsection{Method \code{watch_storage_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$watch_storage_version(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternalApiserverV1alpha1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InternalApiserverV1alpha1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalApiserverV1alpha1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
