% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_IntOrString.R
\name{IntOrString}
\alias{IntOrString}
\title{IntOrString}
\description{
IntOrString

IntOrString
}
\details{
IntOrString is a type that can hold an int32 or a string.  When used in JSON or YAML marshalling and unmarshalling, it produces or consumes the inner type.  This allows you to have, for example, a JSON field that can accept a name or number.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IntOrString-new}{\code{IntOrString$new()}}
\item \href{#method-IntOrString-to_list}{\code{IntOrString$to_list()}}
\item \href{#method-IntOrString-clone}{\code{IntOrString$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntOrString-new"></a>}}
\if{latex}{\out{\hypertarget{method-IntOrString-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntOrString$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntOrString-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-IntOrString-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntOrString$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntOrString-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IntOrString-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntOrString$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
