% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informer.R
\name{Lister}
\alias{Lister}
\title{Read-only indexed view over an Informer's cache}
\description{
Read-only indexed view over an Informer's cache

Read-only indexed view over an Informer's cache
}
\details{
A \code{Lister} is a cheap read-only wrapper over the cache an \link{Informer}
maintains. Use it in controllers to answer "does an object by this name
already exist?" without hitting the API server on every tick.

Mirrors client-go's \code{Lister} / \code{NamespaceLister} interfaces.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lister-new}{\code{Lister$new()}}
\item \href{#method-Lister-list}{\code{Lister$list()}}
\item \href{#method-Lister-namespace}{\code{Lister$namespace()}}
\item \href{#method-Lister-get}{\code{Lister$get()}}
\item \href{#method-Lister-clone}{\code{Lister$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lister-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lister-new}{}}}
\subsection{Method \code{new()}}{
Construct (usually via \code{Informer$lister()}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lister$new(store)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{store}}{An environment holding \code{namespace/name -> object}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lister-list"></a>}}
\if{latex}{\out{\hypertarget{method-Lister-list}{}}}
\subsection{Method \code{list()}}{
All objects in the cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lister$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lister-namespace"></a>}}
\if{latex}{\out{\hypertarget{method-Lister-namespace}{}}}
\subsection{Method \code{namespace()}}{
Objects scoped to a namespace.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lister$namespace(namespace)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{Namespace to filter on.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lister-get"></a>}}
\if{latex}{\out{\hypertarget{method-Lister-get}{}}}
\subsection{Method \code{get()}}{
Fetch a single object by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lister$get(name, namespace = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Object name.}

\item{\code{namespace}}{Optional namespace (for namespaced resources).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lister-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lister-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lister$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
