% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_LogsApi.R
\name{LogsApi}
\alias{LogsApi}
\title{LogsApi}
\description{
LogsApi

LogsApi
}
\details{
Typed Kubernetes API client for the `Logs` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogsApi-new}{\code{LogsApi$new()}}
\item \href{#method-LogsApi-log_file_list_handler}{\code{LogsApi$log_file_list_handler()}}
\item \href{#method-LogsApi-log_file_handler}{\code{LogsApi$log_file_handler()}}
\item \href{#method-LogsApi-clone}{\code{LogsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogsApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogsApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
logFileListHandler
logFileHandler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogsApi-log_file_list_handler"></a>}}
\if{latex}{\out{\hypertarget{method-LogsApi-log_file_list_handler}{}}}
\subsection{Method \code{log_file_list_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogsApi$log_file_list_handler()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogsApi-log_file_handler"></a>}}
\if{latex}{\out{\hypertarget{method-LogsApi-log_file_handler}{}}}
\subsection{Method \code{log_file_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogsApi$log_file_handler()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
