% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_NetworkingApi.R
\name{NetworkingApi}
\alias{NetworkingApi}
\title{NetworkingApi}
\description{
NetworkingApi

NetworkingApi
}
\details{
Typed Kubernetes API client for the `Networking` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NetworkingApi-new}{\code{NetworkingApi$new()}}
\item \href{#method-NetworkingApi-get_api_group}{\code{NetworkingApi$get_api_group()}}
\item \href{#method-NetworkingApi-clone}{\code{NetworkingApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getNetworkingAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
