% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_NetworkingV1Api.R
\name{NetworkingV1Api}
\alias{NetworkingV1Api}
\title{NetworkingV1Api}
\description{
NetworkingV1Api

NetworkingV1Api
}
\details{
Typed Kubernetes API client for the `NetworkingV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NetworkingV1Api-new}{\code{NetworkingV1Api$new()}}
\item \href{#method-NetworkingV1Api-get_api_resources}{\code{NetworkingV1Api$get_api_resources()}}
\item \href{#method-NetworkingV1Api-list_ingress_class}{\code{NetworkingV1Api$list_ingress_class()}}
\item \href{#method-NetworkingV1Api-create_ingress_class}{\code{NetworkingV1Api$create_ingress_class()}}
\item \href{#method-NetworkingV1Api-delete_collection_ingress_class}{\code{NetworkingV1Api$delete_collection_ingress_class()}}
\item \href{#method-NetworkingV1Api-read_ingress_class}{\code{NetworkingV1Api$read_ingress_class()}}
\item \href{#method-NetworkingV1Api-replace_ingress_class}{\code{NetworkingV1Api$replace_ingress_class()}}
\item \href{#method-NetworkingV1Api-patch_ingress_class}{\code{NetworkingV1Api$patch_ingress_class()}}
\item \href{#method-NetworkingV1Api-delete_ingress_class}{\code{NetworkingV1Api$delete_ingress_class()}}
\item \href{#method-NetworkingV1Api-list_ingress_for_all_namespaces}{\code{NetworkingV1Api$list_ingress_for_all_namespaces()}}
\item \href{#method-NetworkingV1Api-list_namespaced_ingress}{\code{NetworkingV1Api$list_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-create_namespaced_ingress}{\code{NetworkingV1Api$create_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-delete_collection_namespaced_ingress}{\code{NetworkingV1Api$delete_collection_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-read_namespaced_ingress}{\code{NetworkingV1Api$read_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-replace_namespaced_ingress}{\code{NetworkingV1Api$replace_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-patch_namespaced_ingress}{\code{NetworkingV1Api$patch_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-delete_namespaced_ingress}{\code{NetworkingV1Api$delete_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-read_namespaced_ingress_status}{\code{NetworkingV1Api$read_namespaced_ingress_status()}}
\item \href{#method-NetworkingV1Api-replace_namespaced_ingress_status}{\code{NetworkingV1Api$replace_namespaced_ingress_status()}}
\item \href{#method-NetworkingV1Api-patch_namespaced_ingress_status}{\code{NetworkingV1Api$patch_namespaced_ingress_status()}}
\item \href{#method-NetworkingV1Api-list_namespaced_network_policy}{\code{NetworkingV1Api$list_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-create_namespaced_network_policy}{\code{NetworkingV1Api$create_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-delete_collection_namespaced_network_policy}{\code{NetworkingV1Api$delete_collection_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-read_namespaced_network_policy}{\code{NetworkingV1Api$read_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-replace_namespaced_network_policy}{\code{NetworkingV1Api$replace_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-patch_namespaced_network_policy}{\code{NetworkingV1Api$patch_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-delete_namespaced_network_policy}{\code{NetworkingV1Api$delete_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-list_network_policy_for_all_namespaces}{\code{NetworkingV1Api$list_network_policy_for_all_namespaces()}}
\item \href{#method-NetworkingV1Api-watch_ingress_class_list}{\code{NetworkingV1Api$watch_ingress_class_list()}}
\item \href{#method-NetworkingV1Api-watch_ingress_class}{\code{NetworkingV1Api$watch_ingress_class()}}
\item \href{#method-NetworkingV1Api-watch_ingress_list_for_all_namespaces}{\code{NetworkingV1Api$watch_ingress_list_for_all_namespaces()}}
\item \href{#method-NetworkingV1Api-watch_namespaced_ingress_list}{\code{NetworkingV1Api$watch_namespaced_ingress_list()}}
\item \href{#method-NetworkingV1Api-watch_namespaced_ingress}{\code{NetworkingV1Api$watch_namespaced_ingress()}}
\item \href{#method-NetworkingV1Api-watch_namespaced_network_policy_list}{\code{NetworkingV1Api$watch_namespaced_network_policy_list()}}
\item \href{#method-NetworkingV1Api-watch_namespaced_network_policy}{\code{NetworkingV1Api$watch_namespaced_network_policy()}}
\item \href{#method-NetworkingV1Api-watch_network_policy_list_for_all_namespaces}{\code{NetworkingV1Api$watch_network_policy_list_for_all_namespaces()}}
\item \href{#method-NetworkingV1Api-clone}{\code{NetworkingV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getNetworkingV1APIResources
get available resources
listNetworkingV1IngressClass
list or watch objects of kind IngressClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-list_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-list_ingress_class}{}}}
\subsection{Method \code{list_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$list_ingress_class(
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createNetworkingV1IngressClass
create an IngressClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-create_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-create_ingress_class}{}}}
\subsection{Method \code{create_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$create_ingress_class(
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNetworkingV1CollectionIngressClass
delete collection of IngressClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-delete_collection_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-delete_collection_ingress_class}{}}}
\subsection{Method \code{delete_collection_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$delete_collection_ingress_class(
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readNetworkingV1IngressClass
read the specified IngressClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-read_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-read_ingress_class}{}}}
\subsection{Method \code{read_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$read_ingress_class(name, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceNetworkingV1IngressClass
replace the specified IngressClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-replace_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-replace_ingress_class}{}}}
\subsection{Method \code{replace_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$replace_ingress_class(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNetworkingV1IngressClass
partially update the specified IngressClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-patch_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-patch_ingress_class}{}}}
\subsection{Method \code{patch_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$patch_ingress_class(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteNetworkingV1IngressClass
delete an IngressClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-delete_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-delete_ingress_class}{}}}
\subsection{Method \code{delete_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$delete_ingress_class(
  name,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
listNetworkingV1IngressForAllNamespaces
list or watch objects of kind Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-list_ingress_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-list_ingress_for_all_namespaces}{}}}
\subsection{Method \code{list_ingress_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$list_ingress_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listNetworkingV1NamespacedIngress
list or watch objects of kind Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-list_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-list_namespaced_ingress}{}}}
\subsection{Method \code{list_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$list_namespaced_ingress(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createNetworkingV1NamespacedIngress
create an Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-create_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-create_namespaced_ingress}{}}}
\subsection{Method \code{create_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$create_namespaced_ingress(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNetworkingV1CollectionNamespacedIngress
delete collection of Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-delete_collection_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-delete_collection_namespaced_ingress}{}}}
\subsection{Method \code{delete_collection_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$delete_collection_namespaced_ingress(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readNetworkingV1NamespacedIngress
read the specified Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-read_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-read_namespaced_ingress}{}}}
\subsection{Method \code{read_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$read_namespaced_ingress(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceNetworkingV1NamespacedIngress
replace the specified Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-replace_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-replace_namespaced_ingress}{}}}
\subsection{Method \code{replace_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$replace_namespaced_ingress(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNetworkingV1NamespacedIngress
partially update the specified Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-patch_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-patch_namespaced_ingress}{}}}
\subsection{Method \code{patch_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$patch_namespaced_ingress(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteNetworkingV1NamespacedIngress
delete an Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-delete_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-delete_namespaced_ingress}{}}}
\subsection{Method \code{delete_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$delete_namespaced_ingress(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readNetworkingV1NamespacedIngressStatus
read status of the specified Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-read_namespaced_ingress_status"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-read_namespaced_ingress_status}{}}}
\subsection{Method \code{read_namespaced_ingress_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$read_namespaced_ingress_status(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceNetworkingV1NamespacedIngressStatus
replace status of the specified Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-replace_namespaced_ingress_status"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-replace_namespaced_ingress_status}{}}}
\subsection{Method \code{replace_namespaced_ingress_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$replace_namespaced_ingress_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNetworkingV1NamespacedIngressStatus
partially update status of the specified Ingress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-patch_namespaced_ingress_status"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-patch_namespaced_ingress_status}{}}}
\subsection{Method \code{patch_namespaced_ingress_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$patch_namespaced_ingress_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listNetworkingV1NamespacedNetworkPolicy
list or watch objects of kind NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-list_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-list_namespaced_network_policy}{}}}
\subsection{Method \code{list_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$list_namespaced_network_policy(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createNetworkingV1NamespacedNetworkPolicy
create a NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-create_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-create_namespaced_network_policy}{}}}
\subsection{Method \code{create_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$create_namespaced_network_policy(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNetworkingV1CollectionNamespacedNetworkPolicy
delete collection of NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-delete_collection_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-delete_collection_namespaced_network_policy}{}}}
\subsection{Method \code{delete_collection_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$delete_collection_namespaced_network_policy(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readNetworkingV1NamespacedNetworkPolicy
read the specified NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-read_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-read_namespaced_network_policy}{}}}
\subsection{Method \code{read_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$read_namespaced_network_policy(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceNetworkingV1NamespacedNetworkPolicy
replace the specified NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-replace_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-replace_namespaced_network_policy}{}}}
\subsection{Method \code{replace_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$replace_namespaced_network_policy(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNetworkingV1NamespacedNetworkPolicy
partially update the specified NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-patch_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-patch_namespaced_network_policy}{}}}
\subsection{Method \code{patch_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$patch_namespaced_network_policy(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteNetworkingV1NamespacedNetworkPolicy
delete a NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-delete_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-delete_namespaced_network_policy}{}}}
\subsection{Method \code{delete_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$delete_namespaced_network_policy(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
listNetworkingV1NetworkPolicyForAllNamespaces
list or watch objects of kind NetworkPolicy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-list_network_policy_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-list_network_policy_for_all_namespaces}{}}}
\subsection{Method \code{list_network_policy_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$list_network_policy_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1IngressClassList
watch individual changes to a list of IngressClass. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_ingress_class_list"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_ingress_class_list}{}}}
\subsection{Method \code{watch_ingress_class_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_ingress_class_list(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1IngressClass
watch changes to an object of kind IngressClass. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_ingress_class"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_ingress_class}{}}}
\subsection{Method \code{watch_ingress_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_ingress_class(
  name,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1IngressListForAllNamespaces
watch individual changes to a list of Ingress. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_ingress_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_ingress_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_ingress_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_ingress_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1NamespacedIngressList
watch individual changes to a list of Ingress. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_namespaced_ingress_list"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_namespaced_ingress_list}{}}}
\subsection{Method \code{watch_namespaced_ingress_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_namespaced_ingress_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1NamespacedIngress
watch changes to an object of kind Ingress. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_namespaced_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_namespaced_ingress}{}}}
\subsection{Method \code{watch_namespaced_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_namespaced_ingress(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1NamespacedNetworkPolicyList
watch individual changes to a list of NetworkPolicy. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_namespaced_network_policy_list"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_namespaced_network_policy_list}{}}}
\subsection{Method \code{watch_namespaced_network_policy_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_namespaced_network_policy_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1NamespacedNetworkPolicy
watch changes to an object of kind NetworkPolicy. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_namespaced_network_policy"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_namespaced_network_policy}{}}}
\subsection{Method \code{watch_namespaced_network_policy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_namespaced_network_policy(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchNetworkingV1NetworkPolicyListForAllNamespaces
watch individual changes to a list of NetworkPolicy. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-watch_network_policy_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-watch_network_policy_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_network_policy_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$watch_network_policy_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
