% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_NetworkingV1beta1Api.R
\name{NetworkingV1beta1Api}
\alias{NetworkingV1beta1Api}
\title{NetworkingV1beta1Api}
\description{
NetworkingV1beta1Api

NetworkingV1beta1Api
}
\details{
Typed Kubernetes API client for the `NetworkingV1beta1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NetworkingV1beta1Api-new}{\code{NetworkingV1beta1Api$new()}}
\item \href{#method-NetworkingV1beta1Api-get_api_resources}{\code{NetworkingV1beta1Api$get_api_resources()}}
\item \href{#method-NetworkingV1beta1Api-list_ip_address}{\code{NetworkingV1beta1Api$list_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-create_ip_address}{\code{NetworkingV1beta1Api$create_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-delete_collection_ip_address}{\code{NetworkingV1beta1Api$delete_collection_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-read_ip_address}{\code{NetworkingV1beta1Api$read_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-replace_ip_address}{\code{NetworkingV1beta1Api$replace_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-patch_ip_address}{\code{NetworkingV1beta1Api$patch_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-delete_ip_address}{\code{NetworkingV1beta1Api$delete_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-list_service_cidr}{\code{NetworkingV1beta1Api$list_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-create_service_cidr}{\code{NetworkingV1beta1Api$create_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-delete_collection_service_cidr}{\code{NetworkingV1beta1Api$delete_collection_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-read_service_cidr}{\code{NetworkingV1beta1Api$read_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-replace_service_cidr}{\code{NetworkingV1beta1Api$replace_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-patch_service_cidr}{\code{NetworkingV1beta1Api$patch_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-delete_service_cidr}{\code{NetworkingV1beta1Api$delete_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-read_service_cidr_status}{\code{NetworkingV1beta1Api$read_service_cidr_status()}}
\item \href{#method-NetworkingV1beta1Api-replace_service_cidr_status}{\code{NetworkingV1beta1Api$replace_service_cidr_status()}}
\item \href{#method-NetworkingV1beta1Api-patch_service_cidr_status}{\code{NetworkingV1beta1Api$patch_service_cidr_status()}}
\item \href{#method-NetworkingV1beta1Api-watch_ip_address_list}{\code{NetworkingV1beta1Api$watch_ip_address_list()}}
\item \href{#method-NetworkingV1beta1Api-watch_ip_address}{\code{NetworkingV1beta1Api$watch_ip_address()}}
\item \href{#method-NetworkingV1beta1Api-watch_service_cidr_list}{\code{NetworkingV1beta1Api$watch_service_cidr_list()}}
\item \href{#method-NetworkingV1beta1Api-watch_service_cidr}{\code{NetworkingV1beta1Api$watch_service_cidr()}}
\item \href{#method-NetworkingV1beta1Api-clone}{\code{NetworkingV1beta1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getNetworkingV1beta1APIResources
get available resources
listNetworkingV1beta1IPAddress
list or watch objects of kind IPAddress
createNetworkingV1beta1IPAddress
create an IPAddress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-list_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-list_ip_address}{}}}
\subsection{Method \code{list_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$list_ip_address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-create_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-create_ip_address}{}}}
\subsection{Method \code{create_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$create_ip_address(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNetworkingV1beta1CollectionIPAddress
delete collection of IPAddress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-delete_collection_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-delete_collection_ip_address}{}}}
\subsection{Method \code{delete_collection_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$delete_collection_ip_address(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readNetworkingV1beta1IPAddress
read the specified IPAddress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-read_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-read_ip_address}{}}}
\subsection{Method \code{read_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$read_ip_address(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceNetworkingV1beta1IPAddress
replace the specified IPAddress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-replace_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-replace_ip_address}{}}}
\subsection{Method \code{replace_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$replace_ip_address(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNetworkingV1beta1IPAddress
partially update the specified IPAddress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-patch_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-patch_ip_address}{}}}
\subsection{Method \code{patch_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$patch_ip_address(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNetworkingV1beta1IPAddress
delete an IPAddress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-delete_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-delete_ip_address}{}}}
\subsection{Method \code{delete_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$delete_ip_address(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listNetworkingV1beta1ServiceCIDR
list or watch objects of kind ServiceCIDR
createNetworkingV1beta1ServiceCIDR
create a ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-list_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-list_service_cidr}{}}}
\subsection{Method \code{list_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$list_service_cidr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-create_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-create_service_cidr}{}}}
\subsection{Method \code{create_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$create_service_cidr(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNetworkingV1beta1CollectionServiceCIDR
delete collection of ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-delete_collection_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-delete_collection_service_cidr}{}}}
\subsection{Method \code{delete_collection_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$delete_collection_service_cidr(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readNetworkingV1beta1ServiceCIDR
read the specified ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-read_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-read_service_cidr}{}}}
\subsection{Method \code{read_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$read_service_cidr(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceNetworkingV1beta1ServiceCIDR
replace the specified ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-replace_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-replace_service_cidr}{}}}
\subsection{Method \code{replace_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$replace_service_cidr(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNetworkingV1beta1ServiceCIDR
partially update the specified ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-patch_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-patch_service_cidr}{}}}
\subsection{Method \code{patch_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$patch_service_cidr(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNetworkingV1beta1ServiceCIDR
delete a ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-delete_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-delete_service_cidr}{}}}
\subsection{Method \code{delete_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$delete_service_cidr(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readNetworkingV1beta1ServiceCIDRStatus
read status of the specified ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-read_service_cidr_status"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-read_service_cidr_status}{}}}
\subsection{Method \code{read_service_cidr_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$read_service_cidr_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceNetworkingV1beta1ServiceCIDRStatus
replace status of the specified ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-replace_service_cidr_status"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-replace_service_cidr_status}{}}}
\subsection{Method \code{replace_service_cidr_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$replace_service_cidr_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNetworkingV1beta1ServiceCIDRStatus
partially update status of the specified ServiceCIDR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-patch_service_cidr_status"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-patch_service_cidr_status}{}}}
\subsection{Method \code{patch_service_cidr_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$patch_service_cidr_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchNetworkingV1beta1IPAddressList
watch individual changes to a list of IPAddress. deprecated: use the 'watch' parameter with a list operation instead.
watchNetworkingV1beta1IPAddress
watch changes to an object of kind IPAddress. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-watch_ip_address_list"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-watch_ip_address_list}{}}}
\subsection{Method \code{watch_ip_address_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$watch_ip_address_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-watch_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-watch_ip_address}{}}}
\subsection{Method \code{watch_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$watch_ip_address(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchNetworkingV1beta1ServiceCIDRList
watch individual changes to a list of ServiceCIDR. deprecated: use the 'watch' parameter with a list operation instead.
watchNetworkingV1beta1ServiceCIDR
watch changes to an object of kind ServiceCIDR. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-watch_service_cidr_list"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-watch_service_cidr_list}{}}}
\subsection{Method \code{watch_service_cidr_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$watch_service_cidr_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-watch_service_cidr"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-watch_service_cidr}{}}}
\subsection{Method \code{watch_service_cidr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$watch_service_cidr(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetworkingV1beta1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NetworkingV1beta1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetworkingV1beta1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
