% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{new_client_from_config}
\alias{new_client_from_config}
\title{Shorthand for loading config + constructing an \link{ApiClient}.}
\usage{
new_client_from_config(config_file = NULL, context = NULL)
}
\arguments{
\item{config_file}{Path to a kubeconfig. Defaults to \verb{$KUBECONFIG} (the
first entry if colon-separated) or \verb{~/.kube/config}.}

\item{context}{Context name. Defaults to the kubeconfig's \code{current-context}.}
}
\value{
An \link{ApiClient}.
}
\description{
Shorthand for loading config + constructing an \link{ApiClient}.
}
\examples{
\dontrun{
client <- new_client_from_config()
CoreV1Api$new(client)$list_namespace()
}
}
