% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_NodeApi.R
\name{NodeApi}
\alias{NodeApi}
\title{NodeApi}
\description{
NodeApi

NodeApi
}
\details{
Typed Kubernetes API client for the `Node` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NodeApi-new}{\code{NodeApi$new()}}
\item \href{#method-NodeApi-get_api_group}{\code{NodeApi$get_api_group()}}
\item \href{#method-NodeApi-clone}{\code{NodeApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-NodeApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getNodeAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-NodeApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NodeApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
