% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_NodeV1Api.R
\name{NodeV1Api}
\alias{NodeV1Api}
\title{NodeV1Api}
\description{
NodeV1Api

NodeV1Api
}
\details{
Typed Kubernetes API client for the `NodeV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NodeV1Api-new}{\code{NodeV1Api$new()}}
\item \href{#method-NodeV1Api-get_api_resources}{\code{NodeV1Api$get_api_resources()}}
\item \href{#method-NodeV1Api-list_runtime_class}{\code{NodeV1Api$list_runtime_class()}}
\item \href{#method-NodeV1Api-create_runtime_class}{\code{NodeV1Api$create_runtime_class()}}
\item \href{#method-NodeV1Api-delete_collection_runtime_class}{\code{NodeV1Api$delete_collection_runtime_class()}}
\item \href{#method-NodeV1Api-read_runtime_class}{\code{NodeV1Api$read_runtime_class()}}
\item \href{#method-NodeV1Api-replace_runtime_class}{\code{NodeV1Api$replace_runtime_class()}}
\item \href{#method-NodeV1Api-patch_runtime_class}{\code{NodeV1Api$patch_runtime_class()}}
\item \href{#method-NodeV1Api-delete_runtime_class}{\code{NodeV1Api$delete_runtime_class()}}
\item \href{#method-NodeV1Api-watch_runtime_class_list}{\code{NodeV1Api$watch_runtime_class_list()}}
\item \href{#method-NodeV1Api-watch_runtime_class}{\code{NodeV1Api$watch_runtime_class()}}
\item \href{#method-NodeV1Api-clone}{\code{NodeV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getNodeV1APIResources
get available resources
listNodeV1RuntimeClass
list or watch objects of kind RuntimeClass
createNodeV1RuntimeClass
create a RuntimeClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-list_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-list_runtime_class}{}}}
\subsection{Method \code{list_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$list_runtime_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-create_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-create_runtime_class}{}}}
\subsection{Method \code{create_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$create_runtime_class(body, dry_run = NULL, field_validation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNodeV1CollectionRuntimeClass
delete collection of RuntimeClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-delete_collection_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-delete_collection_runtime_class}{}}}
\subsection{Method \code{delete_collection_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$delete_collection_runtime_class(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readNodeV1RuntimeClass
read the specified RuntimeClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-read_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-read_runtime_class}{}}}
\subsection{Method \code{read_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$read_runtime_class(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceNodeV1RuntimeClass
replace the specified RuntimeClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-replace_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-replace_runtime_class}{}}}
\subsection{Method \code{replace_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$replace_runtime_class(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchNodeV1RuntimeClass
partially update the specified RuntimeClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-patch_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-patch_runtime_class}{}}}
\subsection{Method \code{patch_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$patch_runtime_class(name, dry_run = NULL, field_validation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteNodeV1RuntimeClass
delete a RuntimeClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-delete_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-delete_runtime_class}{}}}
\subsection{Method \code{delete_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$delete_runtime_class(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchNodeV1RuntimeClassList
watch individual changes to a list of RuntimeClass. deprecated: use the 'watch' parameter with a list operation instead.
watchNodeV1RuntimeClass
watch changes to an object of kind RuntimeClass. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-watch_runtime_class_list"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-watch_runtime_class_list}{}}}
\subsection{Method \code{watch_runtime_class_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$watch_runtime_class_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-watch_runtime_class"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-watch_runtime_class}{}}}
\subsection{Method \code{watch_runtime_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$watch_runtime_class(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NodeV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NodeV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
