% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_OpenidApi.R
\name{OpenidApi}
\alias{OpenidApi}
\title{OpenidApi}
\description{
OpenidApi

OpenidApi
}
\details{
Typed Kubernetes API client for the `Openid` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpenidApi-new}{\code{OpenidApi$new()}}
\item \href{#method-OpenidApi-get_service_account_issuer_open_id_keyset}{\code{OpenidApi$get_service_account_issuer_open_id_keyset()}}
\item \href{#method-OpenidApi-clone}{\code{OpenidApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenidApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpenidApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenidApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getServiceAccountIssuerOpenIDKeyset
get service account issuer OpenID JSON Web Key Set (contains public token verification keys)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenidApi-get_service_account_issuer_open_id_keyset"></a>}}
\if{latex}{\out{\hypertarget{method-OpenidApi-get_service_account_issuer_open_id_keyset}{}}}
\subsection{Method \code{get_service_account_issuer_open_id_keyset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenidApi$get_service_account_issuer_open_id_keyset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenidApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpenidApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenidApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
