% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_PolicyApi.R
\name{PolicyApi}
\alias{PolicyApi}
\title{PolicyApi}
\description{
PolicyApi

PolicyApi
}
\details{
Typed Kubernetes API client for the `Policy` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PolicyApi-new}{\code{PolicyApi$new()}}
\item \href{#method-PolicyApi-get_api_group}{\code{PolicyApi$get_api_group()}}
\item \href{#method-PolicyApi-clone}{\code{PolicyApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getPolicyAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
