% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_RbacAuthorizationApi.R
\name{RbacAuthorizationApi}
\alias{RbacAuthorizationApi}
\title{RbacAuthorizationApi}
\description{
RbacAuthorizationApi

RbacAuthorizationApi
}
\details{
Typed Kubernetes API client for the `RbacAuthorization` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RbacAuthorizationApi-new}{\code{RbacAuthorizationApi$new()}}
\item \href{#method-RbacAuthorizationApi-get_api_group}{\code{RbacAuthorizationApi$get_api_group()}}
\item \href{#method-RbacAuthorizationApi-clone}{\code{RbacAuthorizationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getRbacAuthorizationAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
