% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_RbacAuthorizationV1Api.R
\name{RbacAuthorizationV1Api}
\alias{RbacAuthorizationV1Api}
\title{RbacAuthorizationV1Api}
\description{
RbacAuthorizationV1Api

RbacAuthorizationV1Api
}
\details{
Typed Kubernetes API client for the `RbacAuthorizationV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RbacAuthorizationV1Api-new}{\code{RbacAuthorizationV1Api$new()}}
\item \href{#method-RbacAuthorizationV1Api-get_api_resources}{\code{RbacAuthorizationV1Api$get_api_resources()}}
\item \href{#method-RbacAuthorizationV1Api-list_cluster_role_binding}{\code{RbacAuthorizationV1Api$list_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-create_cluster_role_binding}{\code{RbacAuthorizationV1Api$create_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-delete_collection_cluster_role_binding}{\code{RbacAuthorizationV1Api$delete_collection_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-read_cluster_role_binding}{\code{RbacAuthorizationV1Api$read_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-replace_cluster_role_binding}{\code{RbacAuthorizationV1Api$replace_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-patch_cluster_role_binding}{\code{RbacAuthorizationV1Api$patch_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-delete_cluster_role_binding}{\code{RbacAuthorizationV1Api$delete_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-list_cluster_role}{\code{RbacAuthorizationV1Api$list_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-create_cluster_role}{\code{RbacAuthorizationV1Api$create_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-delete_collection_cluster_role}{\code{RbacAuthorizationV1Api$delete_collection_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-read_cluster_role}{\code{RbacAuthorizationV1Api$read_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-replace_cluster_role}{\code{RbacAuthorizationV1Api$replace_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-patch_cluster_role}{\code{RbacAuthorizationV1Api$patch_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-delete_cluster_role}{\code{RbacAuthorizationV1Api$delete_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-list_namespaced_role_binding}{\code{RbacAuthorizationV1Api$list_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-create_namespaced_role_binding}{\code{RbacAuthorizationV1Api$create_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-delete_collection_namespaced_role_binding}{\code{RbacAuthorizationV1Api$delete_collection_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-read_namespaced_role_binding}{\code{RbacAuthorizationV1Api$read_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-replace_namespaced_role_binding}{\code{RbacAuthorizationV1Api$replace_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-patch_namespaced_role_binding}{\code{RbacAuthorizationV1Api$patch_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-delete_namespaced_role_binding}{\code{RbacAuthorizationV1Api$delete_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-list_namespaced_role}{\code{RbacAuthorizationV1Api$list_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-create_namespaced_role}{\code{RbacAuthorizationV1Api$create_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-delete_collection_namespaced_role}{\code{RbacAuthorizationV1Api$delete_collection_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-read_namespaced_role}{\code{RbacAuthorizationV1Api$read_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-replace_namespaced_role}{\code{RbacAuthorizationV1Api$replace_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-patch_namespaced_role}{\code{RbacAuthorizationV1Api$patch_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-delete_namespaced_role}{\code{RbacAuthorizationV1Api$delete_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-list_role_binding_for_all_namespaces}{\code{RbacAuthorizationV1Api$list_role_binding_for_all_namespaces()}}
\item \href{#method-RbacAuthorizationV1Api-list_role_for_all_namespaces}{\code{RbacAuthorizationV1Api$list_role_for_all_namespaces()}}
\item \href{#method-RbacAuthorizationV1Api-watch_cluster_role_binding_list}{\code{RbacAuthorizationV1Api$watch_cluster_role_binding_list()}}
\item \href{#method-RbacAuthorizationV1Api-watch_cluster_role_binding}{\code{RbacAuthorizationV1Api$watch_cluster_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-watch_cluster_role_list}{\code{RbacAuthorizationV1Api$watch_cluster_role_list()}}
\item \href{#method-RbacAuthorizationV1Api-watch_cluster_role}{\code{RbacAuthorizationV1Api$watch_cluster_role()}}
\item \href{#method-RbacAuthorizationV1Api-watch_namespaced_role_binding_list}{\code{RbacAuthorizationV1Api$watch_namespaced_role_binding_list()}}
\item \href{#method-RbacAuthorizationV1Api-watch_namespaced_role_binding}{\code{RbacAuthorizationV1Api$watch_namespaced_role_binding()}}
\item \href{#method-RbacAuthorizationV1Api-watch_namespaced_role_list}{\code{RbacAuthorizationV1Api$watch_namespaced_role_list()}}
\item \href{#method-RbacAuthorizationV1Api-watch_namespaced_role}{\code{RbacAuthorizationV1Api$watch_namespaced_role()}}
\item \href{#method-RbacAuthorizationV1Api-watch_role_binding_list_for_all_namespaces}{\code{RbacAuthorizationV1Api$watch_role_binding_list_for_all_namespaces()}}
\item \href{#method-RbacAuthorizationV1Api-watch_role_list_for_all_namespaces}{\code{RbacAuthorizationV1Api$watch_role_list_for_all_namespaces()}}
\item \href{#method-RbacAuthorizationV1Api-clone}{\code{RbacAuthorizationV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getRbacAuthorizationV1APIResources
get available resources
listRbacAuthorizationV1ClusterRoleBinding
list or watch objects of kind ClusterRoleBinding
createRbacAuthorizationV1ClusterRoleBinding
create a ClusterRoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-list_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-list_cluster_role_binding}{}}}
\subsection{Method \code{list_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$list_cluster_role_binding()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-create_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-create_cluster_role_binding}{}}}
\subsection{Method \code{create_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$create_cluster_role_binding(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1CollectionClusterRoleBinding
delete collection of ClusterRoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_collection_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_collection_cluster_role_binding}{}}}
\subsection{Method \code{delete_collection_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_collection_cluster_role_binding(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readRbacAuthorizationV1ClusterRoleBinding
read the specified ClusterRoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-read_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-read_cluster_role_binding}{}}}
\subsection{Method \code{read_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$read_cluster_role_binding(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceRbacAuthorizationV1ClusterRoleBinding
replace the specified ClusterRoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-replace_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-replace_cluster_role_binding}{}}}
\subsection{Method \code{replace_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$replace_cluster_role_binding(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchRbacAuthorizationV1ClusterRoleBinding
partially update the specified ClusterRoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-patch_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-patch_cluster_role_binding}{}}}
\subsection{Method \code{patch_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$patch_cluster_role_binding(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1ClusterRoleBinding
delete a ClusterRoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_cluster_role_binding}{}}}
\subsection{Method \code{delete_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_cluster_role_binding(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listRbacAuthorizationV1ClusterRole
list or watch objects of kind ClusterRole
createRbacAuthorizationV1ClusterRole
create a ClusterRole}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-list_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-list_cluster_role}{}}}
\subsection{Method \code{list_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$list_cluster_role()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-create_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-create_cluster_role}{}}}
\subsection{Method \code{create_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$create_cluster_role(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1CollectionClusterRole
delete collection of ClusterRole}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_collection_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_collection_cluster_role}{}}}
\subsection{Method \code{delete_collection_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_collection_cluster_role(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readRbacAuthorizationV1ClusterRole
read the specified ClusterRole}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-read_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-read_cluster_role}{}}}
\subsection{Method \code{read_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$read_cluster_role(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceRbacAuthorizationV1ClusterRole
replace the specified ClusterRole}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-replace_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-replace_cluster_role}{}}}
\subsection{Method \code{replace_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$replace_cluster_role(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchRbacAuthorizationV1ClusterRole
partially update the specified ClusterRole}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-patch_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-patch_cluster_role}{}}}
\subsection{Method \code{patch_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$patch_cluster_role(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1ClusterRole
delete a ClusterRole}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_cluster_role}{}}}
\subsection{Method \code{delete_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_cluster_role(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listRbacAuthorizationV1NamespacedRoleBinding
list or watch objects of kind RoleBinding
createRbacAuthorizationV1NamespacedRoleBinding
create a RoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-list_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-list_namespaced_role_binding}{}}}
\subsection{Method \code{list_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$list_namespaced_role_binding()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-create_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-create_namespaced_role_binding}{}}}
\subsection{Method \code{create_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$create_namespaced_role_binding(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1CollectionNamespacedRoleBinding
delete collection of RoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_collection_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_collection_namespaced_role_binding}{}}}
\subsection{Method \code{delete_collection_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_collection_namespaced_role_binding(
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readRbacAuthorizationV1NamespacedRoleBinding
read the specified RoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-read_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-read_namespaced_role_binding}{}}}
\subsection{Method \code{read_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$read_namespaced_role_binding(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceRbacAuthorizationV1NamespacedRoleBinding
replace the specified RoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-replace_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-replace_namespaced_role_binding}{}}}
\subsection{Method \code{replace_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$replace_namespaced_role_binding(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchRbacAuthorizationV1NamespacedRoleBinding
partially update the specified RoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-patch_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-patch_namespaced_role_binding}{}}}
\subsection{Method \code{patch_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$patch_namespaced_role_binding(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1NamespacedRoleBinding
delete a RoleBinding}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_namespaced_role_binding}{}}}
\subsection{Method \code{delete_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_namespaced_role_binding(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listRbacAuthorizationV1NamespacedRole
list or watch objects of kind Role
createRbacAuthorizationV1NamespacedRole
create a Role}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-list_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-list_namespaced_role}{}}}
\subsection{Method \code{list_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$list_namespaced_role()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-create_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-create_namespaced_role}{}}}
\subsection{Method \code{create_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$create_namespaced_role(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1CollectionNamespacedRole
delete collection of Role}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_collection_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_collection_namespaced_role}{}}}
\subsection{Method \code{delete_collection_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_collection_namespaced_role(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readRbacAuthorizationV1NamespacedRole
read the specified Role}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-read_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-read_namespaced_role}{}}}
\subsection{Method \code{read_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$read_namespaced_role(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceRbacAuthorizationV1NamespacedRole
replace the specified Role}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-replace_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-replace_namespaced_role}{}}}
\subsection{Method \code{replace_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$replace_namespaced_role(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchRbacAuthorizationV1NamespacedRole
partially update the specified Role}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-patch_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-patch_namespaced_role}{}}}
\subsection{Method \code{patch_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$patch_namespaced_role(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteRbacAuthorizationV1NamespacedRole
delete a Role}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-delete_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-delete_namespaced_role}{}}}
\subsection{Method \code{delete_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$delete_namespaced_role(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listRbacAuthorizationV1RoleBindingForAllNamespaces
list or watch objects of kind RoleBinding
listRbacAuthorizationV1RoleForAllNamespaces
list or watch objects of kind Role
watchRbacAuthorizationV1ClusterRoleBindingList
watch individual changes to a list of ClusterRoleBinding. deprecated: use the 'watch' parameter with a list operation instead.
watchRbacAuthorizationV1ClusterRoleBinding
watch changes to an object of kind ClusterRoleBinding. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-list_role_binding_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-list_role_binding_for_all_namespaces}{}}}
\subsection{Method \code{list_role_binding_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$list_role_binding_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-list_role_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-list_role_for_all_namespaces}{}}}
\subsection{Method \code{list_role_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$list_role_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_cluster_role_binding_list"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_cluster_role_binding_list}{}}}
\subsection{Method \code{watch_cluster_role_binding_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_cluster_role_binding_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_cluster_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_cluster_role_binding}{}}}
\subsection{Method \code{watch_cluster_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_cluster_role_binding(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchRbacAuthorizationV1ClusterRoleList
watch individual changes to a list of ClusterRole. deprecated: use the 'watch' parameter with a list operation instead.
watchRbacAuthorizationV1ClusterRole
watch changes to an object of kind ClusterRole. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_cluster_role_list"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_cluster_role_list}{}}}
\subsection{Method \code{watch_cluster_role_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_cluster_role_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_cluster_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_cluster_role}{}}}
\subsection{Method \code{watch_cluster_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_cluster_role(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchRbacAuthorizationV1NamespacedRoleBindingList
watch individual changes to a list of RoleBinding. deprecated: use the 'watch' parameter with a list operation instead.
watchRbacAuthorizationV1NamespacedRoleBinding
watch changes to an object of kind RoleBinding. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_namespaced_role_binding_list"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_namespaced_role_binding_list}{}}}
\subsection{Method \code{watch_namespaced_role_binding_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_namespaced_role_binding_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_namespaced_role_binding"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_namespaced_role_binding}{}}}
\subsection{Method \code{watch_namespaced_role_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_namespaced_role_binding(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchRbacAuthorizationV1NamespacedRoleList
watch individual changes to a list of Role. deprecated: use the 'watch' parameter with a list operation instead.
watchRbacAuthorizationV1NamespacedRole
watch changes to an object of kind Role. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_namespaced_role_list"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_namespaced_role_list}{}}}
\subsection{Method \code{watch_namespaced_role_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_namespaced_role_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_namespaced_role"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_namespaced_role}{}}}
\subsection{Method \code{watch_namespaced_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_namespaced_role(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchRbacAuthorizationV1RoleBindingListForAllNamespaces
watch individual changes to a list of RoleBinding. deprecated: use the 'watch' parameter with a list operation instead.
watchRbacAuthorizationV1RoleListForAllNamespaces
watch individual changes to a list of Role. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_role_binding_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_role_binding_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_role_binding_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_role_binding_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-watch_role_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-watch_role_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_role_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$watch_role_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacAuthorizationV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RbacAuthorizationV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacAuthorizationV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
