% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_RbacV1Subject.R
\name{RbacV1Subject}
\alias{RbacV1Subject}
\title{RbacV1Subject}
\description{
RbacV1Subject

RbacV1Subject
}
\details{
Subject contains a reference to the object or user identities a role binding applies to.  This can either hold a direct API object reference, or a value for non-objects such as user and group names.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RbacV1Subject-new}{\code{RbacV1Subject$new()}}
\item \href{#method-RbacV1Subject-to_list}{\code{RbacV1Subject$to_list()}}
\item \href{#method-RbacV1Subject-clone}{\code{RbacV1Subject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacV1Subject-new"></a>}}
\if{latex}{\out{\hypertarget{method-RbacV1Subject-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacV1Subject$new(api_group = NULL, kind = NULL, name = NULL, namespace = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacV1Subject-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RbacV1Subject-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacV1Subject$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RbacV1Subject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RbacV1Subject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RbacV1Subject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
