% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_ResourceApi.R
\name{ResourceApi}
\alias{ResourceApi}
\title{ResourceApi}
\description{
ResourceApi

ResourceApi
}
\details{
Typed Kubernetes API client for the `Resource` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceApi-new}{\code{ResourceApi$new()}}
\item \href{#method-ResourceApi-get_api_group}{\code{ResourceApi$get_api_group()}}
\item \href{#method-ResourceApi-clone}{\code{ResourceApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getResourceAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
