% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_ResourceV1alpha3Api.R
\name{ResourceV1alpha3Api}
\alias{ResourceV1alpha3Api}
\title{ResourceV1alpha3Api}
\description{
ResourceV1alpha3Api

ResourceV1alpha3Api
}
\details{
Typed Kubernetes API client for the `ResourceV1alpha3` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceV1alpha3Api-new}{\code{ResourceV1alpha3Api$new()}}
\item \href{#method-ResourceV1alpha3Api-get_api_resources}{\code{ResourceV1alpha3Api$get_api_resources()}}
\item \href{#method-ResourceV1alpha3Api-list_device_class}{\code{ResourceV1alpha3Api$list_device_class()}}
\item \href{#method-ResourceV1alpha3Api-create_device_class}{\code{ResourceV1alpha3Api$create_device_class()}}
\item \href{#method-ResourceV1alpha3Api-delete_collection_device_class}{\code{ResourceV1alpha3Api$delete_collection_device_class()}}
\item \href{#method-ResourceV1alpha3Api-read_device_class}{\code{ResourceV1alpha3Api$read_device_class()}}
\item \href{#method-ResourceV1alpha3Api-replace_device_class}{\code{ResourceV1alpha3Api$replace_device_class()}}
\item \href{#method-ResourceV1alpha3Api-patch_device_class}{\code{ResourceV1alpha3Api$patch_device_class()}}
\item \href{#method-ResourceV1alpha3Api-delete_device_class}{\code{ResourceV1alpha3Api$delete_device_class()}}
\item \href{#method-ResourceV1alpha3Api-list_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$list_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-create_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$create_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-delete_collection_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$delete_collection_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-read_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$read_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-replace_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$replace_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-patch_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$patch_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-delete_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$delete_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-read_namespaced_pod_scheduling_context_status}{\code{ResourceV1alpha3Api$read_namespaced_pod_scheduling_context_status()}}
\item \href{#method-ResourceV1alpha3Api-replace_namespaced_pod_scheduling_context_status}{\code{ResourceV1alpha3Api$replace_namespaced_pod_scheduling_context_status()}}
\item \href{#method-ResourceV1alpha3Api-patch_namespaced_pod_scheduling_context_status}{\code{ResourceV1alpha3Api$patch_namespaced_pod_scheduling_context_status()}}
\item \href{#method-ResourceV1alpha3Api-list_namespaced_resource_claim}{\code{ResourceV1alpha3Api$list_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-create_namespaced_resource_claim}{\code{ResourceV1alpha3Api$create_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-delete_collection_namespaced_resource_claim}{\code{ResourceV1alpha3Api$delete_collection_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-read_namespaced_resource_claim}{\code{ResourceV1alpha3Api$read_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-replace_namespaced_resource_claim}{\code{ResourceV1alpha3Api$replace_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-patch_namespaced_resource_claim}{\code{ResourceV1alpha3Api$patch_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-delete_namespaced_resource_claim}{\code{ResourceV1alpha3Api$delete_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-read_namespaced_resource_claim_status}{\code{ResourceV1alpha3Api$read_namespaced_resource_claim_status()}}
\item \href{#method-ResourceV1alpha3Api-replace_namespaced_resource_claim_status}{\code{ResourceV1alpha3Api$replace_namespaced_resource_claim_status()}}
\item \href{#method-ResourceV1alpha3Api-patch_namespaced_resource_claim_status}{\code{ResourceV1alpha3Api$patch_namespaced_resource_claim_status()}}
\item \href{#method-ResourceV1alpha3Api-list_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$list_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-create_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$create_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-delete_collection_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$delete_collection_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-read_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$read_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-replace_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$replace_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-patch_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$patch_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-delete_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$delete_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-list_pod_scheduling_context_for_all_namespaces}{\code{ResourceV1alpha3Api$list_pod_scheduling_context_for_all_namespaces()}}
\item \href{#method-ResourceV1alpha3Api-list_resource_claim_for_all_namespaces}{\code{ResourceV1alpha3Api$list_resource_claim_for_all_namespaces()}}
\item \href{#method-ResourceV1alpha3Api-list_resource_claim_template_for_all_namespaces}{\code{ResourceV1alpha3Api$list_resource_claim_template_for_all_namespaces()}}
\item \href{#method-ResourceV1alpha3Api-list_resource_slice}{\code{ResourceV1alpha3Api$list_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-create_resource_slice}{\code{ResourceV1alpha3Api$create_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-delete_collection_resource_slice}{\code{ResourceV1alpha3Api$delete_collection_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-read_resource_slice}{\code{ResourceV1alpha3Api$read_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-replace_resource_slice}{\code{ResourceV1alpha3Api$replace_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-patch_resource_slice}{\code{ResourceV1alpha3Api$patch_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-delete_resource_slice}{\code{ResourceV1alpha3Api$delete_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-watch_device_class_list}{\code{ResourceV1alpha3Api$watch_device_class_list()}}
\item \href{#method-ResourceV1alpha3Api-watch_device_class}{\code{ResourceV1alpha3Api$watch_device_class()}}
\item \href{#method-ResourceV1alpha3Api-watch_namespaced_pod_scheduling_context_list}{\code{ResourceV1alpha3Api$watch_namespaced_pod_scheduling_context_list()}}
\item \href{#method-ResourceV1alpha3Api-watch_namespaced_pod_scheduling_context}{\code{ResourceV1alpha3Api$watch_namespaced_pod_scheduling_context()}}
\item \href{#method-ResourceV1alpha3Api-watch_namespaced_resource_claim_list}{\code{ResourceV1alpha3Api$watch_namespaced_resource_claim_list()}}
\item \href{#method-ResourceV1alpha3Api-watch_namespaced_resource_claim}{\code{ResourceV1alpha3Api$watch_namespaced_resource_claim()}}
\item \href{#method-ResourceV1alpha3Api-watch_namespaced_resource_claim_template_list}{\code{ResourceV1alpha3Api$watch_namespaced_resource_claim_template_list()}}
\item \href{#method-ResourceV1alpha3Api-watch_namespaced_resource_claim_template}{\code{ResourceV1alpha3Api$watch_namespaced_resource_claim_template()}}
\item \href{#method-ResourceV1alpha3Api-watch_pod_scheduling_context_list_for_all_namespaces}{\code{ResourceV1alpha3Api$watch_pod_scheduling_context_list_for_all_namespaces()}}
\item \href{#method-ResourceV1alpha3Api-watch_resource_claim_list_for_all_namespaces}{\code{ResourceV1alpha3Api$watch_resource_claim_list_for_all_namespaces()}}
\item \href{#method-ResourceV1alpha3Api-watch_resource_claim_template_list_for_all_namespaces}{\code{ResourceV1alpha3Api$watch_resource_claim_template_list_for_all_namespaces()}}
\item \href{#method-ResourceV1alpha3Api-watch_resource_slice_list}{\code{ResourceV1alpha3Api$watch_resource_slice_list()}}
\item \href{#method-ResourceV1alpha3Api-watch_resource_slice}{\code{ResourceV1alpha3Api$watch_resource_slice()}}
\item \href{#method-ResourceV1alpha3Api-clone}{\code{ResourceV1alpha3Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getResourceV1alpha3APIResources
get available resources
listResourceV1alpha3DeviceClass
list or watch objects of kind DeviceClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_device_class}{}}}
\subsection{Method \code{list_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_device_class(
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createResourceV1alpha3DeviceClass
create a DeviceClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-create_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-create_device_class}{}}}
\subsection{Method \code{create_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$create_device_class(
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteResourceV1alpha3CollectionDeviceClass
delete collection of DeviceClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_collection_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_collection_device_class}{}}}
\subsection{Method \code{delete_collection_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_collection_device_class(
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readResourceV1alpha3DeviceClass
read the specified DeviceClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-read_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-read_device_class}{}}}
\subsection{Method \code{read_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$read_device_class(name, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceResourceV1alpha3DeviceClass
replace the specified DeviceClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-replace_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-replace_device_class}{}}}
\subsection{Method \code{replace_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$replace_device_class(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchResourceV1alpha3DeviceClass
partially update the specified DeviceClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-patch_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-patch_device_class}{}}}
\subsection{Method \code{patch_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$patch_device_class(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteResourceV1alpha3DeviceClass
delete a DeviceClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_device_class}{}}}
\subsection{Method \code{delete_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_device_class(
  name,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
listResourceV1alpha3NamespacedPodSchedulingContext
list or watch objects of kind PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{list_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_namespaced_pod_scheduling_context(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createResourceV1alpha3NamespacedPodSchedulingContext
create a PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-create_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-create_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{create_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$create_namespaced_pod_scheduling_context(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteResourceV1alpha3CollectionNamespacedPodSchedulingContext
delete collection of PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_collection_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_collection_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{delete_collection_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_collection_namespaced_pod_scheduling_context(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readResourceV1alpha3NamespacedPodSchedulingContext
read the specified PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-read_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-read_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{read_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$read_namespaced_pod_scheduling_context(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceResourceV1alpha3NamespacedPodSchedulingContext
replace the specified PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-replace_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-replace_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{replace_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$replace_namespaced_pod_scheduling_context(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchResourceV1alpha3NamespacedPodSchedulingContext
partially update the specified PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-patch_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-patch_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{patch_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$patch_namespaced_pod_scheduling_context(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteResourceV1alpha3NamespacedPodSchedulingContext
delete a PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{delete_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_namespaced_pod_scheduling_context(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readResourceV1alpha3NamespacedPodSchedulingContextStatus
read status of the specified PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-read_namespaced_pod_scheduling_context_status"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-read_namespaced_pod_scheduling_context_status}{}}}
\subsection{Method \code{read_namespaced_pod_scheduling_context_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$read_namespaced_pod_scheduling_context_status(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceResourceV1alpha3NamespacedPodSchedulingContextStatus
replace status of the specified PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-replace_namespaced_pod_scheduling_context_status"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-replace_namespaced_pod_scheduling_context_status}{}}}
\subsection{Method \code{replace_namespaced_pod_scheduling_context_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$replace_namespaced_pod_scheduling_context_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchResourceV1alpha3NamespacedPodSchedulingContextStatus
partially update status of the specified PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-patch_namespaced_pod_scheduling_context_status"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-patch_namespaced_pod_scheduling_context_status}{}}}
\subsection{Method \code{patch_namespaced_pod_scheduling_context_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$patch_namespaced_pod_scheduling_context_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listResourceV1alpha3NamespacedResourceClaim
list or watch objects of kind ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_namespaced_resource_claim}{}}}
\subsection{Method \code{list_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_namespaced_resource_claim(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createResourceV1alpha3NamespacedResourceClaim
create a ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-create_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-create_namespaced_resource_claim}{}}}
\subsection{Method \code{create_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$create_namespaced_resource_claim(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteResourceV1alpha3CollectionNamespacedResourceClaim
delete collection of ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_collection_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_collection_namespaced_resource_claim}{}}}
\subsection{Method \code{delete_collection_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_collection_namespaced_resource_claim(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readResourceV1alpha3NamespacedResourceClaim
read the specified ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-read_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-read_namespaced_resource_claim}{}}}
\subsection{Method \code{read_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$read_namespaced_resource_claim(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceResourceV1alpha3NamespacedResourceClaim
replace the specified ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-replace_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-replace_namespaced_resource_claim}{}}}
\subsection{Method \code{replace_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$replace_namespaced_resource_claim(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchResourceV1alpha3NamespacedResourceClaim
partially update the specified ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-patch_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-patch_namespaced_resource_claim}{}}}
\subsection{Method \code{patch_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$patch_namespaced_resource_claim(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteResourceV1alpha3NamespacedResourceClaim
delete a ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_namespaced_resource_claim}{}}}
\subsection{Method \code{delete_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_namespaced_resource_claim(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readResourceV1alpha3NamespacedResourceClaimStatus
read status of the specified ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-read_namespaced_resource_claim_status"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-read_namespaced_resource_claim_status}{}}}
\subsection{Method \code{read_namespaced_resource_claim_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$read_namespaced_resource_claim_status(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceResourceV1alpha3NamespacedResourceClaimStatus
replace status of the specified ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-replace_namespaced_resource_claim_status"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-replace_namespaced_resource_claim_status}{}}}
\subsection{Method \code{replace_namespaced_resource_claim_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$replace_namespaced_resource_claim_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchResourceV1alpha3NamespacedResourceClaimStatus
partially update status of the specified ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-patch_namespaced_resource_claim_status"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-patch_namespaced_resource_claim_status}{}}}
\subsection{Method \code{patch_namespaced_resource_claim_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$patch_namespaced_resource_claim_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listResourceV1alpha3NamespacedResourceClaimTemplate
list or watch objects of kind ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_namespaced_resource_claim_template}{}}}
\subsection{Method \code{list_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_namespaced_resource_claim_template(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createResourceV1alpha3NamespacedResourceClaimTemplate
create a ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-create_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-create_namespaced_resource_claim_template}{}}}
\subsection{Method \code{create_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$create_namespaced_resource_claim_template(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteResourceV1alpha3CollectionNamespacedResourceClaimTemplate
delete collection of ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_collection_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_collection_namespaced_resource_claim_template}{}}}
\subsection{Method \code{delete_collection_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_collection_namespaced_resource_claim_template(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readResourceV1alpha3NamespacedResourceClaimTemplate
read the specified ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-read_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-read_namespaced_resource_claim_template}{}}}
\subsection{Method \code{read_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$read_namespaced_resource_claim_template(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceResourceV1alpha3NamespacedResourceClaimTemplate
replace the specified ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-replace_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-replace_namespaced_resource_claim_template}{}}}
\subsection{Method \code{replace_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$replace_namespaced_resource_claim_template(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchResourceV1alpha3NamespacedResourceClaimTemplate
partially update the specified ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-patch_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-patch_namespaced_resource_claim_template}{}}}
\subsection{Method \code{patch_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$patch_namespaced_resource_claim_template(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteResourceV1alpha3NamespacedResourceClaimTemplate
delete a ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_namespaced_resource_claim_template}{}}}
\subsection{Method \code{delete_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_namespaced_resource_claim_template(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
listResourceV1alpha3PodSchedulingContextForAllNamespaces
list or watch objects of kind PodSchedulingContext}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_pod_scheduling_context_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_pod_scheduling_context_for_all_namespaces}{}}}
\subsection{Method \code{list_pod_scheduling_context_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_pod_scheduling_context_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listResourceV1alpha3ResourceClaimForAllNamespaces
list or watch objects of kind ResourceClaim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_resource_claim_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_resource_claim_for_all_namespaces}{}}}
\subsection{Method \code{list_resource_claim_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_resource_claim_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listResourceV1alpha3ResourceClaimTemplateForAllNamespaces
list or watch objects of kind ResourceClaimTemplate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_resource_claim_template_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_resource_claim_template_for_all_namespaces}{}}}
\subsection{Method \code{list_resource_claim_template_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_resource_claim_template_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listResourceV1alpha3ResourceSlice
list or watch objects of kind ResourceSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-list_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-list_resource_slice}{}}}
\subsection{Method \code{list_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$list_resource_slice(
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createResourceV1alpha3ResourceSlice
create a ResourceSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-create_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-create_resource_slice}{}}}
\subsection{Method \code{create_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$create_resource_slice(
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteResourceV1alpha3CollectionResourceSlice
delete collection of ResourceSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_collection_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_collection_resource_slice}{}}}
\subsection{Method \code{delete_collection_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_collection_resource_slice(
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readResourceV1alpha3ResourceSlice
read the specified ResourceSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-read_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-read_resource_slice}{}}}
\subsection{Method \code{read_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$read_resource_slice(name, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceResourceV1alpha3ResourceSlice
replace the specified ResourceSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-replace_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-replace_resource_slice}{}}}
\subsection{Method \code{replace_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$replace_resource_slice(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchResourceV1alpha3ResourceSlice
partially update the specified ResourceSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-patch_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-patch_resource_slice}{}}}
\subsection{Method \code{patch_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$patch_resource_slice(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteResourceV1alpha3ResourceSlice
delete a ResourceSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-delete_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-delete_resource_slice}{}}}
\subsection{Method \code{delete_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$delete_resource_slice(
  name,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
watchResourceV1alpha3DeviceClassList
watch individual changes to a list of DeviceClass. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_device_class_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_device_class_list}{}}}
\subsection{Method \code{watch_device_class_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_device_class_list(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3DeviceClass
watch changes to an object of kind DeviceClass. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_device_class"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_device_class}{}}}
\subsection{Method \code{watch_device_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_device_class(
  name,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3NamespacedPodSchedulingContextList
watch individual changes to a list of PodSchedulingContext. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_namespaced_pod_scheduling_context_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_namespaced_pod_scheduling_context_list}{}}}
\subsection{Method \code{watch_namespaced_pod_scheduling_context_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_namespaced_pod_scheduling_context_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3NamespacedPodSchedulingContext
watch changes to an object of kind PodSchedulingContext. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_namespaced_pod_scheduling_context"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_namespaced_pod_scheduling_context}{}}}
\subsection{Method \code{watch_namespaced_pod_scheduling_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_namespaced_pod_scheduling_context(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3NamespacedResourceClaimList
watch individual changes to a list of ResourceClaim. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_namespaced_resource_claim_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_namespaced_resource_claim_list}{}}}
\subsection{Method \code{watch_namespaced_resource_claim_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_namespaced_resource_claim_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3NamespacedResourceClaim
watch changes to an object of kind ResourceClaim. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_namespaced_resource_claim"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_namespaced_resource_claim}{}}}
\subsection{Method \code{watch_namespaced_resource_claim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_namespaced_resource_claim(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3NamespacedResourceClaimTemplateList
watch individual changes to a list of ResourceClaimTemplate. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_namespaced_resource_claim_template_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_namespaced_resource_claim_template_list}{}}}
\subsection{Method \code{watch_namespaced_resource_claim_template_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_namespaced_resource_claim_template_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3NamespacedResourceClaimTemplate
watch changes to an object of kind ResourceClaimTemplate. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_namespaced_resource_claim_template"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_namespaced_resource_claim_template}{}}}
\subsection{Method \code{watch_namespaced_resource_claim_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_namespaced_resource_claim_template(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3PodSchedulingContextListForAllNamespaces
watch individual changes to a list of PodSchedulingContext. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_pod_scheduling_context_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_pod_scheduling_context_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_pod_scheduling_context_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_pod_scheduling_context_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3ResourceClaimListForAllNamespaces
watch individual changes to a list of ResourceClaim. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_resource_claim_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_resource_claim_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_resource_claim_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_resource_claim_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3ResourceClaimTemplateListForAllNamespaces
watch individual changes to a list of ResourceClaimTemplate. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_resource_claim_template_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_resource_claim_template_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_resource_claim_template_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_resource_claim_template_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3ResourceSliceList
watch individual changes to a list of ResourceSlice. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_resource_slice_list"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_resource_slice_list}{}}}
\subsection{Method \code{watch_resource_slice_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_resource_slice_list(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchResourceV1alpha3ResourceSlice
watch changes to an object of kind ResourceSlice. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-watch_resource_slice"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-watch_resource_slice}{}}}
\subsection{Method \code{watch_resource_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$watch_resource_slice(
  name,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceV1alpha3Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceV1alpha3Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceV1alpha3Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
