% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_SchedulingApi.R
\name{SchedulingApi}
\alias{SchedulingApi}
\title{SchedulingApi}
\description{
SchedulingApi

SchedulingApi
}
\details{
Typed Kubernetes API client for the `Scheduling` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SchedulingApi-new}{\code{SchedulingApi$new()}}
\item \href{#method-SchedulingApi-get_api_group}{\code{SchedulingApi$get_api_group()}}
\item \href{#method-SchedulingApi-clone}{\code{SchedulingApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getSchedulingAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
