% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_StorageApi.R
\name{StorageApi}
\alias{StorageApi}
\title{StorageApi}
\description{
StorageApi

StorageApi
}
\details{
Typed Kubernetes API client for the `Storage` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StorageApi-new}{\code{StorageApi$new()}}
\item \href{#method-StorageApi-get_api_group}{\code{StorageApi$get_api_group()}}
\item \href{#method-StorageApi-clone}{\code{StorageApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-StorageApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getStorageAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-StorageApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StorageApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
