% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_StoragemigrationApi.R
\name{StoragemigrationApi}
\alias{StoragemigrationApi}
\title{StoragemigrationApi}
\description{
StoragemigrationApi

StoragemigrationApi
}
\details{
Typed Kubernetes API client for the `Storagemigration` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StoragemigrationApi-new}{\code{StoragemigrationApi$new()}}
\item \href{#method-StoragemigrationApi-get_api_group}{\code{StoragemigrationApi$get_api_group()}}
\item \href{#method-StoragemigrationApi-clone}{\code{StoragemigrationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getStoragemigrationAPIGroup
get information of a group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationApi-get_api_group"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationApi-get_api_group}{}}}
\subsection{Method \code{get_api_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationApi$get_api_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
