% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_StoragemigrationV1alpha1Api.R
\name{StoragemigrationV1alpha1Api}
\alias{StoragemigrationV1alpha1Api}
\title{StoragemigrationV1alpha1Api}
\description{
StoragemigrationV1alpha1Api

StoragemigrationV1alpha1Api
}
\details{
Typed Kubernetes API client for the `StoragemigrationV1alpha1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StoragemigrationV1alpha1Api-new}{\code{StoragemigrationV1alpha1Api$new()}}
\item \href{#method-StoragemigrationV1alpha1Api-get_api_resources}{\code{StoragemigrationV1alpha1Api$get_api_resources()}}
\item \href{#method-StoragemigrationV1alpha1Api-list_storage_version_migration}{\code{StoragemigrationV1alpha1Api$list_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-create_storage_version_migration}{\code{StoragemigrationV1alpha1Api$create_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-delete_collection_storage_version_migration}{\code{StoragemigrationV1alpha1Api$delete_collection_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-read_storage_version_migration}{\code{StoragemigrationV1alpha1Api$read_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-replace_storage_version_migration}{\code{StoragemigrationV1alpha1Api$replace_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-patch_storage_version_migration}{\code{StoragemigrationV1alpha1Api$patch_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-delete_storage_version_migration}{\code{StoragemigrationV1alpha1Api$delete_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-read_storage_version_migration_status}{\code{StoragemigrationV1alpha1Api$read_storage_version_migration_status()}}
\item \href{#method-StoragemigrationV1alpha1Api-replace_storage_version_migration_status}{\code{StoragemigrationV1alpha1Api$replace_storage_version_migration_status()}}
\item \href{#method-StoragemigrationV1alpha1Api-patch_storage_version_migration_status}{\code{StoragemigrationV1alpha1Api$patch_storage_version_migration_status()}}
\item \href{#method-StoragemigrationV1alpha1Api-watch_storage_version_migration_list}{\code{StoragemigrationV1alpha1Api$watch_storage_version_migration_list()}}
\item \href{#method-StoragemigrationV1alpha1Api-watch_storage_version_migration}{\code{StoragemigrationV1alpha1Api$watch_storage_version_migration()}}
\item \href{#method-StoragemigrationV1alpha1Api-clone}{\code{StoragemigrationV1alpha1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getStoragemigrationV1alpha1APIResources
get available resources
listStoragemigrationV1alpha1StorageVersionMigration
list or watch objects of kind StorageVersionMigration
createStoragemigrationV1alpha1StorageVersionMigration
create a StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-list_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-list_storage_version_migration}{}}}
\subsection{Method \code{list_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$list_storage_version_migration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-create_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-create_storage_version_migration}{}}}
\subsection{Method \code{create_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$create_storage_version_migration(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStoragemigrationV1alpha1CollectionStorageVersionMigration
delete collection of StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-delete_collection_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-delete_collection_storage_version_migration}{}}}
\subsection{Method \code{delete_collection_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$delete_collection_storage_version_migration(
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStoragemigrationV1alpha1StorageVersionMigration
read the specified StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-read_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-read_storage_version_migration}{}}}
\subsection{Method \code{read_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$read_storage_version_migration(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStoragemigrationV1alpha1StorageVersionMigration
replace the specified StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-replace_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-replace_storage_version_migration}{}}}
\subsection{Method \code{replace_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$replace_storage_version_migration(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStoragemigrationV1alpha1StorageVersionMigration
partially update the specified StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-patch_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-patch_storage_version_migration}{}}}
\subsection{Method \code{patch_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$patch_storage_version_migration(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStoragemigrationV1alpha1StorageVersionMigration
delete a StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-delete_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-delete_storage_version_migration}{}}}
\subsection{Method \code{delete_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$delete_storage_version_migration(
  name,
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStoragemigrationV1alpha1StorageVersionMigrationStatus
read status of the specified StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-read_storage_version_migration_status"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-read_storage_version_migration_status}{}}}
\subsection{Method \code{read_storage_version_migration_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$read_storage_version_migration_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStoragemigrationV1alpha1StorageVersionMigrationStatus
replace status of the specified StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-replace_storage_version_migration_status"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-replace_storage_version_migration_status}{}}}
\subsection{Method \code{replace_storage_version_migration_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$replace_storage_version_migration_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStoragemigrationV1alpha1StorageVersionMigrationStatus
partially update status of the specified StorageVersionMigration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-patch_storage_version_migration_status"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-patch_storage_version_migration_status}{}}}
\subsection{Method \code{patch_storage_version_migration_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$patch_storage_version_migration_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchStoragemigrationV1alpha1StorageVersionMigrationList
watch individual changes to a list of StorageVersionMigration. deprecated: use the 'watch' parameter with a list operation instead.
watchStoragemigrationV1alpha1StorageVersionMigration
watch changes to an object of kind StorageVersionMigration. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-watch_storage_version_migration_list"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-watch_storage_version_migration_list}{}}}
\subsection{Method \code{watch_storage_version_migration_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$watch_storage_version_migration_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-watch_storage_version_migration"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-watch_storage_version_migration}{}}}
\subsection{Method \code{watch_storage_version_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$watch_storage_version_migration(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StoragemigrationV1alpha1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StoragemigrationV1alpha1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StoragemigrationV1alpha1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
