% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_StorageV1alpha1Api.R
\name{StorageV1alpha1Api}
\alias{StorageV1alpha1Api}
\title{StorageV1alpha1Api}
\description{
StorageV1alpha1Api

StorageV1alpha1Api
}
\details{
Typed Kubernetes API client for the `StorageV1alpha1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StorageV1alpha1Api-new}{\code{StorageV1alpha1Api$new()}}
\item \href{#method-StorageV1alpha1Api-get_api_resources}{\code{StorageV1alpha1Api$get_api_resources()}}
\item \href{#method-StorageV1alpha1Api-list_volume_attributes_class}{\code{StorageV1alpha1Api$list_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-create_volume_attributes_class}{\code{StorageV1alpha1Api$create_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-delete_collection_volume_attributes_class}{\code{StorageV1alpha1Api$delete_collection_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-read_volume_attributes_class}{\code{StorageV1alpha1Api$read_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-replace_volume_attributes_class}{\code{StorageV1alpha1Api$replace_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-patch_volume_attributes_class}{\code{StorageV1alpha1Api$patch_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-delete_volume_attributes_class}{\code{StorageV1alpha1Api$delete_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-watch_volume_attributes_class_list}{\code{StorageV1alpha1Api$watch_volume_attributes_class_list()}}
\item \href{#method-StorageV1alpha1Api-watch_volume_attributes_class}{\code{StorageV1alpha1Api$watch_volume_attributes_class()}}
\item \href{#method-StorageV1alpha1Api-clone}{\code{StorageV1alpha1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getStorageV1alpha1APIResources
get available resources
listStorageV1alpha1VolumeAttributesClass
list or watch objects of kind VolumeAttributesClass
createStorageV1alpha1VolumeAttributesClass
create a VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-list_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-list_volume_attributes_class}{}}}
\subsection{Method \code{list_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$list_volume_attributes_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-create_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-create_volume_attributes_class}{}}}
\subsection{Method \code{create_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$create_volume_attributes_class(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1alpha1CollectionVolumeAttributesClass
delete collection of VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-delete_collection_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-delete_collection_volume_attributes_class}{}}}
\subsection{Method \code{delete_collection_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$delete_collection_volume_attributes_class(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStorageV1alpha1VolumeAttributesClass
read the specified VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-read_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-read_volume_attributes_class}{}}}
\subsection{Method \code{read_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$read_volume_attributes_class(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStorageV1alpha1VolumeAttributesClass
replace the specified VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-replace_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-replace_volume_attributes_class}{}}}
\subsection{Method \code{replace_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$replace_volume_attributes_class(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1alpha1VolumeAttributesClass
partially update the specified VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-patch_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-patch_volume_attributes_class}{}}}
\subsection{Method \code{patch_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$patch_volume_attributes_class(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1alpha1VolumeAttributesClass
delete a VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-delete_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-delete_volume_attributes_class}{}}}
\subsection{Method \code{delete_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$delete_volume_attributes_class(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchStorageV1alpha1VolumeAttributesClassList
watch individual changes to a list of VolumeAttributesClass. deprecated: use the 'watch' parameter with a list operation instead.
watchStorageV1alpha1VolumeAttributesClass
watch changes to an object of kind VolumeAttributesClass. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-watch_volume_attributes_class_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-watch_volume_attributes_class_list}{}}}
\subsection{Method \code{watch_volume_attributes_class_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$watch_volume_attributes_class_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-watch_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-watch_volume_attributes_class}{}}}
\subsection{Method \code{watch_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$watch_volume_attributes_class(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1alpha1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1alpha1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1alpha1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
