% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_StorageV1Api.R
\name{StorageV1Api}
\alias{StorageV1Api}
\title{StorageV1Api}
\description{
StorageV1Api

StorageV1Api
}
\details{
Typed Kubernetes API client for the `StorageV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StorageV1Api-new}{\code{StorageV1Api$new()}}
\item \href{#method-StorageV1Api-get_api_resources}{\code{StorageV1Api$get_api_resources()}}
\item \href{#method-StorageV1Api-list_csi_driver}{\code{StorageV1Api$list_csi_driver()}}
\item \href{#method-StorageV1Api-create_csi_driver}{\code{StorageV1Api$create_csi_driver()}}
\item \href{#method-StorageV1Api-delete_collection_csi_driver}{\code{StorageV1Api$delete_collection_csi_driver()}}
\item \href{#method-StorageV1Api-read_csi_driver}{\code{StorageV1Api$read_csi_driver()}}
\item \href{#method-StorageV1Api-replace_csi_driver}{\code{StorageV1Api$replace_csi_driver()}}
\item \href{#method-StorageV1Api-patch_csi_driver}{\code{StorageV1Api$patch_csi_driver()}}
\item \href{#method-StorageV1Api-delete_csi_driver}{\code{StorageV1Api$delete_csi_driver()}}
\item \href{#method-StorageV1Api-list_csi_node}{\code{StorageV1Api$list_csi_node()}}
\item \href{#method-StorageV1Api-create_csi_node}{\code{StorageV1Api$create_csi_node()}}
\item \href{#method-StorageV1Api-delete_collection_csi_node}{\code{StorageV1Api$delete_collection_csi_node()}}
\item \href{#method-StorageV1Api-read_csi_node}{\code{StorageV1Api$read_csi_node()}}
\item \href{#method-StorageV1Api-replace_csi_node}{\code{StorageV1Api$replace_csi_node()}}
\item \href{#method-StorageV1Api-patch_csi_node}{\code{StorageV1Api$patch_csi_node()}}
\item \href{#method-StorageV1Api-delete_csi_node}{\code{StorageV1Api$delete_csi_node()}}
\item \href{#method-StorageV1Api-list_csi_storage_capacity_for_all_namespaces}{\code{StorageV1Api$list_csi_storage_capacity_for_all_namespaces()}}
\item \href{#method-StorageV1Api-list_namespaced_csi_storage_capacity}{\code{StorageV1Api$list_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-create_namespaced_csi_storage_capacity}{\code{StorageV1Api$create_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-delete_collection_namespaced_csi_storage_capacity}{\code{StorageV1Api$delete_collection_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-read_namespaced_csi_storage_capacity}{\code{StorageV1Api$read_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-replace_namespaced_csi_storage_capacity}{\code{StorageV1Api$replace_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-patch_namespaced_csi_storage_capacity}{\code{StorageV1Api$patch_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-delete_namespaced_csi_storage_capacity}{\code{StorageV1Api$delete_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-list_storage_class}{\code{StorageV1Api$list_storage_class()}}
\item \href{#method-StorageV1Api-create_storage_class}{\code{StorageV1Api$create_storage_class()}}
\item \href{#method-StorageV1Api-delete_collection_storage_class}{\code{StorageV1Api$delete_collection_storage_class()}}
\item \href{#method-StorageV1Api-read_storage_class}{\code{StorageV1Api$read_storage_class()}}
\item \href{#method-StorageV1Api-replace_storage_class}{\code{StorageV1Api$replace_storage_class()}}
\item \href{#method-StorageV1Api-patch_storage_class}{\code{StorageV1Api$patch_storage_class()}}
\item \href{#method-StorageV1Api-delete_storage_class}{\code{StorageV1Api$delete_storage_class()}}
\item \href{#method-StorageV1Api-list_volume_attachment}{\code{StorageV1Api$list_volume_attachment()}}
\item \href{#method-StorageV1Api-create_volume_attachment}{\code{StorageV1Api$create_volume_attachment()}}
\item \href{#method-StorageV1Api-delete_collection_volume_attachment}{\code{StorageV1Api$delete_collection_volume_attachment()}}
\item \href{#method-StorageV1Api-read_volume_attachment}{\code{StorageV1Api$read_volume_attachment()}}
\item \href{#method-StorageV1Api-replace_volume_attachment}{\code{StorageV1Api$replace_volume_attachment()}}
\item \href{#method-StorageV1Api-patch_volume_attachment}{\code{StorageV1Api$patch_volume_attachment()}}
\item \href{#method-StorageV1Api-delete_volume_attachment}{\code{StorageV1Api$delete_volume_attachment()}}
\item \href{#method-StorageV1Api-read_volume_attachment_status}{\code{StorageV1Api$read_volume_attachment_status()}}
\item \href{#method-StorageV1Api-replace_volume_attachment_status}{\code{StorageV1Api$replace_volume_attachment_status()}}
\item \href{#method-StorageV1Api-patch_volume_attachment_status}{\code{StorageV1Api$patch_volume_attachment_status()}}
\item \href{#method-StorageV1Api-watch_csi_driver_list}{\code{StorageV1Api$watch_csi_driver_list()}}
\item \href{#method-StorageV1Api-watch_csi_driver}{\code{StorageV1Api$watch_csi_driver()}}
\item \href{#method-StorageV1Api-watch_csi_node_list}{\code{StorageV1Api$watch_csi_node_list()}}
\item \href{#method-StorageV1Api-watch_csi_node}{\code{StorageV1Api$watch_csi_node()}}
\item \href{#method-StorageV1Api-watch_csi_storage_capacity_list_for_all_namespaces}{\code{StorageV1Api$watch_csi_storage_capacity_list_for_all_namespaces()}}
\item \href{#method-StorageV1Api-watch_namespaced_csi_storage_capacity_list}{\code{StorageV1Api$watch_namespaced_csi_storage_capacity_list()}}
\item \href{#method-StorageV1Api-watch_namespaced_csi_storage_capacity}{\code{StorageV1Api$watch_namespaced_csi_storage_capacity()}}
\item \href{#method-StorageV1Api-watch_storage_class_list}{\code{StorageV1Api$watch_storage_class_list()}}
\item \href{#method-StorageV1Api-watch_storage_class}{\code{StorageV1Api$watch_storage_class()}}
\item \href{#method-StorageV1Api-watch_volume_attachment_list}{\code{StorageV1Api$watch_volume_attachment_list()}}
\item \href{#method-StorageV1Api-watch_volume_attachment}{\code{StorageV1Api$watch_volume_attachment()}}
\item \href{#method-StorageV1Api-clone}{\code{StorageV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getStorageV1APIResources
get available resources
listStorageV1CSIDriver
list or watch objects of kind CSIDriver
createStorageV1CSIDriver
create a CSIDriver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-list_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-list_csi_driver}{}}}
\subsection{Method \code{list_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$list_csi_driver()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-create_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-create_csi_driver}{}}}
\subsection{Method \code{create_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$create_csi_driver(body, dry_run = NULL, field_validation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1CollectionCSIDriver
delete collection of CSIDriver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_collection_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_collection_csi_driver}{}}}
\subsection{Method \code{delete_collection_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_collection_csi_driver(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStorageV1CSIDriver
read the specified CSIDriver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-read_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-read_csi_driver}{}}}
\subsection{Method \code{read_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$read_csi_driver(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStorageV1CSIDriver
replace the specified CSIDriver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-replace_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-replace_csi_driver}{}}}
\subsection{Method \code{replace_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$replace_csi_driver(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1CSIDriver
partially update the specified CSIDriver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-patch_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-patch_csi_driver}{}}}
\subsection{Method \code{patch_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$patch_csi_driver(name, dry_run = NULL, field_validation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1CSIDriver
delete a CSIDriver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_csi_driver}{}}}
\subsection{Method \code{delete_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_csi_driver(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listStorageV1CSINode
list or watch objects of kind CSINode
createStorageV1CSINode
create a CSINode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-list_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-list_csi_node}{}}}
\subsection{Method \code{list_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$list_csi_node()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-create_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-create_csi_node}{}}}
\subsection{Method \code{create_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$create_csi_node(body, dry_run = NULL, field_validation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1CollectionCSINode
delete collection of CSINode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_collection_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_collection_csi_node}{}}}
\subsection{Method \code{delete_collection_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_collection_csi_node(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStorageV1CSINode
read the specified CSINode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-read_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-read_csi_node}{}}}
\subsection{Method \code{read_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$read_csi_node(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStorageV1CSINode
replace the specified CSINode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-replace_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-replace_csi_node}{}}}
\subsection{Method \code{replace_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$replace_csi_node(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1CSINode
partially update the specified CSINode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-patch_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-patch_csi_node}{}}}
\subsection{Method \code{patch_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$patch_csi_node(name, dry_run = NULL, field_validation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1CSINode
delete a CSINode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_csi_node}{}}}
\subsection{Method \code{delete_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_csi_node(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listStorageV1CSIStorageCapacityForAllNamespaces
list or watch objects of kind CSIStorageCapacity
listStorageV1NamespacedCSIStorageCapacity
list or watch objects of kind CSIStorageCapacity
createStorageV1NamespacedCSIStorageCapacity
create a CSIStorageCapacity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-list_csi_storage_capacity_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-list_csi_storage_capacity_for_all_namespaces}{}}}
\subsection{Method \code{list_csi_storage_capacity_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$list_csi_storage_capacity_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-list_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-list_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{list_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$list_namespaced_csi_storage_capacity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-create_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-create_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{create_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$create_namespaced_csi_storage_capacity(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1CollectionNamespacedCSIStorageCapacity
delete collection of CSIStorageCapacity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_collection_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_collection_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{delete_collection_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_collection_namespaced_csi_storage_capacity(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStorageV1NamespacedCSIStorageCapacity
read the specified CSIStorageCapacity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-read_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-read_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{read_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$read_namespaced_csi_storage_capacity(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStorageV1NamespacedCSIStorageCapacity
replace the specified CSIStorageCapacity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-replace_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-replace_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{replace_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$replace_namespaced_csi_storage_capacity(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1NamespacedCSIStorageCapacity
partially update the specified CSIStorageCapacity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-patch_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-patch_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{patch_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$patch_namespaced_csi_storage_capacity(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1NamespacedCSIStorageCapacity
delete a CSIStorageCapacity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{delete_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_namespaced_csi_storage_capacity(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listStorageV1StorageClass
list or watch objects of kind StorageClass
createStorageV1StorageClass
create a StorageClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-list_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-list_storage_class}{}}}
\subsection{Method \code{list_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$list_storage_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-create_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-create_storage_class}{}}}
\subsection{Method \code{create_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$create_storage_class(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1CollectionStorageClass
delete collection of StorageClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_collection_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_collection_storage_class}{}}}
\subsection{Method \code{delete_collection_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_collection_storage_class(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStorageV1StorageClass
read the specified StorageClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-read_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-read_storage_class}{}}}
\subsection{Method \code{read_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$read_storage_class(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStorageV1StorageClass
replace the specified StorageClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-replace_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-replace_storage_class}{}}}
\subsection{Method \code{replace_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$replace_storage_class(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1StorageClass
partially update the specified StorageClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-patch_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-patch_storage_class}{}}}
\subsection{Method \code{patch_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$patch_storage_class(name, dry_run = NULL, field_validation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1StorageClass
delete a StorageClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_storage_class}{}}}
\subsection{Method \code{delete_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_storage_class(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listStorageV1VolumeAttachment
list or watch objects of kind VolumeAttachment
createStorageV1VolumeAttachment
create a VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-list_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-list_volume_attachment}{}}}
\subsection{Method \code{list_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$list_volume_attachment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-create_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-create_volume_attachment}{}}}
\subsection{Method \code{create_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$create_volume_attachment(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1CollectionVolumeAttachment
delete collection of VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_collection_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_collection_volume_attachment}{}}}
\subsection{Method \code{delete_collection_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_collection_volume_attachment(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStorageV1VolumeAttachment
read the specified VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-read_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-read_volume_attachment}{}}}
\subsection{Method \code{read_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$read_volume_attachment(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStorageV1VolumeAttachment
replace the specified VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-replace_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-replace_volume_attachment}{}}}
\subsection{Method \code{replace_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$replace_volume_attachment(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1VolumeAttachment
partially update the specified VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-patch_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-patch_volume_attachment}{}}}
\subsection{Method \code{patch_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$patch_volume_attachment(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1VolumeAttachment
delete a VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-delete_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-delete_volume_attachment}{}}}
\subsection{Method \code{delete_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$delete_volume_attachment(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readStorageV1VolumeAttachmentStatus
read status of the specified VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-read_volume_attachment_status"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-read_volume_attachment_status}{}}}
\subsection{Method \code{read_volume_attachment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$read_volume_attachment_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceStorageV1VolumeAttachmentStatus
replace status of the specified VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-replace_volume_attachment_status"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-replace_volume_attachment_status}{}}}
\subsection{Method \code{replace_volume_attachment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$replace_volume_attachment_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1VolumeAttachmentStatus
partially update status of the specified VolumeAttachment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-patch_volume_attachment_status"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-patch_volume_attachment_status}{}}}
\subsection{Method \code{patch_volume_attachment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$patch_volume_attachment_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
watchStorageV1CSIDriverList
watch individual changes to a list of CSIDriver. deprecated: use the 'watch' parameter with a list operation instead.
watchStorageV1CSIDriver
watch changes to an object of kind CSIDriver. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_csi_driver_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_csi_driver_list}{}}}
\subsection{Method \code{watch_csi_driver_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_csi_driver_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_csi_driver"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_csi_driver}{}}}
\subsection{Method \code{watch_csi_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_csi_driver(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchStorageV1CSINodeList
watch individual changes to a list of CSINode. deprecated: use the 'watch' parameter with a list operation instead.
watchStorageV1CSINode
watch changes to an object of kind CSINode. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_csi_node_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_csi_node_list}{}}}
\subsection{Method \code{watch_csi_node_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_csi_node_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_csi_node"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_csi_node}{}}}
\subsection{Method \code{watch_csi_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_csi_node(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchStorageV1CSIStorageCapacityListForAllNamespaces
watch individual changes to a list of CSIStorageCapacity. deprecated: use the 'watch' parameter with a list operation instead.
watchStorageV1NamespacedCSIStorageCapacityList
watch individual changes to a list of CSIStorageCapacity. deprecated: use the 'watch' parameter with a list operation instead.
watchStorageV1NamespacedCSIStorageCapacity
watch changes to an object of kind CSIStorageCapacity. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_csi_storage_capacity_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_csi_storage_capacity_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_csi_storage_capacity_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_csi_storage_capacity_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_namespaced_csi_storage_capacity_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_namespaced_csi_storage_capacity_list}{}}}
\subsection{Method \code{watch_namespaced_csi_storage_capacity_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_namespaced_csi_storage_capacity_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_namespaced_csi_storage_capacity"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_namespaced_csi_storage_capacity}{}}}
\subsection{Method \code{watch_namespaced_csi_storage_capacity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_namespaced_csi_storage_capacity(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchStorageV1StorageClassList
watch individual changes to a list of StorageClass. deprecated: use the 'watch' parameter with a list operation instead.
watchStorageV1StorageClass
watch changes to an object of kind StorageClass. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_storage_class_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_storage_class_list}{}}}
\subsection{Method \code{watch_storage_class_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_storage_class_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_storage_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_storage_class}{}}}
\subsection{Method \code{watch_storage_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_storage_class(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchStorageV1VolumeAttachmentList
watch individual changes to a list of VolumeAttachment. deprecated: use the 'watch' parameter with a list operation instead.
watchStorageV1VolumeAttachment
watch changes to an object of kind VolumeAttachment. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_volume_attachment_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_volume_attachment_list}{}}}
\subsection{Method \code{watch_volume_attachment_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_volume_attachment_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-watch_volume_attachment"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-watch_volume_attachment}{}}}
\subsection{Method \code{watch_volume_attachment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$watch_volume_attachment(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
