% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_StorageV1beta1Api.R
\name{StorageV1beta1Api}
\alias{StorageV1beta1Api}
\title{StorageV1beta1Api}
\description{
StorageV1beta1Api

StorageV1beta1Api
}
\details{
Typed Kubernetes API client for the `StorageV1beta1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StorageV1beta1Api-new}{\code{StorageV1beta1Api$new()}}
\item \href{#method-StorageV1beta1Api-get_api_resources}{\code{StorageV1beta1Api$get_api_resources()}}
\item \href{#method-StorageV1beta1Api-list_volume_attributes_class}{\code{StorageV1beta1Api$list_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-create_volume_attributes_class}{\code{StorageV1beta1Api$create_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-delete_collection_volume_attributes_class}{\code{StorageV1beta1Api$delete_collection_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-read_volume_attributes_class}{\code{StorageV1beta1Api$read_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-replace_volume_attributes_class}{\code{StorageV1beta1Api$replace_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-patch_volume_attributes_class}{\code{StorageV1beta1Api$patch_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-delete_volume_attributes_class}{\code{StorageV1beta1Api$delete_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-watch_volume_attributes_class_list}{\code{StorageV1beta1Api$watch_volume_attributes_class_list()}}
\item \href{#method-StorageV1beta1Api-watch_volume_attributes_class}{\code{StorageV1beta1Api$watch_volume_attributes_class()}}
\item \href{#method-StorageV1beta1Api-clone}{\code{StorageV1beta1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getStorageV1beta1APIResources
get available resources
listStorageV1beta1VolumeAttributesClass
list or watch objects of kind VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-list_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-list_volume_attributes_class}{}}}
\subsection{Method \code{list_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$list_volume_attributes_class(
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createStorageV1beta1VolumeAttributesClass
create a VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-create_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-create_volume_attributes_class}{}}}
\subsection{Method \code{create_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$create_volume_attributes_class(
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteStorageV1beta1CollectionVolumeAttributesClass
delete collection of VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-delete_collection_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-delete_collection_volume_attributes_class}{}}}
\subsection{Method \code{delete_collection_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$delete_collection_volume_attributes_class(
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readStorageV1beta1VolumeAttributesClass
read the specified VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-read_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-read_volume_attributes_class}{}}}
\subsection{Method \code{read_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$read_volume_attributes_class(name, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceStorageV1beta1VolumeAttributesClass
replace the specified VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-replace_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-replace_volume_attributes_class}{}}}
\subsection{Method \code{replace_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$replace_volume_attributes_class(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchStorageV1beta1VolumeAttributesClass
partially update the specified VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-patch_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-patch_volume_attributes_class}{}}}
\subsection{Method \code{patch_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$patch_volume_attributes_class(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteStorageV1beta1VolumeAttributesClass
delete a VolumeAttributesClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-delete_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-delete_volume_attributes_class}{}}}
\subsection{Method \code{delete_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$delete_volume_attributes_class(
  name,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
watchStorageV1beta1VolumeAttributesClassList
watch individual changes to a list of VolumeAttributesClass. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-watch_volume_attributes_class_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-watch_volume_attributes_class_list}{}}}
\subsection{Method \code{watch_volume_attributes_class_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$watch_volume_attributes_class_list(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchStorageV1beta1VolumeAttributesClass
watch changes to an object of kind VolumeAttributesClass. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-watch_volume_attributes_class"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-watch_volume_attributes_class}{}}}
\subsection{Method \code{watch_volume_attributes_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$watch_volume_attributes_class(
  name,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1beta1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1beta1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1beta1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
