% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_StorageV1TokenRequest.R
\name{StorageV1TokenRequest}
\alias{StorageV1TokenRequest}
\title{StorageV1TokenRequest}
\description{
StorageV1TokenRequest

StorageV1TokenRequest
}
\details{
TokenRequest contains parameters of a service account token.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StorageV1TokenRequest-new}{\code{StorageV1TokenRequest$new()}}
\item \href{#method-StorageV1TokenRequest-to_list}{\code{StorageV1TokenRequest$to_list()}}
\item \href{#method-StorageV1TokenRequest-clone}{\code{StorageV1TokenRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1TokenRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1TokenRequest-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1TokenRequest$new(audience = NULL, expiration_seconds = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1TokenRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1TokenRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1TokenRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageV1TokenRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StorageV1TokenRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageV1TokenRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
