% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Affinity.R
\name{V1Affinity}
\alias{V1Affinity}
\title{V1Affinity}
\description{
V1Affinity

V1Affinity
}
\details{
Affinity is a group of affinity scheduling rules.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Affinity-new}{\code{V1Affinity$new()}}
\item \href{#method-V1Affinity-to_list}{\code{V1Affinity$to_list()}}
\item \href{#method-V1Affinity-clone}{\code{V1Affinity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Affinity-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Affinity-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Affinity$new(
  node_affinity = NULL,
  pod_affinity = NULL,
  pod_anti_affinity = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Affinity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Affinity-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Affinity$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Affinity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Affinity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Affinity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
