% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1AggregationRule.R
\name{V1AggregationRule}
\alias{V1AggregationRule}
\title{V1AggregationRule}
\description{
V1AggregationRule

V1AggregationRule
}
\details{
AggregationRule describes how to locate ClusterRoles to aggregate into the ClusterRole
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1AggregationRule-new}{\code{V1AggregationRule$new()}}
\item \href{#method-V1AggregationRule-to_list}{\code{V1AggregationRule$to_list()}}
\item \href{#method-V1AggregationRule-clone}{\code{V1AggregationRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AggregationRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1AggregationRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AggregationRule$new(cluster_role_selectors = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AggregationRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1AggregationRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AggregationRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AggregationRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1AggregationRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AggregationRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
