% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1ClusterTrustBundle.R
\name{V1alpha1ClusterTrustBundle}
\alias{V1alpha1ClusterTrustBundle}
\title{V1alpha1ClusterTrustBundle}
\description{
V1alpha1ClusterTrustBundle

V1alpha1ClusterTrustBundle
}
\details{
ClusterTrustBundle is a cluster-scoped container for X.509 trust anchors (root certificates).

ClusterTrustBundle objects are considered to be readable by any authenticated user in the cluster, because they can be mounted by pods using the `clusterTrustBundle` projection.  All service accounts have read access to ClusterTrustBundles by default.  Users who only have namespace-level access to a cluster can read ClusterTrustBundles by impersonating a serviceaccount that they have access to.

It can be optionally associated with a particular assigner, in which case it contains one valid set of trust anchors for that signer. Signers may have multiple associated ClusterTrustBundles; each is an independent set of trust anchors for that signer. Admission control is used to enforce that only users with permissions on the signer can create or modify the corresponding bundle.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1ClusterTrustBundle-new}{\code{V1alpha1ClusterTrustBundle$new()}}
\item \href{#method-V1alpha1ClusterTrustBundle-to_list}{\code{V1alpha1ClusterTrustBundle$to_list()}}
\item \href{#method-V1alpha1ClusterTrustBundle-clone}{\code{V1alpha1ClusterTrustBundle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ClusterTrustBundle-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ClusterTrustBundle-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ClusterTrustBundle$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ClusterTrustBundle-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ClusterTrustBundle-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ClusterTrustBundle$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ClusterTrustBundle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ClusterTrustBundle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ClusterTrustBundle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
