% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1ClusterTrustBundleSpec.R
\name{V1alpha1ClusterTrustBundleSpec}
\alias{V1alpha1ClusterTrustBundleSpec}
\title{V1alpha1ClusterTrustBundleSpec}
\description{
V1alpha1ClusterTrustBundleSpec

V1alpha1ClusterTrustBundleSpec
}
\details{
ClusterTrustBundleSpec contains the signer and trust anchors.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1ClusterTrustBundleSpec-new}{\code{V1alpha1ClusterTrustBundleSpec$new()}}
\item \href{#method-V1alpha1ClusterTrustBundleSpec-to_list}{\code{V1alpha1ClusterTrustBundleSpec$to_list()}}
\item \href{#method-V1alpha1ClusterTrustBundleSpec-clone}{\code{V1alpha1ClusterTrustBundleSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ClusterTrustBundleSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ClusterTrustBundleSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ClusterTrustBundleSpec$new(signer_name = NULL, trust_bundle = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ClusterTrustBundleSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ClusterTrustBundleSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ClusterTrustBundleSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ClusterTrustBundleSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ClusterTrustBundleSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ClusterTrustBundleSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
