% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1ExpressionWarning.R
\name{V1alpha1ExpressionWarning}
\alias{V1alpha1ExpressionWarning}
\title{V1alpha1ExpressionWarning}
\description{
V1alpha1ExpressionWarning

V1alpha1ExpressionWarning
}
\details{
ExpressionWarning is a warning information that targets a specific expression.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1ExpressionWarning-new}{\code{V1alpha1ExpressionWarning$new()}}
\item \href{#method-V1alpha1ExpressionWarning-to_list}{\code{V1alpha1ExpressionWarning$to_list()}}
\item \href{#method-V1alpha1ExpressionWarning-clone}{\code{V1alpha1ExpressionWarning$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ExpressionWarning-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ExpressionWarning-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ExpressionWarning$new(field_ref = NULL, warning = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ExpressionWarning-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ExpressionWarning-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ExpressionWarning$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ExpressionWarning-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ExpressionWarning-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ExpressionWarning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
