% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1GroupVersionResource.R
\name{V1alpha1GroupVersionResource}
\alias{V1alpha1GroupVersionResource}
\title{V1alpha1GroupVersionResource}
\description{
V1alpha1GroupVersionResource

V1alpha1GroupVersionResource
}
\details{
The names of the group, the version, and the resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1GroupVersionResource-new}{\code{V1alpha1GroupVersionResource$new()}}
\item \href{#method-V1alpha1GroupVersionResource-to_list}{\code{V1alpha1GroupVersionResource$to_list()}}
\item \href{#method-V1alpha1GroupVersionResource-clone}{\code{V1alpha1GroupVersionResource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1GroupVersionResource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1GroupVersionResource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1GroupVersionResource$new(group = NULL, resource = NULL, version = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1GroupVersionResource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1GroupVersionResource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1GroupVersionResource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1GroupVersionResource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1GroupVersionResource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1GroupVersionResource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
