% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1LeaseCandidate.R
\name{V1alpha1LeaseCandidate}
\alias{V1alpha1LeaseCandidate}
\title{V1alpha1LeaseCandidate}
\description{
V1alpha1LeaseCandidate

V1alpha1LeaseCandidate
}
\details{
LeaseCandidate defines a candidate for a Lease object. Candidates are created such that coordinated leader election will pick the best leader from the list of candidates.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1LeaseCandidate-new}{\code{V1alpha1LeaseCandidate$new()}}
\item \href{#method-V1alpha1LeaseCandidate-to_list}{\code{V1alpha1LeaseCandidate$to_list()}}
\item \href{#method-V1alpha1LeaseCandidate-clone}{\code{V1alpha1LeaseCandidate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1LeaseCandidate-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1LeaseCandidate-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1LeaseCandidate$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1LeaseCandidate-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1LeaseCandidate-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1LeaseCandidate$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1LeaseCandidate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1LeaseCandidate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1LeaseCandidate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
