% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1LeaseCandidateSpec.R
\name{V1alpha1LeaseCandidateSpec}
\alias{V1alpha1LeaseCandidateSpec}
\title{V1alpha1LeaseCandidateSpec}
\description{
V1alpha1LeaseCandidateSpec

V1alpha1LeaseCandidateSpec
}
\details{
LeaseCandidateSpec is a specification of a Lease.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1LeaseCandidateSpec-new}{\code{V1alpha1LeaseCandidateSpec$new()}}
\item \href{#method-V1alpha1LeaseCandidateSpec-to_list}{\code{V1alpha1LeaseCandidateSpec$to_list()}}
\item \href{#method-V1alpha1LeaseCandidateSpec-clone}{\code{V1alpha1LeaseCandidateSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1LeaseCandidateSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1LeaseCandidateSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1LeaseCandidateSpec$new(
  binary_version = NULL,
  emulation_version = NULL,
  lease_name = NULL,
  ping_time = NULL,
  preferred_strategies = NULL,
  renew_time = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1LeaseCandidateSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1LeaseCandidateSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1LeaseCandidateSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1LeaseCandidateSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1LeaseCandidateSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1LeaseCandidateSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
