% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1MigrationCondition.R
\name{V1alpha1MigrationCondition}
\alias{V1alpha1MigrationCondition}
\title{V1alpha1MigrationCondition}
\description{
V1alpha1MigrationCondition

V1alpha1MigrationCondition
}
\details{
Describes the state of a migration at a certain point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1MigrationCondition-new}{\code{V1alpha1MigrationCondition$new()}}
\item \href{#method-V1alpha1MigrationCondition-to_list}{\code{V1alpha1MigrationCondition$to_list()}}
\item \href{#method-V1alpha1MigrationCondition-clone}{\code{V1alpha1MigrationCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1MigrationCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1MigrationCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1MigrationCondition$new(
  last_update_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1MigrationCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1MigrationCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1MigrationCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1MigrationCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1MigrationCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1MigrationCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
