% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1NamedRuleWithOperations.R
\name{V1alpha1NamedRuleWithOperations}
\alias{V1alpha1NamedRuleWithOperations}
\title{V1alpha1NamedRuleWithOperations}
\description{
V1alpha1NamedRuleWithOperations

V1alpha1NamedRuleWithOperations
}
\details{
NamedRuleWithOperations is a tuple of Operations and Resources with ResourceNames.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1NamedRuleWithOperations-new}{\code{V1alpha1NamedRuleWithOperations$new()}}
\item \href{#method-V1alpha1NamedRuleWithOperations-to_list}{\code{V1alpha1NamedRuleWithOperations$to_list()}}
\item \href{#method-V1alpha1NamedRuleWithOperations-clone}{\code{V1alpha1NamedRuleWithOperations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1NamedRuleWithOperations-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1NamedRuleWithOperations-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1NamedRuleWithOperations$new(
  api_groups = NULL,
  api_versions = NULL,
  operations = NULL,
  resource_names = NULL,
  resources = NULL,
  scope = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1NamedRuleWithOperations-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1NamedRuleWithOperations-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1NamedRuleWithOperations$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1NamedRuleWithOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1NamedRuleWithOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1NamedRuleWithOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
