% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1ServerStorageVersion.R
\name{V1alpha1ServerStorageVersion}
\alias{V1alpha1ServerStorageVersion}
\title{V1alpha1ServerStorageVersion}
\description{
V1alpha1ServerStorageVersion

V1alpha1ServerStorageVersion
}
\details{
An API server instance reports the version it can decode and the version it encodes objects to when persisting objects in the backend.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1ServerStorageVersion-new}{\code{V1alpha1ServerStorageVersion$new()}}
\item \href{#method-V1alpha1ServerStorageVersion-to_list}{\code{V1alpha1ServerStorageVersion$to_list()}}
\item \href{#method-V1alpha1ServerStorageVersion-clone}{\code{V1alpha1ServerStorageVersion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ServerStorageVersion-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ServerStorageVersion-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ServerStorageVersion$new(
  api_server_id = NULL,
  decodable_versions = NULL,
  encoding_version = NULL,
  served_versions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ServerStorageVersion-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ServerStorageVersion-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ServerStorageVersion$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ServerStorageVersion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ServerStorageVersion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ServerStorageVersion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
