% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1StorageVersion.R
\name{V1alpha1StorageVersion}
\alias{V1alpha1StorageVersion}
\title{V1alpha1StorageVersion}
\description{
V1alpha1StorageVersion

V1alpha1StorageVersion
}
\details{
Storage version of a specific resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1StorageVersion-new}{\code{V1alpha1StorageVersion$new()}}
\item \href{#method-V1alpha1StorageVersion-to_list}{\code{V1alpha1StorageVersion$to_list()}}
\item \href{#method-V1alpha1StorageVersion-clone}{\code{V1alpha1StorageVersion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersion-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersion-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersion$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersion-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersion-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersion$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
