% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1StorageVersionCondition.R
\name{V1alpha1StorageVersionCondition}
\alias{V1alpha1StorageVersionCondition}
\title{V1alpha1StorageVersionCondition}
\description{
V1alpha1StorageVersionCondition

V1alpha1StorageVersionCondition
}
\details{
Describes the state of the storageVersion at a certain point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1StorageVersionCondition-new}{\code{V1alpha1StorageVersionCondition$new()}}
\item \href{#method-V1alpha1StorageVersionCondition-to_list}{\code{V1alpha1StorageVersionCondition$to_list()}}
\item \href{#method-V1alpha1StorageVersionCondition-clone}{\code{V1alpha1StorageVersionCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionCondition$new(
  last_transition_time = NULL,
  message = NULL,
  observed_generation = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
