% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1StorageVersionList.R
\name{V1alpha1StorageVersionList}
\alias{V1alpha1StorageVersionList}
\title{V1alpha1StorageVersionList}
\description{
V1alpha1StorageVersionList

V1alpha1StorageVersionList
}
\details{
A list of StorageVersions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1StorageVersionList-new}{\code{V1alpha1StorageVersionList$new()}}
\item \href{#method-V1alpha1StorageVersionList-to_list}{\code{V1alpha1StorageVersionList$to_list()}}
\item \href{#method-V1alpha1StorageVersionList-clone}{\code{V1alpha1StorageVersionList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
