% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1StorageVersionMigrationSpec.R
\name{V1alpha1StorageVersionMigrationSpec}
\alias{V1alpha1StorageVersionMigrationSpec}
\title{V1alpha1StorageVersionMigrationSpec}
\description{
V1alpha1StorageVersionMigrationSpec

V1alpha1StorageVersionMigrationSpec
}
\details{
Spec of the storage version migration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1StorageVersionMigrationSpec-new}{\code{V1alpha1StorageVersionMigrationSpec$new()}}
\item \href{#method-V1alpha1StorageVersionMigrationSpec-to_list}{\code{V1alpha1StorageVersionMigrationSpec$to_list()}}
\item \href{#method-V1alpha1StorageVersionMigrationSpec-clone}{\code{V1alpha1StorageVersionMigrationSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionMigrationSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionMigrationSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionMigrationSpec$new(continue_token = NULL, resource = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionMigrationSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionMigrationSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionMigrationSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionMigrationSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionMigrationSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionMigrationSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
