% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1alpha1StorageVersionMigrationStatus.R
\name{V1alpha1StorageVersionMigrationStatus}
\alias{V1alpha1StorageVersionMigrationStatus}
\title{V1alpha1StorageVersionMigrationStatus}
\description{
V1alpha1StorageVersionMigrationStatus

V1alpha1StorageVersionMigrationStatus
}
\details{
Status of the storage version migration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1StorageVersionMigrationStatus-new}{\code{V1alpha1StorageVersionMigrationStatus$new()}}
\item \href{#method-V1alpha1StorageVersionMigrationStatus-to_list}{\code{V1alpha1StorageVersionMigrationStatus$to_list()}}
\item \href{#method-V1alpha1StorageVersionMigrationStatus-clone}{\code{V1alpha1StorageVersionMigrationStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionMigrationStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionMigrationStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionMigrationStatus$new(
  conditions = NULL,
  resource_version = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionMigrationStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionMigrationStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionMigrationStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionMigrationStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionMigrationStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionMigrationStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
