% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1StorageVersionSpec.R
\name{V1alpha1StorageVersionSpec}
\alias{V1alpha1StorageVersionSpec}
\title{V1alpha1StorageVersionSpec}
\description{
V1alpha1StorageVersionSpec

V1alpha1StorageVersionSpec
}
\details{
StorageVersionSpec is an empty spec.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1StorageVersionSpec-new}{\code{V1alpha1StorageVersionSpec$new()}}
\item \href{#method-V1alpha1StorageVersionSpec-to_list}{\code{V1alpha1StorageVersionSpec$to_list()}}
\item \href{#method-V1alpha1StorageVersionSpec-clone}{\code{V1alpha1StorageVersionSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionSpec$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
