% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha1StorageVersionStatus.R
\name{V1alpha1StorageVersionStatus}
\alias{V1alpha1StorageVersionStatus}
\title{V1alpha1StorageVersionStatus}
\description{
V1alpha1StorageVersionStatus

V1alpha1StorageVersionStatus
}
\details{
API server instances report the versions they can decode and the version they encode objects to when persisting objects in the backend.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1StorageVersionStatus-new}{\code{V1alpha1StorageVersionStatus$new()}}
\item \href{#method-V1alpha1StorageVersionStatus-to_list}{\code{V1alpha1StorageVersionStatus$to_list()}}
\item \href{#method-V1alpha1StorageVersionStatus-clone}{\code{V1alpha1StorageVersionStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionStatus$new(
  common_encoding_version = NULL,
  conditions = NULL,
  storage_versions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1StorageVersionStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1StorageVersionStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1StorageVersionStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
