% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1alpha1ValidatingAdmissionPolicySpec.R
\name{V1alpha1ValidatingAdmissionPolicySpec}
\alias{V1alpha1ValidatingAdmissionPolicySpec}
\title{V1alpha1ValidatingAdmissionPolicySpec}
\description{
V1alpha1ValidatingAdmissionPolicySpec

V1alpha1ValidatingAdmissionPolicySpec
}
\details{
ValidatingAdmissionPolicySpec is the specification of the desired behavior of the AdmissionPolicy.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha1ValidatingAdmissionPolicySpec-new}{\code{V1alpha1ValidatingAdmissionPolicySpec$new()}}
\item \href{#method-V1alpha1ValidatingAdmissionPolicySpec-to_list}{\code{V1alpha1ValidatingAdmissionPolicySpec$to_list()}}
\item \href{#method-V1alpha1ValidatingAdmissionPolicySpec-clone}{\code{V1alpha1ValidatingAdmissionPolicySpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ValidatingAdmissionPolicySpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ValidatingAdmissionPolicySpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ValidatingAdmissionPolicySpec$new(
  audit_annotations = NULL,
  failure_policy = NULL,
  match_conditions = NULL,
  match_constraints = NULL,
  param_kind = NULL,
  validations = NULL,
  variables = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ValidatingAdmissionPolicySpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ValidatingAdmissionPolicySpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ValidatingAdmissionPolicySpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha1ValidatingAdmissionPolicySpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha1ValidatingAdmissionPolicySpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha1ValidatingAdmissionPolicySpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
