% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3BasicDevice.R
\name{V1alpha3BasicDevice}
\alias{V1alpha3BasicDevice}
\title{V1alpha3BasicDevice}
\description{
V1alpha3BasicDevice

V1alpha3BasicDevice
}
\details{
BasicDevice defines one device instance.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3BasicDevice-new}{\code{V1alpha3BasicDevice$new()}}
\item \href{#method-V1alpha3BasicDevice-to_list}{\code{V1alpha3BasicDevice$to_list()}}
\item \href{#method-V1alpha3BasicDevice-clone}{\code{V1alpha3BasicDevice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3BasicDevice-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3BasicDevice-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3BasicDevice$new(attributes = NULL, capacity = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3BasicDevice-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3BasicDevice-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3BasicDevice$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3BasicDevice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3BasicDevice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3BasicDevice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
