% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3CELDeviceSelector.R
\name{V1alpha3CELDeviceSelector}
\alias{V1alpha3CELDeviceSelector}
\title{V1alpha3CELDeviceSelector}
\description{
V1alpha3CELDeviceSelector

V1alpha3CELDeviceSelector
}
\details{
CELDeviceSelector contains a CEL expression for selecting a device.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3CELDeviceSelector-new}{\code{V1alpha3CELDeviceSelector$new()}}
\item \href{#method-V1alpha3CELDeviceSelector-to_list}{\code{V1alpha3CELDeviceSelector$to_list()}}
\item \href{#method-V1alpha3CELDeviceSelector-clone}{\code{V1alpha3CELDeviceSelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3CELDeviceSelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3CELDeviceSelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3CELDeviceSelector$new(expression = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3CELDeviceSelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3CELDeviceSelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3CELDeviceSelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3CELDeviceSelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3CELDeviceSelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3CELDeviceSelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
