% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3Device.R
\name{V1alpha3Device}
\alias{V1alpha3Device}
\title{V1alpha3Device}
\description{
V1alpha3Device

V1alpha3Device
}
\details{
Device represents one individual hardware instance that can be selected based on its attributes. Besides the name, exactly one field must be set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3Device-new}{\code{V1alpha3Device$new()}}
\item \href{#method-V1alpha3Device-to_list}{\code{V1alpha3Device$to_list()}}
\item \href{#method-V1alpha3Device-clone}{\code{V1alpha3Device$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3Device-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3Device-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3Device$new(basic = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3Device-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3Device-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3Device$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3Device-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3Device-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3Device$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
