% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceAttribute.R
\name{V1alpha3DeviceAttribute}
\alias{V1alpha3DeviceAttribute}
\title{V1alpha3DeviceAttribute}
\description{
V1alpha3DeviceAttribute

V1alpha3DeviceAttribute
}
\details{
DeviceAttribute must have exactly one field set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceAttribute-new}{\code{V1alpha3DeviceAttribute$new()}}
\item \href{#method-V1alpha3DeviceAttribute-to_list}{\code{V1alpha3DeviceAttribute$to_list()}}
\item \href{#method-V1alpha3DeviceAttribute-clone}{\code{V1alpha3DeviceAttribute$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceAttribute-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceAttribute-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceAttribute$new(
  bool = NULL,
  int = NULL,
  string = NULL,
  version = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceAttribute-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceAttribute-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceAttribute$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceAttribute-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceAttribute-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceAttribute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
