% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceClaim.R
\name{V1alpha3DeviceClaim}
\alias{V1alpha3DeviceClaim}
\title{V1alpha3DeviceClaim}
\description{
V1alpha3DeviceClaim

V1alpha3DeviceClaim
}
\details{
DeviceClaim defines how to request devices with a ResourceClaim.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceClaim-new}{\code{V1alpha3DeviceClaim$new()}}
\item \href{#method-V1alpha3DeviceClaim-to_list}{\code{V1alpha3DeviceClaim$to_list()}}
\item \href{#method-V1alpha3DeviceClaim-clone}{\code{V1alpha3DeviceClaim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClaim-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClaim-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClaim$new(config = NULL, constraints = NULL, requests = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClaim-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClaim-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClaim$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClaim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClaim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClaim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
