% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceClaimConfiguration.R
\name{V1alpha3DeviceClaimConfiguration}
\alias{V1alpha3DeviceClaimConfiguration}
\title{V1alpha3DeviceClaimConfiguration}
\description{
V1alpha3DeviceClaimConfiguration

V1alpha3DeviceClaimConfiguration
}
\details{
DeviceClaimConfiguration is used for configuration parameters in DeviceClaim.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceClaimConfiguration-new}{\code{V1alpha3DeviceClaimConfiguration$new()}}
\item \href{#method-V1alpha3DeviceClaimConfiguration-to_list}{\code{V1alpha3DeviceClaimConfiguration$to_list()}}
\item \href{#method-V1alpha3DeviceClaimConfiguration-clone}{\code{V1alpha3DeviceClaimConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClaimConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClaimConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClaimConfiguration$new(opaque = NULL, requests = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClaimConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClaimConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClaimConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClaimConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClaimConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClaimConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
