% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceClass.R
\name{V1alpha3DeviceClass}
\alias{V1alpha3DeviceClass}
\title{V1alpha3DeviceClass}
\description{
V1alpha3DeviceClass

V1alpha3DeviceClass
}
\details{
DeviceClass is a vendor- or admin-provided resource that contains device configuration and selectors. It can be referenced in the device requests of a claim to apply these presets. Cluster scoped.

This is an alpha type and requires enabling the DynamicResourceAllocation feature gate.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceClass-new}{\code{V1alpha3DeviceClass$new()}}
\item \href{#method-V1alpha3DeviceClass-to_list}{\code{V1alpha3DeviceClass$to_list()}}
\item \href{#method-V1alpha3DeviceClass-clone}{\code{V1alpha3DeviceClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClass-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClass$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClass-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClass-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClass$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
