% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceClassConfiguration.R
\name{V1alpha3DeviceClassConfiguration}
\alias{V1alpha3DeviceClassConfiguration}
\title{V1alpha3DeviceClassConfiguration}
\description{
V1alpha3DeviceClassConfiguration

V1alpha3DeviceClassConfiguration
}
\details{
DeviceClassConfiguration is used in DeviceClass.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceClassConfiguration-new}{\code{V1alpha3DeviceClassConfiguration$new()}}
\item \href{#method-V1alpha3DeviceClassConfiguration-to_list}{\code{V1alpha3DeviceClassConfiguration$to_list()}}
\item \href{#method-V1alpha3DeviceClassConfiguration-clone}{\code{V1alpha3DeviceClassConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClassConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClassConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClassConfiguration$new(opaque = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClassConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClassConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClassConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClassConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClassConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClassConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
