% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceClassSpec.R
\name{V1alpha3DeviceClassSpec}
\alias{V1alpha3DeviceClassSpec}
\title{V1alpha3DeviceClassSpec}
\description{
V1alpha3DeviceClassSpec

V1alpha3DeviceClassSpec
}
\details{
DeviceClassSpec is used in a DeviceClass to define what can be allocated and how to configure it.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceClassSpec-new}{\code{V1alpha3DeviceClassSpec$new()}}
\item \href{#method-V1alpha3DeviceClassSpec-to_list}{\code{V1alpha3DeviceClassSpec$to_list()}}
\item \href{#method-V1alpha3DeviceClassSpec-clone}{\code{V1alpha3DeviceClassSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClassSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClassSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClassSpec$new(
  config = NULL,
  selectors = NULL,
  suitable_nodes = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClassSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClassSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClassSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceClassSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceClassSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceClassSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
