% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceConstraint.R
\name{V1alpha3DeviceConstraint}
\alias{V1alpha3DeviceConstraint}
\title{V1alpha3DeviceConstraint}
\description{
V1alpha3DeviceConstraint

V1alpha3DeviceConstraint
}
\details{
DeviceConstraint must have exactly one field set besides Requests.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceConstraint-new}{\code{V1alpha3DeviceConstraint$new()}}
\item \href{#method-V1alpha3DeviceConstraint-to_list}{\code{V1alpha3DeviceConstraint$to_list()}}
\item \href{#method-V1alpha3DeviceConstraint-clone}{\code{V1alpha3DeviceConstraint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceConstraint-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceConstraint-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceConstraint$new(match_attribute = NULL, requests = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceConstraint-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceConstraint-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceConstraint$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceConstraint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceConstraint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceConstraint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
