% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceRequest.R
\name{V1alpha3DeviceRequest}
\alias{V1alpha3DeviceRequest}
\title{V1alpha3DeviceRequest}
\description{
V1alpha3DeviceRequest

V1alpha3DeviceRequest
}
\details{
DeviceRequest is a request for devices required for a claim. This is typically a request for a single resource like a device, but can also ask for several identical devices.

A DeviceClassName is currently required. Clients must check that it is indeed set. It's absence indicates that something changed in a way that is not supported by the client yet, in which case it must refuse to handle the request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceRequest-new}{\code{V1alpha3DeviceRequest$new()}}
\item \href{#method-V1alpha3DeviceRequest-to_list}{\code{V1alpha3DeviceRequest$to_list()}}
\item \href{#method-V1alpha3DeviceRequest-clone}{\code{V1alpha3DeviceRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceRequest-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceRequest$new(
  admin_access = NULL,
  allocation_mode = NULL,
  count = NULL,
  device_class_name = NULL,
  name = NULL,
  selectors = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
