% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1alpha3DeviceRequestAllocationResult.R
\name{V1alpha3DeviceRequestAllocationResult}
\alias{V1alpha3DeviceRequestAllocationResult}
\title{V1alpha3DeviceRequestAllocationResult}
\description{
V1alpha3DeviceRequestAllocationResult

V1alpha3DeviceRequestAllocationResult
}
\details{
DeviceRequestAllocationResult contains the allocation result for one request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceRequestAllocationResult-new}{\code{V1alpha3DeviceRequestAllocationResult$new()}}
\item \href{#method-V1alpha3DeviceRequestAllocationResult-to_list}{\code{V1alpha3DeviceRequestAllocationResult$to_list()}}
\item \href{#method-V1alpha3DeviceRequestAllocationResult-clone}{\code{V1alpha3DeviceRequestAllocationResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceRequestAllocationResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceRequestAllocationResult-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceRequestAllocationResult$new(
  device = NULL,
  driver = NULL,
  pool = NULL,
  request = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceRequestAllocationResult-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceRequestAllocationResult-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceRequestAllocationResult$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceRequestAllocationResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceRequestAllocationResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceRequestAllocationResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
