% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3DeviceSelector.R
\name{V1alpha3DeviceSelector}
\alias{V1alpha3DeviceSelector}
\title{V1alpha3DeviceSelector}
\description{
V1alpha3DeviceSelector

V1alpha3DeviceSelector
}
\details{
DeviceSelector must have exactly one field set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3DeviceSelector-new}{\code{V1alpha3DeviceSelector$new()}}
\item \href{#method-V1alpha3DeviceSelector-to_list}{\code{V1alpha3DeviceSelector$to_list()}}
\item \href{#method-V1alpha3DeviceSelector-clone}{\code{V1alpha3DeviceSelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceSelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceSelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceSelector$new(cel = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceSelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceSelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceSelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3DeviceSelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3DeviceSelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3DeviceSelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
