% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3OpaqueDeviceConfiguration.R
\name{V1alpha3OpaqueDeviceConfiguration}
\alias{V1alpha3OpaqueDeviceConfiguration}
\title{V1alpha3OpaqueDeviceConfiguration}
\description{
V1alpha3OpaqueDeviceConfiguration

V1alpha3OpaqueDeviceConfiguration
}
\details{
OpaqueDeviceConfiguration contains configuration parameters for a driver in a format defined by the driver vendor.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3OpaqueDeviceConfiguration-new}{\code{V1alpha3OpaqueDeviceConfiguration$new()}}
\item \href{#method-V1alpha3OpaqueDeviceConfiguration-to_list}{\code{V1alpha3OpaqueDeviceConfiguration$to_list()}}
\item \href{#method-V1alpha3OpaqueDeviceConfiguration-clone}{\code{V1alpha3OpaqueDeviceConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3OpaqueDeviceConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3OpaqueDeviceConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3OpaqueDeviceConfiguration$new(driver = NULL, parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3OpaqueDeviceConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3OpaqueDeviceConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3OpaqueDeviceConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3OpaqueDeviceConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3OpaqueDeviceConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3OpaqueDeviceConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
