% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3PodSchedulingContext.R
\name{V1alpha3PodSchedulingContext}
\alias{V1alpha3PodSchedulingContext}
\title{V1alpha3PodSchedulingContext}
\description{
V1alpha3PodSchedulingContext

V1alpha3PodSchedulingContext
}
\details{
PodSchedulingContext objects hold information that is needed to schedule a Pod with ResourceClaims that use "WaitForFirstConsumer" allocation mode.

This is an alpha type and requires enabling the DRAControlPlaneController feature gate.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3PodSchedulingContext-new}{\code{V1alpha3PodSchedulingContext$new()}}
\item \href{#method-V1alpha3PodSchedulingContext-to_list}{\code{V1alpha3PodSchedulingContext$to_list()}}
\item \href{#method-V1alpha3PodSchedulingContext-clone}{\code{V1alpha3PodSchedulingContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContext-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContext$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContext-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContext-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContext$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
