% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3PodSchedulingContextList.R
\name{V1alpha3PodSchedulingContextList}
\alias{V1alpha3PodSchedulingContextList}
\title{V1alpha3PodSchedulingContextList}
\description{
V1alpha3PodSchedulingContextList

V1alpha3PodSchedulingContextList
}
\details{
PodSchedulingContextList is a collection of Pod scheduling objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3PodSchedulingContextList-new}{\code{V1alpha3PodSchedulingContextList$new()}}
\item \href{#method-V1alpha3PodSchedulingContextList-to_list}{\code{V1alpha3PodSchedulingContextList$to_list()}}
\item \href{#method-V1alpha3PodSchedulingContextList-clone}{\code{V1alpha3PodSchedulingContextList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
