% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3PodSchedulingContextSpec.R
\name{V1alpha3PodSchedulingContextSpec}
\alias{V1alpha3PodSchedulingContextSpec}
\title{V1alpha3PodSchedulingContextSpec}
\description{
V1alpha3PodSchedulingContextSpec

V1alpha3PodSchedulingContextSpec
}
\details{
PodSchedulingContextSpec describes where resources for the Pod are needed.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3PodSchedulingContextSpec-new}{\code{V1alpha3PodSchedulingContextSpec$new()}}
\item \href{#method-V1alpha3PodSchedulingContextSpec-to_list}{\code{V1alpha3PodSchedulingContextSpec$to_list()}}
\item \href{#method-V1alpha3PodSchedulingContextSpec-clone}{\code{V1alpha3PodSchedulingContextSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextSpec$new(
  potential_nodes = NULL,
  selected_node = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
