% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3PodSchedulingContextStatus.R
\name{V1alpha3PodSchedulingContextStatus}
\alias{V1alpha3PodSchedulingContextStatus}
\title{V1alpha3PodSchedulingContextStatus}
\description{
V1alpha3PodSchedulingContextStatus

V1alpha3PodSchedulingContextStatus
}
\details{
PodSchedulingContextStatus describes where resources for the Pod can be allocated.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3PodSchedulingContextStatus-new}{\code{V1alpha3PodSchedulingContextStatus$new()}}
\item \href{#method-V1alpha3PodSchedulingContextStatus-to_list}{\code{V1alpha3PodSchedulingContextStatus$to_list()}}
\item \href{#method-V1alpha3PodSchedulingContextStatus-clone}{\code{V1alpha3PodSchedulingContextStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextStatus$new(resource_claims = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3PodSchedulingContextStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3PodSchedulingContextStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3PodSchedulingContextStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
