% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceClaim.R
\name{V1alpha3ResourceClaim}
\alias{V1alpha3ResourceClaim}
\title{V1alpha3ResourceClaim}
\description{
V1alpha3ResourceClaim

V1alpha3ResourceClaim
}
\details{
ResourceClaim describes a request for access to resources in the cluster, for use by workloads. For example, if a workload needs an accelerator device with specific properties, this is how that request is expressed. The status stanza tracks whether this claim has been satisfied and what specific resources have been allocated.

This is an alpha type and requires enabling the DynamicResourceAllocation feature gate.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceClaim-new}{\code{V1alpha3ResourceClaim$new()}}
\item \href{#method-V1alpha3ResourceClaim-to_list}{\code{V1alpha3ResourceClaim$to_list()}}
\item \href{#method-V1alpha3ResourceClaim-clone}{\code{V1alpha3ResourceClaim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaim-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaim-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaim$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaim-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaim-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaim$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
