% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1alpha3ResourceClaimConsumerReference.R
\name{V1alpha3ResourceClaimConsumerReference}
\alias{V1alpha3ResourceClaimConsumerReference}
\title{V1alpha3ResourceClaimConsumerReference}
\description{
V1alpha3ResourceClaimConsumerReference

V1alpha3ResourceClaimConsumerReference
}
\details{
ResourceClaimConsumerReference contains enough information to let you locate the consumer of a ResourceClaim. The user must be a resource in the same namespace as the ResourceClaim.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceClaimConsumerReference-new}{\code{V1alpha3ResourceClaimConsumerReference$new()}}
\item \href{#method-V1alpha3ResourceClaimConsumerReference-to_list}{\code{V1alpha3ResourceClaimConsumerReference$to_list()}}
\item \href{#method-V1alpha3ResourceClaimConsumerReference-clone}{\code{V1alpha3ResourceClaimConsumerReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimConsumerReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimConsumerReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimConsumerReference$new(
  api_group = NULL,
  name = NULL,
  resource = NULL,
  uid = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimConsumerReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimConsumerReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimConsumerReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimConsumerReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimConsumerReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimConsumerReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
