% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1alpha3ResourceClaimSchedulingStatus.R
\name{V1alpha3ResourceClaimSchedulingStatus}
\alias{V1alpha3ResourceClaimSchedulingStatus}
\title{V1alpha3ResourceClaimSchedulingStatus}
\description{
V1alpha3ResourceClaimSchedulingStatus

V1alpha3ResourceClaimSchedulingStatus
}
\details{
ResourceClaimSchedulingStatus contains information about one particular ResourceClaim with "WaitForFirstConsumer" allocation mode.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceClaimSchedulingStatus-new}{\code{V1alpha3ResourceClaimSchedulingStatus$new()}}
\item \href{#method-V1alpha3ResourceClaimSchedulingStatus-to_list}{\code{V1alpha3ResourceClaimSchedulingStatus$to_list()}}
\item \href{#method-V1alpha3ResourceClaimSchedulingStatus-clone}{\code{V1alpha3ResourceClaimSchedulingStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimSchedulingStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimSchedulingStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimSchedulingStatus$new(name = NULL, unsuitable_nodes = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimSchedulingStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimSchedulingStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimSchedulingStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimSchedulingStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimSchedulingStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimSchedulingStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
