% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceClaimSpec.R
\name{V1alpha3ResourceClaimSpec}
\alias{V1alpha3ResourceClaimSpec}
\title{V1alpha3ResourceClaimSpec}
\description{
V1alpha3ResourceClaimSpec

V1alpha3ResourceClaimSpec
}
\details{
ResourceClaimSpec defines what is being requested in a ResourceClaim and how to configure it.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceClaimSpec-new}{\code{V1alpha3ResourceClaimSpec$new()}}
\item \href{#method-V1alpha3ResourceClaimSpec-to_list}{\code{V1alpha3ResourceClaimSpec$to_list()}}
\item \href{#method-V1alpha3ResourceClaimSpec-clone}{\code{V1alpha3ResourceClaimSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimSpec$new(controller = NULL, devices = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
