% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceClaimStatus.R
\name{V1alpha3ResourceClaimStatus}
\alias{V1alpha3ResourceClaimStatus}
\title{V1alpha3ResourceClaimStatus}
\description{
V1alpha3ResourceClaimStatus

V1alpha3ResourceClaimStatus
}
\details{
ResourceClaimStatus tracks whether the resource has been allocated and what the result of that was.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceClaimStatus-new}{\code{V1alpha3ResourceClaimStatus$new()}}
\item \href{#method-V1alpha3ResourceClaimStatus-to_list}{\code{V1alpha3ResourceClaimStatus$to_list()}}
\item \href{#method-V1alpha3ResourceClaimStatus-clone}{\code{V1alpha3ResourceClaimStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimStatus$new(
  allocation = NULL,
  deallocation_requested = NULL,
  reserved_for = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
