% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceClaimTemplate.R
\name{V1alpha3ResourceClaimTemplate}
\alias{V1alpha3ResourceClaimTemplate}
\title{V1alpha3ResourceClaimTemplate}
\description{
V1alpha3ResourceClaimTemplate

V1alpha3ResourceClaimTemplate
}
\details{
ResourceClaimTemplate is used to produce ResourceClaim objects.

This is an alpha type and requires enabling the DynamicResourceAllocation feature gate.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceClaimTemplate-new}{\code{V1alpha3ResourceClaimTemplate$new()}}
\item \href{#method-V1alpha3ResourceClaimTemplate-to_list}{\code{V1alpha3ResourceClaimTemplate$to_list()}}
\item \href{#method-V1alpha3ResourceClaimTemplate-clone}{\code{V1alpha3ResourceClaimTemplate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimTemplate-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimTemplate-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimTemplate$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimTemplate-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimTemplate-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimTemplate$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimTemplate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimTemplate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
