% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceClaimTemplateSpec.R
\name{V1alpha3ResourceClaimTemplateSpec}
\alias{V1alpha3ResourceClaimTemplateSpec}
\title{V1alpha3ResourceClaimTemplateSpec}
\description{
V1alpha3ResourceClaimTemplateSpec

V1alpha3ResourceClaimTemplateSpec
}
\details{
ResourceClaimTemplateSpec contains the metadata and fields for a ResourceClaim.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceClaimTemplateSpec-new}{\code{V1alpha3ResourceClaimTemplateSpec$new()}}
\item \href{#method-V1alpha3ResourceClaimTemplateSpec-to_list}{\code{V1alpha3ResourceClaimTemplateSpec$to_list()}}
\item \href{#method-V1alpha3ResourceClaimTemplateSpec-clone}{\code{V1alpha3ResourceClaimTemplateSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimTemplateSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimTemplateSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimTemplateSpec$new(metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimTemplateSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimTemplateSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimTemplateSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceClaimTemplateSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceClaimTemplateSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceClaimTemplateSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
