% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourcePool.R
\name{V1alpha3ResourcePool}
\alias{V1alpha3ResourcePool}
\title{V1alpha3ResourcePool}
\description{
V1alpha3ResourcePool

V1alpha3ResourcePool
}
\details{
ResourcePool describes the pool that ResourceSlices belong to.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourcePool-new}{\code{V1alpha3ResourcePool$new()}}
\item \href{#method-V1alpha3ResourcePool-to_list}{\code{V1alpha3ResourcePool$to_list()}}
\item \href{#method-V1alpha3ResourcePool-clone}{\code{V1alpha3ResourcePool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourcePool-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourcePool-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourcePool$new(
  generation = NULL,
  name = NULL,
  resource_slice_count = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourcePool-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourcePool-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourcePool$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourcePool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourcePool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourcePool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
