% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceSlice.R
\name{V1alpha3ResourceSlice}
\alias{V1alpha3ResourceSlice}
\title{V1alpha3ResourceSlice}
\description{
V1alpha3ResourceSlice

V1alpha3ResourceSlice
}
\details{
ResourceSlice represents one or more resources in a pool of similar resources, managed by a common driver. A pool may span more than one ResourceSlice, and exactly how many ResourceSlices comprise a pool is determined by the driver.

At the moment, the only supported resources are devices with attributes and capacities. Each device in a given pool, regardless of how many ResourceSlices, must have a unique name. The ResourceSlice in which a device gets published may change over time. The unique identifier for a device is the tuple <driver name>, <pool name>, <device name>.

Whenever a driver needs to update a pool, it increments the pool.Spec.Pool.Generation number and updates all ResourceSlices with that new number and new resource definitions. A consumer must only use ResourceSlices with the highest generation number and ignore all others.

When allocating all resources in a pool matching certain criteria or when looking for the best solution among several different alternatives, a consumer should check the number of ResourceSlices in a pool (included in each ResourceSlice) to determine whether its view of a pool is complete and if not, should wait until the driver has completed updating the pool.

For resources that are not local to a node, the node name is not set. Instead, the driver may use a node selector to specify where the devices are available.

This is an alpha type and requires enabling the DynamicResourceAllocation feature gate.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceSlice-new}{\code{V1alpha3ResourceSlice$new()}}
\item \href{#method-V1alpha3ResourceSlice-to_list}{\code{V1alpha3ResourceSlice$to_list()}}
\item \href{#method-V1alpha3ResourceSlice-clone}{\code{V1alpha3ResourceSlice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSlice-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSlice-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSlice$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSlice-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSlice-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSlice$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSlice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSlice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSlice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
