% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceSliceList.R
\name{V1alpha3ResourceSliceList}
\alias{V1alpha3ResourceSliceList}
\title{V1alpha3ResourceSliceList}
\description{
V1alpha3ResourceSliceList

V1alpha3ResourceSliceList
}
\details{
ResourceSliceList is a collection of ResourceSlices.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceSliceList-new}{\code{V1alpha3ResourceSliceList$new()}}
\item \href{#method-V1alpha3ResourceSliceList-to_list}{\code{V1alpha3ResourceSliceList$to_list()}}
\item \href{#method-V1alpha3ResourceSliceList-clone}{\code{V1alpha3ResourceSliceList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSliceList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSliceList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSliceList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSliceList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSliceList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSliceList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSliceList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSliceList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSliceList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
