% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1alpha3ResourceSliceSpec.R
\name{V1alpha3ResourceSliceSpec}
\alias{V1alpha3ResourceSliceSpec}
\title{V1alpha3ResourceSliceSpec}
\description{
V1alpha3ResourceSliceSpec

V1alpha3ResourceSliceSpec
}
\details{
ResourceSliceSpec contains the information published by the driver in one ResourceSlice.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1alpha3ResourceSliceSpec-new}{\code{V1alpha3ResourceSliceSpec$new()}}
\item \href{#method-V1alpha3ResourceSliceSpec-to_list}{\code{V1alpha3ResourceSliceSpec$to_list()}}
\item \href{#method-V1alpha3ResourceSliceSpec-clone}{\code{V1alpha3ResourceSliceSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSliceSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSliceSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSliceSpec$new(
  all_nodes = NULL,
  devices = NULL,
  driver = NULL,
  node_name = NULL,
  node_selector = NULL,
  pool = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSliceSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSliceSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSliceSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1alpha3ResourceSliceSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1alpha3ResourceSliceSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1alpha3ResourceSliceSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
