% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIGroup.R
\name{V1APIGroup}
\alias{V1APIGroup}
\title{V1APIGroup}
\description{
V1APIGroup

V1APIGroup
}
\details{
APIGroup contains the name, the supported versions, and the preferred version of a group.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIGroup-new}{\code{V1APIGroup$new()}}
\item \href{#method-V1APIGroup-to_list}{\code{V1APIGroup$to_list()}}
\item \href{#method-V1APIGroup-clone}{\code{V1APIGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIGroup-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIGroup$new(
  api_version = NULL,
  kind = NULL,
  name = NULL,
  preferred_version = NULL,
  server_address_by_client_cid_rs = NULL,
  versions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIGroup-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIGroup-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIGroup$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
