% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIGroupList.R
\name{V1APIGroupList}
\alias{V1APIGroupList}
\title{V1APIGroupList}
\description{
V1APIGroupList

V1APIGroupList
}
\details{
APIGroupList is a list of APIGroup, to allow clients to discover the API at /apis.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIGroupList-new}{\code{V1APIGroupList$new()}}
\item \href{#method-V1APIGroupList-to_list}{\code{V1APIGroupList$to_list()}}
\item \href{#method-V1APIGroupList-clone}{\code{V1APIGroupList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIGroupList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIGroupList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIGroupList$new(api_version = NULL, groups = NULL, kind = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIGroupList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIGroupList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIGroupList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIGroupList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIGroupList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIGroupList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
