% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIResource.R
\name{V1APIResource}
\alias{V1APIResource}
\title{V1APIResource}
\description{
V1APIResource

V1APIResource
}
\details{
APIResource specifies the name of a resource and whether it is namespaced.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIResource-new}{\code{V1APIResource$new()}}
\item \href{#method-V1APIResource-to_list}{\code{V1APIResource$to_list()}}
\item \href{#method-V1APIResource-clone}{\code{V1APIResource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIResource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIResource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIResource$new(
  categories = NULL,
  group = NULL,
  kind = NULL,
  name = NULL,
  namespaced = NULL,
  short_names = NULL,
  singular_name = NULL,
  storage_version_hash = NULL,
  verbs = NULL,
  version = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIResource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIResource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIResource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIResource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIResource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIResource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
