% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIResourceList.R
\name{V1APIResourceList}
\alias{V1APIResourceList}
\title{V1APIResourceList}
\description{
V1APIResourceList

V1APIResourceList
}
\details{
APIResourceList is a list of APIResource, it is used to expose the name of the resources supported in a specific group and version, and if the resource is namespaced.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIResourceList-new}{\code{V1APIResourceList$new()}}
\item \href{#method-V1APIResourceList-to_list}{\code{V1APIResourceList$to_list()}}
\item \href{#method-V1APIResourceList-clone}{\code{V1APIResourceList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIResourceList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIResourceList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIResourceList$new(
  api_version = NULL,
  group_version = NULL,
  kind = NULL,
  resources = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIResourceList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIResourceList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIResourceList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIResourceList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIResourceList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIResourceList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
