% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIService.R
\name{V1APIService}
\alias{V1APIService}
\title{V1APIService}
\description{
V1APIService

V1APIService
}
\details{
APIService represents a server for a particular GroupVersion. Name must be "version.group".
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIService-new}{\code{V1APIService$new()}}
\item \href{#method-V1APIService-to_list}{\code{V1APIService$to_list()}}
\item \href{#method-V1APIService-clone}{\code{V1APIService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIService-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIService-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIService$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIService-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIService-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIService$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
