% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIServiceCondition.R
\name{V1APIServiceCondition}
\alias{V1APIServiceCondition}
\title{V1APIServiceCondition}
\description{
V1APIServiceCondition

V1APIServiceCondition
}
\details{
APIServiceCondition describes the state of an APIService at a particular point
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIServiceCondition-new}{\code{V1APIServiceCondition$new()}}
\item \href{#method-V1APIServiceCondition-to_list}{\code{V1APIServiceCondition$to_list()}}
\item \href{#method-V1APIServiceCondition-clone}{\code{V1APIServiceCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
