% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIServiceList.R
\name{V1APIServiceList}
\alias{V1APIServiceList}
\title{V1APIServiceList}
\description{
V1APIServiceList

V1APIServiceList
}
\details{
APIServiceList is a list of APIService objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIServiceList-new}{\code{V1APIServiceList$new()}}
\item \href{#method-V1APIServiceList-to_list}{\code{V1APIServiceList$to_list()}}
\item \href{#method-V1APIServiceList-clone}{\code{V1APIServiceList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
