% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIServiceSpec.R
\name{V1APIServiceSpec}
\alias{V1APIServiceSpec}
\title{V1APIServiceSpec}
\description{
V1APIServiceSpec

V1APIServiceSpec
}
\details{
APIServiceSpec contains information for locating and communicating with a server. Only https is supported, though you are able to disable certificate verification.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIServiceSpec-new}{\code{V1APIServiceSpec$new()}}
\item \href{#method-V1APIServiceSpec-to_list}{\code{V1APIServiceSpec$to_list()}}
\item \href{#method-V1APIServiceSpec-clone}{\code{V1APIServiceSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceSpec$new(
  ca_bundle = NULL,
  group = NULL,
  group_priority_minimum = NULL,
  insecure_skip_tls_verify = NULL,
  service = NULL,
  version = NULL,
  version_priority = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
