% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIServiceStatus.R
\name{V1APIServiceStatus}
\alias{V1APIServiceStatus}
\title{V1APIServiceStatus}
\description{
V1APIServiceStatus

V1APIServiceStatus
}
\details{
APIServiceStatus contains derived information about an API server
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIServiceStatus-new}{\code{V1APIServiceStatus$new()}}
\item \href{#method-V1APIServiceStatus-to_list}{\code{V1APIServiceStatus$to_list()}}
\item \href{#method-V1APIServiceStatus-clone}{\code{V1APIServiceStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceStatus$new(conditions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIServiceStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIServiceStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIServiceStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
