% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1APIVersions.R
\name{V1APIVersions}
\alias{V1APIVersions}
\title{V1APIVersions}
\description{
V1APIVersions

V1APIVersions
}
\details{
APIVersions lists the versions that are available, to allow clients to discover the API at /api, which is the root path of the legacy v1 API.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1APIVersions-new}{\code{V1APIVersions$new()}}
\item \href{#method-V1APIVersions-to_list}{\code{V1APIVersions$to_list()}}
\item \href{#method-V1APIVersions-clone}{\code{V1APIVersions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIVersions-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIVersions-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIVersions$new(
  api_version = NULL,
  kind = NULL,
  server_address_by_client_cid_rs = NULL,
  versions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIVersions-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIVersions-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIVersions$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1APIVersions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1APIVersions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1APIVersions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
