% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1AppArmorProfile.R
\name{V1AppArmorProfile}
\alias{V1AppArmorProfile}
\title{V1AppArmorProfile}
\description{
V1AppArmorProfile

V1AppArmorProfile
}
\details{
AppArmorProfile defines a pod or container's AppArmor settings.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1AppArmorProfile-new}{\code{V1AppArmorProfile$new()}}
\item \href{#method-V1AppArmorProfile-to_list}{\code{V1AppArmorProfile$to_list()}}
\item \href{#method-V1AppArmorProfile-clone}{\code{V1AppArmorProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AppArmorProfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1AppArmorProfile-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AppArmorProfile$new(localhost_profile = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AppArmorProfile-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1AppArmorProfile-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AppArmorProfile$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AppArmorProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1AppArmorProfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AppArmorProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
