% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1AttachedVolume.R
\name{V1AttachedVolume}
\alias{V1AttachedVolume}
\title{V1AttachedVolume}
\description{
V1AttachedVolume

V1AttachedVolume
}
\details{
AttachedVolume describes a volume attached to a node
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1AttachedVolume-new}{\code{V1AttachedVolume$new()}}
\item \href{#method-V1AttachedVolume-to_list}{\code{V1AttachedVolume$to_list()}}
\item \href{#method-V1AttachedVolume-clone}{\code{V1AttachedVolume$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AttachedVolume-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1AttachedVolume-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AttachedVolume$new(device_path = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AttachedVolume-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1AttachedVolume-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AttachedVolume$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AttachedVolume-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1AttachedVolume-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AttachedVolume$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
