% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1AuditAnnotation.R
\name{V1AuditAnnotation}
\alias{V1AuditAnnotation}
\title{V1AuditAnnotation}
\description{
V1AuditAnnotation

V1AuditAnnotation
}
\details{
AuditAnnotation describes how to produce an audit annotation for an API request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1AuditAnnotation-new}{\code{V1AuditAnnotation$new()}}
\item \href{#method-V1AuditAnnotation-to_list}{\code{V1AuditAnnotation$to_list()}}
\item \href{#method-V1AuditAnnotation-clone}{\code{V1AuditAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AuditAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1AuditAnnotation-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AuditAnnotation$new(key = NULL, value_expression = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AuditAnnotation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1AuditAnnotation-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AuditAnnotation$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AuditAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1AuditAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AuditAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
