% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1AWSElasticBlockStoreVolumeSource.R
\name{V1AWSElasticBlockStoreVolumeSource}
\alias{V1AWSElasticBlockStoreVolumeSource}
\title{V1AWSElasticBlockStoreVolumeSource}
\description{
V1AWSElasticBlockStoreVolumeSource

V1AWSElasticBlockStoreVolumeSource
}
\details{
Represents a Persistent Disk resource in AWS.

An AWS EBS disk must exist before mounting to a container. The disk must also be in the same AWS zone as the kubelet. An AWS EBS disk can only be mounted as read/write once. AWS EBS volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1AWSElasticBlockStoreVolumeSource-new}{\code{V1AWSElasticBlockStoreVolumeSource$new()}}
\item \href{#method-V1AWSElasticBlockStoreVolumeSource-to_list}{\code{V1AWSElasticBlockStoreVolumeSource$to_list()}}
\item \href{#method-V1AWSElasticBlockStoreVolumeSource-clone}{\code{V1AWSElasticBlockStoreVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AWSElasticBlockStoreVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1AWSElasticBlockStoreVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AWSElasticBlockStoreVolumeSource$new(
  fs_type = NULL,
  partition = NULL,
  read_only = NULL,
  volume_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AWSElasticBlockStoreVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1AWSElasticBlockStoreVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AWSElasticBlockStoreVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AWSElasticBlockStoreVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1AWSElasticBlockStoreVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AWSElasticBlockStoreVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
