% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1AzureDiskVolumeSource.R
\name{V1AzureDiskVolumeSource}
\alias{V1AzureDiskVolumeSource}
\title{V1AzureDiskVolumeSource}
\description{
V1AzureDiskVolumeSource

V1AzureDiskVolumeSource
}
\details{
AzureDisk represents an Azure Data Disk mount on the host and bind mount to the pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1AzureDiskVolumeSource-new}{\code{V1AzureDiskVolumeSource$new()}}
\item \href{#method-V1AzureDiskVolumeSource-to_list}{\code{V1AzureDiskVolumeSource$to_list()}}
\item \href{#method-V1AzureDiskVolumeSource-clone}{\code{V1AzureDiskVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AzureDiskVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1AzureDiskVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AzureDiskVolumeSource$new(
  caching_mode = NULL,
  disk_name = NULL,
  disk_uri = NULL,
  fs_type = NULL,
  kind = NULL,
  read_only = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AzureDiskVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1AzureDiskVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AzureDiskVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AzureDiskVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1AzureDiskVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AzureDiskVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
