% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1AzureFilePersistentVolumeSource.R
\name{V1AzureFilePersistentVolumeSource}
\alias{V1AzureFilePersistentVolumeSource}
\title{V1AzureFilePersistentVolumeSource}
\description{
V1AzureFilePersistentVolumeSource

V1AzureFilePersistentVolumeSource
}
\details{
AzureFile represents an Azure File Service mount on the host and bind mount to the pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1AzureFilePersistentVolumeSource-new}{\code{V1AzureFilePersistentVolumeSource$new()}}
\item \href{#method-V1AzureFilePersistentVolumeSource-to_list}{\code{V1AzureFilePersistentVolumeSource$to_list()}}
\item \href{#method-V1AzureFilePersistentVolumeSource-clone}{\code{V1AzureFilePersistentVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AzureFilePersistentVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1AzureFilePersistentVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AzureFilePersistentVolumeSource$new(
  read_only = NULL,
  secret_name = NULL,
  secret_namespace = NULL,
  share_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AzureFilePersistentVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1AzureFilePersistentVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AzureFilePersistentVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1AzureFilePersistentVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1AzureFilePersistentVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1AzureFilePersistentVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
