% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1IPAddress.R
\name{V1beta1IPAddress}
\alias{V1beta1IPAddress}
\title{V1beta1IPAddress}
\description{
V1beta1IPAddress

V1beta1IPAddress
}
\details{
IPAddress represents a single IP of a single IP Family. The object is designed to be used by APIs that operate on IP addresses. The object is used by the Service core API for allocation of IP addresses. An IP address can be represented in different formats, to guarantee the uniqueness of the IP, the name of the object is the IP address in canonical format, four decimal digits separated by dots suppressing leading zeros for IPv4 and the representation defined by RFC 5952 for IPv6. Valid: 192.168.1.5 or 2001:db8::1 or 2001:db8:aaaa:bbbb:cccc:dddd:eeee:1 Invalid: 10.01.2.3 or 2001:db8:0:0:0::1
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1IPAddress-new}{\code{V1beta1IPAddress$new()}}
\item \href{#method-V1beta1IPAddress-to_list}{\code{V1beta1IPAddress$to_list()}}
\item \href{#method-V1beta1IPAddress-clone}{\code{V1beta1IPAddress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddress-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddress-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddress$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddress-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddress-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddress$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
