% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1IPAddressList.R
\name{V1beta1IPAddressList}
\alias{V1beta1IPAddressList}
\title{V1beta1IPAddressList}
\description{
V1beta1IPAddressList

V1beta1IPAddressList
}
\details{
IPAddressList contains a list of IPAddress.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1IPAddressList-new}{\code{V1beta1IPAddressList$new()}}
\item \href{#method-V1beta1IPAddressList-to_list}{\code{V1beta1IPAddressList$to_list()}}
\item \href{#method-V1beta1IPAddressList-clone}{\code{V1beta1IPAddressList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddressList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddressList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddressList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddressList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddressList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddressList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddressList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddressList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddressList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
