% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1IPAddressSpec.R
\name{V1beta1IPAddressSpec}
\alias{V1beta1IPAddressSpec}
\title{V1beta1IPAddressSpec}
\description{
V1beta1IPAddressSpec

V1beta1IPAddressSpec
}
\details{
IPAddressSpec describe the attributes in an IP Address.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1IPAddressSpec-new}{\code{V1beta1IPAddressSpec$new()}}
\item \href{#method-V1beta1IPAddressSpec-to_list}{\code{V1beta1IPAddressSpec$to_list()}}
\item \href{#method-V1beta1IPAddressSpec-clone}{\code{V1beta1IPAddressSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddressSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddressSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddressSpec$new(parent_ref = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddressSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddressSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddressSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1IPAddressSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1IPAddressSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1IPAddressSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
