% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1ParamKind.R
\name{V1beta1ParamKind}
\alias{V1beta1ParamKind}
\title{V1beta1ParamKind}
\description{
V1beta1ParamKind

V1beta1ParamKind
}
\details{
ParamKind is a tuple of Group Kind and Version.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ParamKind-new}{\code{V1beta1ParamKind$new()}}
\item \href{#method-V1beta1ParamKind-to_list}{\code{V1beta1ParamKind$to_list()}}
\item \href{#method-V1beta1ParamKind-clone}{\code{V1beta1ParamKind$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParamKind-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParamKind-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParamKind$new(api_version = NULL, kind = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParamKind-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParamKind-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParamKind$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParamKind-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParamKind-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParamKind$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
