% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1ParamRef.R
\name{V1beta1ParamRef}
\alias{V1beta1ParamRef}
\title{V1beta1ParamRef}
\description{
V1beta1ParamRef

V1beta1ParamRef
}
\details{
ParamRef describes how to locate the params to be used as input to expressions of rules applied by a policy binding.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ParamRef-new}{\code{V1beta1ParamRef$new()}}
\item \href{#method-V1beta1ParamRef-to_list}{\code{V1beta1ParamRef$to_list()}}
\item \href{#method-V1beta1ParamRef-clone}{\code{V1beta1ParamRef$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParamRef-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParamRef-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParamRef$new(
  name = NULL,
  namespace = NULL,
  parameter_not_found_action = NULL,
  selector = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParamRef-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParamRef-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParamRef$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParamRef-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParamRef-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParamRef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
