% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1ParentReference.R
\name{V1beta1ParentReference}
\alias{V1beta1ParentReference}
\title{V1beta1ParentReference}
\description{
V1beta1ParentReference

V1beta1ParentReference
}
\details{
ParentReference describes a reference to a parent object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ParentReference-new}{\code{V1beta1ParentReference$new()}}
\item \href{#method-V1beta1ParentReference-to_list}{\code{V1beta1ParentReference$to_list()}}
\item \href{#method-V1beta1ParentReference-clone}{\code{V1beta1ParentReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParentReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParentReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParentReference$new(
  group = NULL,
  name = NULL,
  namespace = NULL,
  resource = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParentReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParentReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParentReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ParentReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ParentReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ParentReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
