% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1ServiceCIDRList.R
\name{V1beta1ServiceCIDRList}
\alias{V1beta1ServiceCIDRList}
\title{V1beta1ServiceCIDRList}
\description{
V1beta1ServiceCIDRList

V1beta1ServiceCIDRList
}
\details{
ServiceCIDRList contains a list of ServiceCIDR objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ServiceCIDRList-new}{\code{V1beta1ServiceCIDRList$new()}}
\item \href{#method-V1beta1ServiceCIDRList-to_list}{\code{V1beta1ServiceCIDRList$to_list()}}
\item \href{#method-V1beta1ServiceCIDRList-clone}{\code{V1beta1ServiceCIDRList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
