% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1ServiceCIDRSpec.R
\name{V1beta1ServiceCIDRSpec}
\alias{V1beta1ServiceCIDRSpec}
\title{V1beta1ServiceCIDRSpec}
\description{
V1beta1ServiceCIDRSpec

V1beta1ServiceCIDRSpec
}
\details{
ServiceCIDRSpec define the CIDRs the user wants to use for allocating ClusterIPs for Services.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ServiceCIDRSpec-new}{\code{V1beta1ServiceCIDRSpec$new()}}
\item \href{#method-V1beta1ServiceCIDRSpec-to_list}{\code{V1beta1ServiceCIDRSpec$to_list()}}
\item \href{#method-V1beta1ServiceCIDRSpec-clone}{\code{V1beta1ServiceCIDRSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRSpec$new(cidrs = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
