% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1ServiceCIDRStatus.R
\name{V1beta1ServiceCIDRStatus}
\alias{V1beta1ServiceCIDRStatus}
\title{V1beta1ServiceCIDRStatus}
\description{
V1beta1ServiceCIDRStatus

V1beta1ServiceCIDRStatus
}
\details{
ServiceCIDRStatus describes the current state of the ServiceCIDR.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ServiceCIDRStatus-new}{\code{V1beta1ServiceCIDRStatus$new()}}
\item \href{#method-V1beta1ServiceCIDRStatus-to_list}{\code{V1beta1ServiceCIDRStatus$to_list()}}
\item \href{#method-V1beta1ServiceCIDRStatus-clone}{\code{V1beta1ServiceCIDRStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRStatus$new(conditions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ServiceCIDRStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ServiceCIDRStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ServiceCIDRStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
