% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1TypeChecking.R
\name{V1beta1TypeChecking}
\alias{V1beta1TypeChecking}
\title{V1beta1TypeChecking}
\description{
V1beta1TypeChecking

V1beta1TypeChecking
}
\details{
TypeChecking contains results of type checking the expressions in the ValidatingAdmissionPolicy
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1TypeChecking-new}{\code{V1beta1TypeChecking$new()}}
\item \href{#method-V1beta1TypeChecking-to_list}{\code{V1beta1TypeChecking$to_list()}}
\item \href{#method-V1beta1TypeChecking-clone}{\code{V1beta1TypeChecking$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1TypeChecking-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1TypeChecking-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1TypeChecking$new(expression_warnings = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1TypeChecking-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1TypeChecking-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1TypeChecking$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1TypeChecking-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1TypeChecking-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1TypeChecking$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
