% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1ValidatingAdmissionPolicy.R
\name{V1beta1ValidatingAdmissionPolicy}
\alias{V1beta1ValidatingAdmissionPolicy}
\title{V1beta1ValidatingAdmissionPolicy}
\description{
V1beta1ValidatingAdmissionPolicy

V1beta1ValidatingAdmissionPolicy
}
\details{
ValidatingAdmissionPolicy describes the definition of an admission validation policy that accepts or rejects an object without changing it.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ValidatingAdmissionPolicy-new}{\code{V1beta1ValidatingAdmissionPolicy$new()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicy-to_list}{\code{V1beta1ValidatingAdmissionPolicy$to_list()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicy-clone}{\code{V1beta1ValidatingAdmissionPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicy$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
