% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1beta1ValidatingAdmissionPolicyBinding.R
\name{V1beta1ValidatingAdmissionPolicyBinding}
\alias{V1beta1ValidatingAdmissionPolicyBinding}
\title{V1beta1ValidatingAdmissionPolicyBinding}
\description{
V1beta1ValidatingAdmissionPolicyBinding

V1beta1ValidatingAdmissionPolicyBinding
}
\details{
ValidatingAdmissionPolicyBinding binds the ValidatingAdmissionPolicy with paramerized resources. ValidatingAdmissionPolicyBinding and parameter CRDs together define how cluster administrators configure policies for clusters.

For a given admission request, each binding will cause its policy to be evaluated N times, where N is 1 for policies/bindings that don't use params, otherwise N is the number of parameters selected by the binding.

The CEL expressions of a policy must have a computed CEL cost below the maximum CEL budget. Each evaluation of the policy is given an independent CEL cost budget. Adding/removing policies, bindings, or params can not affect whether a given (policy, binding, param) combination is within its own CEL budget.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ValidatingAdmissionPolicyBinding-new}{\code{V1beta1ValidatingAdmissionPolicyBinding$new()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicyBinding-to_list}{\code{V1beta1ValidatingAdmissionPolicyBinding$to_list()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicyBinding-clone}{\code{V1beta1ValidatingAdmissionPolicyBinding$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyBinding-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyBinding-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyBinding$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyBinding-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyBinding-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyBinding$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyBinding-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyBinding-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyBinding$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
