% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1beta1ValidatingAdmissionPolicyBindingSpec.R
\name{V1beta1ValidatingAdmissionPolicyBindingSpec}
\alias{V1beta1ValidatingAdmissionPolicyBindingSpec}
\title{V1beta1ValidatingAdmissionPolicyBindingSpec}
\description{
V1beta1ValidatingAdmissionPolicyBindingSpec

V1beta1ValidatingAdmissionPolicyBindingSpec
}
\details{
ValidatingAdmissionPolicyBindingSpec is the specification of the ValidatingAdmissionPolicyBinding.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ValidatingAdmissionPolicyBindingSpec-new}{\code{V1beta1ValidatingAdmissionPolicyBindingSpec$new()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicyBindingSpec-to_list}{\code{V1beta1ValidatingAdmissionPolicyBindingSpec$to_list()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicyBindingSpec-clone}{\code{V1beta1ValidatingAdmissionPolicyBindingSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyBindingSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyBindingSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyBindingSpec$new(
  match_resources = NULL,
  param_ref = NULL,
  policy_name = NULL,
  validation_actions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyBindingSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyBindingSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyBindingSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyBindingSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyBindingSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyBindingSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
