% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1beta1ValidatingAdmissionPolicyStatus.R
\name{V1beta1ValidatingAdmissionPolicyStatus}
\alias{V1beta1ValidatingAdmissionPolicyStatus}
\title{V1beta1ValidatingAdmissionPolicyStatus}
\description{
V1beta1ValidatingAdmissionPolicyStatus

V1beta1ValidatingAdmissionPolicyStatus
}
\details{
ValidatingAdmissionPolicyStatus represents the status of an admission validation policy.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1ValidatingAdmissionPolicyStatus-new}{\code{V1beta1ValidatingAdmissionPolicyStatus$new()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicyStatus-to_list}{\code{V1beta1ValidatingAdmissionPolicyStatus$to_list()}}
\item \href{#method-V1beta1ValidatingAdmissionPolicyStatus-clone}{\code{V1beta1ValidatingAdmissionPolicyStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyStatus$new(
  conditions = NULL,
  observed_generation = NULL,
  type_checking = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1ValidatingAdmissionPolicyStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1ValidatingAdmissionPolicyStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1ValidatingAdmissionPolicyStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
