% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1Variable.R
\name{V1beta1Variable}
\alias{V1beta1Variable}
\title{V1beta1Variable}
\description{
V1beta1Variable

V1beta1Variable
}
\details{
Variable is the definition of a variable that is used for composition. A variable is defined as a named expression.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1Variable-new}{\code{V1beta1Variable$new()}}
\item \href{#method-V1beta1Variable-to_list}{\code{V1beta1Variable$to_list()}}
\item \href{#method-V1beta1Variable-clone}{\code{V1beta1Variable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1Variable-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1Variable-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1Variable$new(expression = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1Variable-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1Variable-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1Variable$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1Variable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1Variable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1Variable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
