% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta1VolumeAttributesClass.R
\name{V1beta1VolumeAttributesClass}
\alias{V1beta1VolumeAttributesClass}
\title{V1beta1VolumeAttributesClass}
\description{
V1beta1VolumeAttributesClass

V1beta1VolumeAttributesClass
}
\details{
VolumeAttributesClass represents a specification of mutable volume attributes defined by the CSI driver. The class can be specified during dynamic provisioning of PersistentVolumeClaims, and changed in the PersistentVolumeClaim spec after provisioning.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta1VolumeAttributesClass-new}{\code{V1beta1VolumeAttributesClass$new()}}
\item \href{#method-V1beta1VolumeAttributesClass-to_list}{\code{V1beta1VolumeAttributesClass$to_list()}}
\item \href{#method-V1beta1VolumeAttributesClass-clone}{\code{V1beta1VolumeAttributesClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1VolumeAttributesClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1VolumeAttributesClass-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1VolumeAttributesClass$new(
  api_version = NULL,
  driver_name = NULL,
  kind = NULL,
  metadata = NULL,
  parameters = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1VolumeAttributesClass-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1VolumeAttributesClass-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1VolumeAttributesClass$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta1VolumeAttributesClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta1VolumeAttributesClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta1VolumeAttributesClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
