% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta3FlowSchemaSpec.R
\name{V1beta3FlowSchemaSpec}
\alias{V1beta3FlowSchemaSpec}
\title{V1beta3FlowSchemaSpec}
\description{
V1beta3FlowSchemaSpec

V1beta3FlowSchemaSpec
}
\details{
FlowSchemaSpec describes how the FlowSchema's specification looks like.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta3FlowSchemaSpec-new}{\code{V1beta3FlowSchemaSpec$new()}}
\item \href{#method-V1beta3FlowSchemaSpec-to_list}{\code{V1beta3FlowSchemaSpec$to_list()}}
\item \href{#method-V1beta3FlowSchemaSpec-clone}{\code{V1beta3FlowSchemaSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3FlowSchemaSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3FlowSchemaSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3FlowSchemaSpec$new(
  distinguisher_method = NULL,
  matching_precedence = NULL,
  priority_level_configuration = NULL,
  rules = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3FlowSchemaSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3FlowSchemaSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3FlowSchemaSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3FlowSchemaSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3FlowSchemaSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3FlowSchemaSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
