% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1beta3LimitedPriorityLevelConfiguration.R
\name{V1beta3LimitedPriorityLevelConfiguration}
\alias{V1beta3LimitedPriorityLevelConfiguration}
\title{V1beta3LimitedPriorityLevelConfiguration}
\description{
V1beta3LimitedPriorityLevelConfiguration

V1beta3LimitedPriorityLevelConfiguration
}
\details{
LimitedPriorityLevelConfiguration specifies how to handle requests that are subject to limits. It addresses two issues:
  - How are requests for this priority level limited?
  - What should be done with requests that exceed the limit?
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta3LimitedPriorityLevelConfiguration-new}{\code{V1beta3LimitedPriorityLevelConfiguration$new()}}
\item \href{#method-V1beta3LimitedPriorityLevelConfiguration-to_list}{\code{V1beta3LimitedPriorityLevelConfiguration$to_list()}}
\item \href{#method-V1beta3LimitedPriorityLevelConfiguration-clone}{\code{V1beta3LimitedPriorityLevelConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3LimitedPriorityLevelConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3LimitedPriorityLevelConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3LimitedPriorityLevelConfiguration$new(
  borrowing_limit_percent = NULL,
  lendable_percent = NULL,
  limit_response = NULL,
  nominal_concurrency_shares = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3LimitedPriorityLevelConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3LimitedPriorityLevelConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3LimitedPriorityLevelConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3LimitedPriorityLevelConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3LimitedPriorityLevelConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3LimitedPriorityLevelConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
