% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta3PriorityLevelConfiguration.R
\name{V1beta3PriorityLevelConfiguration}
\alias{V1beta3PriorityLevelConfiguration}
\title{V1beta3PriorityLevelConfiguration}
\description{
V1beta3PriorityLevelConfiguration

V1beta3PriorityLevelConfiguration
}
\details{
PriorityLevelConfiguration represents the configuration of a priority level.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta3PriorityLevelConfiguration-new}{\code{V1beta3PriorityLevelConfiguration$new()}}
\item \href{#method-V1beta3PriorityLevelConfiguration-to_list}{\code{V1beta3PriorityLevelConfiguration$to_list()}}
\item \href{#method-V1beta3PriorityLevelConfiguration-clone}{\code{V1beta3PriorityLevelConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3PriorityLevelConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3PriorityLevelConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3PriorityLevelConfiguration$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3PriorityLevelConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3PriorityLevelConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3PriorityLevelConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3PriorityLevelConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3PriorityLevelConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3PriorityLevelConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
