% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1beta3QueuingConfiguration.R
\name{V1beta3QueuingConfiguration}
\alias{V1beta3QueuingConfiguration}
\title{V1beta3QueuingConfiguration}
\description{
V1beta3QueuingConfiguration

V1beta3QueuingConfiguration
}
\details{
QueuingConfiguration holds the configuration parameters for queuing
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1beta3QueuingConfiguration-new}{\code{V1beta3QueuingConfiguration$new()}}
\item \href{#method-V1beta3QueuingConfiguration-to_list}{\code{V1beta3QueuingConfiguration$to_list()}}
\item \href{#method-V1beta3QueuingConfiguration-clone}{\code{V1beta3QueuingConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3QueuingConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3QueuingConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3QueuingConfiguration$new(
  hand_size = NULL,
  queue_length_limit = NULL,
  queues = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3QueuingConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3QueuingConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3QueuingConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1beta3QueuingConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1beta3QueuingConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1beta3QueuingConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
