% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Binding.R
\name{V1Binding}
\alias{V1Binding}
\title{V1Binding}
\description{
V1Binding

V1Binding
}
\details{
Binding ties one object to another; for example, a pod is bound to a node by a scheduler. Deprecated in 1.7, please use the bindings subresource of pods instead.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Binding-new}{\code{V1Binding$new()}}
\item \href{#method-V1Binding-to_list}{\code{V1Binding$to_list()}}
\item \href{#method-V1Binding-clone}{\code{V1Binding$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Binding-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Binding-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Binding$new(api_version = NULL, kind = NULL, metadata = NULL, target = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Binding-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Binding-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Binding$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Binding-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Binding-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Binding$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
