% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Capabilities.R
\name{V1Capabilities}
\alias{V1Capabilities}
\title{V1Capabilities}
\description{
V1Capabilities

V1Capabilities
}
\details{
Adds and removes POSIX capabilities from running containers.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Capabilities-new}{\code{V1Capabilities$new()}}
\item \href{#method-V1Capabilities-to_list}{\code{V1Capabilities$to_list()}}
\item \href{#method-V1Capabilities-clone}{\code{V1Capabilities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Capabilities-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Capabilities-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Capabilities$new(add = NULL, drop = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Capabilities-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Capabilities-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Capabilities$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Capabilities-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Capabilities-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Capabilities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
