% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CephFSPersistentVolumeSource.R
\name{V1CephFSPersistentVolumeSource}
\alias{V1CephFSPersistentVolumeSource}
\title{V1CephFSPersistentVolumeSource}
\description{
V1CephFSPersistentVolumeSource

V1CephFSPersistentVolumeSource
}
\details{
Represents a Ceph Filesystem mount that lasts the lifetime of a pod Cephfs volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CephFSPersistentVolumeSource-new}{\code{V1CephFSPersistentVolumeSource$new()}}
\item \href{#method-V1CephFSPersistentVolumeSource-to_list}{\code{V1CephFSPersistentVolumeSource$to_list()}}
\item \href{#method-V1CephFSPersistentVolumeSource-clone}{\code{V1CephFSPersistentVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CephFSPersistentVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CephFSPersistentVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CephFSPersistentVolumeSource$new(
  monitors = NULL,
  path = NULL,
  read_only = NULL,
  secret_file = NULL,
  secret_ref = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CephFSPersistentVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CephFSPersistentVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CephFSPersistentVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CephFSPersistentVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CephFSPersistentVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CephFSPersistentVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
