% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CephFSVolumeSource.R
\name{V1CephFSVolumeSource}
\alias{V1CephFSVolumeSource}
\title{V1CephFSVolumeSource}
\description{
V1CephFSVolumeSource

V1CephFSVolumeSource
}
\details{
Represents a Ceph Filesystem mount that lasts the lifetime of a pod Cephfs volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CephFSVolumeSource-new}{\code{V1CephFSVolumeSource$new()}}
\item \href{#method-V1CephFSVolumeSource-to_list}{\code{V1CephFSVolumeSource$to_list()}}
\item \href{#method-V1CephFSVolumeSource-clone}{\code{V1CephFSVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CephFSVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CephFSVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CephFSVolumeSource$new(
  monitors = NULL,
  path = NULL,
  read_only = NULL,
  secret_file = NULL,
  secret_ref = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CephFSVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CephFSVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CephFSVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CephFSVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CephFSVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CephFSVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
