% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CertificateSigningRequest.R
\name{V1CertificateSigningRequest}
\alias{V1CertificateSigningRequest}
\title{V1CertificateSigningRequest}
\description{
V1CertificateSigningRequest

V1CertificateSigningRequest
}
\details{
CertificateSigningRequest objects provide a mechanism to obtain x509 certificates by submitting a certificate signing request, and having it asynchronously approved and issued.

Kubelets use this API to obtain:
 1. client certificates to authenticate to kube-apiserver (with the "kubernetes.io/kube-apiserver-client-kubelet" signerName).
 2. serving certificates for TLS endpoints kube-apiserver can connect to securely (with the "kubernetes.io/kubelet-serving" signerName).

This API can be used to request client certificates to authenticate to kube-apiserver (with the "kubernetes.io/kube-apiserver-client" signerName), or to obtain certificates from custom non-Kubernetes signers.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CertificateSigningRequest-new}{\code{V1CertificateSigningRequest$new()}}
\item \href{#method-V1CertificateSigningRequest-to_list}{\code{V1CertificateSigningRequest$to_list()}}
\item \href{#method-V1CertificateSigningRequest-clone}{\code{V1CertificateSigningRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequest-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequest$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
