% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1CertificateSigningRequestCondition.R
\name{V1CertificateSigningRequestCondition}
\alias{V1CertificateSigningRequestCondition}
\title{V1CertificateSigningRequestCondition}
\description{
V1CertificateSigningRequestCondition

V1CertificateSigningRequestCondition
}
\details{
CertificateSigningRequestCondition describes a condition of a CertificateSigningRequest object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CertificateSigningRequestCondition-new}{\code{V1CertificateSigningRequestCondition$new()}}
\item \href{#method-V1CertificateSigningRequestCondition-to_list}{\code{V1CertificateSigningRequestCondition$to_list()}}
\item \href{#method-V1CertificateSigningRequestCondition-clone}{\code{V1CertificateSigningRequestCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestCondition$new(
  last_transition_time = NULL,
  last_update_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
