% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CertificateSigningRequestList.R
\name{V1CertificateSigningRequestList}
\alias{V1CertificateSigningRequestList}
\title{V1CertificateSigningRequestList}
\description{
V1CertificateSigningRequestList

V1CertificateSigningRequestList
}
\details{
CertificateSigningRequestList is a collection of CertificateSigningRequest objects
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CertificateSigningRequestList-new}{\code{V1CertificateSigningRequestList$new()}}
\item \href{#method-V1CertificateSigningRequestList-to_list}{\code{V1CertificateSigningRequestList$to_list()}}
\item \href{#method-V1CertificateSigningRequestList-clone}{\code{V1CertificateSigningRequestList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
