% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CertificateSigningRequestSpec.R
\name{V1CertificateSigningRequestSpec}
\alias{V1CertificateSigningRequestSpec}
\title{V1CertificateSigningRequestSpec}
\description{
V1CertificateSigningRequestSpec

V1CertificateSigningRequestSpec
}
\details{
CertificateSigningRequestSpec contains the certificate request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CertificateSigningRequestSpec-new}{\code{V1CertificateSigningRequestSpec$new()}}
\item \href{#method-V1CertificateSigningRequestSpec-to_list}{\code{V1CertificateSigningRequestSpec$to_list()}}
\item \href{#method-V1CertificateSigningRequestSpec-clone}{\code{V1CertificateSigningRequestSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestSpec$new(
  expiration_seconds = NULL,
  extra = NULL,
  groups = NULL,
  request = NULL,
  signer_name = NULL,
  uid = NULL,
  usages = NULL,
  username = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CertificateSigningRequestSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CertificateSigningRequestSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CertificateSigningRequestSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
