% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CinderVolumeSource.R
\name{V1CinderVolumeSource}
\alias{V1CinderVolumeSource}
\title{V1CinderVolumeSource}
\description{
V1CinderVolumeSource

V1CinderVolumeSource
}
\details{
Represents a cinder volume resource in Openstack. A Cinder volume must exist before mounting to a container. The volume must also be in the same region as the kubelet. Cinder volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CinderVolumeSource-new}{\code{V1CinderVolumeSource$new()}}
\item \href{#method-V1CinderVolumeSource-to_list}{\code{V1CinderVolumeSource$to_list()}}
\item \href{#method-V1CinderVolumeSource-clone}{\code{V1CinderVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CinderVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CinderVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CinderVolumeSource$new(
  fs_type = NULL,
  read_only = NULL,
  secret_ref = NULL,
  volume_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CinderVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CinderVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CinderVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CinderVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CinderVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CinderVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
