% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ClientIPConfig.R
\name{V1ClientIPConfig}
\alias{V1ClientIPConfig}
\title{V1ClientIPConfig}
\description{
V1ClientIPConfig

V1ClientIPConfig
}
\details{
ClientIPConfig represents the configurations of Client IP based session affinity.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ClientIPConfig-new}{\code{V1ClientIPConfig$new()}}
\item \href{#method-V1ClientIPConfig-to_list}{\code{V1ClientIPConfig$to_list()}}
\item \href{#method-V1ClientIPConfig-clone}{\code{V1ClientIPConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClientIPConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClientIPConfig-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClientIPConfig$new(timeout_seconds = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClientIPConfig-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClientIPConfig-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClientIPConfig$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClientIPConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClientIPConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClientIPConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
