% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ClusterRole.R
\name{V1ClusterRole}
\alias{V1ClusterRole}
\title{V1ClusterRole}
\description{
V1ClusterRole

V1ClusterRole
}
\details{
ClusterRole is a cluster level, logical grouping of PolicyRules that can be referenced as a unit by a RoleBinding or ClusterRoleBinding.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ClusterRole-new}{\code{V1ClusterRole$new()}}
\item \href{#method-V1ClusterRole-to_list}{\code{V1ClusterRole$to_list()}}
\item \href{#method-V1ClusterRole-clone}{\code{V1ClusterRole$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRole-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRole-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRole$new(
  aggregation_rule = NULL,
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  rules = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRole-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRole-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRole$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRole-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRole-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRole$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
