% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ClusterRoleBinding.R
\name{V1ClusterRoleBinding}
\alias{V1ClusterRoleBinding}
\title{V1ClusterRoleBinding}
\description{
V1ClusterRoleBinding

V1ClusterRoleBinding
}
\details{
ClusterRoleBinding references a ClusterRole, but not contain it.  It can reference a ClusterRole in the global namespace, and adds who information via Subject.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ClusterRoleBinding-new}{\code{V1ClusterRoleBinding$new()}}
\item \href{#method-V1ClusterRoleBinding-to_list}{\code{V1ClusterRoleBinding$to_list()}}
\item \href{#method-V1ClusterRoleBinding-clone}{\code{V1ClusterRoleBinding$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRoleBinding-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRoleBinding-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRoleBinding$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  role_ref = NULL,
  subjects = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRoleBinding-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRoleBinding-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRoleBinding$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRoleBinding-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRoleBinding-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRoleBinding$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
