% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ClusterRoleBindingList.R
\name{V1ClusterRoleBindingList}
\alias{V1ClusterRoleBindingList}
\title{V1ClusterRoleBindingList}
\description{
V1ClusterRoleBindingList

V1ClusterRoleBindingList
}
\details{
ClusterRoleBindingList is a collection of ClusterRoleBindings
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ClusterRoleBindingList-new}{\code{V1ClusterRoleBindingList$new()}}
\item \href{#method-V1ClusterRoleBindingList-to_list}{\code{V1ClusterRoleBindingList$to_list()}}
\item \href{#method-V1ClusterRoleBindingList-clone}{\code{V1ClusterRoleBindingList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRoleBindingList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRoleBindingList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRoleBindingList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRoleBindingList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRoleBindingList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRoleBindingList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterRoleBindingList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterRoleBindingList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterRoleBindingList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
