% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ClusterTrustBundleProjection.R
\name{V1ClusterTrustBundleProjection}
\alias{V1ClusterTrustBundleProjection}
\title{V1ClusterTrustBundleProjection}
\description{
V1ClusterTrustBundleProjection

V1ClusterTrustBundleProjection
}
\details{
ClusterTrustBundleProjection describes how to select a set of ClusterTrustBundle objects and project their contents into the pod filesystem.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ClusterTrustBundleProjection-new}{\code{V1ClusterTrustBundleProjection$new()}}
\item \href{#method-V1ClusterTrustBundleProjection-to_list}{\code{V1ClusterTrustBundleProjection$to_list()}}
\item \href{#method-V1ClusterTrustBundleProjection-clone}{\code{V1ClusterTrustBundleProjection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterTrustBundleProjection-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterTrustBundleProjection-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterTrustBundleProjection$new(
  label_selector = NULL,
  name = NULL,
  optional = NULL,
  path = NULL,
  signer_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterTrustBundleProjection-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterTrustBundleProjection-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterTrustBundleProjection$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ClusterTrustBundleProjection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ClusterTrustBundleProjection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ClusterTrustBundleProjection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
