% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ComponentCondition.R
\name{V1ComponentCondition}
\alias{V1ComponentCondition}
\title{V1ComponentCondition}
\description{
V1ComponentCondition

V1ComponentCondition
}
\details{
Information about the condition of a component.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ComponentCondition-new}{\code{V1ComponentCondition$new()}}
\item \href{#method-V1ComponentCondition-to_list}{\code{V1ComponentCondition$to_list()}}
\item \href{#method-V1ComponentCondition-clone}{\code{V1ComponentCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentCondition$new(
  error = NULL,
  message = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
