% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ComponentStatus.R
\name{V1ComponentStatus}
\alias{V1ComponentStatus}
\title{V1ComponentStatus}
\description{
V1ComponentStatus

V1ComponentStatus
}
\details{
ComponentStatus (and ComponentStatusList) holds the cluster validation info. Deprecated: This API is deprecated in v1.19+
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ComponentStatus-new}{\code{V1ComponentStatus$new()}}
\item \href{#method-V1ComponentStatus-to_list}{\code{V1ComponentStatus$to_list()}}
\item \href{#method-V1ComponentStatus-clone}{\code{V1ComponentStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentStatus$new(
  api_version = NULL,
  conditions = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
