% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ComponentStatusList.R
\name{V1ComponentStatusList}
\alias{V1ComponentStatusList}
\title{V1ComponentStatusList}
\description{
V1ComponentStatusList

V1ComponentStatusList
}
\details{
Status of all the conditions for the component as a list of ComponentStatus objects. Deprecated: This API is deprecated in v1.19+
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ComponentStatusList-new}{\code{V1ComponentStatusList$new()}}
\item \href{#method-V1ComponentStatusList-to_list}{\code{V1ComponentStatusList$to_list()}}
\item \href{#method-V1ComponentStatusList-clone}{\code{V1ComponentStatusList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentStatusList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentStatusList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentStatusList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentStatusList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentStatusList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentStatusList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ComponentStatusList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ComponentStatusList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ComponentStatusList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
