% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Condition.R
\name{V1Condition}
\alias{V1Condition}
\title{V1Condition}
\description{
V1Condition

V1Condition
}
\details{
Condition contains details for one aspect of the current state of this API Resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Condition-new}{\code{V1Condition$new()}}
\item \href{#method-V1Condition-to_list}{\code{V1Condition$to_list()}}
\item \href{#method-V1Condition-clone}{\code{V1Condition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Condition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Condition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Condition$new(
  last_transition_time = NULL,
  message = NULL,
  observed_generation = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Condition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Condition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Condition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Condition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Condition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Condition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
