% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ConfigMap.R
\name{V1ConfigMap}
\alias{V1ConfigMap}
\title{V1ConfigMap}
\description{
V1ConfigMap

V1ConfigMap
}
\details{
ConfigMap holds configuration data for pods to consume.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ConfigMap-new}{\code{V1ConfigMap$new()}}
\item \href{#method-V1ConfigMap-to_list}{\code{V1ConfigMap$to_list()}}
\item \href{#method-V1ConfigMap-clone}{\code{V1ConfigMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMap-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMap-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMap$new(
  api_version = NULL,
  binary_data = NULL,
  data = NULL,
  immutable = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMap-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMap-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMap$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
