% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ConfigMapEnvSource.R
\name{V1ConfigMapEnvSource}
\alias{V1ConfigMapEnvSource}
\title{V1ConfigMapEnvSource}
\description{
V1ConfigMapEnvSource

V1ConfigMapEnvSource
}
\details{
ConfigMapEnvSource selects a ConfigMap to populate the environment variables with.

The contents of the target ConfigMap's Data field will represent the key-value pairs as environment variables.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ConfigMapEnvSource-new}{\code{V1ConfigMapEnvSource$new()}}
\item \href{#method-V1ConfigMapEnvSource-to_list}{\code{V1ConfigMapEnvSource$to_list()}}
\item \href{#method-V1ConfigMapEnvSource-clone}{\code{V1ConfigMapEnvSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapEnvSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapEnvSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapEnvSource$new(name = NULL, optional = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapEnvSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapEnvSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapEnvSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapEnvSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapEnvSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapEnvSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
