% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ConfigMapKeySelector.R
\name{V1ConfigMapKeySelector}
\alias{V1ConfigMapKeySelector}
\title{V1ConfigMapKeySelector}
\description{
V1ConfigMapKeySelector

V1ConfigMapKeySelector
}
\details{
Selects a key from a ConfigMap.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ConfigMapKeySelector-new}{\code{V1ConfigMapKeySelector$new()}}
\item \href{#method-V1ConfigMapKeySelector-to_list}{\code{V1ConfigMapKeySelector$to_list()}}
\item \href{#method-V1ConfigMapKeySelector-clone}{\code{V1ConfigMapKeySelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapKeySelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapKeySelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapKeySelector$new(key = NULL, name = NULL, optional = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapKeySelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapKeySelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapKeySelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapKeySelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapKeySelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapKeySelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
