% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ConfigMapList.R
\name{V1ConfigMapList}
\alias{V1ConfigMapList}
\title{V1ConfigMapList}
\description{
V1ConfigMapList

V1ConfigMapList
}
\details{
ConfigMapList is a resource containing a list of ConfigMap objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ConfigMapList-new}{\code{V1ConfigMapList$new()}}
\item \href{#method-V1ConfigMapList-to_list}{\code{V1ConfigMapList$to_list()}}
\item \href{#method-V1ConfigMapList-clone}{\code{V1ConfigMapList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
