% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ConfigMapNodeConfigSource.R
\name{V1ConfigMapNodeConfigSource}
\alias{V1ConfigMapNodeConfigSource}
\title{V1ConfigMapNodeConfigSource}
\description{
V1ConfigMapNodeConfigSource

V1ConfigMapNodeConfigSource
}
\details{
ConfigMapNodeConfigSource contains the information to reference a ConfigMap as a config source for the Node. This API is deprecated since 1.22: https://git.k8s.io/enhancements/keps/sig-node/281-dynamic-kubelet-configuration
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ConfigMapNodeConfigSource-new}{\code{V1ConfigMapNodeConfigSource$new()}}
\item \href{#method-V1ConfigMapNodeConfigSource-to_list}{\code{V1ConfigMapNodeConfigSource$to_list()}}
\item \href{#method-V1ConfigMapNodeConfigSource-clone}{\code{V1ConfigMapNodeConfigSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapNodeConfigSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapNodeConfigSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapNodeConfigSource$new(
  kubelet_config_key = NULL,
  name = NULL,
  namespace = NULL,
  resource_version = NULL,
  uid = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapNodeConfigSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapNodeConfigSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapNodeConfigSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapNodeConfigSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapNodeConfigSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapNodeConfigSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
