% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ConfigMapProjection.R
\name{V1ConfigMapProjection}
\alias{V1ConfigMapProjection}
\title{V1ConfigMapProjection}
\description{
V1ConfigMapProjection

V1ConfigMapProjection
}
\details{
Adapts a ConfigMap into a projected volume.

The contents of the target ConfigMap's Data field will be presented in a projected volume as files using the keys in the Data field as the file names, unless the items element is populated with specific mappings of keys to paths. Note that this is identical to a configmap volume source without the default mode.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ConfigMapProjection-new}{\code{V1ConfigMapProjection$new()}}
\item \href{#method-V1ConfigMapProjection-to_list}{\code{V1ConfigMapProjection$to_list()}}
\item \href{#method-V1ConfigMapProjection-clone}{\code{V1ConfigMapProjection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapProjection-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapProjection-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapProjection$new(items = NULL, name = NULL, optional = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapProjection-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapProjection-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapProjection$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ConfigMapProjection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ConfigMapProjection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ConfigMapProjection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
