% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Container.R
\name{V1Container}
\alias{V1Container}
\title{V1Container}
\description{
V1Container

V1Container
}
\details{
A single application container that you want to run within a pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Container-new}{\code{V1Container$new()}}
\item \href{#method-V1Container-to_list}{\code{V1Container$to_list()}}
\item \href{#method-V1Container-clone}{\code{V1Container$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Container-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Container-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Container$new(
  args = NULL,
  command = NULL,
  env = NULL,
  env_from = NULL,
  image = NULL,
  image_pull_policy = NULL,
  lifecycle = NULL,
  liveness_probe = NULL,
  name = NULL,
  ports = NULL,
  readiness_probe = NULL,
  resize_policy = NULL,
  resources = NULL,
  restart_policy = NULL,
  security_context = NULL,
  startup_probe = NULL,
  stdin = NULL,
  stdin_once = NULL,
  termination_message_path = NULL,
  termination_message_policy = NULL,
  tty = NULL,
  volume_devices = NULL,
  volume_mounts = NULL,
  working_dir = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Container-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Container-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Container$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Container-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Container-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Container$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
