% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerImage.R
\name{V1ContainerImage}
\alias{V1ContainerImage}
\title{V1ContainerImage}
\description{
V1ContainerImage

V1ContainerImage
}
\details{
Describe a container image
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerImage-new}{\code{V1ContainerImage$new()}}
\item \href{#method-V1ContainerImage-to_list}{\code{V1ContainerImage$to_list()}}
\item \href{#method-V1ContainerImage-clone}{\code{V1ContainerImage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerImage-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerImage-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerImage$new(names = NULL, size_bytes = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerImage-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerImage-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerImage$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerImage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerImage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerImage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
