% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerPort.R
\name{V1ContainerPort}
\alias{V1ContainerPort}
\title{V1ContainerPort}
\description{
V1ContainerPort

V1ContainerPort
}
\details{
ContainerPort represents a network port in a single container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerPort-new}{\code{V1ContainerPort$new()}}
\item \href{#method-V1ContainerPort-to_list}{\code{V1ContainerPort$to_list()}}
\item \href{#method-V1ContainerPort-clone}{\code{V1ContainerPort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerPort-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerPort-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerPort$new(
  container_port = NULL,
  host_ip = NULL,
  host_port = NULL,
  name = NULL,
  protocol = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerPort-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerPort-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerPort$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerPort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerPort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerPort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
