% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerState.R
\name{V1ContainerState}
\alias{V1ContainerState}
\title{V1ContainerState}
\description{
V1ContainerState

V1ContainerState
}
\details{
ContainerState holds a possible state of container. Only one of its members may be specified. If none of them is specified, the default one is ContainerStateWaiting.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerState-new}{\code{V1ContainerState$new()}}
\item \href{#method-V1ContainerState-to_list}{\code{V1ContainerState$to_list()}}
\item \href{#method-V1ContainerState-clone}{\code{V1ContainerState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerState-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerState-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerState$new(running = NULL, terminated = NULL, waiting = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerState-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerState-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerState$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
