% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerStateRunning.R
\name{V1ContainerStateRunning}
\alias{V1ContainerStateRunning}
\title{V1ContainerStateRunning}
\description{
V1ContainerStateRunning

V1ContainerStateRunning
}
\details{
ContainerStateRunning is a running state of a container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerStateRunning-new}{\code{V1ContainerStateRunning$new()}}
\item \href{#method-V1ContainerStateRunning-to_list}{\code{V1ContainerStateRunning$to_list()}}
\item \href{#method-V1ContainerStateRunning-clone}{\code{V1ContainerStateRunning$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateRunning-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateRunning-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateRunning$new(started_at = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateRunning-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateRunning-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateRunning$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateRunning-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateRunning-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateRunning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
