% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerStateTerminated.R
\name{V1ContainerStateTerminated}
\alias{V1ContainerStateTerminated}
\title{V1ContainerStateTerminated}
\description{
V1ContainerStateTerminated

V1ContainerStateTerminated
}
\details{
ContainerStateTerminated is a terminated state of a container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerStateTerminated-new}{\code{V1ContainerStateTerminated$new()}}
\item \href{#method-V1ContainerStateTerminated-to_list}{\code{V1ContainerStateTerminated$to_list()}}
\item \href{#method-V1ContainerStateTerminated-clone}{\code{V1ContainerStateTerminated$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateTerminated-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateTerminated-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateTerminated$new(
  container_id = NULL,
  exit_code = NULL,
  finished_at = NULL,
  message = NULL,
  reason = NULL,
  signal = NULL,
  started_at = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateTerminated-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateTerminated-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateTerminated$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateTerminated-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateTerminated-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateTerminated$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
