% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerStateWaiting.R
\name{V1ContainerStateWaiting}
\alias{V1ContainerStateWaiting}
\title{V1ContainerStateWaiting}
\description{
V1ContainerStateWaiting

V1ContainerStateWaiting
}
\details{
ContainerStateWaiting is a waiting state of a container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerStateWaiting-new}{\code{V1ContainerStateWaiting$new()}}
\item \href{#method-V1ContainerStateWaiting-to_list}{\code{V1ContainerStateWaiting$to_list()}}
\item \href{#method-V1ContainerStateWaiting-clone}{\code{V1ContainerStateWaiting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateWaiting-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateWaiting-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateWaiting$new(message = NULL, reason = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateWaiting-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateWaiting-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateWaiting$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStateWaiting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStateWaiting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStateWaiting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
