% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerStatus.R
\name{V1ContainerStatus}
\alias{V1ContainerStatus}
\title{V1ContainerStatus}
\description{
V1ContainerStatus

V1ContainerStatus
}
\details{
ContainerStatus contains details for the current status of this container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerStatus-new}{\code{V1ContainerStatus$new()}}
\item \href{#method-V1ContainerStatus-to_list}{\code{V1ContainerStatus$to_list()}}
\item \href{#method-V1ContainerStatus-clone}{\code{V1ContainerStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStatus$new(
  allocated_resources = NULL,
  allocated_resources_status = NULL,
  container_id = NULL,
  image = NULL,
  image_id = NULL,
  last_state = NULL,
  name = NULL,
  ready = NULL,
  resources = NULL,
  restart_count = NULL,
  started = NULL,
  state = NULL,
  user = NULL,
  volume_mounts = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
