% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ContainerUser.R
\name{V1ContainerUser}
\alias{V1ContainerUser}
\title{V1ContainerUser}
\description{
V1ContainerUser

V1ContainerUser
}
\details{
ContainerUser represents user identity information
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ContainerUser-new}{\code{V1ContainerUser$new()}}
\item \href{#method-V1ContainerUser-to_list}{\code{V1ContainerUser$to_list()}}
\item \href{#method-V1ContainerUser-clone}{\code{V1ContainerUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerUser-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerUser$new(linux = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerUser-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerUser-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerUser$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ContainerUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ContainerUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ContainerUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
