% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ControllerRevision.R
\name{V1ControllerRevision}
\alias{V1ControllerRevision}
\title{V1ControllerRevision}
\description{
V1ControllerRevision

V1ControllerRevision
}
\details{
ControllerRevision implements an immutable snapshot of state data. Clients are responsible for serializing and deserializing the objects that contain their internal state. Once a ControllerRevision has been successfully created, it can not be updated. The API Server will fail validation of all requests that attempt to mutate the Data field. ControllerRevisions may, however, be deleted. Note that, due to its use by both the DaemonSet and StatefulSet controllers for update and rollback, this object is beta. However, it may be subject to name and representation changes in future releases, and clients should not depend on its stability. It is primarily for internal use by controllers.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ControllerRevision-new}{\code{V1ControllerRevision$new()}}
\item \href{#method-V1ControllerRevision-to_list}{\code{V1ControllerRevision$to_list()}}
\item \href{#method-V1ControllerRevision-clone}{\code{V1ControllerRevision$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ControllerRevision-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ControllerRevision-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ControllerRevision$new(
  api_version = NULL,
  data = NULL,
  kind = NULL,
  metadata = NULL,
  revision = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ControllerRevision-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ControllerRevision-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ControllerRevision$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ControllerRevision-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ControllerRevision-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ControllerRevision$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
