% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ControllerRevisionList.R
\name{V1ControllerRevisionList}
\alias{V1ControllerRevisionList}
\title{V1ControllerRevisionList}
\description{
V1ControllerRevisionList

V1ControllerRevisionList
}
\details{
ControllerRevisionList is a resource containing a list of ControllerRevision objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ControllerRevisionList-new}{\code{V1ControllerRevisionList$new()}}
\item \href{#method-V1ControllerRevisionList-to_list}{\code{V1ControllerRevisionList$to_list()}}
\item \href{#method-V1ControllerRevisionList-clone}{\code{V1ControllerRevisionList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ControllerRevisionList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ControllerRevisionList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ControllerRevisionList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ControllerRevisionList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ControllerRevisionList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ControllerRevisionList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ControllerRevisionList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ControllerRevisionList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ControllerRevisionList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
