% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CronJob.R
\name{V1CronJob}
\alias{V1CronJob}
\title{V1CronJob}
\description{
V1CronJob

V1CronJob
}
\details{
CronJob represents the configuration of a single cron job.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CronJob-new}{\code{V1CronJob$new()}}
\item \href{#method-V1CronJob-to_list}{\code{V1CronJob$to_list()}}
\item \href{#method-V1CronJob-clone}{\code{V1CronJob$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJob-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJob-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJob$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJob-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJob-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJob$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJob-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJob-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
