% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CronJobSpec.R
\name{V1CronJobSpec}
\alias{V1CronJobSpec}
\title{V1CronJobSpec}
\description{
V1CronJobSpec

V1CronJobSpec
}
\details{
CronJobSpec describes how the job execution will look like and when it will actually run.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CronJobSpec-new}{\code{V1CronJobSpec$new()}}
\item \href{#method-V1CronJobSpec-to_list}{\code{V1CronJobSpec$to_list()}}
\item \href{#method-V1CronJobSpec-clone}{\code{V1CronJobSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJobSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJobSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJobSpec$new(
  concurrency_policy = NULL,
  failed_jobs_history_limit = NULL,
  job_template = NULL,
  schedule = NULL,
  starting_deadline_seconds = NULL,
  successful_jobs_history_limit = NULL,
  suspend = NULL,
  time_zone = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJobSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJobSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJobSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJobSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJobSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJobSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
