% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CronJobStatus.R
\name{V1CronJobStatus}
\alias{V1CronJobStatus}
\title{V1CronJobStatus}
\description{
V1CronJobStatus

V1CronJobStatus
}
\details{
CronJobStatus represents the current state of a cron job.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CronJobStatus-new}{\code{V1CronJobStatus$new()}}
\item \href{#method-V1CronJobStatus-to_list}{\code{V1CronJobStatus$to_list()}}
\item \href{#method-V1CronJobStatus-clone}{\code{V1CronJobStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJobStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJobStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJobStatus$new(
  active = NULL,
  last_schedule_time = NULL,
  last_successful_time = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJobStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJobStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJobStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CronJobStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CronJobStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CronJobStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
