% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CrossVersionObjectReference.R
\name{V1CrossVersionObjectReference}
\alias{V1CrossVersionObjectReference}
\title{V1CrossVersionObjectReference}
\description{
V1CrossVersionObjectReference

V1CrossVersionObjectReference
}
\details{
CrossVersionObjectReference contains enough information to let you identify the referred resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CrossVersionObjectReference-new}{\code{V1CrossVersionObjectReference$new()}}
\item \href{#method-V1CrossVersionObjectReference-to_list}{\code{V1CrossVersionObjectReference$to_list()}}
\item \href{#method-V1CrossVersionObjectReference-clone}{\code{V1CrossVersionObjectReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CrossVersionObjectReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CrossVersionObjectReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CrossVersionObjectReference$new(api_version = NULL, kind = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CrossVersionObjectReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CrossVersionObjectReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CrossVersionObjectReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CrossVersionObjectReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CrossVersionObjectReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CrossVersionObjectReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
