% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSIDriver.R
\name{V1CSIDriver}
\alias{V1CSIDriver}
\title{V1CSIDriver}
\description{
V1CSIDriver

V1CSIDriver
}
\details{
CSIDriver captures information about a Container Storage Interface (CSI) volume driver deployed on the cluster. Kubernetes attach detach controller uses this object to determine whether attach is required. Kubelet uses this object to determine whether pod information needs to be passed on mount. CSIDriver objects are non-namespaced.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSIDriver-new}{\code{V1CSIDriver$new()}}
\item \href{#method-V1CSIDriver-to_list}{\code{V1CSIDriver$to_list()}}
\item \href{#method-V1CSIDriver-clone}{\code{V1CSIDriver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriver-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriver-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriver$new(api_version = NULL, kind = NULL, metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriver-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriver-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriver$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
