% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSIDriverList.R
\name{V1CSIDriverList}
\alias{V1CSIDriverList}
\title{V1CSIDriverList}
\description{
V1CSIDriverList

V1CSIDriverList
}
\details{
CSIDriverList is a collection of CSIDriver objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSIDriverList-new}{\code{V1CSIDriverList$new()}}
\item \href{#method-V1CSIDriverList-to_list}{\code{V1CSIDriverList$to_list()}}
\item \href{#method-V1CSIDriverList-clone}{\code{V1CSIDriverList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriverList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriverList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriverList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriverList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriverList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriverList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriverList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriverList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriverList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
