% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSIDriverSpec.R
\name{V1CSIDriverSpec}
\alias{V1CSIDriverSpec}
\title{V1CSIDriverSpec}
\description{
V1CSIDriverSpec

V1CSIDriverSpec
}
\details{
CSIDriverSpec is the specification of a CSIDriver.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSIDriverSpec-new}{\code{V1CSIDriverSpec$new()}}
\item \href{#method-V1CSIDriverSpec-to_list}{\code{V1CSIDriverSpec$to_list()}}
\item \href{#method-V1CSIDriverSpec-clone}{\code{V1CSIDriverSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriverSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriverSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriverSpec$new(
  attach_required = NULL,
  fs_group_policy = NULL,
  pod_info_on_mount = NULL,
  requires_republish = NULL,
  se_linux_mount = NULL,
  storage_capacity = NULL,
  token_requests = NULL,
  volume_lifecycle_modes = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriverSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriverSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriverSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIDriverSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIDriverSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIDriverSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
