% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSINode.R
\name{V1CSINode}
\alias{V1CSINode}
\title{V1CSINode}
\description{
V1CSINode

V1CSINode
}
\details{
CSINode holds information about all CSI drivers installed on a node. CSI drivers do not need to create the CSINode object directly. As long as they use the node-driver-registrar sidecar container, the kubelet will automatically populate the CSINode object for the CSI driver as part of kubelet plugin registration. CSINode has the same name as a node. If the object is missing, it means either there are no CSI Drivers available on the node, or the Kubelet version is low enough that it doesn't create this object. CSINode has an OwnerReference that points to the corresponding node object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSINode-new}{\code{V1CSINode$new()}}
\item \href{#method-V1CSINode-to_list}{\code{V1CSINode$to_list()}}
\item \href{#method-V1CSINode-clone}{\code{V1CSINode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINode-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINode-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINode$new(api_version = NULL, kind = NULL, metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINode-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINode-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINode$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
