% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSINodeDriver.R
\name{V1CSINodeDriver}
\alias{V1CSINodeDriver}
\title{V1CSINodeDriver}
\description{
V1CSINodeDriver

V1CSINodeDriver
}
\details{
CSINodeDriver holds information about the specification of one CSI driver installed on a node
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSINodeDriver-new}{\code{V1CSINodeDriver$new()}}
\item \href{#method-V1CSINodeDriver-to_list}{\code{V1CSINodeDriver$to_list()}}
\item \href{#method-V1CSINodeDriver-clone}{\code{V1CSINodeDriver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeDriver-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeDriver-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeDriver$new(
  allocatable = NULL,
  name = NULL,
  node_id = NULL,
  topology_keys = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeDriver-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeDriver-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeDriver$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeDriver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeDriver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeDriver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
