% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSINodeList.R
\name{V1CSINodeList}
\alias{V1CSINodeList}
\title{V1CSINodeList}
\description{
V1CSINodeList

V1CSINodeList
}
\details{
CSINodeList is a collection of CSINode objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSINodeList-new}{\code{V1CSINodeList$new()}}
\item \href{#method-V1CSINodeList-to_list}{\code{V1CSINodeList$to_list()}}
\item \href{#method-V1CSINodeList-clone}{\code{V1CSINodeList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
