% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSINodeSpec.R
\name{V1CSINodeSpec}
\alias{V1CSINodeSpec}
\title{V1CSINodeSpec}
\description{
V1CSINodeSpec

V1CSINodeSpec
}
\details{
CSINodeSpec holds information about the specification of all CSI drivers installed on a node
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSINodeSpec-new}{\code{V1CSINodeSpec$new()}}
\item \href{#method-V1CSINodeSpec-to_list}{\code{V1CSINodeSpec$to_list()}}
\item \href{#method-V1CSINodeSpec-clone}{\code{V1CSINodeSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeSpec$new(drivers = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSINodeSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSINodeSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSINodeSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
