% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSIPersistentVolumeSource.R
\name{V1CSIPersistentVolumeSource}
\alias{V1CSIPersistentVolumeSource}
\title{V1CSIPersistentVolumeSource}
\description{
V1CSIPersistentVolumeSource

V1CSIPersistentVolumeSource
}
\details{
Represents storage that is managed by an external CSI volume driver (Beta feature)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSIPersistentVolumeSource-new}{\code{V1CSIPersistentVolumeSource$new()}}
\item \href{#method-V1CSIPersistentVolumeSource-to_list}{\code{V1CSIPersistentVolumeSource$to_list()}}
\item \href{#method-V1CSIPersistentVolumeSource-clone}{\code{V1CSIPersistentVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIPersistentVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIPersistentVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIPersistentVolumeSource$new(
  controller_expand_secret_ref = NULL,
  controller_publish_secret_ref = NULL,
  driver = NULL,
  fs_type = NULL,
  node_expand_secret_ref = NULL,
  node_publish_secret_ref = NULL,
  node_stage_secret_ref = NULL,
  read_only = NULL,
  volume_attributes = NULL,
  volume_handle = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIPersistentVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIPersistentVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIPersistentVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIPersistentVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIPersistentVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIPersistentVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
