% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSIStorageCapacity.R
\name{V1CSIStorageCapacity}
\alias{V1CSIStorageCapacity}
\title{V1CSIStorageCapacity}
\description{
V1CSIStorageCapacity

V1CSIStorageCapacity
}
\details{
CSIStorageCapacity stores the result of one CSI GetCapacity call. For a given StorageClass, this describes the available capacity in a particular topology segment.  This can be used when considering where to instantiate new PersistentVolumes.

For example this can express things like: - StorageClass "standard" has "1234 GiB" available in "topology.kubernetes.io/zone=us-east1" - StorageClass "localssd" has "10 GiB" available in "kubernetes.io/hostname=knode-abc123"

The following three cases all imply that no capacity is available for a certain combination: - no object exists with suitable topology and storage class name - such an object exists, but the capacity is unset - such an object exists, but the capacity is zero

The producer of these objects can decide which approach is more suitable.

They are consumed by the kube-scheduler when a CSI driver opts into capacity-aware scheduling with CSIDriverSpec.StorageCapacity. The scheduler compares the MaximumVolumeSize against the requested size of pending volumes to filter out unsuitable nodes. If MaximumVolumeSize is unset, it falls back to a comparison against the less precise Capacity. If that is also unset, the scheduler assumes that capacity is insufficient and tries some other node.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSIStorageCapacity-new}{\code{V1CSIStorageCapacity$new()}}
\item \href{#method-V1CSIStorageCapacity-to_list}{\code{V1CSIStorageCapacity$to_list()}}
\item \href{#method-V1CSIStorageCapacity-clone}{\code{V1CSIStorageCapacity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIStorageCapacity-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIStorageCapacity-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIStorageCapacity$new(
  api_version = NULL,
  capacity = NULL,
  kind = NULL,
  maximum_volume_size = NULL,
  metadata = NULL,
  node_topology = NULL,
  storage_class_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIStorageCapacity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIStorageCapacity-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIStorageCapacity$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIStorageCapacity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIStorageCapacity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIStorageCapacity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
