% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSIStorageCapacityList.R
\name{V1CSIStorageCapacityList}
\alias{V1CSIStorageCapacityList}
\title{V1CSIStorageCapacityList}
\description{
V1CSIStorageCapacityList

V1CSIStorageCapacityList
}
\details{
CSIStorageCapacityList is a collection of CSIStorageCapacity objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSIStorageCapacityList-new}{\code{V1CSIStorageCapacityList$new()}}
\item \href{#method-V1CSIStorageCapacityList-to_list}{\code{V1CSIStorageCapacityList$to_list()}}
\item \href{#method-V1CSIStorageCapacityList-clone}{\code{V1CSIStorageCapacityList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIStorageCapacityList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIStorageCapacityList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIStorageCapacityList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIStorageCapacityList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIStorageCapacityList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIStorageCapacityList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIStorageCapacityList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIStorageCapacityList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIStorageCapacityList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
