% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CSIVolumeSource.R
\name{V1CSIVolumeSource}
\alias{V1CSIVolumeSource}
\title{V1CSIVolumeSource}
\description{
V1CSIVolumeSource

V1CSIVolumeSource
}
\details{
Represents a source location of a volume to mount, managed by an external CSI driver
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CSIVolumeSource-new}{\code{V1CSIVolumeSource$new()}}
\item \href{#method-V1CSIVolumeSource-to_list}{\code{V1CSIVolumeSource$to_list()}}
\item \href{#method-V1CSIVolumeSource-clone}{\code{V1CSIVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIVolumeSource$new(
  driver = NULL,
  fs_type = NULL,
  node_publish_secret_ref = NULL,
  read_only = NULL,
  volume_attributes = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CSIVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CSIVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CSIVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
