% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceColumnDefinition.R
\name{V1CustomResourceColumnDefinition}
\alias{V1CustomResourceColumnDefinition}
\title{V1CustomResourceColumnDefinition}
\description{
V1CustomResourceColumnDefinition

V1CustomResourceColumnDefinition
}
\details{
CustomResourceColumnDefinition specifies a column for server side printing.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceColumnDefinition-new}{\code{V1CustomResourceColumnDefinition$new()}}
\item \href{#method-V1CustomResourceColumnDefinition-to_list}{\code{V1CustomResourceColumnDefinition$to_list()}}
\item \href{#method-V1CustomResourceColumnDefinition-clone}{\code{V1CustomResourceColumnDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceColumnDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceColumnDefinition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceColumnDefinition$new(
  description = NULL,
  format = NULL,
  json_path = NULL,
  name = NULL,
  priority = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceColumnDefinition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceColumnDefinition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceColumnDefinition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceColumnDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceColumnDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceColumnDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
