% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceConversion.R
\name{V1CustomResourceConversion}
\alias{V1CustomResourceConversion}
\title{V1CustomResourceConversion}
\description{
V1CustomResourceConversion

V1CustomResourceConversion
}
\details{
CustomResourceConversion describes how to convert different versions of a CR.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceConversion-new}{\code{V1CustomResourceConversion$new()}}
\item \href{#method-V1CustomResourceConversion-to_list}{\code{V1CustomResourceConversion$to_list()}}
\item \href{#method-V1CustomResourceConversion-clone}{\code{V1CustomResourceConversion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceConversion-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceConversion-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceConversion$new(strategy = NULL, webhook = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceConversion-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceConversion-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceConversion$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceConversion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceConversion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceConversion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
