% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceDefinition.R
\name{V1CustomResourceDefinition}
\alias{V1CustomResourceDefinition}
\title{V1CustomResourceDefinition}
\description{
V1CustomResourceDefinition

V1CustomResourceDefinition
}
\details{
CustomResourceDefinition represents a resource that should be exposed on the API server.  Its name MUST be in the format <.spec.name>.<.spec.group>.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceDefinition-new}{\code{V1CustomResourceDefinition$new()}}
\item \href{#method-V1CustomResourceDefinition-to_list}{\code{V1CustomResourceDefinition$to_list()}}
\item \href{#method-V1CustomResourceDefinition-clone}{\code{V1CustomResourceDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinition$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
