% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceDefinitionCondition.R
\name{V1CustomResourceDefinitionCondition}
\alias{V1CustomResourceDefinitionCondition}
\title{V1CustomResourceDefinitionCondition}
\description{
V1CustomResourceDefinitionCondition

V1CustomResourceDefinitionCondition
}
\details{
CustomResourceDefinitionCondition contains details for the current condition of this pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceDefinitionCondition-new}{\code{V1CustomResourceDefinitionCondition$new()}}
\item \href{#method-V1CustomResourceDefinitionCondition-to_list}{\code{V1CustomResourceDefinitionCondition$to_list()}}
\item \href{#method-V1CustomResourceDefinitionCondition-clone}{\code{V1CustomResourceDefinitionCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
