% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceDefinitionList.R
\name{V1CustomResourceDefinitionList}
\alias{V1CustomResourceDefinitionList}
\title{V1CustomResourceDefinitionList}
\description{
V1CustomResourceDefinitionList

V1CustomResourceDefinitionList
}
\details{
CustomResourceDefinitionList is a list of CustomResourceDefinition objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceDefinitionList-new}{\code{V1CustomResourceDefinitionList$new()}}
\item \href{#method-V1CustomResourceDefinitionList-to_list}{\code{V1CustomResourceDefinitionList$to_list()}}
\item \href{#method-V1CustomResourceDefinitionList-clone}{\code{V1CustomResourceDefinitionList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
