% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceDefinitionNames.R
\name{V1CustomResourceDefinitionNames}
\alias{V1CustomResourceDefinitionNames}
\title{V1CustomResourceDefinitionNames}
\description{
V1CustomResourceDefinitionNames

V1CustomResourceDefinitionNames
}
\details{
CustomResourceDefinitionNames indicates the names to serve this CustomResourceDefinition
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceDefinitionNames-new}{\code{V1CustomResourceDefinitionNames$new()}}
\item \href{#method-V1CustomResourceDefinitionNames-to_list}{\code{V1CustomResourceDefinitionNames$to_list()}}
\item \href{#method-V1CustomResourceDefinitionNames-clone}{\code{V1CustomResourceDefinitionNames$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionNames-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionNames-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionNames$new(
  categories = NULL,
  kind = NULL,
  list_kind = NULL,
  plural = NULL,
  short_names = NULL,
  singular = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionNames-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionNames-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionNames$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionNames-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionNames-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionNames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
