% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceDefinitionSpec.R
\name{V1CustomResourceDefinitionSpec}
\alias{V1CustomResourceDefinitionSpec}
\title{V1CustomResourceDefinitionSpec}
\description{
V1CustomResourceDefinitionSpec

V1CustomResourceDefinitionSpec
}
\details{
CustomResourceDefinitionSpec describes how a user wants their resource to appear
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceDefinitionSpec-new}{\code{V1CustomResourceDefinitionSpec$new()}}
\item \href{#method-V1CustomResourceDefinitionSpec-to_list}{\code{V1CustomResourceDefinitionSpec$to_list()}}
\item \href{#method-V1CustomResourceDefinitionSpec-clone}{\code{V1CustomResourceDefinitionSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionSpec$new(
  conversion = NULL,
  group = NULL,
  names = NULL,
  preserve_unknown_fields = NULL,
  scope = NULL,
  versions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
