% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceDefinitionStatus.R
\name{V1CustomResourceDefinitionStatus}
\alias{V1CustomResourceDefinitionStatus}
\title{V1CustomResourceDefinitionStatus}
\description{
V1CustomResourceDefinitionStatus

V1CustomResourceDefinitionStatus
}
\details{
CustomResourceDefinitionStatus indicates the state of the CustomResourceDefinition
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceDefinitionStatus-new}{\code{V1CustomResourceDefinitionStatus$new()}}
\item \href{#method-V1CustomResourceDefinitionStatus-to_list}{\code{V1CustomResourceDefinitionStatus$to_list()}}
\item \href{#method-V1CustomResourceDefinitionStatus-clone}{\code{V1CustomResourceDefinitionStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionStatus$new(
  accepted_names = NULL,
  conditions = NULL,
  stored_versions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
