% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceDefinitionVersion.R
\name{V1CustomResourceDefinitionVersion}
\alias{V1CustomResourceDefinitionVersion}
\title{V1CustomResourceDefinitionVersion}
\description{
V1CustomResourceDefinitionVersion

V1CustomResourceDefinitionVersion
}
\details{
CustomResourceDefinitionVersion describes a version for CRD.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceDefinitionVersion-new}{\code{V1CustomResourceDefinitionVersion$new()}}
\item \href{#method-V1CustomResourceDefinitionVersion-to_list}{\code{V1CustomResourceDefinitionVersion$to_list()}}
\item \href{#method-V1CustomResourceDefinitionVersion-clone}{\code{V1CustomResourceDefinitionVersion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionVersion-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionVersion-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionVersion$new(
  additional_printer_columns = NULL,
  deprecated = NULL,
  deprecation_warning = NULL,
  name = NULL,
  schema = NULL,
  selectable_fields = NULL,
  served = NULL,
  storage = NULL,
  subresources = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionVersion-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionVersion-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionVersion$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceDefinitionVersion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceDefinitionVersion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceDefinitionVersion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
