% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceSubresources.R
\name{V1CustomResourceSubresources}
\alias{V1CustomResourceSubresources}
\title{V1CustomResourceSubresources}
\description{
V1CustomResourceSubresources

V1CustomResourceSubresources
}
\details{
CustomResourceSubresources defines the status and scale subresources for CustomResources.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceSubresources-new}{\code{V1CustomResourceSubresources$new()}}
\item \href{#method-V1CustomResourceSubresources-to_list}{\code{V1CustomResourceSubresources$to_list()}}
\item \href{#method-V1CustomResourceSubresources-clone}{\code{V1CustomResourceSubresources$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresources-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresources-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresources$new(scale = NULL, status = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresources-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresources-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresources$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresources-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresources-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresources$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
