% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceSubresourceScale.R
\name{V1CustomResourceSubresourceScale}
\alias{V1CustomResourceSubresourceScale}
\title{V1CustomResourceSubresourceScale}
\description{
V1CustomResourceSubresourceScale

V1CustomResourceSubresourceScale
}
\details{
CustomResourceSubresourceScale defines how to serve the scale subresource for CustomResources.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceSubresourceScale-new}{\code{V1CustomResourceSubresourceScale$new()}}
\item \href{#method-V1CustomResourceSubresourceScale-to_list}{\code{V1CustomResourceSubresourceScale$to_list()}}
\item \href{#method-V1CustomResourceSubresourceScale-clone}{\code{V1CustomResourceSubresourceScale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresourceScale-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresourceScale-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresourceScale$new(
  label_selector_path = NULL,
  spec_replicas_path = NULL,
  status_replicas_path = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresourceScale-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresourceScale-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresourceScale$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresourceScale-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresourceScale-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresourceScale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
