% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceSubresourceStatus.R
\name{V1CustomResourceSubresourceStatus}
\alias{V1CustomResourceSubresourceStatus}
\title{V1CustomResourceSubresourceStatus}
\description{
V1CustomResourceSubresourceStatus

V1CustomResourceSubresourceStatus
}
\details{
CustomResourceSubresourceStatus defines how to serve the status subresource for CustomResources. Status is represented by the `.status` JSON path inside of a CustomResource. When set, * exposes a /status subresource for the custom resource * PUT requests to the /status subresource take a custom resource object, and ignore changes to anything except the status stanza * PUT/POST/PATCH requests to the custom resource ignore changes to the status stanza
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceSubresourceStatus-new}{\code{V1CustomResourceSubresourceStatus$new()}}
\item \href{#method-V1CustomResourceSubresourceStatus-to_list}{\code{V1CustomResourceSubresourceStatus$to_list()}}
\item \href{#method-V1CustomResourceSubresourceStatus-clone}{\code{V1CustomResourceSubresourceStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresourceStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresourceStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresourceStatus$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresourceStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresourceStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresourceStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceSubresourceStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceSubresourceStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceSubresourceStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
