% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1CustomResourceValidation.R
\name{V1CustomResourceValidation}
\alias{V1CustomResourceValidation}
\title{V1CustomResourceValidation}
\description{
V1CustomResourceValidation

V1CustomResourceValidation
}
\details{
CustomResourceValidation is a list of validation methods for CustomResources.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1CustomResourceValidation-new}{\code{V1CustomResourceValidation$new()}}
\item \href{#method-V1CustomResourceValidation-to_list}{\code{V1CustomResourceValidation$to_list()}}
\item \href{#method-V1CustomResourceValidation-clone}{\code{V1CustomResourceValidation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceValidation-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceValidation-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceValidation$new(open_apiv3_schema = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceValidation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceValidation-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceValidation$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1CustomResourceValidation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1CustomResourceValidation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1CustomResourceValidation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
