% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DaemonEndpoint.R
\name{V1DaemonEndpoint}
\alias{V1DaemonEndpoint}
\title{V1DaemonEndpoint}
\description{
V1DaemonEndpoint

V1DaemonEndpoint
}
\details{
DaemonEndpoint contains information about a single Daemon endpoint.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DaemonEndpoint-new}{\code{V1DaemonEndpoint$new()}}
\item \href{#method-V1DaemonEndpoint-to_list}{\code{V1DaemonEndpoint$to_list()}}
\item \href{#method-V1DaemonEndpoint-clone}{\code{V1DaemonEndpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonEndpoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonEndpoint-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonEndpoint$new(port = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonEndpoint-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonEndpoint-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonEndpoint$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonEndpoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonEndpoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonEndpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
