% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DaemonSet.R
\name{V1DaemonSet}
\alias{V1DaemonSet}
\title{V1DaemonSet}
\description{
V1DaemonSet

V1DaemonSet
}
\details{
DaemonSet represents the configuration of a daemon set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DaemonSet-new}{\code{V1DaemonSet$new()}}
\item \href{#method-V1DaemonSet-to_list}{\code{V1DaemonSet$to_list()}}
\item \href{#method-V1DaemonSet-clone}{\code{V1DaemonSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSet-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSet$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSet-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSet-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSet$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
