% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DaemonSetCondition.R
\name{V1DaemonSetCondition}
\alias{V1DaemonSetCondition}
\title{V1DaemonSetCondition}
\description{
V1DaemonSetCondition

V1DaemonSetCondition
}
\details{
DaemonSetCondition describes the state of a DaemonSet at a certain point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DaemonSetCondition-new}{\code{V1DaemonSetCondition$new()}}
\item \href{#method-V1DaemonSetCondition-to_list}{\code{V1DaemonSetCondition$to_list()}}
\item \href{#method-V1DaemonSetCondition-clone}{\code{V1DaemonSetCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
