% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DaemonSetList.R
\name{V1DaemonSetList}
\alias{V1DaemonSetList}
\title{V1DaemonSetList}
\description{
V1DaemonSetList

V1DaemonSetList
}
\details{
DaemonSetList is a collection of daemon sets.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DaemonSetList-new}{\code{V1DaemonSetList$new()}}
\item \href{#method-V1DaemonSetList-to_list}{\code{V1DaemonSetList$to_list()}}
\item \href{#method-V1DaemonSetList-clone}{\code{V1DaemonSetList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
