% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DaemonSetSpec.R
\name{V1DaemonSetSpec}
\alias{V1DaemonSetSpec}
\title{V1DaemonSetSpec}
\description{
V1DaemonSetSpec

V1DaemonSetSpec
}
\details{
DaemonSetSpec is the specification of a daemon set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DaemonSetSpec-new}{\code{V1DaemonSetSpec$new()}}
\item \href{#method-V1DaemonSetSpec-to_list}{\code{V1DaemonSetSpec$to_list()}}
\item \href{#method-V1DaemonSetSpec-clone}{\code{V1DaemonSetSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetSpec$new(
  min_ready_seconds = NULL,
  revision_history_limit = NULL,
  selector = NULL,
  template = NULL,
  update_strategy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
