% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DaemonSetStatus.R
\name{V1DaemonSetStatus}
\alias{V1DaemonSetStatus}
\title{V1DaemonSetStatus}
\description{
V1DaemonSetStatus

V1DaemonSetStatus
}
\details{
DaemonSetStatus represents the current status of a daemon set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DaemonSetStatus-new}{\code{V1DaemonSetStatus$new()}}
\item \href{#method-V1DaemonSetStatus-to_list}{\code{V1DaemonSetStatus$to_list()}}
\item \href{#method-V1DaemonSetStatus-clone}{\code{V1DaemonSetStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetStatus$new(
  collision_count = NULL,
  conditions = NULL,
  current_number_scheduled = NULL,
  desired_number_scheduled = NULL,
  number_available = NULL,
  number_misscheduled = NULL,
  number_ready = NULL,
  number_unavailable = NULL,
  observed_generation = NULL,
  updated_number_scheduled = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
