% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DaemonSetUpdateStrategy.R
\name{V1DaemonSetUpdateStrategy}
\alias{V1DaemonSetUpdateStrategy}
\title{V1DaemonSetUpdateStrategy}
\description{
V1DaemonSetUpdateStrategy

V1DaemonSetUpdateStrategy
}
\details{
DaemonSetUpdateStrategy is a struct used to control the update strategy for a DaemonSet.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DaemonSetUpdateStrategy-new}{\code{V1DaemonSetUpdateStrategy$new()}}
\item \href{#method-V1DaemonSetUpdateStrategy-to_list}{\code{V1DaemonSetUpdateStrategy$to_list()}}
\item \href{#method-V1DaemonSetUpdateStrategy-clone}{\code{V1DaemonSetUpdateStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetUpdateStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetUpdateStrategy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetUpdateStrategy$new(rolling_update = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetUpdateStrategy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetUpdateStrategy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetUpdateStrategy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DaemonSetUpdateStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DaemonSetUpdateStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DaemonSetUpdateStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
