% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DeleteOptions.R
\name{V1DeleteOptions}
\alias{V1DeleteOptions}
\title{V1DeleteOptions}
\description{
V1DeleteOptions

V1DeleteOptions
}
\details{
DeleteOptions may be provided when deleting an API object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DeleteOptions-new}{\code{V1DeleteOptions$new()}}
\item \href{#method-V1DeleteOptions-to_list}{\code{V1DeleteOptions$to_list()}}
\item \href{#method-V1DeleteOptions-clone}{\code{V1DeleteOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeleteOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeleteOptions-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeleteOptions$new(
  api_version = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  kind = NULL,
  orphan_dependents = NULL,
  preconditions = NULL,
  propagation_policy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeleteOptions-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeleteOptions-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeleteOptions$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeleteOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeleteOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeleteOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
