% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Deployment.R
\name{V1Deployment}
\alias{V1Deployment}
\title{V1Deployment}
\description{
V1Deployment

V1Deployment
}
\details{
Deployment enables declarative updates for Pods and ReplicaSets.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Deployment-new}{\code{V1Deployment$new()}}
\item \href{#method-V1Deployment-to_list}{\code{V1Deployment$to_list()}}
\item \href{#method-V1Deployment-clone}{\code{V1Deployment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Deployment-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Deployment-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Deployment$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Deployment-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Deployment-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Deployment$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Deployment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Deployment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Deployment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
