% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DeploymentList.R
\name{V1DeploymentList}
\alias{V1DeploymentList}
\title{V1DeploymentList}
\description{
V1DeploymentList

V1DeploymentList
}
\details{
DeploymentList is a list of Deployments.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DeploymentList-new}{\code{V1DeploymentList$new()}}
\item \href{#method-V1DeploymentList-to_list}{\code{V1DeploymentList$to_list()}}
\item \href{#method-V1DeploymentList-clone}{\code{V1DeploymentList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
