% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DeploymentSpec.R
\name{V1DeploymentSpec}
\alias{V1DeploymentSpec}
\title{V1DeploymentSpec}
\description{
V1DeploymentSpec

V1DeploymentSpec
}
\details{
DeploymentSpec is the specification of the desired behavior of the Deployment.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DeploymentSpec-new}{\code{V1DeploymentSpec$new()}}
\item \href{#method-V1DeploymentSpec-to_list}{\code{V1DeploymentSpec$to_list()}}
\item \href{#method-V1DeploymentSpec-clone}{\code{V1DeploymentSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentSpec$new(
  min_ready_seconds = NULL,
  paused = NULL,
  progress_deadline_seconds = NULL,
  replicas = NULL,
  revision_history_limit = NULL,
  selector = NULL,
  strategy = NULL,
  template = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
