% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DeploymentStatus.R
\name{V1DeploymentStatus}
\alias{V1DeploymentStatus}
\title{V1DeploymentStatus}
\description{
V1DeploymentStatus

V1DeploymentStatus
}
\details{
DeploymentStatus is the most recently observed status of the Deployment.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DeploymentStatus-new}{\code{V1DeploymentStatus$new()}}
\item \href{#method-V1DeploymentStatus-to_list}{\code{V1DeploymentStatus$to_list()}}
\item \href{#method-V1DeploymentStatus-clone}{\code{V1DeploymentStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentStatus$new(
  available_replicas = NULL,
  collision_count = NULL,
  conditions = NULL,
  observed_generation = NULL,
  ready_replicas = NULL,
  replicas = NULL,
  unavailable_replicas = NULL,
  updated_replicas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
