% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DeploymentStrategy.R
\name{V1DeploymentStrategy}
\alias{V1DeploymentStrategy}
\title{V1DeploymentStrategy}
\description{
V1DeploymentStrategy

V1DeploymentStrategy
}
\details{
DeploymentStrategy describes how to replace existing pods with new ones.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DeploymentStrategy-new}{\code{V1DeploymentStrategy$new()}}
\item \href{#method-V1DeploymentStrategy-to_list}{\code{V1DeploymentStrategy$to_list()}}
\item \href{#method-V1DeploymentStrategy-clone}{\code{V1DeploymentStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentStrategy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentStrategy$new(rolling_update = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentStrategy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentStrategy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentStrategy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DeploymentStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DeploymentStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DeploymentStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
