% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DownwardAPIProjection.R
\name{V1DownwardAPIProjection}
\alias{V1DownwardAPIProjection}
\title{V1DownwardAPIProjection}
\description{
V1DownwardAPIProjection

V1DownwardAPIProjection
}
\details{
Represents downward API info for projecting into a projected volume. Note that this is identical to a downwardAPI volume source without the default mode.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DownwardAPIProjection-new}{\code{V1DownwardAPIProjection$new()}}
\item \href{#method-V1DownwardAPIProjection-to_list}{\code{V1DownwardAPIProjection$to_list()}}
\item \href{#method-V1DownwardAPIProjection-clone}{\code{V1DownwardAPIProjection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIProjection-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIProjection-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIProjection$new(items = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIProjection-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIProjection-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIProjection$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIProjection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIProjection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIProjection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
