% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DownwardAPIVolumeFile.R
\name{V1DownwardAPIVolumeFile}
\alias{V1DownwardAPIVolumeFile}
\title{V1DownwardAPIVolumeFile}
\description{
V1DownwardAPIVolumeFile

V1DownwardAPIVolumeFile
}
\details{
DownwardAPIVolumeFile represents information to create the file containing the pod field
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DownwardAPIVolumeFile-new}{\code{V1DownwardAPIVolumeFile$new()}}
\item \href{#method-V1DownwardAPIVolumeFile-to_list}{\code{V1DownwardAPIVolumeFile$to_list()}}
\item \href{#method-V1DownwardAPIVolumeFile-clone}{\code{V1DownwardAPIVolumeFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIVolumeFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIVolumeFile-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIVolumeFile$new(
  field_ref = NULL,
  mode = NULL,
  path = NULL,
  resource_field_ref = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIVolumeFile-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIVolumeFile-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIVolumeFile$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIVolumeFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIVolumeFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIVolumeFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
