% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1DownwardAPIVolumeSource.R
\name{V1DownwardAPIVolumeSource}
\alias{V1DownwardAPIVolumeSource}
\title{V1DownwardAPIVolumeSource}
\description{
V1DownwardAPIVolumeSource

V1DownwardAPIVolumeSource
}
\details{
DownwardAPIVolumeSource represents a volume containing downward API info. Downward API volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1DownwardAPIVolumeSource-new}{\code{V1DownwardAPIVolumeSource$new()}}
\item \href{#method-V1DownwardAPIVolumeSource-to_list}{\code{V1DownwardAPIVolumeSource$to_list()}}
\item \href{#method-V1DownwardAPIVolumeSource-clone}{\code{V1DownwardAPIVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIVolumeSource$new(default_mode = NULL, items = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1DownwardAPIVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1DownwardAPIVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1DownwardAPIVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
