% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EmptyDirVolumeSource.R
\name{V1EmptyDirVolumeSource}
\alias{V1EmptyDirVolumeSource}
\title{V1EmptyDirVolumeSource}
\description{
V1EmptyDirVolumeSource

V1EmptyDirVolumeSource
}
\details{
Represents an empty directory for a pod. Empty directory volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EmptyDirVolumeSource-new}{\code{V1EmptyDirVolumeSource$new()}}
\item \href{#method-V1EmptyDirVolumeSource-to_list}{\code{V1EmptyDirVolumeSource$to_list()}}
\item \href{#method-V1EmptyDirVolumeSource-clone}{\code{V1EmptyDirVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EmptyDirVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EmptyDirVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EmptyDirVolumeSource$new(medium = NULL, size_limit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EmptyDirVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EmptyDirVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EmptyDirVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EmptyDirVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EmptyDirVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EmptyDirVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
