% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EndpointAddress.R
\name{V1EndpointAddress}
\alias{V1EndpointAddress}
\title{V1EndpointAddress}
\description{
V1EndpointAddress

V1EndpointAddress
}
\details{
EndpointAddress is a tuple that describes single IP address.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EndpointAddress-new}{\code{V1EndpointAddress$new()}}
\item \href{#method-V1EndpointAddress-to_list}{\code{V1EndpointAddress$to_list()}}
\item \href{#method-V1EndpointAddress-clone}{\code{V1EndpointAddress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointAddress-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointAddress-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointAddress$new(
  hostname = NULL,
  ip = NULL,
  node_name = NULL,
  target_ref = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointAddress-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointAddress-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointAddress$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointAddress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointAddress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointAddress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
