% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EndpointConditions.R
\name{V1EndpointConditions}
\alias{V1EndpointConditions}
\title{V1EndpointConditions}
\description{
V1EndpointConditions

V1EndpointConditions
}
\details{
EndpointConditions represents the current condition of an endpoint.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EndpointConditions-new}{\code{V1EndpointConditions$new()}}
\item \href{#method-V1EndpointConditions-to_list}{\code{V1EndpointConditions$to_list()}}
\item \href{#method-V1EndpointConditions-clone}{\code{V1EndpointConditions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointConditions-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointConditions-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointConditions$new(ready = NULL, serving = NULL, terminating = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointConditions-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointConditions-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointConditions$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointConditions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointConditions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointConditions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
