% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EndpointHints.R
\name{V1EndpointHints}
\alias{V1EndpointHints}
\title{V1EndpointHints}
\description{
V1EndpointHints

V1EndpointHints
}
\details{
EndpointHints provides hints describing how an endpoint should be consumed.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EndpointHints-new}{\code{V1EndpointHints$new()}}
\item \href{#method-V1EndpointHints-to_list}{\code{V1EndpointHints$to_list()}}
\item \href{#method-V1EndpointHints-clone}{\code{V1EndpointHints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointHints-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointHints-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointHints$new(for_zones = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointHints-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointHints-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointHints$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointHints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointHints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointHints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
