% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Endpoints.R
\name{V1Endpoints}
\alias{V1Endpoints}
\title{V1Endpoints}
\description{
V1Endpoints

V1Endpoints
}
\details{
Endpoints is a collection of endpoints that implement the actual service. Example:

	 Name: "mysvc",
	 Subsets: 
	   {
	     Addresses: [{"ip": "10.10.1.1"}, {"ip": "10.10.2.2"},
	     Ports: {"name": "a", "port": 8675}, {"name": "b", "port": 309}
	   },
	   {
	     Addresses: {"ip": "10.10.3.3"},
	     Ports: {"name": "a", "port": 93}, {"name": "b", "port": 76}
	   },
	]
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Endpoints-new}{\code{V1Endpoints$new()}}
\item \href{#method-V1Endpoints-to_list}{\code{V1Endpoints$to_list()}}
\item \href{#method-V1Endpoints-clone}{\code{V1Endpoints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Endpoints-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Endpoints-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Endpoints$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  subsets = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Endpoints-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Endpoints-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Endpoints$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Endpoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Endpoints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Endpoints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
