% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EndpointSlice.R
\name{V1EndpointSlice}
\alias{V1EndpointSlice}
\title{V1EndpointSlice}
\description{
V1EndpointSlice

V1EndpointSlice
}
\details{
EndpointSlice represents a subset of the endpoints that implement a service. For a given service there may be multiple EndpointSlice objects, selected by labels, which must be joined to produce the full set of endpoints.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EndpointSlice-new}{\code{V1EndpointSlice$new()}}
\item \href{#method-V1EndpointSlice-to_list}{\code{V1EndpointSlice$to_list()}}
\item \href{#method-V1EndpointSlice-clone}{\code{V1EndpointSlice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSlice-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSlice-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSlice$new(
  address_type = NULL,
  api_version = NULL,
  endpoints = NULL,
  kind = NULL,
  metadata = NULL,
  ports = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSlice-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSlice-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSlice$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSlice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSlice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSlice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
