% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EndpointSliceList.R
\name{V1EndpointSliceList}
\alias{V1EndpointSliceList}
\title{V1EndpointSliceList}
\description{
V1EndpointSliceList

V1EndpointSliceList
}
\details{
EndpointSliceList represents a list of endpoint slices
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EndpointSliceList-new}{\code{V1EndpointSliceList$new()}}
\item \href{#method-V1EndpointSliceList-to_list}{\code{V1EndpointSliceList$to_list()}}
\item \href{#method-V1EndpointSliceList-clone}{\code{V1EndpointSliceList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSliceList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSliceList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSliceList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSliceList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSliceList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSliceList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSliceList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSliceList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSliceList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
