% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EndpointsList.R
\name{V1EndpointsList}
\alias{V1EndpointsList}
\title{V1EndpointsList}
\description{
V1EndpointsList

V1EndpointsList
}
\details{
EndpointsList is a list of endpoints.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EndpointsList-new}{\code{V1EndpointsList$new()}}
\item \href{#method-V1EndpointsList-to_list}{\code{V1EndpointsList$to_list()}}
\item \href{#method-V1EndpointsList-clone}{\code{V1EndpointsList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointsList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointsList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointsList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointsList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointsList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointsList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointsList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointsList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointsList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
