% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EndpointSubset.R
\name{V1EndpointSubset}
\alias{V1EndpointSubset}
\title{V1EndpointSubset}
\description{
V1EndpointSubset

V1EndpointSubset
}
\details{
EndpointSubset is a group of addresses with a common set of ports. The expanded set of endpoints is the Cartesian product of Addresses x Ports. For example, given:

	{
	  Addresses: {"ip": "10.10.1.1"}, {"ip": "10.10.2.2"},
	  Ports:     {"name": "a", "port": 8675}, {"name": "b", "port": 309}
	}

The resulting set of endpoints can be viewed as:

	a:  10.10.1.1:8675, 10.10.2.2:8675 ,
	b:  10.10.1.1:309, 10.10.2.2:309
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EndpointSubset-new}{\code{V1EndpointSubset$new()}}
\item \href{#method-V1EndpointSubset-to_list}{\code{V1EndpointSubset$to_list()}}
\item \href{#method-V1EndpointSubset-clone}{\code{V1EndpointSubset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSubset-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSubset-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSubset$new(
  addresses = NULL,
  not_ready_addresses = NULL,
  ports = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSubset-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSubset-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSubset$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EndpointSubset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EndpointSubset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EndpointSubset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
