% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EnvFromSource.R
\name{V1EnvFromSource}
\alias{V1EnvFromSource}
\title{V1EnvFromSource}
\description{
V1EnvFromSource

V1EnvFromSource
}
\details{
EnvFromSource represents the source of a set of ConfigMaps
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EnvFromSource-new}{\code{V1EnvFromSource$new()}}
\item \href{#method-V1EnvFromSource-to_list}{\code{V1EnvFromSource$to_list()}}
\item \href{#method-V1EnvFromSource-clone}{\code{V1EnvFromSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EnvFromSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EnvFromSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EnvFromSource$new(config_map_ref = NULL, prefix = NULL, secret_ref = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EnvFromSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EnvFromSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EnvFromSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EnvFromSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EnvFromSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EnvFromSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
