% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EnvVarSource.R
\name{V1EnvVarSource}
\alias{V1EnvVarSource}
\title{V1EnvVarSource}
\description{
V1EnvVarSource

V1EnvVarSource
}
\details{
EnvVarSource represents a source for the value of an EnvVar.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EnvVarSource-new}{\code{V1EnvVarSource$new()}}
\item \href{#method-V1EnvVarSource-to_list}{\code{V1EnvVarSource$to_list()}}
\item \href{#method-V1EnvVarSource-clone}{\code{V1EnvVarSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EnvVarSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EnvVarSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EnvVarSource$new(
  config_map_key_ref = NULL,
  field_ref = NULL,
  resource_field_ref = NULL,
  secret_key_ref = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EnvVarSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EnvVarSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EnvVarSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EnvVarSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EnvVarSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EnvVarSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
