% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EphemeralContainer.R
\name{V1EphemeralContainer}
\alias{V1EphemeralContainer}
\title{V1EphemeralContainer}
\description{
V1EphemeralContainer

V1EphemeralContainer
}
\details{
An EphemeralContainer is a temporary container that you may add to an existing Pod for user-initiated activities such as debugging. Ephemeral containers have no resource or scheduling guarantees, and they will not be restarted when they exit or when a Pod is removed or restarted. The kubelet may evict a Pod if an ephemeral container causes the Pod to exceed its resource allocation.

To add an ephemeral container, use the ephemeralcontainers subresource of an existing Pod. Ephemeral containers may not be removed or restarted.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EphemeralContainer-new}{\code{V1EphemeralContainer$new()}}
\item \href{#method-V1EphemeralContainer-to_list}{\code{V1EphemeralContainer$to_list()}}
\item \href{#method-V1EphemeralContainer-clone}{\code{V1EphemeralContainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EphemeralContainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EphemeralContainer-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EphemeralContainer$new(
  args = NULL,
  command = NULL,
  env = NULL,
  env_from = NULL,
  image = NULL,
  image_pull_policy = NULL,
  lifecycle = NULL,
  liveness_probe = NULL,
  name = NULL,
  ports = NULL,
  readiness_probe = NULL,
  resize_policy = NULL,
  resources = NULL,
  restart_policy = NULL,
  security_context = NULL,
  startup_probe = NULL,
  stdin = NULL,
  stdin_once = NULL,
  target_container_name = NULL,
  termination_message_path = NULL,
  termination_message_policy = NULL,
  tty = NULL,
  volume_devices = NULL,
  volume_mounts = NULL,
  working_dir = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EphemeralContainer-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EphemeralContainer-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EphemeralContainer$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EphemeralContainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EphemeralContainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EphemeralContainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
