% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EphemeralVolumeSource.R
\name{V1EphemeralVolumeSource}
\alias{V1EphemeralVolumeSource}
\title{V1EphemeralVolumeSource}
\description{
V1EphemeralVolumeSource

V1EphemeralVolumeSource
}
\details{
Represents an ephemeral volume that is handled by a normal storage driver.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EphemeralVolumeSource-new}{\code{V1EphemeralVolumeSource$new()}}
\item \href{#method-V1EphemeralVolumeSource-to_list}{\code{V1EphemeralVolumeSource$to_list()}}
\item \href{#method-V1EphemeralVolumeSource-clone}{\code{V1EphemeralVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EphemeralVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EphemeralVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EphemeralVolumeSource$new(volume_claim_template = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EphemeralVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EphemeralVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EphemeralVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EphemeralVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EphemeralVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EphemeralVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
