% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1EventSource.R
\name{V1EventSource}
\alias{V1EventSource}
\title{V1EventSource}
\description{
V1EventSource

V1EventSource
}
\details{
EventSource contains information for an event.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1EventSource-new}{\code{V1EventSource$new()}}
\item \href{#method-V1EventSource-to_list}{\code{V1EventSource$to_list()}}
\item \href{#method-V1EventSource-clone}{\code{V1EventSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EventSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1EventSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EventSource$new(component = NULL, host = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EventSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1EventSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EventSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1EventSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1EventSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1EventSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
