% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Eviction.R
\name{V1Eviction}
\alias{V1Eviction}
\title{V1Eviction}
\description{
V1Eviction

V1Eviction
}
\details{
Eviction evicts a pod from its node subject to certain policies and safety constraints. This is a subresource of Pod.  A request to cause such an eviction is created by POSTing to .../pods/<pod name>/evictions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Eviction-new}{\code{V1Eviction$new()}}
\item \href{#method-V1Eviction-to_list}{\code{V1Eviction$to_list()}}
\item \href{#method-V1Eviction-clone}{\code{V1Eviction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Eviction-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Eviction-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Eviction$new(
  api_version = NULL,
  delete_options = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Eviction-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Eviction-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Eviction$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Eviction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Eviction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Eviction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
