% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ExecAction.R
\name{V1ExecAction}
\alias{V1ExecAction}
\title{V1ExecAction}
\description{
V1ExecAction

V1ExecAction
}
\details{
ExecAction describes a "run in container" action.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ExecAction-new}{\code{V1ExecAction$new()}}
\item \href{#method-V1ExecAction-to_list}{\code{V1ExecAction$to_list()}}
\item \href{#method-V1ExecAction-clone}{\code{V1ExecAction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExecAction-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExecAction-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExecAction$new(command = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExecAction-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExecAction-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExecAction$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExecAction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExecAction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExecAction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
