% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ExemptPriorityLevelConfiguration.R
\name{V1ExemptPriorityLevelConfiguration}
\alias{V1ExemptPriorityLevelConfiguration}
\title{V1ExemptPriorityLevelConfiguration}
\description{
V1ExemptPriorityLevelConfiguration

V1ExemptPriorityLevelConfiguration
}
\details{
ExemptPriorityLevelConfiguration describes the configurable aspects of the handling of exempt requests. In the mandatory exempt configuration object the values in the fields here can be modified by authorized users, unlike the rest of the `spec`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ExemptPriorityLevelConfiguration-new}{\code{V1ExemptPriorityLevelConfiguration$new()}}
\item \href{#method-V1ExemptPriorityLevelConfiguration-to_list}{\code{V1ExemptPriorityLevelConfiguration$to_list()}}
\item \href{#method-V1ExemptPriorityLevelConfiguration-clone}{\code{V1ExemptPriorityLevelConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExemptPriorityLevelConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExemptPriorityLevelConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExemptPriorityLevelConfiguration$new(
  lendable_percent = NULL,
  nominal_concurrency_shares = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExemptPriorityLevelConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExemptPriorityLevelConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExemptPriorityLevelConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExemptPriorityLevelConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExemptPriorityLevelConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExemptPriorityLevelConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
