% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ExternalDocumentation.R
\name{V1ExternalDocumentation}
\alias{V1ExternalDocumentation}
\title{V1ExternalDocumentation}
\description{
V1ExternalDocumentation

V1ExternalDocumentation
}
\details{
ExternalDocumentation allows referencing an external resource for extended documentation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ExternalDocumentation-new}{\code{V1ExternalDocumentation$new()}}
\item \href{#method-V1ExternalDocumentation-to_list}{\code{V1ExternalDocumentation$to_list()}}
\item \href{#method-V1ExternalDocumentation-clone}{\code{V1ExternalDocumentation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExternalDocumentation-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExternalDocumentation-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExternalDocumentation$new(description = NULL, url = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExternalDocumentation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExternalDocumentation-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExternalDocumentation$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ExternalDocumentation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ExternalDocumentation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ExternalDocumentation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
