% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FCVolumeSource.R
\name{V1FCVolumeSource}
\alias{V1FCVolumeSource}
\title{V1FCVolumeSource}
\description{
V1FCVolumeSource

V1FCVolumeSource
}
\details{
Represents a Fibre Channel volume. Fibre Channel volumes can only be mounted as read/write once. Fibre Channel volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FCVolumeSource-new}{\code{V1FCVolumeSource$new()}}
\item \href{#method-V1FCVolumeSource-to_list}{\code{V1FCVolumeSource$to_list()}}
\item \href{#method-V1FCVolumeSource-clone}{\code{V1FCVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FCVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FCVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FCVolumeSource$new(
  fs_type = NULL,
  lun = NULL,
  read_only = NULL,
  target_ww_ns = NULL,
  wwids = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FCVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FCVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FCVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FCVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FCVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FCVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
