% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FieldSelectorRequirement.R
\name{V1FieldSelectorRequirement}
\alias{V1FieldSelectorRequirement}
\title{V1FieldSelectorRequirement}
\description{
V1FieldSelectorRequirement

V1FieldSelectorRequirement
}
\details{
FieldSelectorRequirement is a selector that contains values, a key, and an operator that relates the key and values.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FieldSelectorRequirement-new}{\code{V1FieldSelectorRequirement$new()}}
\item \href{#method-V1FieldSelectorRequirement-to_list}{\code{V1FieldSelectorRequirement$to_list()}}
\item \href{#method-V1FieldSelectorRequirement-clone}{\code{V1FieldSelectorRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FieldSelectorRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FieldSelectorRequirement-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FieldSelectorRequirement$new(key = NULL, operator = NULL, values = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FieldSelectorRequirement-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FieldSelectorRequirement-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FieldSelectorRequirement$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FieldSelectorRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FieldSelectorRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FieldSelectorRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
