% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FieldsV1.R
\name{V1FieldsV1}
\alias{V1FieldsV1}
\title{V1FieldsV1}
\description{
V1FieldsV1

V1FieldsV1
}
\details{
FieldsV1 stores a set of fields in a data structure like a Trie, in JSON format.

Each key is either a '.' representing the field itself, and will always map to an empty set, or a string representing a sub-field or item. The string will follow one of these four formats: 'f:<name>', where <name> is the name of a field in a struct, or key in a map 'v:<value>', where <value> is the exact json formatted value of a list item 'i:<index>', where <index> is position of a item in a list 'k:<keys>', where <keys> is a map of  a list item's key fields to their unique values If a key maps to an empty Fields value, the field that key represents is part of the set.

The exact format is defined in sigs.k8s.io/structured-merge-diff
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FieldsV1-new}{\code{V1FieldsV1$new()}}
\item \href{#method-V1FieldsV1-to_list}{\code{V1FieldsV1$to_list()}}
\item \href{#method-V1FieldsV1-clone}{\code{V1FieldsV1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FieldsV1-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FieldsV1-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FieldsV1$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FieldsV1-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FieldsV1-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FieldsV1$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FieldsV1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FieldsV1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FieldsV1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
