% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlexPersistentVolumeSource.R
\name{V1FlexPersistentVolumeSource}
\alias{V1FlexPersistentVolumeSource}
\title{V1FlexPersistentVolumeSource}
\description{
V1FlexPersistentVolumeSource

V1FlexPersistentVolumeSource
}
\details{
FlexPersistentVolumeSource represents a generic persistent volume resource that is provisioned/attached using an exec based plugin.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlexPersistentVolumeSource-new}{\code{V1FlexPersistentVolumeSource$new()}}
\item \href{#method-V1FlexPersistentVolumeSource-to_list}{\code{V1FlexPersistentVolumeSource$to_list()}}
\item \href{#method-V1FlexPersistentVolumeSource-clone}{\code{V1FlexPersistentVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlexPersistentVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlexPersistentVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlexPersistentVolumeSource$new(
  driver = NULL,
  fs_type = NULL,
  options = NULL,
  read_only = NULL,
  secret_ref = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlexPersistentVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlexPersistentVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlexPersistentVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlexPersistentVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlexPersistentVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlexPersistentVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
