% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlexVolumeSource.R
\name{V1FlexVolumeSource}
\alias{V1FlexVolumeSource}
\title{V1FlexVolumeSource}
\description{
V1FlexVolumeSource

V1FlexVolumeSource
}
\details{
FlexVolume represents a generic volume resource that is provisioned/attached using an exec based plugin.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlexVolumeSource-new}{\code{V1FlexVolumeSource$new()}}
\item \href{#method-V1FlexVolumeSource-to_list}{\code{V1FlexVolumeSource$to_list()}}
\item \href{#method-V1FlexVolumeSource-clone}{\code{V1FlexVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlexVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlexVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlexVolumeSource$new(
  driver = NULL,
  fs_type = NULL,
  options = NULL,
  read_only = NULL,
  secret_ref = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlexVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlexVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlexVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlexVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlexVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlexVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
