% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlockerVolumeSource.R
\name{V1FlockerVolumeSource}
\alias{V1FlockerVolumeSource}
\title{V1FlockerVolumeSource}
\description{
V1FlockerVolumeSource

V1FlockerVolumeSource
}
\details{
Represents a Flocker volume mounted by the Flocker agent. One and only one of datasetName and datasetUUID should be set. Flocker volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlockerVolumeSource-new}{\code{V1FlockerVolumeSource$new()}}
\item \href{#method-V1FlockerVolumeSource-to_list}{\code{V1FlockerVolumeSource$to_list()}}
\item \href{#method-V1FlockerVolumeSource-clone}{\code{V1FlockerVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlockerVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlockerVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlockerVolumeSource$new(dataset_name = NULL, dataset_uuid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlockerVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlockerVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlockerVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlockerVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlockerVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlockerVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
