% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlowDistinguisherMethod.R
\name{V1FlowDistinguisherMethod}
\alias{V1FlowDistinguisherMethod}
\title{V1FlowDistinguisherMethod}
\description{
V1FlowDistinguisherMethod

V1FlowDistinguisherMethod
}
\details{
FlowDistinguisherMethod specifies the method of a flow distinguisher.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlowDistinguisherMethod-new}{\code{V1FlowDistinguisherMethod$new()}}
\item \href{#method-V1FlowDistinguisherMethod-to_list}{\code{V1FlowDistinguisherMethod$to_list()}}
\item \href{#method-V1FlowDistinguisherMethod-clone}{\code{V1FlowDistinguisherMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowDistinguisherMethod-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowDistinguisherMethod-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowDistinguisherMethod$new(type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowDistinguisherMethod-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowDistinguisherMethod-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowDistinguisherMethod$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowDistinguisherMethod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowDistinguisherMethod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowDistinguisherMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
