% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlowSchema.R
\name{V1FlowSchema}
\alias{V1FlowSchema}
\title{V1FlowSchema}
\description{
V1FlowSchema

V1FlowSchema
}
\details{
FlowSchema defines the schema of a group of flows. Note that a flow is made up of a set of inbound API requests with similar attributes and is identified by a pair of strings: the name of the FlowSchema and a "flow distinguisher".
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlowSchema-new}{\code{V1FlowSchema$new()}}
\item \href{#method-V1FlowSchema-to_list}{\code{V1FlowSchema$to_list()}}
\item \href{#method-V1FlowSchema-clone}{\code{V1FlowSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchema-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchema$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchema-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchema-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchema$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
