% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlowSchemaCondition.R
\name{V1FlowSchemaCondition}
\alias{V1FlowSchemaCondition}
\title{V1FlowSchemaCondition}
\description{
V1FlowSchemaCondition

V1FlowSchemaCondition
}
\details{
FlowSchemaCondition describes conditions for a FlowSchema.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlowSchemaCondition-new}{\code{V1FlowSchemaCondition$new()}}
\item \href{#method-V1FlowSchemaCondition-to_list}{\code{V1FlowSchemaCondition$to_list()}}
\item \href{#method-V1FlowSchemaCondition-clone}{\code{V1FlowSchemaCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
