% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlowSchemaList.R
\name{V1FlowSchemaList}
\alias{V1FlowSchemaList}
\title{V1FlowSchemaList}
\description{
V1FlowSchemaList

V1FlowSchemaList
}
\details{
FlowSchemaList is a list of FlowSchema objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlowSchemaList-new}{\code{V1FlowSchemaList$new()}}
\item \href{#method-V1FlowSchemaList-to_list}{\code{V1FlowSchemaList$to_list()}}
\item \href{#method-V1FlowSchemaList-clone}{\code{V1FlowSchemaList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
