% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1FlowSchemaStatus.R
\name{V1FlowSchemaStatus}
\alias{V1FlowSchemaStatus}
\title{V1FlowSchemaStatus}
\description{
V1FlowSchemaStatus

V1FlowSchemaStatus
}
\details{
FlowSchemaStatus represents the current state of a FlowSchema.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1FlowSchemaStatus-new}{\code{V1FlowSchemaStatus$new()}}
\item \href{#method-V1FlowSchemaStatus-to_list}{\code{V1FlowSchemaStatus$to_list()}}
\item \href{#method-V1FlowSchemaStatus-clone}{\code{V1FlowSchemaStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaStatus$new(conditions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1FlowSchemaStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1FlowSchemaStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1FlowSchemaStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
