% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ForZone.R
\name{V1ForZone}
\alias{V1ForZone}
\title{V1ForZone}
\description{
V1ForZone

V1ForZone
}
\details{
ForZone provides information about which zones should consume this endpoint.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ForZone-new}{\code{V1ForZone$new()}}
\item \href{#method-V1ForZone-to_list}{\code{V1ForZone$to_list()}}
\item \href{#method-V1ForZone-clone}{\code{V1ForZone$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ForZone-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ForZone-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ForZone$new(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ForZone-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ForZone-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ForZone$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ForZone-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ForZone-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ForZone$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
