% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1GCEPersistentDiskVolumeSource.R
\name{V1GCEPersistentDiskVolumeSource}
\alias{V1GCEPersistentDiskVolumeSource}
\title{V1GCEPersistentDiskVolumeSource}
\description{
V1GCEPersistentDiskVolumeSource

V1GCEPersistentDiskVolumeSource
}
\details{
Represents a Persistent Disk resource in Google Compute Engine.

A GCE PD must exist before mounting to a container. The disk must also be in the same GCE project and zone as the kubelet. A GCE PD can only be mounted as read/write once or read-only many times. GCE PDs support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1GCEPersistentDiskVolumeSource-new}{\code{V1GCEPersistentDiskVolumeSource$new()}}
\item \href{#method-V1GCEPersistentDiskVolumeSource-to_list}{\code{V1GCEPersistentDiskVolumeSource$to_list()}}
\item \href{#method-V1GCEPersistentDiskVolumeSource-clone}{\code{V1GCEPersistentDiskVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GCEPersistentDiskVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1GCEPersistentDiskVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GCEPersistentDiskVolumeSource$new(
  fs_type = NULL,
  partition = NULL,
  pd_name = NULL,
  read_only = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GCEPersistentDiskVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1GCEPersistentDiskVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GCEPersistentDiskVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GCEPersistentDiskVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1GCEPersistentDiskVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GCEPersistentDiskVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
