% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1GitRepoVolumeSource.R
\name{V1GitRepoVolumeSource}
\alias{V1GitRepoVolumeSource}
\title{V1GitRepoVolumeSource}
\description{
V1GitRepoVolumeSource

V1GitRepoVolumeSource
}
\details{
Represents a volume that is populated with the contents of a git repository. Git repo volumes do not support ownership management. Git repo volumes support SELinux relabeling.

DEPRECATED: GitRepo is deprecated. To provision a container with a git repo, mount an EmptyDir into an InitContainer that clones the repo using git, then mount the EmptyDir into the Pod's container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1GitRepoVolumeSource-new}{\code{V1GitRepoVolumeSource$new()}}
\item \href{#method-V1GitRepoVolumeSource-to_list}{\code{V1GitRepoVolumeSource$to_list()}}
\item \href{#method-V1GitRepoVolumeSource-clone}{\code{V1GitRepoVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GitRepoVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1GitRepoVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GitRepoVolumeSource$new(directory = NULL, repository = NULL, revision = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GitRepoVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1GitRepoVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GitRepoVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GitRepoVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1GitRepoVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GitRepoVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
