% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1GlusterfsPersistentVolumeSource.R
\name{V1GlusterfsPersistentVolumeSource}
\alias{V1GlusterfsPersistentVolumeSource}
\title{V1GlusterfsPersistentVolumeSource}
\description{
V1GlusterfsPersistentVolumeSource

V1GlusterfsPersistentVolumeSource
}
\details{
Represents a Glusterfs mount that lasts the lifetime of a pod. Glusterfs volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1GlusterfsPersistentVolumeSource-new}{\code{V1GlusterfsPersistentVolumeSource$new()}}
\item \href{#method-V1GlusterfsPersistentVolumeSource-to_list}{\code{V1GlusterfsPersistentVolumeSource$to_list()}}
\item \href{#method-V1GlusterfsPersistentVolumeSource-clone}{\code{V1GlusterfsPersistentVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GlusterfsPersistentVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1GlusterfsPersistentVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GlusterfsPersistentVolumeSource$new(
  endpoints = NULL,
  endpoints_namespace = NULL,
  path = NULL,
  read_only = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GlusterfsPersistentVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1GlusterfsPersistentVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GlusterfsPersistentVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GlusterfsPersistentVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1GlusterfsPersistentVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GlusterfsPersistentVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
