% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1GlusterfsVolumeSource.R
\name{V1GlusterfsVolumeSource}
\alias{V1GlusterfsVolumeSource}
\title{V1GlusterfsVolumeSource}
\description{
V1GlusterfsVolumeSource

V1GlusterfsVolumeSource
}
\details{
Represents a Glusterfs mount that lasts the lifetime of a pod. Glusterfs volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1GlusterfsVolumeSource-new}{\code{V1GlusterfsVolumeSource$new()}}
\item \href{#method-V1GlusterfsVolumeSource-to_list}{\code{V1GlusterfsVolumeSource$to_list()}}
\item \href{#method-V1GlusterfsVolumeSource-clone}{\code{V1GlusterfsVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GlusterfsVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1GlusterfsVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GlusterfsVolumeSource$new(endpoints = NULL, path = NULL, read_only = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GlusterfsVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1GlusterfsVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GlusterfsVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GlusterfsVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1GlusterfsVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GlusterfsVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
