% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1GroupSubject.R
\name{V1GroupSubject}
\alias{V1GroupSubject}
\title{V1GroupSubject}
\description{
V1GroupSubject

V1GroupSubject
}
\details{
GroupSubject holds detailed information for group-kind subject.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1GroupSubject-new}{\code{V1GroupSubject$new()}}
\item \href{#method-V1GroupSubject-to_list}{\code{V1GroupSubject$to_list()}}
\item \href{#method-V1GroupSubject-clone}{\code{V1GroupSubject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GroupSubject-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1GroupSubject-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GroupSubject$new(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GroupSubject-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1GroupSubject-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GroupSubject$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GroupSubject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1GroupSubject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GroupSubject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
