% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1GroupVersionForDiscovery.R
\name{V1GroupVersionForDiscovery}
\alias{V1GroupVersionForDiscovery}
\title{V1GroupVersionForDiscovery}
\description{
V1GroupVersionForDiscovery

V1GroupVersionForDiscovery
}
\details{
GroupVersion contains the "group/version" and "version" string of a version. It is made a struct to keep extensibility.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1GroupVersionForDiscovery-new}{\code{V1GroupVersionForDiscovery$new()}}
\item \href{#method-V1GroupVersionForDiscovery-to_list}{\code{V1GroupVersionForDiscovery$to_list()}}
\item \href{#method-V1GroupVersionForDiscovery-clone}{\code{V1GroupVersionForDiscovery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GroupVersionForDiscovery-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1GroupVersionForDiscovery-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GroupVersionForDiscovery$new(group_version = NULL, version = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GroupVersionForDiscovery-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1GroupVersionForDiscovery-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GroupVersionForDiscovery$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GroupVersionForDiscovery-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1GroupVersionForDiscovery-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GroupVersionForDiscovery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
