% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1GRPCAction.R
\name{V1GRPCAction}
\alias{V1GRPCAction}
\title{V1GRPCAction}
\description{
V1GRPCAction

V1GRPCAction
}
\details{
(no description)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1GRPCAction-new}{\code{V1GRPCAction$new()}}
\item \href{#method-V1GRPCAction-to_list}{\code{V1GRPCAction$to_list()}}
\item \href{#method-V1GRPCAction-clone}{\code{V1GRPCAction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GRPCAction-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1GRPCAction-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GRPCAction$new(port = NULL, service = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GRPCAction-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1GRPCAction-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GRPCAction$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1GRPCAction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1GRPCAction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1GRPCAction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
