% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HorizontalPodAutoscaler.R
\name{V1HorizontalPodAutoscaler}
\alias{V1HorizontalPodAutoscaler}
\title{V1HorizontalPodAutoscaler}
\description{
V1HorizontalPodAutoscaler

V1HorizontalPodAutoscaler
}
\details{
configuration of a horizontal pod autoscaler.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HorizontalPodAutoscaler-new}{\code{V1HorizontalPodAutoscaler$new()}}
\item \href{#method-V1HorizontalPodAutoscaler-to_list}{\code{V1HorizontalPodAutoscaler$to_list()}}
\item \href{#method-V1HorizontalPodAutoscaler-clone}{\code{V1HorizontalPodAutoscaler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscaler-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscaler-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscaler$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscaler-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscaler-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscaler$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscaler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscaler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
