% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HorizontalPodAutoscalerList.R
\name{V1HorizontalPodAutoscalerList}
\alias{V1HorizontalPodAutoscalerList}
\title{V1HorizontalPodAutoscalerList}
\description{
V1HorizontalPodAutoscalerList

V1HorizontalPodAutoscalerList
}
\details{
list of horizontal pod autoscaler objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HorizontalPodAutoscalerList-new}{\code{V1HorizontalPodAutoscalerList$new()}}
\item \href{#method-V1HorizontalPodAutoscalerList-to_list}{\code{V1HorizontalPodAutoscalerList$to_list()}}
\item \href{#method-V1HorizontalPodAutoscalerList-clone}{\code{V1HorizontalPodAutoscalerList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
