% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HorizontalPodAutoscalerSpec.R
\name{V1HorizontalPodAutoscalerSpec}
\alias{V1HorizontalPodAutoscalerSpec}
\title{V1HorizontalPodAutoscalerSpec}
\description{
V1HorizontalPodAutoscalerSpec

V1HorizontalPodAutoscalerSpec
}
\details{
specification of a horizontal pod autoscaler.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HorizontalPodAutoscalerSpec-new}{\code{V1HorizontalPodAutoscalerSpec$new()}}
\item \href{#method-V1HorizontalPodAutoscalerSpec-to_list}{\code{V1HorizontalPodAutoscalerSpec$to_list()}}
\item \href{#method-V1HorizontalPodAutoscalerSpec-clone}{\code{V1HorizontalPodAutoscalerSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerSpec$new(
  max_replicas = NULL,
  min_replicas = NULL,
  scale_target_ref = NULL,
  target_cpu_utilization_percentage = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
