% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HorizontalPodAutoscalerStatus.R
\name{V1HorizontalPodAutoscalerStatus}
\alias{V1HorizontalPodAutoscalerStatus}
\title{V1HorizontalPodAutoscalerStatus}
\description{
V1HorizontalPodAutoscalerStatus

V1HorizontalPodAutoscalerStatus
}
\details{
current status of a horizontal pod autoscaler
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HorizontalPodAutoscalerStatus-new}{\code{V1HorizontalPodAutoscalerStatus$new()}}
\item \href{#method-V1HorizontalPodAutoscalerStatus-to_list}{\code{V1HorizontalPodAutoscalerStatus$to_list()}}
\item \href{#method-V1HorizontalPodAutoscalerStatus-clone}{\code{V1HorizontalPodAutoscalerStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerStatus$new(
  current_cpu_utilization_percentage = NULL,
  current_replicas = NULL,
  desired_replicas = NULL,
  last_scale_time = NULL,
  observed_generation = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HorizontalPodAutoscalerStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HorizontalPodAutoscalerStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HorizontalPodAutoscalerStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
