% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HostAlias.R
\name{V1HostAlias}
\alias{V1HostAlias}
\title{V1HostAlias}
\description{
V1HostAlias

V1HostAlias
}
\details{
HostAlias holds the mapping between IP and hostnames that will be injected as an entry in the pod's hosts file.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HostAlias-new}{\code{V1HostAlias$new()}}
\item \href{#method-V1HostAlias-to_list}{\code{V1HostAlias$to_list()}}
\item \href{#method-V1HostAlias-clone}{\code{V1HostAlias$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostAlias-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostAlias-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostAlias$new(hostnames = NULL, ip = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostAlias-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostAlias-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostAlias$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostAlias-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostAlias-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostAlias$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
