% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HostIP.R
\name{V1HostIP}
\alias{V1HostIP}
\title{V1HostIP}
\description{
V1HostIP

V1HostIP
}
\details{
HostIP represents a single IP address allocated to the host.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HostIP-new}{\code{V1HostIP$new()}}
\item \href{#method-V1HostIP-to_list}{\code{V1HostIP$to_list()}}
\item \href{#method-V1HostIP-clone}{\code{V1HostIP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostIP-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostIP-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostIP$new(ip = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostIP-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostIP-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostIP$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostIP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostIP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostIP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
