% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HostPathVolumeSource.R
\name{V1HostPathVolumeSource}
\alias{V1HostPathVolumeSource}
\title{V1HostPathVolumeSource}
\description{
V1HostPathVolumeSource

V1HostPathVolumeSource
}
\details{
Represents a host path mapped into a pod. Host path volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HostPathVolumeSource-new}{\code{V1HostPathVolumeSource$new()}}
\item \href{#method-V1HostPathVolumeSource-to_list}{\code{V1HostPathVolumeSource$to_list()}}
\item \href{#method-V1HostPathVolumeSource-clone}{\code{V1HostPathVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostPathVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostPathVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostPathVolumeSource$new(path = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostPathVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostPathVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostPathVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HostPathVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HostPathVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HostPathVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
