% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HTTPGetAction.R
\name{V1HTTPGetAction}
\alias{V1HTTPGetAction}
\title{V1HTTPGetAction}
\description{
V1HTTPGetAction

V1HTTPGetAction
}
\details{
HTTPGetAction describes an action based on HTTP Get requests.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HTTPGetAction-new}{\code{V1HTTPGetAction$new()}}
\item \href{#method-V1HTTPGetAction-to_list}{\code{V1HTTPGetAction$to_list()}}
\item \href{#method-V1HTTPGetAction-clone}{\code{V1HTTPGetAction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPGetAction-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPGetAction-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPGetAction$new(
  host = NULL,
  http_headers = NULL,
  path = NULL,
  port = NULL,
  scheme = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPGetAction-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPGetAction-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPGetAction$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPGetAction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPGetAction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPGetAction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
