% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HTTPHeader.R
\name{V1HTTPHeader}
\alias{V1HTTPHeader}
\title{V1HTTPHeader}
\description{
V1HTTPHeader

V1HTTPHeader
}
\details{
HTTPHeader describes a custom header to be used in HTTP probes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HTTPHeader-new}{\code{V1HTTPHeader$new()}}
\item \href{#method-V1HTTPHeader-to_list}{\code{V1HTTPHeader$to_list()}}
\item \href{#method-V1HTTPHeader-clone}{\code{V1HTTPHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPHeader-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPHeader$new(name = NULL, value = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPHeader-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPHeader-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPHeader$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
