% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HTTPIngressPath.R
\name{V1HTTPIngressPath}
\alias{V1HTTPIngressPath}
\title{V1HTTPIngressPath}
\description{
V1HTTPIngressPath

V1HTTPIngressPath
}
\details{
HTTPIngressPath associates a path with a backend. Incoming urls matching the path are forwarded to the backend.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HTTPIngressPath-new}{\code{V1HTTPIngressPath$new()}}
\item \href{#method-V1HTTPIngressPath-to_list}{\code{V1HTTPIngressPath$to_list()}}
\item \href{#method-V1HTTPIngressPath-clone}{\code{V1HTTPIngressPath$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPIngressPath-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPIngressPath-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPIngressPath$new(backend = NULL, path = NULL, path_type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPIngressPath-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPIngressPath-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPIngressPath$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPIngressPath-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPIngressPath-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPIngressPath$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
