% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1HTTPIngressRuleValue.R
\name{V1HTTPIngressRuleValue}
\alias{V1HTTPIngressRuleValue}
\title{V1HTTPIngressRuleValue}
\description{
V1HTTPIngressRuleValue

V1HTTPIngressRuleValue
}
\details{
HTTPIngressRuleValue is a list of http selectors pointing to backends. In the example: http://<host>/<path>?<searchpart> -> backend where where parts of the url correspond to RFC 3986, this resource will be used to match against everything after the last '/' and before the first '?' or '#'.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1HTTPIngressRuleValue-new}{\code{V1HTTPIngressRuleValue$new()}}
\item \href{#method-V1HTTPIngressRuleValue-to_list}{\code{V1HTTPIngressRuleValue$to_list()}}
\item \href{#method-V1HTTPIngressRuleValue-clone}{\code{V1HTTPIngressRuleValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPIngressRuleValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPIngressRuleValue-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPIngressRuleValue$new(paths = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPIngressRuleValue-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPIngressRuleValue-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPIngressRuleValue$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1HTTPIngressRuleValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1HTTPIngressRuleValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1HTTPIngressRuleValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
