% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ImageVolumeSource.R
\name{V1ImageVolumeSource}
\alias{V1ImageVolumeSource}
\title{V1ImageVolumeSource}
\description{
V1ImageVolumeSource

V1ImageVolumeSource
}
\details{
ImageVolumeSource represents a image volume resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ImageVolumeSource-new}{\code{V1ImageVolumeSource$new()}}
\item \href{#method-V1ImageVolumeSource-to_list}{\code{V1ImageVolumeSource$to_list()}}
\item \href{#method-V1ImageVolumeSource-clone}{\code{V1ImageVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ImageVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ImageVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ImageVolumeSource$new(pull_policy = NULL, reference = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ImageVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ImageVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ImageVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ImageVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ImageVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ImageVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
