% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Ingress.R
\name{V1Ingress}
\alias{V1Ingress}
\title{V1Ingress}
\description{
V1Ingress

V1Ingress
}
\details{
Ingress is a collection of rules that allow inbound connections to reach the endpoints defined by a backend. An Ingress can be configured to give services externally-reachable urls, load balance traffic, terminate SSL, offer name based virtual hosting etc.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Ingress-new}{\code{V1Ingress$new()}}
\item \href{#method-V1Ingress-to_list}{\code{V1Ingress$to_list()}}
\item \href{#method-V1Ingress-clone}{\code{V1Ingress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Ingress-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Ingress-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Ingress$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Ingress-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Ingress-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Ingress$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Ingress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Ingress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Ingress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
