% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressClassList.R
\name{V1IngressClassList}
\alias{V1IngressClassList}
\title{V1IngressClassList}
\description{
V1IngressClassList

V1IngressClassList
}
\details{
IngressClassList is a collection of IngressClasses.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressClassList-new}{\code{V1IngressClassList$new()}}
\item \href{#method-V1IngressClassList-to_list}{\code{V1IngressClassList$to_list()}}
\item \href{#method-V1IngressClassList-clone}{\code{V1IngressClassList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
