% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressClassParametersReference.R
\name{V1IngressClassParametersReference}
\alias{V1IngressClassParametersReference}
\title{V1IngressClassParametersReference}
\description{
V1IngressClassParametersReference

V1IngressClassParametersReference
}
\details{
IngressClassParametersReference identifies an API object. This can be used to specify a cluster or namespace-scoped resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressClassParametersReference-new}{\code{V1IngressClassParametersReference$new()}}
\item \href{#method-V1IngressClassParametersReference-to_list}{\code{V1IngressClassParametersReference$to_list()}}
\item \href{#method-V1IngressClassParametersReference-clone}{\code{V1IngressClassParametersReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassParametersReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassParametersReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassParametersReference$new(
  api_group = NULL,
  kind = NULL,
  name = NULL,
  namespace = NULL,
  scope = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassParametersReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassParametersReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassParametersReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassParametersReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassParametersReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassParametersReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
