% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressClassSpec.R
\name{V1IngressClassSpec}
\alias{V1IngressClassSpec}
\title{V1IngressClassSpec}
\description{
V1IngressClassSpec

V1IngressClassSpec
}
\details{
IngressClassSpec provides information about the class of an Ingress.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressClassSpec-new}{\code{V1IngressClassSpec$new()}}
\item \href{#method-V1IngressClassSpec-to_list}{\code{V1IngressClassSpec$to_list()}}
\item \href{#method-V1IngressClassSpec-clone}{\code{V1IngressClassSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassSpec$new(controller = NULL, parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressClassSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressClassSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressClassSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
