% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressLoadBalancerIngress.R
\name{V1IngressLoadBalancerIngress}
\alias{V1IngressLoadBalancerIngress}
\title{V1IngressLoadBalancerIngress}
\description{
V1IngressLoadBalancerIngress

V1IngressLoadBalancerIngress
}
\details{
IngressLoadBalancerIngress represents the status of a load-balancer ingress point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressLoadBalancerIngress-new}{\code{V1IngressLoadBalancerIngress$new()}}
\item \href{#method-V1IngressLoadBalancerIngress-to_list}{\code{V1IngressLoadBalancerIngress$to_list()}}
\item \href{#method-V1IngressLoadBalancerIngress-clone}{\code{V1IngressLoadBalancerIngress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressLoadBalancerIngress-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressLoadBalancerIngress-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressLoadBalancerIngress$new(hostname = NULL, ip = NULL, ports = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressLoadBalancerIngress-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressLoadBalancerIngress-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressLoadBalancerIngress$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressLoadBalancerIngress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressLoadBalancerIngress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressLoadBalancerIngress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
