% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressLoadBalancerStatus.R
\name{V1IngressLoadBalancerStatus}
\alias{V1IngressLoadBalancerStatus}
\title{V1IngressLoadBalancerStatus}
\description{
V1IngressLoadBalancerStatus

V1IngressLoadBalancerStatus
}
\details{
IngressLoadBalancerStatus represents the status of a load-balancer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressLoadBalancerStatus-new}{\code{V1IngressLoadBalancerStatus$new()}}
\item \href{#method-V1IngressLoadBalancerStatus-to_list}{\code{V1IngressLoadBalancerStatus$to_list()}}
\item \href{#method-V1IngressLoadBalancerStatus-clone}{\code{V1IngressLoadBalancerStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressLoadBalancerStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressLoadBalancerStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressLoadBalancerStatus$new(ingress = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressLoadBalancerStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressLoadBalancerStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressLoadBalancerStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressLoadBalancerStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressLoadBalancerStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressLoadBalancerStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
