% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressPortStatus.R
\name{V1IngressPortStatus}
\alias{V1IngressPortStatus}
\title{V1IngressPortStatus}
\description{
V1IngressPortStatus

V1IngressPortStatus
}
\details{
IngressPortStatus represents the error condition of a service port
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressPortStatus-new}{\code{V1IngressPortStatus$new()}}
\item \href{#method-V1IngressPortStatus-to_list}{\code{V1IngressPortStatus$to_list()}}
\item \href{#method-V1IngressPortStatus-clone}{\code{V1IngressPortStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressPortStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressPortStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressPortStatus$new(error = NULL, port = NULL, protocol = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressPortStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressPortStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressPortStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressPortStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressPortStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressPortStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
