% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressRule.R
\name{V1IngressRule}
\alias{V1IngressRule}
\title{V1IngressRule}
\description{
V1IngressRule

V1IngressRule
}
\details{
IngressRule represents the rules mapping the paths under a specified host to the related backend services. Incoming requests are first evaluated for a host match, then routed to the backend associated with the matching IngressRuleValue.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressRule-new}{\code{V1IngressRule$new()}}
\item \href{#method-V1IngressRule-to_list}{\code{V1IngressRule$to_list()}}
\item \href{#method-V1IngressRule-clone}{\code{V1IngressRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressRule$new(host = NULL, http = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
