% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressServiceBackend.R
\name{V1IngressServiceBackend}
\alias{V1IngressServiceBackend}
\title{V1IngressServiceBackend}
\description{
V1IngressServiceBackend

V1IngressServiceBackend
}
\details{
IngressServiceBackend references a Kubernetes Service as a Backend.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressServiceBackend-new}{\code{V1IngressServiceBackend$new()}}
\item \href{#method-V1IngressServiceBackend-to_list}{\code{V1IngressServiceBackend$to_list()}}
\item \href{#method-V1IngressServiceBackend-clone}{\code{V1IngressServiceBackend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressServiceBackend-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressServiceBackend-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressServiceBackend$new(name = NULL, port = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressServiceBackend-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressServiceBackend-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressServiceBackend$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressServiceBackend-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressServiceBackend-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressServiceBackend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
