% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressSpec.R
\name{V1IngressSpec}
\alias{V1IngressSpec}
\title{V1IngressSpec}
\description{
V1IngressSpec

V1IngressSpec
}
\details{
IngressSpec describes the Ingress the user wishes to exist.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressSpec-new}{\code{V1IngressSpec$new()}}
\item \href{#method-V1IngressSpec-to_list}{\code{V1IngressSpec$to_list()}}
\item \href{#method-V1IngressSpec-clone}{\code{V1IngressSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressSpec$new(
  default_backend = NULL,
  ingress_class_name = NULL,
  rules = NULL,
  tls = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
