% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressStatus.R
\name{V1IngressStatus}
\alias{V1IngressStatus}
\title{V1IngressStatus}
\description{
V1IngressStatus

V1IngressStatus
}
\details{
IngressStatus describe the current state of the Ingress.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressStatus-new}{\code{V1IngressStatus$new()}}
\item \href{#method-V1IngressStatus-to_list}{\code{V1IngressStatus$to_list()}}
\item \href{#method-V1IngressStatus-clone}{\code{V1IngressStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressStatus$new(load_balancer = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
