% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IngressTLS.R
\name{V1IngressTLS}
\alias{V1IngressTLS}
\title{V1IngressTLS}
\description{
V1IngressTLS

V1IngressTLS
}
\details{
IngressTLS describes the transport layer security associated with an ingress.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IngressTLS-new}{\code{V1IngressTLS$new()}}
\item \href{#method-V1IngressTLS-to_list}{\code{V1IngressTLS$to_list()}}
\item \href{#method-V1IngressTLS-clone}{\code{V1IngressTLS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressTLS-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressTLS-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressTLS$new(hosts = NULL, secret_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressTLS-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressTLS-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressTLS$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IngressTLS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IngressTLS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IngressTLS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
