% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1IPBlock.R
\name{V1IPBlock}
\alias{V1IPBlock}
\title{V1IPBlock}
\description{
V1IPBlock

V1IPBlock
}
\details{
IPBlock describes a particular CIDR (Ex. "192.168.1.0/24","2001:db8::/64") that is allowed to the pods matched by a NetworkPolicySpec's podSelector. The except entry describes CIDRs that should not be included within this rule.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1IPBlock-new}{\code{V1IPBlock$new()}}
\item \href{#method-V1IPBlock-to_list}{\code{V1IPBlock$to_list()}}
\item \href{#method-V1IPBlock-clone}{\code{V1IPBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IPBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1IPBlock-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IPBlock$new(cidr = NULL, except = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IPBlock-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1IPBlock-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IPBlock$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1IPBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1IPBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1IPBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
