% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ISCSIVolumeSource.R
\name{V1ISCSIVolumeSource}
\alias{V1ISCSIVolumeSource}
\title{V1ISCSIVolumeSource}
\description{
V1ISCSIVolumeSource

V1ISCSIVolumeSource
}
\details{
Represents an ISCSI disk. ISCSI volumes can only be mounted as read/write once. ISCSI volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ISCSIVolumeSource-new}{\code{V1ISCSIVolumeSource$new()}}
\item \href{#method-V1ISCSIVolumeSource-to_list}{\code{V1ISCSIVolumeSource$to_list()}}
\item \href{#method-V1ISCSIVolumeSource-clone}{\code{V1ISCSIVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ISCSIVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ISCSIVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ISCSIVolumeSource$new(
  chap_auth_discovery = NULL,
  chap_auth_session = NULL,
  fs_type = NULL,
  initiator_name = NULL,
  iqn = NULL,
  iscsi_interface = NULL,
  lun = NULL,
  portals = NULL,
  read_only = NULL,
  secret_ref = NULL,
  target_portal = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ISCSIVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ISCSIVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ISCSIVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ISCSIVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ISCSIVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ISCSIVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
