% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Job.R
\name{V1Job}
\alias{V1Job}
\title{V1Job}
\description{
V1Job

V1Job
}
\details{
Job represents the configuration of a single job.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Job-new}{\code{V1Job$new()}}
\item \href{#method-V1Job-to_list}{\code{V1Job$to_list()}}
\item \href{#method-V1Job-clone}{\code{V1Job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Job-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Job-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Job$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Job-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Job-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Job$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Job-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Job-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
