% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JobCondition.R
\name{V1JobCondition}
\alias{V1JobCondition}
\title{V1JobCondition}
\description{
V1JobCondition

V1JobCondition
}
\details{
JobCondition describes current state of a job.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JobCondition-new}{\code{V1JobCondition$new()}}
\item \href{#method-V1JobCondition-to_list}{\code{V1JobCondition$to_list()}}
\item \href{#method-V1JobCondition-clone}{\code{V1JobCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobCondition$new(
  last_probe_time = NULL,
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
