% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JobList.R
\name{V1JobList}
\alias{V1JobList}
\title{V1JobList}
\description{
V1JobList

V1JobList
}
\details{
JobList is a collection of jobs.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JobList-new}{\code{V1JobList$new()}}
\item \href{#method-V1JobList-to_list}{\code{V1JobList$to_list()}}
\item \href{#method-V1JobList-clone}{\code{V1JobList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobList$new(api_version = NULL, items = NULL, kind = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
