% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JobSpec.R
\name{V1JobSpec}
\alias{V1JobSpec}
\title{V1JobSpec}
\description{
V1JobSpec

V1JobSpec
}
\details{
JobSpec describes how the job execution will look like.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JobSpec-new}{\code{V1JobSpec$new()}}
\item \href{#method-V1JobSpec-to_list}{\code{V1JobSpec$to_list()}}
\item \href{#method-V1JobSpec-clone}{\code{V1JobSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobSpec$new(
  active_deadline_seconds = NULL,
  backoff_limit = NULL,
  backoff_limit_per_index = NULL,
  completion_mode = NULL,
  completions = NULL,
  managed_by = NULL,
  manual_selector = NULL,
  max_failed_indexes = NULL,
  parallelism = NULL,
  pod_failure_policy = NULL,
  pod_replacement_policy = NULL,
  selector = NULL,
  success_policy = NULL,
  suspend = NULL,
  template = NULL,
  ttl_seconds_after_finished = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
