% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JobStatus.R
\name{V1JobStatus}
\alias{V1JobStatus}
\title{V1JobStatus}
\description{
V1JobStatus

V1JobStatus
}
\details{
JobStatus represents the current state of a Job.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JobStatus-new}{\code{V1JobStatus$new()}}
\item \href{#method-V1JobStatus-to_list}{\code{V1JobStatus$to_list()}}
\item \href{#method-V1JobStatus-clone}{\code{V1JobStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobStatus$new(
  active = NULL,
  completed_indexes = NULL,
  completion_time = NULL,
  conditions = NULL,
  failed = NULL,
  failed_indexes = NULL,
  ready = NULL,
  start_time = NULL,
  succeeded = NULL,
  terminating = NULL,
  uncounted_terminated_pods = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
