% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JobTemplateSpec.R
\name{V1JobTemplateSpec}
\alias{V1JobTemplateSpec}
\title{V1JobTemplateSpec}
\description{
V1JobTemplateSpec

V1JobTemplateSpec
}
\details{
JobTemplateSpec describes the data a Job should have when created from a template
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JobTemplateSpec-new}{\code{V1JobTemplateSpec$new()}}
\item \href{#method-V1JobTemplateSpec-to_list}{\code{V1JobTemplateSpec$to_list()}}
\item \href{#method-V1JobTemplateSpec-clone}{\code{V1JobTemplateSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobTemplateSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobTemplateSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobTemplateSpec$new(metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobTemplateSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobTemplateSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobTemplateSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JobTemplateSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JobTemplateSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JobTemplateSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
