% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JSON.R
\name{V1JSON}
\alias{V1JSON}
\title{V1JSON}
\description{
V1JSON

V1JSON
}
\details{
JSON represents any valid JSON value. These types are supported: bool, int64, float64, string, interface{}, mapstringinterface{} and nil.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JSON-new}{\code{V1JSON$new()}}
\item \href{#method-V1JSON-to_list}{\code{V1JSON$to_list()}}
\item \href{#method-V1JSON-clone}{\code{V1JSON$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSON-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSON-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSON$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSON-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSON-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSON$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSON-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSON-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSON$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
