% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JSONSchemaProps.R
\name{V1JSONSchemaProps}
\alias{V1JSONSchemaProps}
\title{V1JSONSchemaProps}
\description{
V1JSONSchemaProps

V1JSONSchemaProps
}
\details{
JSONSchemaProps is a JSON-Schema following Specification Draft 4 (http://json-schema.org/).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JSONSchemaProps-new}{\code{V1JSONSchemaProps$new()}}
\item \href{#method-V1JSONSchemaProps-to_list}{\code{V1JSONSchemaProps$to_list()}}
\item \href{#method-V1JSONSchemaProps-clone}{\code{V1JSONSchemaProps$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaProps-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaProps-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaProps$new(
  `_ref` = NULL,
  `_schema` = NULL,
  additional_items = NULL,
  additional_properties = NULL,
  all_of = NULL,
  any_of = NULL,
  default = NULL,
  definitions = NULL,
  dependencies = NULL,
  description = NULL,
  enum = NULL,
  example = NULL,
  exclusive_maximum = NULL,
  exclusive_minimum = NULL,
  external_docs = NULL,
  format = NULL,
  id = NULL,
  items = NULL,
  max_items = NULL,
  max_length = NULL,
  max_properties = NULL,
  maximum = NULL,
  min_items = NULL,
  min_length = NULL,
  min_properties = NULL,
  minimum = NULL,
  multiple_of = NULL,
  not = NULL,
  nullable = NULL,
  one_of = NULL,
  pattern = NULL,
  pattern_properties = NULL,
  properties = NULL,
  required = NULL,
  title = NULL,
  type = NULL,
  unique_items = NULL,
  x_kubernetes_embedded_resource = NULL,
  x_kubernetes_int_or_string = NULL,
  x_kubernetes_list_map_keys = NULL,
  x_kubernetes_list_type = NULL,
  x_kubernetes_map_type = NULL,
  x_kubernetes_preserve_unknown_fields = NULL,
  x_kubernetes_validations = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaProps-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaProps-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaProps$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaProps-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaProps-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaProps$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
