% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JSONSchemaPropsOrBool.R
\name{V1JSONSchemaPropsOrBool}
\alias{V1JSONSchemaPropsOrBool}
\title{V1JSONSchemaPropsOrBool}
\description{
V1JSONSchemaPropsOrBool

V1JSONSchemaPropsOrBool
}
\details{
JSONSchemaPropsOrBool represents JSONSchemaProps or a boolean value. Defaults to true for the boolean property.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JSONSchemaPropsOrBool-new}{\code{V1JSONSchemaPropsOrBool$new()}}
\item \href{#method-V1JSONSchemaPropsOrBool-to_list}{\code{V1JSONSchemaPropsOrBool$to_list()}}
\item \href{#method-V1JSONSchemaPropsOrBool-clone}{\code{V1JSONSchemaPropsOrBool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaPropsOrBool-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaPropsOrBool-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaPropsOrBool$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaPropsOrBool-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaPropsOrBool-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaPropsOrBool$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaPropsOrBool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaPropsOrBool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaPropsOrBool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
