% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1JSONSchemaPropsOrStringArray.R
\name{V1JSONSchemaPropsOrStringArray}
\alias{V1JSONSchemaPropsOrStringArray}
\title{V1JSONSchemaPropsOrStringArray}
\description{
V1JSONSchemaPropsOrStringArray

V1JSONSchemaPropsOrStringArray
}
\details{
JSONSchemaPropsOrStringArray represents a JSONSchemaProps or a string array.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1JSONSchemaPropsOrStringArray-new}{\code{V1JSONSchemaPropsOrStringArray$new()}}
\item \href{#method-V1JSONSchemaPropsOrStringArray-to_list}{\code{V1JSONSchemaPropsOrStringArray$to_list()}}
\item \href{#method-V1JSONSchemaPropsOrStringArray-clone}{\code{V1JSONSchemaPropsOrStringArray$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaPropsOrStringArray-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaPropsOrStringArray-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaPropsOrStringArray$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaPropsOrStringArray-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaPropsOrStringArray-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaPropsOrStringArray$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1JSONSchemaPropsOrStringArray-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1JSONSchemaPropsOrStringArray-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1JSONSchemaPropsOrStringArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
