% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1KeyToPath.R
\name{V1KeyToPath}
\alias{V1KeyToPath}
\title{V1KeyToPath}
\description{
V1KeyToPath

V1KeyToPath
}
\details{
Maps a string key to a path within a volume.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1KeyToPath-new}{\code{V1KeyToPath$new()}}
\item \href{#method-V1KeyToPath-to_list}{\code{V1KeyToPath$to_list()}}
\item \href{#method-V1KeyToPath-clone}{\code{V1KeyToPath$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1KeyToPath-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1KeyToPath-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1KeyToPath$new(key = NULL, mode = NULL, path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1KeyToPath-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1KeyToPath-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1KeyToPath$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1KeyToPath-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1KeyToPath-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1KeyToPath$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
