% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LabelSelector.R
\name{V1LabelSelector}
\alias{V1LabelSelector}
\title{V1LabelSelector}
\description{
V1LabelSelector

V1LabelSelector
}
\details{
A label selector is a label query over a set of resources. The result of matchLabels and matchExpressions are ANDed. An empty label selector matches all objects. A null label selector matches no objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LabelSelector-new}{\code{V1LabelSelector$new()}}
\item \href{#method-V1LabelSelector-to_list}{\code{V1LabelSelector$to_list()}}
\item \href{#method-V1LabelSelector-clone}{\code{V1LabelSelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelector$new(match_expressions = NULL, match_labels = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
