% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LabelSelectorAttributes.R
\name{V1LabelSelectorAttributes}
\alias{V1LabelSelectorAttributes}
\title{V1LabelSelectorAttributes}
\description{
V1LabelSelectorAttributes

V1LabelSelectorAttributes
}
\details{
LabelSelectorAttributes indicates a label limited access. Webhook authors are encouraged to * ensure rawSelector and requirements are not both set * consider the requirements field if set * not try to parse or consider the rawSelector field if set. This is to avoid another CVE-2022-2880 (i.e. getting different systems to agree on how exactly to parse a query is not something we want), see https://www.oxeye.io/resources/golang-parameter-smuggling-attack for more details. For the *SubjectAccessReview endpoints of the kube-apiserver: * If rawSelector is empty and requirements are empty, the request is not limited. * If rawSelector is present and requirements are empty, the rawSelector will be parsed and limited if the parsing succeeds. * If rawSelector is empty and requirements are present, the requirements should be honored * If rawSelector is present and requirements are present, the request is invalid.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LabelSelectorAttributes-new}{\code{V1LabelSelectorAttributes$new()}}
\item \href{#method-V1LabelSelectorAttributes-to_list}{\code{V1LabelSelectorAttributes$to_list()}}
\item \href{#method-V1LabelSelectorAttributes-clone}{\code{V1LabelSelectorAttributes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelectorAttributes-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelectorAttributes-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelectorAttributes$new(raw_selector = NULL, requirements = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelectorAttributes-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelectorAttributes-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelectorAttributes$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelectorAttributes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelectorAttributes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelectorAttributes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
