% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LabelSelectorRequirement.R
\name{V1LabelSelectorRequirement}
\alias{V1LabelSelectorRequirement}
\title{V1LabelSelectorRequirement}
\description{
V1LabelSelectorRequirement

V1LabelSelectorRequirement
}
\details{
A label selector requirement is a selector that contains values, a key, and an operator that relates the key and values.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LabelSelectorRequirement-new}{\code{V1LabelSelectorRequirement$new()}}
\item \href{#method-V1LabelSelectorRequirement-to_list}{\code{V1LabelSelectorRequirement$to_list()}}
\item \href{#method-V1LabelSelectorRequirement-clone}{\code{V1LabelSelectorRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelectorRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelectorRequirement-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelectorRequirement$new(key = NULL, operator = NULL, values = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelectorRequirement-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelectorRequirement-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelectorRequirement$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LabelSelectorRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LabelSelectorRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LabelSelectorRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
