% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Lease.R
\name{V1Lease}
\alias{V1Lease}
\title{V1Lease}
\description{
V1Lease

V1Lease
}
\details{
Lease defines a lease concept.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Lease-new}{\code{V1Lease$new()}}
\item \href{#method-V1Lease-to_list}{\code{V1Lease$to_list()}}
\item \href{#method-V1Lease-clone}{\code{V1Lease$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Lease-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Lease-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Lease$new(api_version = NULL, kind = NULL, metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Lease-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Lease-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Lease$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Lease-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Lease-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Lease$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
