% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LeaseList.R
\name{V1LeaseList}
\alias{V1LeaseList}
\title{V1LeaseList}
\description{
V1LeaseList

V1LeaseList
}
\details{
LeaseList is a list of Lease objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LeaseList-new}{\code{V1LeaseList$new()}}
\item \href{#method-V1LeaseList-to_list}{\code{V1LeaseList$to_list()}}
\item \href{#method-V1LeaseList-clone}{\code{V1LeaseList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LeaseList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LeaseList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LeaseList$new(api_version = NULL, items = NULL, kind = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LeaseList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LeaseList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LeaseList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LeaseList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LeaseList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LeaseList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
