% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LeaseSpec.R
\name{V1LeaseSpec}
\alias{V1LeaseSpec}
\title{V1LeaseSpec}
\description{
V1LeaseSpec

V1LeaseSpec
}
\details{
LeaseSpec is a specification of a Lease.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LeaseSpec-new}{\code{V1LeaseSpec$new()}}
\item \href{#method-V1LeaseSpec-to_list}{\code{V1LeaseSpec$to_list()}}
\item \href{#method-V1LeaseSpec-clone}{\code{V1LeaseSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LeaseSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LeaseSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LeaseSpec$new(
  acquire_time = NULL,
  holder_identity = NULL,
  lease_duration_seconds = NULL,
  lease_transitions = NULL,
  preferred_holder = NULL,
  renew_time = NULL,
  strategy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LeaseSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LeaseSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LeaseSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LeaseSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LeaseSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LeaseSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
