% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Lifecycle.R
\name{V1Lifecycle}
\alias{V1Lifecycle}
\title{V1Lifecycle}
\description{
V1Lifecycle

V1Lifecycle
}
\details{
Lifecycle describes actions that the management system should take in response to container lifecycle events. For the PostStart and PreStop lifecycle handlers, management of the container blocks until the action is complete, unless the container process fails, in which case the handler is aborted.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Lifecycle-new}{\code{V1Lifecycle$new()}}
\item \href{#method-V1Lifecycle-to_list}{\code{V1Lifecycle$to_list()}}
\item \href{#method-V1Lifecycle-clone}{\code{V1Lifecycle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Lifecycle-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Lifecycle-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Lifecycle$new(post_start = NULL, pre_stop = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Lifecycle-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Lifecycle-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Lifecycle$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Lifecycle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Lifecycle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Lifecycle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
