% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LifecycleHandler.R
\name{V1LifecycleHandler}
\alias{V1LifecycleHandler}
\title{V1LifecycleHandler}
\description{
V1LifecycleHandler

V1LifecycleHandler
}
\details{
LifecycleHandler defines a specific action that should be taken in a lifecycle hook. One and only one of the fields, except TCPSocket must be specified.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LifecycleHandler-new}{\code{V1LifecycleHandler$new()}}
\item \href{#method-V1LifecycleHandler-to_list}{\code{V1LifecycleHandler$to_list()}}
\item \href{#method-V1LifecycleHandler-clone}{\code{V1LifecycleHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LifecycleHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LifecycleHandler-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LifecycleHandler$new(
  exec = NULL,
  http_get = NULL,
  sleep = NULL,
  tcp_socket = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LifecycleHandler-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LifecycleHandler-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LifecycleHandler$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LifecycleHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LifecycleHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LifecycleHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
