% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LimitRange.R
\name{V1LimitRange}
\alias{V1LimitRange}
\title{V1LimitRange}
\description{
V1LimitRange

V1LimitRange
}
\details{
LimitRange sets resource usage limits for each kind of resource in a Namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LimitRange-new}{\code{V1LimitRange$new()}}
\item \href{#method-V1LimitRange-to_list}{\code{V1LimitRange$to_list()}}
\item \href{#method-V1LimitRange-clone}{\code{V1LimitRange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRange-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRange-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRange$new(api_version = NULL, kind = NULL, metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRange-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRange-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRange$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRange-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRange-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
