% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LimitRangeItem.R
\name{V1LimitRangeItem}
\alias{V1LimitRangeItem}
\title{V1LimitRangeItem}
\description{
V1LimitRangeItem

V1LimitRangeItem
}
\details{
LimitRangeItem defines a min/max usage limit for any resource that matches on kind.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LimitRangeItem-new}{\code{V1LimitRangeItem$new()}}
\item \href{#method-V1LimitRangeItem-to_list}{\code{V1LimitRangeItem$to_list()}}
\item \href{#method-V1LimitRangeItem-clone}{\code{V1LimitRangeItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRangeItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRangeItem-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRangeItem$new(
  default = NULL,
  default_request = NULL,
  max = NULL,
  max_limit_request_ratio = NULL,
  min = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRangeItem-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRangeItem-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRangeItem$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRangeItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRangeItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRangeItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
