% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LimitRangeSpec.R
\name{V1LimitRangeSpec}
\alias{V1LimitRangeSpec}
\title{V1LimitRangeSpec}
\description{
V1LimitRangeSpec

V1LimitRangeSpec
}
\details{
LimitRangeSpec defines a min/max usage limit for resources that match on kind.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LimitRangeSpec-new}{\code{V1LimitRangeSpec$new()}}
\item \href{#method-V1LimitRangeSpec-to_list}{\code{V1LimitRangeSpec$to_list()}}
\item \href{#method-V1LimitRangeSpec-clone}{\code{V1LimitRangeSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRangeSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRangeSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRangeSpec$new(limits = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRangeSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRangeSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRangeSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitRangeSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitRangeSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitRangeSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
