% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LimitResponse.R
\name{V1LimitResponse}
\alias{V1LimitResponse}
\title{V1LimitResponse}
\description{
V1LimitResponse

V1LimitResponse
}
\details{
LimitResponse defines how to handle requests that can not be executed right now.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LimitResponse-new}{\code{V1LimitResponse$new()}}
\item \href{#method-V1LimitResponse-to_list}{\code{V1LimitResponse$to_list()}}
\item \href{#method-V1LimitResponse-clone}{\code{V1LimitResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitResponse-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitResponse$new(queuing = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitResponse-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitResponse-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitResponse$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LimitResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LimitResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LimitResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
