% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LinuxContainerUser.R
\name{V1LinuxContainerUser}
\alias{V1LinuxContainerUser}
\title{V1LinuxContainerUser}
\description{
V1LinuxContainerUser

V1LinuxContainerUser
}
\details{
LinuxContainerUser represents user identity information in Linux containers
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LinuxContainerUser-new}{\code{V1LinuxContainerUser$new()}}
\item \href{#method-V1LinuxContainerUser-to_list}{\code{V1LinuxContainerUser$to_list()}}
\item \href{#method-V1LinuxContainerUser-clone}{\code{V1LinuxContainerUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LinuxContainerUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LinuxContainerUser-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LinuxContainerUser$new(gid = NULL, supplemental_groups = NULL, uid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LinuxContainerUser-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LinuxContainerUser-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LinuxContainerUser$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LinuxContainerUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LinuxContainerUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LinuxContainerUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
