% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ListMeta.R
\name{V1ListMeta}
\alias{V1ListMeta}
\title{V1ListMeta}
\description{
V1ListMeta

V1ListMeta
}
\details{
ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ListMeta-new}{\code{V1ListMeta$new()}}
\item \href{#method-V1ListMeta-to_list}{\code{V1ListMeta$to_list()}}
\item \href{#method-V1ListMeta-clone}{\code{V1ListMeta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ListMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ListMeta-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ListMeta$new(
  continue = NULL,
  remaining_item_count = NULL,
  resource_version = NULL,
  self_link = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ListMeta-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ListMeta-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ListMeta$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ListMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ListMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ListMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
