% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LoadBalancerIngress.R
\name{V1LoadBalancerIngress}
\alias{V1LoadBalancerIngress}
\title{V1LoadBalancerIngress}
\description{
V1LoadBalancerIngress

V1LoadBalancerIngress
}
\details{
LoadBalancerIngress represents the status of a load-balancer ingress point: traffic intended for the service should be sent to an ingress point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LoadBalancerIngress-new}{\code{V1LoadBalancerIngress$new()}}
\item \href{#method-V1LoadBalancerIngress-to_list}{\code{V1LoadBalancerIngress$to_list()}}
\item \href{#method-V1LoadBalancerIngress-clone}{\code{V1LoadBalancerIngress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LoadBalancerIngress-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LoadBalancerIngress-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LoadBalancerIngress$new(
  hostname = NULL,
  ip = NULL,
  ip_mode = NULL,
  ports = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LoadBalancerIngress-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LoadBalancerIngress-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LoadBalancerIngress$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LoadBalancerIngress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LoadBalancerIngress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LoadBalancerIngress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
