% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LoadBalancerStatus.R
\name{V1LoadBalancerStatus}
\alias{V1LoadBalancerStatus}
\title{V1LoadBalancerStatus}
\description{
V1LoadBalancerStatus

V1LoadBalancerStatus
}
\details{
LoadBalancerStatus represents the status of a load-balancer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LoadBalancerStatus-new}{\code{V1LoadBalancerStatus$new()}}
\item \href{#method-V1LoadBalancerStatus-to_list}{\code{V1LoadBalancerStatus$to_list()}}
\item \href{#method-V1LoadBalancerStatus-clone}{\code{V1LoadBalancerStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LoadBalancerStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LoadBalancerStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LoadBalancerStatus$new(ingress = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LoadBalancerStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LoadBalancerStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LoadBalancerStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LoadBalancerStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LoadBalancerStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LoadBalancerStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
