% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LocalObjectReference.R
\name{V1LocalObjectReference}
\alias{V1LocalObjectReference}
\title{V1LocalObjectReference}
\description{
V1LocalObjectReference

V1LocalObjectReference
}
\details{
LocalObjectReference contains enough information to let you locate the referenced object inside the same namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LocalObjectReference-new}{\code{V1LocalObjectReference$new()}}
\item \href{#method-V1LocalObjectReference-to_list}{\code{V1LocalObjectReference$to_list()}}
\item \href{#method-V1LocalObjectReference-clone}{\code{V1LocalObjectReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalObjectReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalObjectReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalObjectReference$new(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalObjectReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalObjectReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalObjectReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalObjectReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalObjectReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalObjectReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
