% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LocalSubjectAccessReview.R
\name{V1LocalSubjectAccessReview}
\alias{V1LocalSubjectAccessReview}
\title{V1LocalSubjectAccessReview}
\description{
V1LocalSubjectAccessReview

V1LocalSubjectAccessReview
}
\details{
LocalSubjectAccessReview checks whether or not a user or group can perform an action in a given namespace. Having a namespace scoped resource makes it much easier to grant namespace scoped policy that includes permissions checking.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LocalSubjectAccessReview-new}{\code{V1LocalSubjectAccessReview$new()}}
\item \href{#method-V1LocalSubjectAccessReview-to_list}{\code{V1LocalSubjectAccessReview$to_list()}}
\item \href{#method-V1LocalSubjectAccessReview-clone}{\code{V1LocalSubjectAccessReview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalSubjectAccessReview-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalSubjectAccessReview-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalSubjectAccessReview$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalSubjectAccessReview-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalSubjectAccessReview-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalSubjectAccessReview$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalSubjectAccessReview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalSubjectAccessReview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalSubjectAccessReview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
