% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1LocalVolumeSource.R
\name{V1LocalVolumeSource}
\alias{V1LocalVolumeSource}
\title{V1LocalVolumeSource}
\description{
V1LocalVolumeSource

V1LocalVolumeSource
}
\details{
Local represents directly-attached storage with node affinity (Beta feature)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1LocalVolumeSource-new}{\code{V1LocalVolumeSource$new()}}
\item \href{#method-V1LocalVolumeSource-to_list}{\code{V1LocalVolumeSource$to_list()}}
\item \href{#method-V1LocalVolumeSource-clone}{\code{V1LocalVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalVolumeSource$new(fs_type = NULL, path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1LocalVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1LocalVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1LocalVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
