% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ManagedFieldsEntry.R
\name{V1ManagedFieldsEntry}
\alias{V1ManagedFieldsEntry}
\title{V1ManagedFieldsEntry}
\description{
V1ManagedFieldsEntry

V1ManagedFieldsEntry
}
\details{
ManagedFieldsEntry is a workflow-id, a FieldSet and the group version of the resource that the fieldset applies to.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ManagedFieldsEntry-new}{\code{V1ManagedFieldsEntry$new()}}
\item \href{#method-V1ManagedFieldsEntry-to_list}{\code{V1ManagedFieldsEntry$to_list()}}
\item \href{#method-V1ManagedFieldsEntry-clone}{\code{V1ManagedFieldsEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ManagedFieldsEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ManagedFieldsEntry-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ManagedFieldsEntry$new(
  api_version = NULL,
  fields_type = NULL,
  fields_v1 = NULL,
  manager = NULL,
  operation = NULL,
  subresource = NULL,
  time = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ManagedFieldsEntry-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ManagedFieldsEntry-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ManagedFieldsEntry$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ManagedFieldsEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ManagedFieldsEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ManagedFieldsEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
