% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1MatchCondition.R
\name{V1MatchCondition}
\alias{V1MatchCondition}
\title{V1MatchCondition}
\description{
V1MatchCondition

V1MatchCondition
}
\details{
MatchCondition represents a condition which must by fulfilled for a request to be sent to a webhook.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1MatchCondition-new}{\code{V1MatchCondition$new()}}
\item \href{#method-V1MatchCondition-to_list}{\code{V1MatchCondition$to_list()}}
\item \href{#method-V1MatchCondition-clone}{\code{V1MatchCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MatchCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1MatchCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MatchCondition$new(expression = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MatchCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1MatchCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MatchCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MatchCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1MatchCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MatchCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
