% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1MatchResources.R
\name{V1MatchResources}
\alias{V1MatchResources}
\title{V1MatchResources}
\description{
V1MatchResources

V1MatchResources
}
\details{
MatchResources decides whether to run the admission control policy on an object based on whether it meets the match criteria. The exclude rules take precedence over include rules (if a resource matches both, it is excluded)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1MatchResources-new}{\code{V1MatchResources$new()}}
\item \href{#method-V1MatchResources-to_list}{\code{V1MatchResources$to_list()}}
\item \href{#method-V1MatchResources-clone}{\code{V1MatchResources$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MatchResources-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1MatchResources-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MatchResources$new(
  exclude_resource_rules = NULL,
  match_policy = NULL,
  namespace_selector = NULL,
  object_selector = NULL,
  resource_rules = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MatchResources-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1MatchResources-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MatchResources$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MatchResources-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1MatchResources-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MatchResources$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
