% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1MicroTime.R
\name{V1MicroTime}
\alias{V1MicroTime}
\title{V1MicroTime}
\description{
V1MicroTime

V1MicroTime
}
\details{
MicroTime is version of Time with microsecond level precision.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1MicroTime-new}{\code{V1MicroTime$new()}}
\item \href{#method-V1MicroTime-to_list}{\code{V1MicroTime$to_list()}}
\item \href{#method-V1MicroTime-clone}{\code{V1MicroTime$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MicroTime-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1MicroTime-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MicroTime$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MicroTime-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1MicroTime-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MicroTime$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MicroTime-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1MicroTime-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MicroTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
