% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1MutatingWebhookConfiguration.R
\name{V1MutatingWebhookConfiguration}
\alias{V1MutatingWebhookConfiguration}
\title{V1MutatingWebhookConfiguration}
\description{
V1MutatingWebhookConfiguration

V1MutatingWebhookConfiguration
}
\details{
MutatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and may change the object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1MutatingWebhookConfiguration-new}{\code{V1MutatingWebhookConfiguration$new()}}
\item \href{#method-V1MutatingWebhookConfiguration-to_list}{\code{V1MutatingWebhookConfiguration$to_list()}}
\item \href{#method-V1MutatingWebhookConfiguration-clone}{\code{V1MutatingWebhookConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MutatingWebhookConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1MutatingWebhookConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MutatingWebhookConfiguration$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  webhooks = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MutatingWebhookConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1MutatingWebhookConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MutatingWebhookConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MutatingWebhookConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1MutatingWebhookConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MutatingWebhookConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
