% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1MutatingWebhookConfigurationList.R
\name{V1MutatingWebhookConfigurationList}
\alias{V1MutatingWebhookConfigurationList}
\title{V1MutatingWebhookConfigurationList}
\description{
V1MutatingWebhookConfigurationList

V1MutatingWebhookConfigurationList
}
\details{
MutatingWebhookConfigurationList is a list of MutatingWebhookConfiguration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1MutatingWebhookConfigurationList-new}{\code{V1MutatingWebhookConfigurationList$new()}}
\item \href{#method-V1MutatingWebhookConfigurationList-to_list}{\code{V1MutatingWebhookConfigurationList$to_list()}}
\item \href{#method-V1MutatingWebhookConfigurationList-clone}{\code{V1MutatingWebhookConfigurationList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MutatingWebhookConfigurationList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1MutatingWebhookConfigurationList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MutatingWebhookConfigurationList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MutatingWebhookConfigurationList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1MutatingWebhookConfigurationList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MutatingWebhookConfigurationList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1MutatingWebhookConfigurationList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1MutatingWebhookConfigurationList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1MutatingWebhookConfigurationList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
