% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Namespace.R
\name{V1Namespace}
\alias{V1Namespace}
\title{V1Namespace}
\description{
V1Namespace

V1Namespace
}
\details{
Namespace provides a scope for Names. Use of multiple namespaces is optional.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Namespace-new}{\code{V1Namespace$new()}}
\item \href{#method-V1Namespace-to_list}{\code{V1Namespace$to_list()}}
\item \href{#method-V1Namespace-clone}{\code{V1Namespace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Namespace-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Namespace-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Namespace$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Namespace-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Namespace-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Namespace$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Namespace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Namespace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Namespace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
