% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NamespaceCondition.R
\name{V1NamespaceCondition}
\alias{V1NamespaceCondition}
\title{V1NamespaceCondition}
\description{
V1NamespaceCondition

V1NamespaceCondition
}
\details{
NamespaceCondition contains details about state of namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NamespaceCondition-new}{\code{V1NamespaceCondition$new()}}
\item \href{#method-V1NamespaceCondition-to_list}{\code{V1NamespaceCondition$to_list()}}
\item \href{#method-V1NamespaceCondition-clone}{\code{V1NamespaceCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
