% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NamespaceList.R
\name{V1NamespaceList}
\alias{V1NamespaceList}
\title{V1NamespaceList}
\description{
V1NamespaceList

V1NamespaceList
}
\details{
NamespaceList is a list of Namespaces.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NamespaceList-new}{\code{V1NamespaceList$new()}}
\item \href{#method-V1NamespaceList-to_list}{\code{V1NamespaceList$to_list()}}
\item \href{#method-V1NamespaceList-clone}{\code{V1NamespaceList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
