% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NamespaceSpec.R
\name{V1NamespaceSpec}
\alias{V1NamespaceSpec}
\title{V1NamespaceSpec}
\description{
V1NamespaceSpec

V1NamespaceSpec
}
\details{
NamespaceSpec describes the attributes on a Namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NamespaceSpec-new}{\code{V1NamespaceSpec$new()}}
\item \href{#method-V1NamespaceSpec-to_list}{\code{V1NamespaceSpec$to_list()}}
\item \href{#method-V1NamespaceSpec-clone}{\code{V1NamespaceSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceSpec$new(finalizers = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
