% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NamespaceStatus.R
\name{V1NamespaceStatus}
\alias{V1NamespaceStatus}
\title{V1NamespaceStatus}
\description{
V1NamespaceStatus

V1NamespaceStatus
}
\details{
NamespaceStatus is information about the current status of a Namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NamespaceStatus-new}{\code{V1NamespaceStatus$new()}}
\item \href{#method-V1NamespaceStatus-to_list}{\code{V1NamespaceStatus$to_list()}}
\item \href{#method-V1NamespaceStatus-clone}{\code{V1NamespaceStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceStatus$new(conditions = NULL, phase = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NamespaceStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NamespaceStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NamespaceStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
