% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NetworkPolicy.R
\name{V1NetworkPolicy}
\alias{V1NetworkPolicy}
\title{V1NetworkPolicy}
\description{
V1NetworkPolicy

V1NetworkPolicy
}
\details{
NetworkPolicy describes what network traffic is allowed for a set of Pods
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NetworkPolicy-new}{\code{V1NetworkPolicy$new()}}
\item \href{#method-V1NetworkPolicy-to_list}{\code{V1NetworkPolicy$to_list()}}
\item \href{#method-V1NetworkPolicy-clone}{\code{V1NetworkPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicy$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
