% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NetworkPolicyEgressRule.R
\name{V1NetworkPolicyEgressRule}
\alias{V1NetworkPolicyEgressRule}
\title{V1NetworkPolicyEgressRule}
\description{
V1NetworkPolicyEgressRule

V1NetworkPolicyEgressRule
}
\details{
NetworkPolicyEgressRule describes a particular set of traffic that is allowed out of pods matched by a NetworkPolicySpec's podSelector. The traffic must match both ports and to. This type is beta-level in 1.8
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NetworkPolicyEgressRule-new}{\code{V1NetworkPolicyEgressRule$new()}}
\item \href{#method-V1NetworkPolicyEgressRule-to_list}{\code{V1NetworkPolicyEgressRule$to_list()}}
\item \href{#method-V1NetworkPolicyEgressRule-clone}{\code{V1NetworkPolicyEgressRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyEgressRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyEgressRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyEgressRule$new(ports = NULL, to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyEgressRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyEgressRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyEgressRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyEgressRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyEgressRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyEgressRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
