% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NetworkPolicyIngressRule.R
\name{V1NetworkPolicyIngressRule}
\alias{V1NetworkPolicyIngressRule}
\title{V1NetworkPolicyIngressRule}
\description{
V1NetworkPolicyIngressRule

V1NetworkPolicyIngressRule
}
\details{
NetworkPolicyIngressRule describes a particular set of traffic that is allowed to the pods matched by a NetworkPolicySpec's podSelector. The traffic must match both ports and from.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NetworkPolicyIngressRule-new}{\code{V1NetworkPolicyIngressRule$new()}}
\item \href{#method-V1NetworkPolicyIngressRule-to_list}{\code{V1NetworkPolicyIngressRule$to_list()}}
\item \href{#method-V1NetworkPolicyIngressRule-clone}{\code{V1NetworkPolicyIngressRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyIngressRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyIngressRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyIngressRule$new(from = NULL, ports = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyIngressRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyIngressRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyIngressRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyIngressRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyIngressRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyIngressRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
