% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NetworkPolicyList.R
\name{V1NetworkPolicyList}
\alias{V1NetworkPolicyList}
\title{V1NetworkPolicyList}
\description{
V1NetworkPolicyList

V1NetworkPolicyList
}
\details{
NetworkPolicyList is a list of NetworkPolicy objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NetworkPolicyList-new}{\code{V1NetworkPolicyList$new()}}
\item \href{#method-V1NetworkPolicyList-to_list}{\code{V1NetworkPolicyList$to_list()}}
\item \href{#method-V1NetworkPolicyList-clone}{\code{V1NetworkPolicyList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
