% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NetworkPolicyPeer.R
\name{V1NetworkPolicyPeer}
\alias{V1NetworkPolicyPeer}
\title{V1NetworkPolicyPeer}
\description{
V1NetworkPolicyPeer

V1NetworkPolicyPeer
}
\details{
NetworkPolicyPeer describes a peer to allow traffic to/from. Only certain combinations of fields are allowed
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NetworkPolicyPeer-new}{\code{V1NetworkPolicyPeer$new()}}
\item \href{#method-V1NetworkPolicyPeer-to_list}{\code{V1NetworkPolicyPeer$to_list()}}
\item \href{#method-V1NetworkPolicyPeer-clone}{\code{V1NetworkPolicyPeer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyPeer-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyPeer-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyPeer$new(
  ip_block = NULL,
  namespace_selector = NULL,
  pod_selector = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyPeer-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyPeer-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyPeer$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicyPeer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicyPeer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicyPeer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
