% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NetworkPolicySpec.R
\name{V1NetworkPolicySpec}
\alias{V1NetworkPolicySpec}
\title{V1NetworkPolicySpec}
\description{
V1NetworkPolicySpec

V1NetworkPolicySpec
}
\details{
NetworkPolicySpec provides the specification of a NetworkPolicy
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NetworkPolicySpec-new}{\code{V1NetworkPolicySpec$new()}}
\item \href{#method-V1NetworkPolicySpec-to_list}{\code{V1NetworkPolicySpec$to_list()}}
\item \href{#method-V1NetworkPolicySpec-clone}{\code{V1NetworkPolicySpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicySpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicySpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicySpec$new(
  egress = NULL,
  ingress = NULL,
  pod_selector = NULL,
  policy_types = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicySpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicySpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicySpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NetworkPolicySpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NetworkPolicySpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NetworkPolicySpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
