% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NFSVolumeSource.R
\name{V1NFSVolumeSource}
\alias{V1NFSVolumeSource}
\title{V1NFSVolumeSource}
\description{
V1NFSVolumeSource

V1NFSVolumeSource
}
\details{
Represents an NFS mount that lasts the lifetime of a pod. NFS volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NFSVolumeSource-new}{\code{V1NFSVolumeSource$new()}}
\item \href{#method-V1NFSVolumeSource-to_list}{\code{V1NFSVolumeSource$to_list()}}
\item \href{#method-V1NFSVolumeSource-clone}{\code{V1NFSVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NFSVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NFSVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NFSVolumeSource$new(path = NULL, read_only = NULL, server = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NFSVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NFSVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NFSVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NFSVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NFSVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NFSVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
