% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Node.R
\name{V1Node}
\alias{V1Node}
\title{V1Node}
\description{
V1Node

V1Node
}
\details{
Node is a worker node in Kubernetes. Each node will have a unique identifier in the cache (i.e. in etcd).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Node-new}{\code{V1Node$new()}}
\item \href{#method-V1Node-to_list}{\code{V1Node$to_list()}}
\item \href{#method-V1Node-clone}{\code{V1Node$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Node-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Node-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Node$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Node-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Node-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Node$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Node-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Node-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Node$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
