% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeAddress.R
\name{V1NodeAddress}
\alias{V1NodeAddress}
\title{V1NodeAddress}
\description{
V1NodeAddress

V1NodeAddress
}
\details{
NodeAddress contains information for the node's address.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeAddress-new}{\code{V1NodeAddress$new()}}
\item \href{#method-V1NodeAddress-to_list}{\code{V1NodeAddress$to_list()}}
\item \href{#method-V1NodeAddress-clone}{\code{V1NodeAddress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeAddress-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeAddress-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeAddress$new(address = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeAddress-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeAddress-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeAddress$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeAddress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeAddress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeAddress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
