% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeAffinity.R
\name{V1NodeAffinity}
\alias{V1NodeAffinity}
\title{V1NodeAffinity}
\description{
V1NodeAffinity

V1NodeAffinity
}
\details{
Node affinity is a group of node affinity scheduling rules.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeAffinity-new}{\code{V1NodeAffinity$new()}}
\item \href{#method-V1NodeAffinity-to_list}{\code{V1NodeAffinity$to_list()}}
\item \href{#method-V1NodeAffinity-clone}{\code{V1NodeAffinity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeAffinity-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeAffinity-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeAffinity$new(
  preferred_during_scheduling_ignored_during_execution = NULL,
  required_during_scheduling_ignored_during_execution = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeAffinity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeAffinity-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeAffinity$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeAffinity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeAffinity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeAffinity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
