% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeCondition.R
\name{V1NodeCondition}
\alias{V1NodeCondition}
\title{V1NodeCondition}
\description{
V1NodeCondition

V1NodeCondition
}
\details{
NodeCondition contains condition information for a node.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeCondition-new}{\code{V1NodeCondition$new()}}
\item \href{#method-V1NodeCondition-to_list}{\code{V1NodeCondition$to_list()}}
\item \href{#method-V1NodeCondition-clone}{\code{V1NodeCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeCondition$new(
  last_heartbeat_time = NULL,
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
