% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeConfigSource.R
\name{V1NodeConfigSource}
\alias{V1NodeConfigSource}
\title{V1NodeConfigSource}
\description{
V1NodeConfigSource

V1NodeConfigSource
}
\details{
NodeConfigSource specifies a source of node configuration. Exactly one subfield (excluding metadata) must be non-nil. This API is deprecated since 1.22
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeConfigSource-new}{\code{V1NodeConfigSource$new()}}
\item \href{#method-V1NodeConfigSource-to_list}{\code{V1NodeConfigSource$to_list()}}
\item \href{#method-V1NodeConfigSource-clone}{\code{V1NodeConfigSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeConfigSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeConfigSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeConfigSource$new(config_map = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeConfigSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeConfigSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeConfigSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeConfigSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeConfigSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeConfigSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
