% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeConfigStatus.R
\name{V1NodeConfigStatus}
\alias{V1NodeConfigStatus}
\title{V1NodeConfigStatus}
\description{
V1NodeConfigStatus

V1NodeConfigStatus
}
\details{
NodeConfigStatus describes the status of the config assigned by Node.Spec.ConfigSource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeConfigStatus-new}{\code{V1NodeConfigStatus$new()}}
\item \href{#method-V1NodeConfigStatus-to_list}{\code{V1NodeConfigStatus$to_list()}}
\item \href{#method-V1NodeConfigStatus-clone}{\code{V1NodeConfigStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeConfigStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeConfigStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeConfigStatus$new(
  active = NULL,
  assigned = NULL,
  error = NULL,
  last_known_good = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeConfigStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeConfigStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeConfigStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeConfigStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeConfigStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeConfigStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
