% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeDaemonEndpoints.R
\name{V1NodeDaemonEndpoints}
\alias{V1NodeDaemonEndpoints}
\title{V1NodeDaemonEndpoints}
\description{
V1NodeDaemonEndpoints

V1NodeDaemonEndpoints
}
\details{
NodeDaemonEndpoints lists ports opened by daemons running on the Node.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeDaemonEndpoints-new}{\code{V1NodeDaemonEndpoints$new()}}
\item \href{#method-V1NodeDaemonEndpoints-to_list}{\code{V1NodeDaemonEndpoints$to_list()}}
\item \href{#method-V1NodeDaemonEndpoints-clone}{\code{V1NodeDaemonEndpoints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeDaemonEndpoints-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeDaemonEndpoints-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeDaemonEndpoints$new(kubelet_endpoint = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeDaemonEndpoints-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeDaemonEndpoints-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeDaemonEndpoints$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeDaemonEndpoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeDaemonEndpoints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeDaemonEndpoints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
