% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeFeatures.R
\name{V1NodeFeatures}
\alias{V1NodeFeatures}
\title{V1NodeFeatures}
\description{
V1NodeFeatures

V1NodeFeatures
}
\details{
NodeFeatures describes the set of features implemented by the CRI implementation. The features contained in the NodeFeatures should depend only on the cri implementation independent of runtime handlers.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeFeatures-new}{\code{V1NodeFeatures$new()}}
\item \href{#method-V1NodeFeatures-to_list}{\code{V1NodeFeatures$to_list()}}
\item \href{#method-V1NodeFeatures-clone}{\code{V1NodeFeatures$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeFeatures-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeFeatures-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeFeatures$new(supplemental_groups_policy = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeFeatures-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeFeatures-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeFeatures$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeFeatures-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeFeatures-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeFeatures$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
