% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeList.R
\name{V1NodeList}
\alias{V1NodeList}
\title{V1NodeList}
\description{
V1NodeList

V1NodeList
}
\details{
NodeList is the whole list of all Nodes which have been registered with master.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeList-new}{\code{V1NodeList$new()}}
\item \href{#method-V1NodeList-to_list}{\code{V1NodeList$to_list()}}
\item \href{#method-V1NodeList-clone}{\code{V1NodeList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeList$new(api_version = NULL, items = NULL, kind = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
