% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeRuntimeHandler.R
\name{V1NodeRuntimeHandler}
\alias{V1NodeRuntimeHandler}
\title{V1NodeRuntimeHandler}
\description{
V1NodeRuntimeHandler

V1NodeRuntimeHandler
}
\details{
NodeRuntimeHandler is a set of runtime handler information.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeRuntimeHandler-new}{\code{V1NodeRuntimeHandler$new()}}
\item \href{#method-V1NodeRuntimeHandler-to_list}{\code{V1NodeRuntimeHandler$to_list()}}
\item \href{#method-V1NodeRuntimeHandler-clone}{\code{V1NodeRuntimeHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeRuntimeHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeRuntimeHandler-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeRuntimeHandler$new(features = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeRuntimeHandler-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeRuntimeHandler-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeRuntimeHandler$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeRuntimeHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeRuntimeHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeRuntimeHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
