% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeRuntimeHandlerFeatures.R
\name{V1NodeRuntimeHandlerFeatures}
\alias{V1NodeRuntimeHandlerFeatures}
\title{V1NodeRuntimeHandlerFeatures}
\description{
V1NodeRuntimeHandlerFeatures

V1NodeRuntimeHandlerFeatures
}
\details{
NodeRuntimeHandlerFeatures is a set of features implemented by the runtime handler.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeRuntimeHandlerFeatures-new}{\code{V1NodeRuntimeHandlerFeatures$new()}}
\item \href{#method-V1NodeRuntimeHandlerFeatures-to_list}{\code{V1NodeRuntimeHandlerFeatures$to_list()}}
\item \href{#method-V1NodeRuntimeHandlerFeatures-clone}{\code{V1NodeRuntimeHandlerFeatures$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeRuntimeHandlerFeatures-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeRuntimeHandlerFeatures-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeRuntimeHandlerFeatures$new(
  recursive_read_only_mounts = NULL,
  user_namespaces = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeRuntimeHandlerFeatures-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeRuntimeHandlerFeatures-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeRuntimeHandlerFeatures$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeRuntimeHandlerFeatures-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeRuntimeHandlerFeatures-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeRuntimeHandlerFeatures$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
