% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeSelector.R
\name{V1NodeSelector}
\alias{V1NodeSelector}
\title{V1NodeSelector}
\description{
V1NodeSelector

V1NodeSelector
}
\details{
A node selector represents the union of the results of one or more label queries over a set of nodes; that is, it represents the OR of the selectors represented by the node selector terms.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeSelector-new}{\code{V1NodeSelector$new()}}
\item \href{#method-V1NodeSelector-to_list}{\code{V1NodeSelector$to_list()}}
\item \href{#method-V1NodeSelector-clone}{\code{V1NodeSelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelector$new(node_selector_terms = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
