% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeSelectorRequirement.R
\name{V1NodeSelectorRequirement}
\alias{V1NodeSelectorRequirement}
\title{V1NodeSelectorRequirement}
\description{
V1NodeSelectorRequirement

V1NodeSelectorRequirement
}
\details{
A node selector requirement is a selector that contains values, a key, and an operator that relates the key and values.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeSelectorRequirement-new}{\code{V1NodeSelectorRequirement$new()}}
\item \href{#method-V1NodeSelectorRequirement-to_list}{\code{V1NodeSelectorRequirement$to_list()}}
\item \href{#method-V1NodeSelectorRequirement-clone}{\code{V1NodeSelectorRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelectorRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelectorRequirement-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelectorRequirement$new(key = NULL, operator = NULL, values = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelectorRequirement-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelectorRequirement-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelectorRequirement$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelectorRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelectorRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelectorRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
