% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeSelectorTerm.R
\name{V1NodeSelectorTerm}
\alias{V1NodeSelectorTerm}
\title{V1NodeSelectorTerm}
\description{
V1NodeSelectorTerm

V1NodeSelectorTerm
}
\details{
A null or empty node selector term matches no objects. The requirements of them are ANDed. The TopologySelectorTerm type implements a subset of the NodeSelectorTerm.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeSelectorTerm-new}{\code{V1NodeSelectorTerm$new()}}
\item \href{#method-V1NodeSelectorTerm-to_list}{\code{V1NodeSelectorTerm$to_list()}}
\item \href{#method-V1NodeSelectorTerm-clone}{\code{V1NodeSelectorTerm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelectorTerm-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelectorTerm-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelectorTerm$new(match_expressions = NULL, match_fields = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelectorTerm-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelectorTerm-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelectorTerm$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSelectorTerm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSelectorTerm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSelectorTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
