% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeSpec.R
\name{V1NodeSpec}
\alias{V1NodeSpec}
\title{V1NodeSpec}
\description{
V1NodeSpec

V1NodeSpec
}
\details{
NodeSpec describes the attributes that a node is created with.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeSpec-new}{\code{V1NodeSpec$new()}}
\item \href{#method-V1NodeSpec-to_list}{\code{V1NodeSpec$to_list()}}
\item \href{#method-V1NodeSpec-clone}{\code{V1NodeSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSpec$new(
  config_source = NULL,
  external_id = NULL,
  pod_cidr = NULL,
  pod_cid_rs = NULL,
  provider_id = NULL,
  taints = NULL,
  unschedulable = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
