% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeStatus.R
\name{V1NodeStatus}
\alias{V1NodeStatus}
\title{V1NodeStatus}
\description{
V1NodeStatus

V1NodeStatus
}
\details{
NodeStatus is information about the current status of a node.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeStatus-new}{\code{V1NodeStatus$new()}}
\item \href{#method-V1NodeStatus-to_list}{\code{V1NodeStatus$to_list()}}
\item \href{#method-V1NodeStatus-clone}{\code{V1NodeStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeStatus$new(
  addresses = NULL,
  allocatable = NULL,
  capacity = NULL,
  conditions = NULL,
  config = NULL,
  daemon_endpoints = NULL,
  features = NULL,
  images = NULL,
  node_info = NULL,
  phase = NULL,
  runtime_handlers = NULL,
  volumes_attached = NULL,
  volumes_in_use = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
