% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NodeSystemInfo.R
\name{V1NodeSystemInfo}
\alias{V1NodeSystemInfo}
\title{V1NodeSystemInfo}
\description{
V1NodeSystemInfo

V1NodeSystemInfo
}
\details{
NodeSystemInfo is a set of ids/uuids to uniquely identify the node.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NodeSystemInfo-new}{\code{V1NodeSystemInfo$new()}}
\item \href{#method-V1NodeSystemInfo-to_list}{\code{V1NodeSystemInfo$to_list()}}
\item \href{#method-V1NodeSystemInfo-clone}{\code{V1NodeSystemInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSystemInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSystemInfo-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSystemInfo$new(
  architecture = NULL,
  boot_id = NULL,
  container_runtime_version = NULL,
  kernel_version = NULL,
  kube_proxy_version = NULL,
  kubelet_version = NULL,
  machine_id = NULL,
  operating_system = NULL,
  os_image = NULL,
  system_uuid = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSystemInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSystemInfo-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSystemInfo$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NodeSystemInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NodeSystemInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NodeSystemInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
