% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NonResourceAttributes.R
\name{V1NonResourceAttributes}
\alias{V1NonResourceAttributes}
\title{V1NonResourceAttributes}
\description{
V1NonResourceAttributes

V1NonResourceAttributes
}
\details{
NonResourceAttributes includes the authorization attributes available for non-resource requests to the Authorizer interface
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NonResourceAttributes-new}{\code{V1NonResourceAttributes$new()}}
\item \href{#method-V1NonResourceAttributes-to_list}{\code{V1NonResourceAttributes$to_list()}}
\item \href{#method-V1NonResourceAttributes-clone}{\code{V1NonResourceAttributes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourceAttributes-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourceAttributes-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourceAttributes$new(path = NULL, verb = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourceAttributes-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourceAttributes-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourceAttributes$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourceAttributes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourceAttributes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourceAttributes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
