% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NonResourcePolicyRule.R
\name{V1NonResourcePolicyRule}
\alias{V1NonResourcePolicyRule}
\title{V1NonResourcePolicyRule}
\description{
V1NonResourcePolicyRule

V1NonResourcePolicyRule
}
\details{
NonResourcePolicyRule is a predicate that matches non-resource requests according to their verb and the target non-resource URL. A NonResourcePolicyRule matches a request if and only if both (a) at least one member of verbs matches the request and (b) at least one member of nonResourceURLs matches the request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NonResourcePolicyRule-new}{\code{V1NonResourcePolicyRule$new()}}
\item \href{#method-V1NonResourcePolicyRule-to_list}{\code{V1NonResourcePolicyRule$to_list()}}
\item \href{#method-V1NonResourcePolicyRule-clone}{\code{V1NonResourcePolicyRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourcePolicyRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourcePolicyRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourcePolicyRule$new(non_resource_ur_ls = NULL, verbs = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourcePolicyRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourcePolicyRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourcePolicyRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourcePolicyRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourcePolicyRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourcePolicyRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
