% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1NonResourceRule.R
\name{V1NonResourceRule}
\alias{V1NonResourceRule}
\title{V1NonResourceRule}
\description{
V1NonResourceRule

V1NonResourceRule
}
\details{
NonResourceRule holds information that describes a rule for the non-resource
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1NonResourceRule-new}{\code{V1NonResourceRule$new()}}
\item \href{#method-V1NonResourceRule-to_list}{\code{V1NonResourceRule$to_list()}}
\item \href{#method-V1NonResourceRule-clone}{\code{V1NonResourceRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourceRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourceRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourceRule$new(non_resource_ur_ls = NULL, verbs = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourceRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourceRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourceRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1NonResourceRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1NonResourceRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1NonResourceRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
