% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ObjectFieldSelector.R
\name{V1ObjectFieldSelector}
\alias{V1ObjectFieldSelector}
\title{V1ObjectFieldSelector}
\description{
V1ObjectFieldSelector

V1ObjectFieldSelector
}
\details{
ObjectFieldSelector selects an APIVersioned field of an object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ObjectFieldSelector-new}{\code{V1ObjectFieldSelector$new()}}
\item \href{#method-V1ObjectFieldSelector-to_list}{\code{V1ObjectFieldSelector$to_list()}}
\item \href{#method-V1ObjectFieldSelector-clone}{\code{V1ObjectFieldSelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectFieldSelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectFieldSelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectFieldSelector$new(api_version = NULL, field_path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectFieldSelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectFieldSelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectFieldSelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectFieldSelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectFieldSelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectFieldSelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
