% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ObjectMeta.R
\name{V1ObjectMeta}
\alias{V1ObjectMeta}
\title{V1ObjectMeta}
\description{
V1ObjectMeta

V1ObjectMeta
}
\details{
ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ObjectMeta-new}{\code{V1ObjectMeta$new()}}
\item \href{#method-V1ObjectMeta-to_list}{\code{V1ObjectMeta$to_list()}}
\item \href{#method-V1ObjectMeta-clone}{\code{V1ObjectMeta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectMeta-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectMeta$new(
  annotations = NULL,
  creation_timestamp = NULL,
  deletion_grace_period_seconds = NULL,
  deletion_timestamp = NULL,
  finalizers = NULL,
  generate_name = NULL,
  generation = NULL,
  labels = NULL,
  managed_fields = NULL,
  name = NULL,
  namespace = NULL,
  owner_references = NULL,
  resource_version = NULL,
  self_link = NULL,
  uid = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectMeta-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectMeta-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectMeta$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
