% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ObjectReference.R
\name{V1ObjectReference}
\alias{V1ObjectReference}
\title{V1ObjectReference}
\description{
V1ObjectReference

V1ObjectReference
}
\details{
ObjectReference contains enough information to let you inspect or modify the referred object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ObjectReference-new}{\code{V1ObjectReference$new()}}
\item \href{#method-V1ObjectReference-to_list}{\code{V1ObjectReference$to_list()}}
\item \href{#method-V1ObjectReference-clone}{\code{V1ObjectReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectReference$new(
  api_version = NULL,
  field_path = NULL,
  kind = NULL,
  name = NULL,
  namespace = NULL,
  resource_version = NULL,
  uid = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ObjectReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ObjectReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ObjectReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
