% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1OwnerReference.R
\name{V1OwnerReference}
\alias{V1OwnerReference}
\title{V1OwnerReference}
\description{
V1OwnerReference

V1OwnerReference
}
\details{
OwnerReference contains enough information to let you identify an owning object. An owning object must be in the same namespace as the dependent, or be cluster-scoped, so there is no namespace field.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1OwnerReference-new}{\code{V1OwnerReference$new()}}
\item \href{#method-V1OwnerReference-to_list}{\code{V1OwnerReference$to_list()}}
\item \href{#method-V1OwnerReference-clone}{\code{V1OwnerReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1OwnerReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1OwnerReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1OwnerReference$new(
  api_version = NULL,
  block_owner_deletion = NULL,
  controller = NULL,
  kind = NULL,
  name = NULL,
  uid = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1OwnerReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1OwnerReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1OwnerReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1OwnerReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1OwnerReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1OwnerReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
