% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Patch.R
\name{V1Patch}
\alias{V1Patch}
\title{V1Patch}
\description{
V1Patch

V1Patch
}
\details{
Patch is provided to give a concrete name and type to the Kubernetes PATCH request body.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Patch-new}{\code{V1Patch$new()}}
\item \href{#method-V1Patch-to_list}{\code{V1Patch$to_list()}}
\item \href{#method-V1Patch-clone}{\code{V1Patch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Patch-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Patch-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Patch$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Patch-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Patch-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Patch$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Patch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Patch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Patch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
