% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolume.R
\name{V1PersistentVolume}
\alias{V1PersistentVolume}
\title{V1PersistentVolume}
\description{
V1PersistentVolume

V1PersistentVolume
}
\details{
PersistentVolume (PV) is a storage resource provisioned by an administrator. It is analogous to a node. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolume-new}{\code{V1PersistentVolume$new()}}
\item \href{#method-V1PersistentVolume-to_list}{\code{V1PersistentVolume$to_list()}}
\item \href{#method-V1PersistentVolume-clone}{\code{V1PersistentVolume$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolume-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolume-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolume$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolume-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolume-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolume$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolume-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolume-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolume$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
