% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeClaim.R
\name{V1PersistentVolumeClaim}
\alias{V1PersistentVolumeClaim}
\title{V1PersistentVolumeClaim}
\description{
V1PersistentVolumeClaim

V1PersistentVolumeClaim
}
\details{
PersistentVolumeClaim is a user's request for and claim to a persistent volume
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeClaim-new}{\code{V1PersistentVolumeClaim$new()}}
\item \href{#method-V1PersistentVolumeClaim-to_list}{\code{V1PersistentVolumeClaim$to_list()}}
\item \href{#method-V1PersistentVolumeClaim-clone}{\code{V1PersistentVolumeClaim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaim-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaim-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaim$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaim-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaim-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaim$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
