% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeClaimCondition.R
\name{V1PersistentVolumeClaimCondition}
\alias{V1PersistentVolumeClaimCondition}
\title{V1PersistentVolumeClaimCondition}
\description{
V1PersistentVolumeClaimCondition

V1PersistentVolumeClaimCondition
}
\details{
PersistentVolumeClaimCondition contains details about state of pvc
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeClaimCondition-new}{\code{V1PersistentVolumeClaimCondition$new()}}
\item \href{#method-V1PersistentVolumeClaimCondition-to_list}{\code{V1PersistentVolumeClaimCondition$to_list()}}
\item \href{#method-V1PersistentVolumeClaimCondition-clone}{\code{V1PersistentVolumeClaimCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimCondition$new(
  last_probe_time = NULL,
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
