% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeClaimList.R
\name{V1PersistentVolumeClaimList}
\alias{V1PersistentVolumeClaimList}
\title{V1PersistentVolumeClaimList}
\description{
V1PersistentVolumeClaimList

V1PersistentVolumeClaimList
}
\details{
PersistentVolumeClaimList is a list of PersistentVolumeClaim items.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeClaimList-new}{\code{V1PersistentVolumeClaimList$new()}}
\item \href{#method-V1PersistentVolumeClaimList-to_list}{\code{V1PersistentVolumeClaimList$to_list()}}
\item \href{#method-V1PersistentVolumeClaimList-clone}{\code{V1PersistentVolumeClaimList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
