% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeClaimSpec.R
\name{V1PersistentVolumeClaimSpec}
\alias{V1PersistentVolumeClaimSpec}
\title{V1PersistentVolumeClaimSpec}
\description{
V1PersistentVolumeClaimSpec

V1PersistentVolumeClaimSpec
}
\details{
PersistentVolumeClaimSpec describes the common attributes of storage devices and allows a Source for provider-specific attributes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeClaimSpec-new}{\code{V1PersistentVolumeClaimSpec$new()}}
\item \href{#method-V1PersistentVolumeClaimSpec-to_list}{\code{V1PersistentVolumeClaimSpec$to_list()}}
\item \href{#method-V1PersistentVolumeClaimSpec-clone}{\code{V1PersistentVolumeClaimSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimSpec$new(
  access_modes = NULL,
  data_source = NULL,
  data_source_ref = NULL,
  resources = NULL,
  selector = NULL,
  storage_class_name = NULL,
  volume_attributes_class_name = NULL,
  volume_mode = NULL,
  volume_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
