% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeClaimStatus.R
\name{V1PersistentVolumeClaimStatus}
\alias{V1PersistentVolumeClaimStatus}
\title{V1PersistentVolumeClaimStatus}
\description{
V1PersistentVolumeClaimStatus

V1PersistentVolumeClaimStatus
}
\details{
PersistentVolumeClaimStatus is the current status of a persistent volume claim.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeClaimStatus-new}{\code{V1PersistentVolumeClaimStatus$new()}}
\item \href{#method-V1PersistentVolumeClaimStatus-to_list}{\code{V1PersistentVolumeClaimStatus$to_list()}}
\item \href{#method-V1PersistentVolumeClaimStatus-clone}{\code{V1PersistentVolumeClaimStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimStatus$new(
  access_modes = NULL,
  allocated_resource_statuses = NULL,
  allocated_resources = NULL,
  capacity = NULL,
  conditions = NULL,
  current_volume_attributes_class_name = NULL,
  modify_volume_status = NULL,
  phase = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
