% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeClaimTemplate.R
\name{V1PersistentVolumeClaimTemplate}
\alias{V1PersistentVolumeClaimTemplate}
\title{V1PersistentVolumeClaimTemplate}
\description{
V1PersistentVolumeClaimTemplate

V1PersistentVolumeClaimTemplate
}
\details{
PersistentVolumeClaimTemplate is used to produce PersistentVolumeClaim objects as part of an EphemeralVolumeSource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeClaimTemplate-new}{\code{V1PersistentVolumeClaimTemplate$new()}}
\item \href{#method-V1PersistentVolumeClaimTemplate-to_list}{\code{V1PersistentVolumeClaimTemplate$to_list()}}
\item \href{#method-V1PersistentVolumeClaimTemplate-clone}{\code{V1PersistentVolumeClaimTemplate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimTemplate-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimTemplate-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimTemplate$new(metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimTemplate-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimTemplate-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimTemplate$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimTemplate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimTemplate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
