% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeClaimVolumeSource.R
\name{V1PersistentVolumeClaimVolumeSource}
\alias{V1PersistentVolumeClaimVolumeSource}
\title{V1PersistentVolumeClaimVolumeSource}
\description{
V1PersistentVolumeClaimVolumeSource

V1PersistentVolumeClaimVolumeSource
}
\details{
PersistentVolumeClaimVolumeSource references the user's PVC in the same namespace. This volume finds the bound PV and mounts that volume for the pod. A PersistentVolumeClaimVolumeSource is, essentially, a wrapper around another type of volume that is owned by someone else (the system).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeClaimVolumeSource-new}{\code{V1PersistentVolumeClaimVolumeSource$new()}}
\item \href{#method-V1PersistentVolumeClaimVolumeSource-to_list}{\code{V1PersistentVolumeClaimVolumeSource$to_list()}}
\item \href{#method-V1PersistentVolumeClaimVolumeSource-clone}{\code{V1PersistentVolumeClaimVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimVolumeSource$new(claim_name = NULL, read_only = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeClaimVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeClaimVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeClaimVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
