% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeList.R
\name{V1PersistentVolumeList}
\alias{V1PersistentVolumeList}
\title{V1PersistentVolumeList}
\description{
V1PersistentVolumeList

V1PersistentVolumeList
}
\details{
PersistentVolumeList is a list of PersistentVolume items.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeList-new}{\code{V1PersistentVolumeList$new()}}
\item \href{#method-V1PersistentVolumeList-to_list}{\code{V1PersistentVolumeList$to_list()}}
\item \href{#method-V1PersistentVolumeList-clone}{\code{V1PersistentVolumeList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
