% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeSpec.R
\name{V1PersistentVolumeSpec}
\alias{V1PersistentVolumeSpec}
\title{V1PersistentVolumeSpec}
\description{
V1PersistentVolumeSpec

V1PersistentVolumeSpec
}
\details{
PersistentVolumeSpec is the specification of a persistent volume.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeSpec-new}{\code{V1PersistentVolumeSpec$new()}}
\item \href{#method-V1PersistentVolumeSpec-to_list}{\code{V1PersistentVolumeSpec$to_list()}}
\item \href{#method-V1PersistentVolumeSpec-clone}{\code{V1PersistentVolumeSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeSpec$new(
  access_modes = NULL,
  aws_elastic_block_store = NULL,
  azure_disk = NULL,
  azure_file = NULL,
  capacity = NULL,
  cephfs = NULL,
  cinder = NULL,
  claim_ref = NULL,
  csi = NULL,
  fc = NULL,
  flex_volume = NULL,
  flocker = NULL,
  gce_persistent_disk = NULL,
  glusterfs = NULL,
  host_path = NULL,
  iscsi = NULL,
  local = NULL,
  mount_options = NULL,
  nfs = NULL,
  node_affinity = NULL,
  persistent_volume_reclaim_policy = NULL,
  photon_persistent_disk = NULL,
  portworx_volume = NULL,
  quobyte = NULL,
  rbd = NULL,
  scale_io = NULL,
  storage_class_name = NULL,
  storageos = NULL,
  volume_attributes_class_name = NULL,
  volume_mode = NULL,
  vsphere_volume = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
