% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PersistentVolumeStatus.R
\name{V1PersistentVolumeStatus}
\alias{V1PersistentVolumeStatus}
\title{V1PersistentVolumeStatus}
\description{
V1PersistentVolumeStatus

V1PersistentVolumeStatus
}
\details{
PersistentVolumeStatus is the current status of a persistent volume.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PersistentVolumeStatus-new}{\code{V1PersistentVolumeStatus$new()}}
\item \href{#method-V1PersistentVolumeStatus-to_list}{\code{V1PersistentVolumeStatus$to_list()}}
\item \href{#method-V1PersistentVolumeStatus-clone}{\code{V1PersistentVolumeStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeStatus$new(
  last_phase_transition_time = NULL,
  message = NULL,
  phase = NULL,
  reason = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PersistentVolumeStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PersistentVolumeStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PersistentVolumeStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
