% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PhotonPersistentDiskVolumeSource.R
\name{V1PhotonPersistentDiskVolumeSource}
\alias{V1PhotonPersistentDiskVolumeSource}
\title{V1PhotonPersistentDiskVolumeSource}
\description{
V1PhotonPersistentDiskVolumeSource

V1PhotonPersistentDiskVolumeSource
}
\details{
Represents a Photon Controller persistent disk resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PhotonPersistentDiskVolumeSource-new}{\code{V1PhotonPersistentDiskVolumeSource$new()}}
\item \href{#method-V1PhotonPersistentDiskVolumeSource-to_list}{\code{V1PhotonPersistentDiskVolumeSource$to_list()}}
\item \href{#method-V1PhotonPersistentDiskVolumeSource-clone}{\code{V1PhotonPersistentDiskVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PhotonPersistentDiskVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PhotonPersistentDiskVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PhotonPersistentDiskVolumeSource$new(fs_type = NULL, pd_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PhotonPersistentDiskVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PhotonPersistentDiskVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PhotonPersistentDiskVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PhotonPersistentDiskVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PhotonPersistentDiskVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PhotonPersistentDiskVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
