% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Pod.R
\name{V1Pod}
\alias{V1Pod}
\title{V1Pod}
\description{
V1Pod

V1Pod
}
\details{
Pod is a collection of containers that can run on a host. This resource is created by clients and scheduled onto hosts.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Pod-new}{\code{V1Pod$new()}}
\item \href{#method-V1Pod-to_list}{\code{V1Pod$to_list()}}
\item \href{#method-V1Pod-clone}{\code{V1Pod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Pod-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Pod-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Pod$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Pod-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Pod-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Pod$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Pod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Pod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Pod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
