% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodAffinity.R
\name{V1PodAffinity}
\alias{V1PodAffinity}
\title{V1PodAffinity}
\description{
V1PodAffinity

V1PodAffinity
}
\details{
Pod affinity is a group of inter pod affinity scheduling rules.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodAffinity-new}{\code{V1PodAffinity$new()}}
\item \href{#method-V1PodAffinity-to_list}{\code{V1PodAffinity$to_list()}}
\item \href{#method-V1PodAffinity-clone}{\code{V1PodAffinity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAffinity-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAffinity-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAffinity$new(
  preferred_during_scheduling_ignored_during_execution = NULL,
  required_during_scheduling_ignored_during_execution = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAffinity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAffinity-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAffinity$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAffinity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAffinity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAffinity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
