% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodAffinityTerm.R
\name{V1PodAffinityTerm}
\alias{V1PodAffinityTerm}
\title{V1PodAffinityTerm}
\description{
V1PodAffinityTerm

V1PodAffinityTerm
}
\details{
Defines a set of pods (namely those matching the labelSelector relative to the given namespace(s)) that this pod should be co-located (affinity) or not co-located (anti-affinity) with, where co-located is defined as running on a node whose value of the label with key <topologyKey> matches that of any node on which a pod of the set of pods is running
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodAffinityTerm-new}{\code{V1PodAffinityTerm$new()}}
\item \href{#method-V1PodAffinityTerm-to_list}{\code{V1PodAffinityTerm$to_list()}}
\item \href{#method-V1PodAffinityTerm-clone}{\code{V1PodAffinityTerm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAffinityTerm-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAffinityTerm-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAffinityTerm$new(
  label_selector = NULL,
  match_label_keys = NULL,
  mismatch_label_keys = NULL,
  namespace_selector = NULL,
  namespaces = NULL,
  topology_key = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAffinityTerm-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAffinityTerm-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAffinityTerm$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAffinityTerm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAffinityTerm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAffinityTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
