% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodAntiAffinity.R
\name{V1PodAntiAffinity}
\alias{V1PodAntiAffinity}
\title{V1PodAntiAffinity}
\description{
V1PodAntiAffinity

V1PodAntiAffinity
}
\details{
Pod anti affinity is a group of inter pod anti affinity scheduling rules.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodAntiAffinity-new}{\code{V1PodAntiAffinity$new()}}
\item \href{#method-V1PodAntiAffinity-to_list}{\code{V1PodAntiAffinity$to_list()}}
\item \href{#method-V1PodAntiAffinity-clone}{\code{V1PodAntiAffinity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAntiAffinity-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAntiAffinity-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAntiAffinity$new(
  preferred_during_scheduling_ignored_during_execution = NULL,
  required_during_scheduling_ignored_during_execution = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAntiAffinity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAntiAffinity-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAntiAffinity$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodAntiAffinity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodAntiAffinity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodAntiAffinity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
