% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodCondition.R
\name{V1PodCondition}
\alias{V1PodCondition}
\title{V1PodCondition}
\description{
V1PodCondition

V1PodCondition
}
\details{
PodCondition contains details for the current condition of this pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodCondition-new}{\code{V1PodCondition$new()}}
\item \href{#method-V1PodCondition-to_list}{\code{V1PodCondition$to_list()}}
\item \href{#method-V1PodCondition-clone}{\code{V1PodCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodCondition$new(
  last_probe_time = NULL,
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
