% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodDisruptionBudget.R
\name{V1PodDisruptionBudget}
\alias{V1PodDisruptionBudget}
\title{V1PodDisruptionBudget}
\description{
V1PodDisruptionBudget

V1PodDisruptionBudget
}
\details{
PodDisruptionBudget is an object to define the max disruption that can be caused to a collection of pods
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodDisruptionBudget-new}{\code{V1PodDisruptionBudget$new()}}
\item \href{#method-V1PodDisruptionBudget-to_list}{\code{V1PodDisruptionBudget$to_list()}}
\item \href{#method-V1PodDisruptionBudget-clone}{\code{V1PodDisruptionBudget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudget-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudget-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudget$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudget-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudget-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudget$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
