% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodDisruptionBudgetList.R
\name{V1PodDisruptionBudgetList}
\alias{V1PodDisruptionBudgetList}
\title{V1PodDisruptionBudgetList}
\description{
V1PodDisruptionBudgetList

V1PodDisruptionBudgetList
}
\details{
PodDisruptionBudgetList is a collection of PodDisruptionBudgets.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodDisruptionBudgetList-new}{\code{V1PodDisruptionBudgetList$new()}}
\item \href{#method-V1PodDisruptionBudgetList-to_list}{\code{V1PodDisruptionBudgetList$to_list()}}
\item \href{#method-V1PodDisruptionBudgetList-clone}{\code{V1PodDisruptionBudgetList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
