% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodDisruptionBudgetSpec.R
\name{V1PodDisruptionBudgetSpec}
\alias{V1PodDisruptionBudgetSpec}
\title{V1PodDisruptionBudgetSpec}
\description{
V1PodDisruptionBudgetSpec

V1PodDisruptionBudgetSpec
}
\details{
PodDisruptionBudgetSpec is a description of a PodDisruptionBudget.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodDisruptionBudgetSpec-new}{\code{V1PodDisruptionBudgetSpec$new()}}
\item \href{#method-V1PodDisruptionBudgetSpec-to_list}{\code{V1PodDisruptionBudgetSpec$to_list()}}
\item \href{#method-V1PodDisruptionBudgetSpec-clone}{\code{V1PodDisruptionBudgetSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetSpec$new(
  max_unavailable = NULL,
  min_available = NULL,
  selector = NULL,
  unhealthy_pod_eviction_policy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
