% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodDisruptionBudgetStatus.R
\name{V1PodDisruptionBudgetStatus}
\alias{V1PodDisruptionBudgetStatus}
\title{V1PodDisruptionBudgetStatus}
\description{
V1PodDisruptionBudgetStatus

V1PodDisruptionBudgetStatus
}
\details{
PodDisruptionBudgetStatus represents information about the status of a PodDisruptionBudget. Status may trail the actual state of a system.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodDisruptionBudgetStatus-new}{\code{V1PodDisruptionBudgetStatus$new()}}
\item \href{#method-V1PodDisruptionBudgetStatus-to_list}{\code{V1PodDisruptionBudgetStatus$to_list()}}
\item \href{#method-V1PodDisruptionBudgetStatus-clone}{\code{V1PodDisruptionBudgetStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetStatus$new(
  conditions = NULL,
  current_healthy = NULL,
  desired_healthy = NULL,
  disrupted_pods = NULL,
  disruptions_allowed = NULL,
  expected_pods = NULL,
  observed_generation = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDisruptionBudgetStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDisruptionBudgetStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDisruptionBudgetStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
