% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodDNSConfig.R
\name{V1PodDNSConfig}
\alias{V1PodDNSConfig}
\title{V1PodDNSConfig}
\description{
V1PodDNSConfig

V1PodDNSConfig
}
\details{
PodDNSConfig defines the DNS parameters of a pod in addition to those generated from DNSPolicy.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodDNSConfig-new}{\code{V1PodDNSConfig$new()}}
\item \href{#method-V1PodDNSConfig-to_list}{\code{V1PodDNSConfig$to_list()}}
\item \href{#method-V1PodDNSConfig-clone}{\code{V1PodDNSConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDNSConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDNSConfig-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDNSConfig$new(nameservers = NULL, options = NULL, searches = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDNSConfig-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDNSConfig-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDNSConfig$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodDNSConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodDNSConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodDNSConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
