% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodFailurePolicy.R
\name{V1PodFailurePolicy}
\alias{V1PodFailurePolicy}
\title{V1PodFailurePolicy}
\description{
V1PodFailurePolicy

V1PodFailurePolicy
}
\details{
PodFailurePolicy describes how failed pods influence the backoffLimit.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodFailurePolicy-new}{\code{V1PodFailurePolicy$new()}}
\item \href{#method-V1PodFailurePolicy-to_list}{\code{V1PodFailurePolicy$to_list()}}
\item \href{#method-V1PodFailurePolicy-clone}{\code{V1PodFailurePolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicy$new(rules = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
