% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1PodFailurePolicyOnExitCodesRequirement.R
\name{V1PodFailurePolicyOnExitCodesRequirement}
\alias{V1PodFailurePolicyOnExitCodesRequirement}
\title{V1PodFailurePolicyOnExitCodesRequirement}
\description{
V1PodFailurePolicyOnExitCodesRequirement

V1PodFailurePolicyOnExitCodesRequirement
}
\details{
PodFailurePolicyOnExitCodesRequirement describes the requirement for handling a failed pod based on its container exit codes. In particular, it lookups the .state.terminated.exitCode for each app container and init container status, represented by the .status.containerStatuses and .status.initContainerStatuses fields in the Pod status, respectively. Containers completed with success (exit code 0) are excluded from the requirement check.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodFailurePolicyOnExitCodesRequirement-new}{\code{V1PodFailurePolicyOnExitCodesRequirement$new()}}
\item \href{#method-V1PodFailurePolicyOnExitCodesRequirement-to_list}{\code{V1PodFailurePolicyOnExitCodesRequirement$to_list()}}
\item \href{#method-V1PodFailurePolicyOnExitCodesRequirement-clone}{\code{V1PodFailurePolicyOnExitCodesRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyOnExitCodesRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyOnExitCodesRequirement-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyOnExitCodesRequirement$new(
  container_name = NULL,
  operator = NULL,
  values = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyOnExitCodesRequirement-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyOnExitCodesRequirement-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyOnExitCodesRequirement$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyOnExitCodesRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyOnExitCodesRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyOnExitCodesRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
