% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1PodFailurePolicyOnPodConditionsPattern.R
\name{V1PodFailurePolicyOnPodConditionsPattern}
\alias{V1PodFailurePolicyOnPodConditionsPattern}
\title{V1PodFailurePolicyOnPodConditionsPattern}
\description{
V1PodFailurePolicyOnPodConditionsPattern

V1PodFailurePolicyOnPodConditionsPattern
}
\details{
PodFailurePolicyOnPodConditionsPattern describes a pattern for matching an actual pod condition type.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodFailurePolicyOnPodConditionsPattern-new}{\code{V1PodFailurePolicyOnPodConditionsPattern$new()}}
\item \href{#method-V1PodFailurePolicyOnPodConditionsPattern-to_list}{\code{V1PodFailurePolicyOnPodConditionsPattern$to_list()}}
\item \href{#method-V1PodFailurePolicyOnPodConditionsPattern-clone}{\code{V1PodFailurePolicyOnPodConditionsPattern$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyOnPodConditionsPattern-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyOnPodConditionsPattern-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyOnPodConditionsPattern$new(status = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyOnPodConditionsPattern-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyOnPodConditionsPattern-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyOnPodConditionsPattern$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyOnPodConditionsPattern-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyOnPodConditionsPattern-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyOnPodConditionsPattern$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
