% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodFailurePolicyRule.R
\name{V1PodFailurePolicyRule}
\alias{V1PodFailurePolicyRule}
\title{V1PodFailurePolicyRule}
\description{
V1PodFailurePolicyRule

V1PodFailurePolicyRule
}
\details{
PodFailurePolicyRule describes how a pod failure is handled when the requirements are met. One of onExitCodes and onPodConditions, but not both, can be used in each rule.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodFailurePolicyRule-new}{\code{V1PodFailurePolicyRule$new()}}
\item \href{#method-V1PodFailurePolicyRule-to_list}{\code{V1PodFailurePolicyRule$to_list()}}
\item \href{#method-V1PodFailurePolicyRule-clone}{\code{V1PodFailurePolicyRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyRule$new(
  action = NULL,
  on_exit_codes = NULL,
  on_pod_conditions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodFailurePolicyRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodFailurePolicyRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodFailurePolicyRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
