% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodList.R
\name{V1PodList}
\alias{V1PodList}
\title{V1PodList}
\description{
V1PodList

V1PodList
}
\details{
PodList is a list of Pods.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodList-new}{\code{V1PodList$new()}}
\item \href{#method-V1PodList-to_list}{\code{V1PodList$to_list()}}
\item \href{#method-V1PodList-clone}{\code{V1PodList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodList$new(api_version = NULL, items = NULL, kind = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
