% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodOS.R
\name{V1PodOS}
\alias{V1PodOS}
\title{V1PodOS}
\description{
V1PodOS

V1PodOS
}
\details{
PodOS defines the OS parameters of a pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodOS-new}{\code{V1PodOS$new()}}
\item \href{#method-V1PodOS-to_list}{\code{V1PodOS$to_list()}}
\item \href{#method-V1PodOS-clone}{\code{V1PodOS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodOS-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodOS-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodOS$new(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodOS-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodOS-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodOS$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodOS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodOS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodOS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
