% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodReadinessGate.R
\name{V1PodReadinessGate}
\alias{V1PodReadinessGate}
\title{V1PodReadinessGate}
\description{
V1PodReadinessGate

V1PodReadinessGate
}
\details{
PodReadinessGate contains the reference to a pod condition
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodReadinessGate-new}{\code{V1PodReadinessGate$new()}}
\item \href{#method-V1PodReadinessGate-to_list}{\code{V1PodReadinessGate$to_list()}}
\item \href{#method-V1PodReadinessGate-clone}{\code{V1PodReadinessGate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodReadinessGate-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodReadinessGate-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodReadinessGate$new(condition_type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodReadinessGate-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodReadinessGate-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodReadinessGate$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodReadinessGate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodReadinessGate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodReadinessGate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
