% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodResourceClaim.R
\name{V1PodResourceClaim}
\alias{V1PodResourceClaim}
\title{V1PodResourceClaim}
\description{
V1PodResourceClaim

V1PodResourceClaim
}
\details{
PodResourceClaim references exactly one ResourceClaim, either directly or by naming a ResourceClaimTemplate which is then turned into a ResourceClaim for the pod.

It adds a name to it that uniquely identifies the ResourceClaim inside the Pod. Containers that need access to the ResourceClaim reference it with this name.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodResourceClaim-new}{\code{V1PodResourceClaim$new()}}
\item \href{#method-V1PodResourceClaim-to_list}{\code{V1PodResourceClaim$to_list()}}
\item \href{#method-V1PodResourceClaim-clone}{\code{V1PodResourceClaim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodResourceClaim-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodResourceClaim-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodResourceClaim$new(
  name = NULL,
  resource_claim_name = NULL,
  resource_claim_template_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodResourceClaim-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodResourceClaim-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodResourceClaim$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodResourceClaim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodResourceClaim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodResourceClaim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
