% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodResourceClaimStatus.R
\name{V1PodResourceClaimStatus}
\alias{V1PodResourceClaimStatus}
\title{V1PodResourceClaimStatus}
\description{
V1PodResourceClaimStatus

V1PodResourceClaimStatus
}
\details{
PodResourceClaimStatus is stored in the PodStatus for each PodResourceClaim which references a ResourceClaimTemplate. It stores the generated name for the corresponding ResourceClaim.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodResourceClaimStatus-new}{\code{V1PodResourceClaimStatus$new()}}
\item \href{#method-V1PodResourceClaimStatus-to_list}{\code{V1PodResourceClaimStatus$to_list()}}
\item \href{#method-V1PodResourceClaimStatus-clone}{\code{V1PodResourceClaimStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodResourceClaimStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodResourceClaimStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodResourceClaimStatus$new(name = NULL, resource_claim_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodResourceClaimStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodResourceClaimStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodResourceClaimStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodResourceClaimStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodResourceClaimStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodResourceClaimStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
