% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodSchedulingGate.R
\name{V1PodSchedulingGate}
\alias{V1PodSchedulingGate}
\title{V1PodSchedulingGate}
\description{
V1PodSchedulingGate

V1PodSchedulingGate
}
\details{
PodSchedulingGate is associated to a Pod to guard its scheduling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodSchedulingGate-new}{\code{V1PodSchedulingGate$new()}}
\item \href{#method-V1PodSchedulingGate-to_list}{\code{V1PodSchedulingGate$to_list()}}
\item \href{#method-V1PodSchedulingGate-clone}{\code{V1PodSchedulingGate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSchedulingGate-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSchedulingGate-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSchedulingGate$new(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSchedulingGate-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSchedulingGate-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSchedulingGate$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSchedulingGate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSchedulingGate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSchedulingGate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
