% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodSecurityContext.R
\name{V1PodSecurityContext}
\alias{V1PodSecurityContext}
\title{V1PodSecurityContext}
\description{
V1PodSecurityContext

V1PodSecurityContext
}
\details{
PodSecurityContext holds pod-level security attributes and common container settings. Some fields are also present in container.securityContext.  Field values of container.securityContext take precedence over field values of PodSecurityContext.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodSecurityContext-new}{\code{V1PodSecurityContext$new()}}
\item \href{#method-V1PodSecurityContext-to_list}{\code{V1PodSecurityContext$to_list()}}
\item \href{#method-V1PodSecurityContext-clone}{\code{V1PodSecurityContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSecurityContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSecurityContext-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSecurityContext$new(
  app_armor_profile = NULL,
  fs_group = NULL,
  fs_group_change_policy = NULL,
  run_as_group = NULL,
  run_as_non_root = NULL,
  run_as_user = NULL,
  se_linux_options = NULL,
  seccomp_profile = NULL,
  supplemental_groups = NULL,
  supplemental_groups_policy = NULL,
  sysctls = NULL,
  windows_options = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSecurityContext-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSecurityContext-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSecurityContext$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSecurityContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSecurityContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSecurityContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
