% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodSpec.R
\name{V1PodSpec}
\alias{V1PodSpec}
\title{V1PodSpec}
\description{
V1PodSpec

V1PodSpec
}
\details{
PodSpec is a description of a pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodSpec-new}{\code{V1PodSpec$new()}}
\item \href{#method-V1PodSpec-to_list}{\code{V1PodSpec$to_list()}}
\item \href{#method-V1PodSpec-clone}{\code{V1PodSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSpec$new(
  active_deadline_seconds = NULL,
  affinity = NULL,
  automount_service_account_token = NULL,
  containers = NULL,
  dns_config = NULL,
  dns_policy = NULL,
  enable_service_links = NULL,
  ephemeral_containers = NULL,
  host_aliases = NULL,
  host_ipc = NULL,
  host_network = NULL,
  host_pid = NULL,
  host_users = NULL,
  hostname = NULL,
  image_pull_secrets = NULL,
  init_containers = NULL,
  node_name = NULL,
  node_selector = NULL,
  os = NULL,
  overhead = NULL,
  preemption_policy = NULL,
  priority = NULL,
  priority_class_name = NULL,
  readiness_gates = NULL,
  resource_claims = NULL,
  restart_policy = NULL,
  runtime_class_name = NULL,
  scheduler_name = NULL,
  scheduling_gates = NULL,
  security_context = NULL,
  service_account = NULL,
  service_account_name = NULL,
  set_hostname_as_fqdn = NULL,
  share_process_namespace = NULL,
  subdomain = NULL,
  termination_grace_period_seconds = NULL,
  tolerations = NULL,
  topology_spread_constraints = NULL,
  volumes = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
