% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodStatus.R
\name{V1PodStatus}
\alias{V1PodStatus}
\title{V1PodStatus}
\description{
V1PodStatus

V1PodStatus
}
\details{
PodStatus represents information about the status of a pod. Status may trail the actual state of a system, especially if the node that hosts the pod cannot contact the control plane.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodStatus-new}{\code{V1PodStatus$new()}}
\item \href{#method-V1PodStatus-to_list}{\code{V1PodStatus$to_list()}}
\item \href{#method-V1PodStatus-clone}{\code{V1PodStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodStatus$new(
  conditions = NULL,
  container_statuses = NULL,
  ephemeral_container_statuses = NULL,
  host_ip = NULL,
  host_i_ps = NULL,
  init_container_statuses = NULL,
  message = NULL,
  nominated_node_name = NULL,
  phase = NULL,
  pod_ip = NULL,
  pod_i_ps = NULL,
  qos_class = NULL,
  reason = NULL,
  resize = NULL,
  resource_claim_statuses = NULL,
  start_time = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
