% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodTemplate.R
\name{V1PodTemplate}
\alias{V1PodTemplate}
\title{V1PodTemplate}
\description{
V1PodTemplate

V1PodTemplate
}
\details{
PodTemplate describes a template for creating copies of a predefined pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodTemplate-new}{\code{V1PodTemplate$new()}}
\item \href{#method-V1PodTemplate-to_list}{\code{V1PodTemplate$to_list()}}
\item \href{#method-V1PodTemplate-clone}{\code{V1PodTemplate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplate-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplate-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplate$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  template = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplate-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplate-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplate$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
