% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodTemplateList.R
\name{V1PodTemplateList}
\alias{V1PodTemplateList}
\title{V1PodTemplateList}
\description{
V1PodTemplateList

V1PodTemplateList
}
\details{
PodTemplateList is a list of PodTemplates.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodTemplateList-new}{\code{V1PodTemplateList$new()}}
\item \href{#method-V1PodTemplateList-to_list}{\code{V1PodTemplateList$to_list()}}
\item \href{#method-V1PodTemplateList-clone}{\code{V1PodTemplateList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplateList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplateList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplateList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplateList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplateList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplateList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplateList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplateList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplateList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
