% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PodTemplateSpec.R
\name{V1PodTemplateSpec}
\alias{V1PodTemplateSpec}
\title{V1PodTemplateSpec}
\description{
V1PodTemplateSpec

V1PodTemplateSpec
}
\details{
PodTemplateSpec describes the data a pod should have when created from a template
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PodTemplateSpec-new}{\code{V1PodTemplateSpec$new()}}
\item \href{#method-V1PodTemplateSpec-to_list}{\code{V1PodTemplateSpec$to_list()}}
\item \href{#method-V1PodTemplateSpec-clone}{\code{V1PodTemplateSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplateSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplateSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplateSpec$new(metadata = NULL, spec = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplateSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplateSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplateSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PodTemplateSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PodTemplateSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PodTemplateSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
