% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PolicyRule.R
\name{V1PolicyRule}
\alias{V1PolicyRule}
\title{V1PolicyRule}
\description{
V1PolicyRule

V1PolicyRule
}
\details{
PolicyRule holds information that describes a policy rule, but does not contain information about who the rule applies to or which namespace the rule applies to.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PolicyRule-new}{\code{V1PolicyRule$new()}}
\item \href{#method-V1PolicyRule-to_list}{\code{V1PolicyRule$to_list()}}
\item \href{#method-V1PolicyRule-clone}{\code{V1PolicyRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PolicyRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PolicyRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PolicyRule$new(
  api_groups = NULL,
  non_resource_ur_ls = NULL,
  resource_names = NULL,
  resources = NULL,
  verbs = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PolicyRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PolicyRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PolicyRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PolicyRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PolicyRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PolicyRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
