% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PolicyRulesWithSubjects.R
\name{V1PolicyRulesWithSubjects}
\alias{V1PolicyRulesWithSubjects}
\title{V1PolicyRulesWithSubjects}
\description{
V1PolicyRulesWithSubjects

V1PolicyRulesWithSubjects
}
\details{
PolicyRulesWithSubjects prescribes a test that applies to a request to an apiserver. The test considers the subject making the request, the verb being requested, and the resource to be acted upon. This PolicyRulesWithSubjects matches a request if and only if both (a) at least one member of subjects matches the request and (b) at least one member of resourceRules or nonResourceRules matches the request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PolicyRulesWithSubjects-new}{\code{V1PolicyRulesWithSubjects$new()}}
\item \href{#method-V1PolicyRulesWithSubjects-to_list}{\code{V1PolicyRulesWithSubjects$to_list()}}
\item \href{#method-V1PolicyRulesWithSubjects-clone}{\code{V1PolicyRulesWithSubjects$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PolicyRulesWithSubjects-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PolicyRulesWithSubjects-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PolicyRulesWithSubjects$new(
  non_resource_rules = NULL,
  resource_rules = NULL,
  subjects = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PolicyRulesWithSubjects-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PolicyRulesWithSubjects-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PolicyRulesWithSubjects$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PolicyRulesWithSubjects-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PolicyRulesWithSubjects-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PolicyRulesWithSubjects$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
