% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PortStatus.R
\name{V1PortStatus}
\alias{V1PortStatus}
\title{V1PortStatus}
\description{
V1PortStatus

V1PortStatus
}
\details{
(no description)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PortStatus-new}{\code{V1PortStatus$new()}}
\item \href{#method-V1PortStatus-to_list}{\code{V1PortStatus$to_list()}}
\item \href{#method-V1PortStatus-clone}{\code{V1PortStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PortStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PortStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PortStatus$new(error = NULL, port = NULL, protocol = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PortStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PortStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PortStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PortStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PortStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PortStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
