% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Preconditions.R
\name{V1Preconditions}
\alias{V1Preconditions}
\title{V1Preconditions}
\description{
V1Preconditions

V1Preconditions
}
\details{
Preconditions must be fulfilled before an operation (update, delete, etc.) is carried out.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Preconditions-new}{\code{V1Preconditions$new()}}
\item \href{#method-V1Preconditions-to_list}{\code{V1Preconditions$to_list()}}
\item \href{#method-V1Preconditions-clone}{\code{V1Preconditions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Preconditions-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Preconditions-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Preconditions$new(resource_version = NULL, uid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Preconditions-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Preconditions-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Preconditions$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Preconditions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Preconditions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Preconditions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
