% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PreferredSchedulingTerm.R
\name{V1PreferredSchedulingTerm}
\alias{V1PreferredSchedulingTerm}
\title{V1PreferredSchedulingTerm}
\description{
V1PreferredSchedulingTerm

V1PreferredSchedulingTerm
}
\details{
An empty preferred scheduling term matches all objects with implicit weight 0 (i.e. it's a no-op). A null preferred scheduling term matches no objects (i.e. is also a no-op).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PreferredSchedulingTerm-new}{\code{V1PreferredSchedulingTerm$new()}}
\item \href{#method-V1PreferredSchedulingTerm-to_list}{\code{V1PreferredSchedulingTerm$to_list()}}
\item \href{#method-V1PreferredSchedulingTerm-clone}{\code{V1PreferredSchedulingTerm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PreferredSchedulingTerm-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PreferredSchedulingTerm-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PreferredSchedulingTerm$new(preference = NULL, weight = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PreferredSchedulingTerm-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PreferredSchedulingTerm-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PreferredSchedulingTerm$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PreferredSchedulingTerm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PreferredSchedulingTerm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PreferredSchedulingTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
