% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PriorityClass.R
\name{V1PriorityClass}
\alias{V1PriorityClass}
\title{V1PriorityClass}
\description{
V1PriorityClass

V1PriorityClass
}
\details{
PriorityClass defines mapping from a priority class name to the priority integer value. The value can be any valid integer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityClass-new}{\code{V1PriorityClass$new()}}
\item \href{#method-V1PriorityClass-to_list}{\code{V1PriorityClass$to_list()}}
\item \href{#method-V1PriorityClass-clone}{\code{V1PriorityClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityClass-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityClass$new(
  api_version = NULL,
  description = NULL,
  global_default = NULL,
  kind = NULL,
  metadata = NULL,
  preemption_policy = NULL,
  value = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityClass-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityClass-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityClass$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
