% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PriorityClassList.R
\name{V1PriorityClassList}
\alias{V1PriorityClassList}
\title{V1PriorityClassList}
\description{
V1PriorityClassList

V1PriorityClassList
}
\details{
PriorityClassList is a collection of priority classes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityClassList-new}{\code{V1PriorityClassList$new()}}
\item \href{#method-V1PriorityClassList-to_list}{\code{V1PriorityClassList$to_list()}}
\item \href{#method-V1PriorityClassList-clone}{\code{V1PriorityClassList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityClassList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityClassList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityClassList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityClassList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityClassList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityClassList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityClassList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityClassList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityClassList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
