% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PriorityLevelConfiguration.R
\name{V1PriorityLevelConfiguration}
\alias{V1PriorityLevelConfiguration}
\title{V1PriorityLevelConfiguration}
\description{
V1PriorityLevelConfiguration

V1PriorityLevelConfiguration
}
\details{
PriorityLevelConfiguration represents the configuration of a priority level.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityLevelConfiguration-new}{\code{V1PriorityLevelConfiguration$new()}}
\item \href{#method-V1PriorityLevelConfiguration-to_list}{\code{V1PriorityLevelConfiguration$to_list()}}
\item \href{#method-V1PriorityLevelConfiguration-clone}{\code{V1PriorityLevelConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfiguration$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
