% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1PriorityLevelConfigurationCondition.R
\name{V1PriorityLevelConfigurationCondition}
\alias{V1PriorityLevelConfigurationCondition}
\title{V1PriorityLevelConfigurationCondition}
\description{
V1PriorityLevelConfigurationCondition

V1PriorityLevelConfigurationCondition
}
\details{
PriorityLevelConfigurationCondition defines the condition of priority level.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityLevelConfigurationCondition-new}{\code{V1PriorityLevelConfigurationCondition$new()}}
\item \href{#method-V1PriorityLevelConfigurationCondition-to_list}{\code{V1PriorityLevelConfigurationCondition$to_list()}}
\item \href{#method-V1PriorityLevelConfigurationCondition-clone}{\code{V1PriorityLevelConfigurationCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
