% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PriorityLevelConfigurationList.R
\name{V1PriorityLevelConfigurationList}
\alias{V1PriorityLevelConfigurationList}
\title{V1PriorityLevelConfigurationList}
\description{
V1PriorityLevelConfigurationList

V1PriorityLevelConfigurationList
}
\details{
PriorityLevelConfigurationList is a list of PriorityLevelConfiguration objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityLevelConfigurationList-new}{\code{V1PriorityLevelConfigurationList$new()}}
\item \href{#method-V1PriorityLevelConfigurationList-to_list}{\code{V1PriorityLevelConfigurationList$to_list()}}
\item \href{#method-V1PriorityLevelConfigurationList-clone}{\code{V1PriorityLevelConfigurationList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
