% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1PriorityLevelConfigurationReference.R
\name{V1PriorityLevelConfigurationReference}
\alias{V1PriorityLevelConfigurationReference}
\title{V1PriorityLevelConfigurationReference}
\description{
V1PriorityLevelConfigurationReference

V1PriorityLevelConfigurationReference
}
\details{
PriorityLevelConfigurationReference contains information that points to the "request-priority" being used.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityLevelConfigurationReference-new}{\code{V1PriorityLevelConfigurationReference$new()}}
\item \href{#method-V1PriorityLevelConfigurationReference-to_list}{\code{V1PriorityLevelConfigurationReference$to_list()}}
\item \href{#method-V1PriorityLevelConfigurationReference-clone}{\code{V1PriorityLevelConfigurationReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationReference$new(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
