% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PriorityLevelConfigurationSpec.R
\name{V1PriorityLevelConfigurationSpec}
\alias{V1PriorityLevelConfigurationSpec}
\title{V1PriorityLevelConfigurationSpec}
\description{
V1PriorityLevelConfigurationSpec

V1PriorityLevelConfigurationSpec
}
\details{
PriorityLevelConfigurationSpec specifies the configuration of a priority level.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityLevelConfigurationSpec-new}{\code{V1PriorityLevelConfigurationSpec$new()}}
\item \href{#method-V1PriorityLevelConfigurationSpec-to_list}{\code{V1PriorityLevelConfigurationSpec$to_list()}}
\item \href{#method-V1PriorityLevelConfigurationSpec-clone}{\code{V1PriorityLevelConfigurationSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationSpec$new(
  exempt = NULL,
  limited = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
