% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1PriorityLevelConfigurationStatus.R
\name{V1PriorityLevelConfigurationStatus}
\alias{V1PriorityLevelConfigurationStatus}
\title{V1PriorityLevelConfigurationStatus}
\description{
V1PriorityLevelConfigurationStatus

V1PriorityLevelConfigurationStatus
}
\details{
PriorityLevelConfigurationStatus represents the current state of a "request-priority".
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1PriorityLevelConfigurationStatus-new}{\code{V1PriorityLevelConfigurationStatus$new()}}
\item \href{#method-V1PriorityLevelConfigurationStatus-to_list}{\code{V1PriorityLevelConfigurationStatus$to_list()}}
\item \href{#method-V1PriorityLevelConfigurationStatus-clone}{\code{V1PriorityLevelConfigurationStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationStatus$new(conditions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1PriorityLevelConfigurationStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1PriorityLevelConfigurationStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1PriorityLevelConfigurationStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
