% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Probe.R
\name{V1Probe}
\alias{V1Probe}
\title{V1Probe}
\description{
V1Probe

V1Probe
}
\details{
Probe describes a health check to be performed against a container to determine whether it is alive or ready to receive traffic.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Probe-new}{\code{V1Probe$new()}}
\item \href{#method-V1Probe-to_list}{\code{V1Probe$to_list()}}
\item \href{#method-V1Probe-clone}{\code{V1Probe$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Probe-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Probe-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Probe$new(
  exec = NULL,
  failure_threshold = NULL,
  grpc = NULL,
  http_get = NULL,
  initial_delay_seconds = NULL,
  period_seconds = NULL,
  success_threshold = NULL,
  tcp_socket = NULL,
  termination_grace_period_seconds = NULL,
  timeout_seconds = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Probe-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Probe-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Probe$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Probe-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Probe-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Probe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
