% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1QueuingConfiguration.R
\name{V1QueuingConfiguration}
\alias{V1QueuingConfiguration}
\title{V1QueuingConfiguration}
\description{
V1QueuingConfiguration

V1QueuingConfiguration
}
\details{
QueuingConfiguration holds the configuration parameters for queuing
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1QueuingConfiguration-new}{\code{V1QueuingConfiguration$new()}}
\item \href{#method-V1QueuingConfiguration-to_list}{\code{V1QueuingConfiguration$to_list()}}
\item \href{#method-V1QueuingConfiguration-clone}{\code{V1QueuingConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1QueuingConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1QueuingConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1QueuingConfiguration$new(
  hand_size = NULL,
  queue_length_limit = NULL,
  queues = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1QueuingConfiguration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1QueuingConfiguration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1QueuingConfiguration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1QueuingConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1QueuingConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1QueuingConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
