% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1QuobyteVolumeSource.R
\name{V1QuobyteVolumeSource}
\alias{V1QuobyteVolumeSource}
\title{V1QuobyteVolumeSource}
\description{
V1QuobyteVolumeSource

V1QuobyteVolumeSource
}
\details{
Represents a Quobyte mount that lasts the lifetime of a pod. Quobyte volumes do not support ownership management or SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1QuobyteVolumeSource-new}{\code{V1QuobyteVolumeSource$new()}}
\item \href{#method-V1QuobyteVolumeSource-to_list}{\code{V1QuobyteVolumeSource$to_list()}}
\item \href{#method-V1QuobyteVolumeSource-clone}{\code{V1QuobyteVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1QuobyteVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1QuobyteVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1QuobyteVolumeSource$new(
  group = NULL,
  read_only = NULL,
  registry = NULL,
  tenant = NULL,
  user = NULL,
  volume = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1QuobyteVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1QuobyteVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1QuobyteVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1QuobyteVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1QuobyteVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1QuobyteVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
