% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RBDPersistentVolumeSource.R
\name{V1RBDPersistentVolumeSource}
\alias{V1RBDPersistentVolumeSource}
\title{V1RBDPersistentVolumeSource}
\description{
V1RBDPersistentVolumeSource

V1RBDPersistentVolumeSource
}
\details{
Represents a Rados Block Device mount that lasts the lifetime of a pod. RBD volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RBDPersistentVolumeSource-new}{\code{V1RBDPersistentVolumeSource$new()}}
\item \href{#method-V1RBDPersistentVolumeSource-to_list}{\code{V1RBDPersistentVolumeSource$to_list()}}
\item \href{#method-V1RBDPersistentVolumeSource-clone}{\code{V1RBDPersistentVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RBDPersistentVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RBDPersistentVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RBDPersistentVolumeSource$new(
  fs_type = NULL,
  image = NULL,
  keyring = NULL,
  monitors = NULL,
  pool = NULL,
  read_only = NULL,
  secret_ref = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RBDPersistentVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RBDPersistentVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RBDPersistentVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RBDPersistentVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RBDPersistentVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RBDPersistentVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
