% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RBDVolumeSource.R
\name{V1RBDVolumeSource}
\alias{V1RBDVolumeSource}
\title{V1RBDVolumeSource}
\description{
V1RBDVolumeSource

V1RBDVolumeSource
}
\details{
Represents a Rados Block Device mount that lasts the lifetime of a pod. RBD volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RBDVolumeSource-new}{\code{V1RBDVolumeSource$new()}}
\item \href{#method-V1RBDVolumeSource-to_list}{\code{V1RBDVolumeSource$to_list()}}
\item \href{#method-V1RBDVolumeSource-clone}{\code{V1RBDVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RBDVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RBDVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RBDVolumeSource$new(
  fs_type = NULL,
  image = NULL,
  keyring = NULL,
  monitors = NULL,
  pool = NULL,
  read_only = NULL,
  secret_ref = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RBDVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RBDVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RBDVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RBDVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RBDVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RBDVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
