% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicaSet.R
\name{V1ReplicaSet}
\alias{V1ReplicaSet}
\title{V1ReplicaSet}
\description{
V1ReplicaSet

V1ReplicaSet
}
\details{
ReplicaSet ensures that a specified number of pod replicas are running at any given time.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicaSet-new}{\code{V1ReplicaSet$new()}}
\item \href{#method-V1ReplicaSet-to_list}{\code{V1ReplicaSet$to_list()}}
\item \href{#method-V1ReplicaSet-clone}{\code{V1ReplicaSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSet-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSet$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSet-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSet-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSet$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
