% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicaSetCondition.R
\name{V1ReplicaSetCondition}
\alias{V1ReplicaSetCondition}
\title{V1ReplicaSetCondition}
\description{
V1ReplicaSetCondition

V1ReplicaSetCondition
}
\details{
ReplicaSetCondition describes the state of a replica set at a certain point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicaSetCondition-new}{\code{V1ReplicaSetCondition$new()}}
\item \href{#method-V1ReplicaSetCondition-to_list}{\code{V1ReplicaSetCondition$to_list()}}
\item \href{#method-V1ReplicaSetCondition-clone}{\code{V1ReplicaSetCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
