% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicaSetSpec.R
\name{V1ReplicaSetSpec}
\alias{V1ReplicaSetSpec}
\title{V1ReplicaSetSpec}
\description{
V1ReplicaSetSpec

V1ReplicaSetSpec
}
\details{
ReplicaSetSpec is the specification of a ReplicaSet.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicaSetSpec-new}{\code{V1ReplicaSetSpec$new()}}
\item \href{#method-V1ReplicaSetSpec-to_list}{\code{V1ReplicaSetSpec$to_list()}}
\item \href{#method-V1ReplicaSetSpec-clone}{\code{V1ReplicaSetSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetSpec$new(
  min_ready_seconds = NULL,
  replicas = NULL,
  selector = NULL,
  template = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
