% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicaSetStatus.R
\name{V1ReplicaSetStatus}
\alias{V1ReplicaSetStatus}
\title{V1ReplicaSetStatus}
\description{
V1ReplicaSetStatus

V1ReplicaSetStatus
}
\details{
ReplicaSetStatus represents the current status of a ReplicaSet.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicaSetStatus-new}{\code{V1ReplicaSetStatus$new()}}
\item \href{#method-V1ReplicaSetStatus-to_list}{\code{V1ReplicaSetStatus$to_list()}}
\item \href{#method-V1ReplicaSetStatus-clone}{\code{V1ReplicaSetStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetStatus$new(
  available_replicas = NULL,
  conditions = NULL,
  fully_labeled_replicas = NULL,
  observed_generation = NULL,
  ready_replicas = NULL,
  replicas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicaSetStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicaSetStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicaSetStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
