% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicationController.R
\name{V1ReplicationController}
\alias{V1ReplicationController}
\title{V1ReplicationController}
\description{
V1ReplicationController

V1ReplicationController
}
\details{
ReplicationController represents the configuration of a replication controller.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicationController-new}{\code{V1ReplicationController$new()}}
\item \href{#method-V1ReplicationController-to_list}{\code{V1ReplicationController$to_list()}}
\item \href{#method-V1ReplicationController-clone}{\code{V1ReplicationController$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationController-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationController-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationController$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationController-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationController-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationController$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationController-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationController-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationController$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
