% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicationControllerCondition.R
\name{V1ReplicationControllerCondition}
\alias{V1ReplicationControllerCondition}
\title{V1ReplicationControllerCondition}
\description{
V1ReplicationControllerCondition

V1ReplicationControllerCondition
}
\details{
ReplicationControllerCondition describes the state of a replication controller at a certain point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicationControllerCondition-new}{\code{V1ReplicationControllerCondition$new()}}
\item \href{#method-V1ReplicationControllerCondition-to_list}{\code{V1ReplicationControllerCondition$to_list()}}
\item \href{#method-V1ReplicationControllerCondition-clone}{\code{V1ReplicationControllerCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
