% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicationControllerList.R
\name{V1ReplicationControllerList}
\alias{V1ReplicationControllerList}
\title{V1ReplicationControllerList}
\description{
V1ReplicationControllerList

V1ReplicationControllerList
}
\details{
ReplicationControllerList is a collection of replication controllers.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicationControllerList-new}{\code{V1ReplicationControllerList$new()}}
\item \href{#method-V1ReplicationControllerList-to_list}{\code{V1ReplicationControllerList$to_list()}}
\item \href{#method-V1ReplicationControllerList-clone}{\code{V1ReplicationControllerList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
