% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicationControllerSpec.R
\name{V1ReplicationControllerSpec}
\alias{V1ReplicationControllerSpec}
\title{V1ReplicationControllerSpec}
\description{
V1ReplicationControllerSpec

V1ReplicationControllerSpec
}
\details{
ReplicationControllerSpec is the specification of a replication controller.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicationControllerSpec-new}{\code{V1ReplicationControllerSpec$new()}}
\item \href{#method-V1ReplicationControllerSpec-to_list}{\code{V1ReplicationControllerSpec$to_list()}}
\item \href{#method-V1ReplicationControllerSpec-clone}{\code{V1ReplicationControllerSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerSpec$new(
  min_ready_seconds = NULL,
  replicas = NULL,
  selector = NULL,
  template = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
