% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ReplicationControllerStatus.R
\name{V1ReplicationControllerStatus}
\alias{V1ReplicationControllerStatus}
\title{V1ReplicationControllerStatus}
\description{
V1ReplicationControllerStatus

V1ReplicationControllerStatus
}
\details{
ReplicationControllerStatus represents the current status of a replication controller.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ReplicationControllerStatus-new}{\code{V1ReplicationControllerStatus$new()}}
\item \href{#method-V1ReplicationControllerStatus-to_list}{\code{V1ReplicationControllerStatus$to_list()}}
\item \href{#method-V1ReplicationControllerStatus-clone}{\code{V1ReplicationControllerStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerStatus$new(
  available_replicas = NULL,
  conditions = NULL,
  fully_labeled_replicas = NULL,
  observed_generation = NULL,
  ready_replicas = NULL,
  replicas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ReplicationControllerStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ReplicationControllerStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ReplicationControllerStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
