% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceAttributes.R
\name{V1ResourceAttributes}
\alias{V1ResourceAttributes}
\title{V1ResourceAttributes}
\description{
V1ResourceAttributes

V1ResourceAttributes
}
\details{
ResourceAttributes includes the authorization attributes available for resource requests to the Authorizer interface
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceAttributes-new}{\code{V1ResourceAttributes$new()}}
\item \href{#method-V1ResourceAttributes-to_list}{\code{V1ResourceAttributes$to_list()}}
\item \href{#method-V1ResourceAttributes-clone}{\code{V1ResourceAttributes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceAttributes-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceAttributes-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceAttributes$new(
  field_selector = NULL,
  group = NULL,
  label_selector = NULL,
  name = NULL,
  namespace = NULL,
  resource = NULL,
  subresource = NULL,
  verb = NULL,
  version = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceAttributes-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceAttributes-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceAttributes$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceAttributes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceAttributes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceAttributes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
