% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceClaim.R
\name{V1ResourceClaim}
\alias{V1ResourceClaim}
\title{V1ResourceClaim}
\description{
V1ResourceClaim

V1ResourceClaim
}
\details{
ResourceClaim references one entry in PodSpec.ResourceClaims.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceClaim-new}{\code{V1ResourceClaim$new()}}
\item \href{#method-V1ResourceClaim-to_list}{\code{V1ResourceClaim$to_list()}}
\item \href{#method-V1ResourceClaim-clone}{\code{V1ResourceClaim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceClaim-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceClaim-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceClaim$new(name = NULL, request = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceClaim-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceClaim-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceClaim$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceClaim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceClaim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceClaim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
