% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceHealth.R
\name{V1ResourceHealth}
\alias{V1ResourceHealth}
\title{V1ResourceHealth}
\description{
V1ResourceHealth

V1ResourceHealth
}
\details{
ResourceHealth represents the health of a resource. It has the latest device health information. This is a part of KEP https://kep.k8s.io/4680 and historical health changes are planned to be added in future iterations of a KEP.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceHealth-new}{\code{V1ResourceHealth$new()}}
\item \href{#method-V1ResourceHealth-to_list}{\code{V1ResourceHealth$to_list()}}
\item \href{#method-V1ResourceHealth-clone}{\code{V1ResourceHealth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceHealth-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceHealth-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceHealth$new(health = NULL, resource_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceHealth-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceHealth-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceHealth$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceHealth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceHealth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceHealth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
