% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourcePolicyRule.R
\name{V1ResourcePolicyRule}
\alias{V1ResourcePolicyRule}
\title{V1ResourcePolicyRule}
\description{
V1ResourcePolicyRule

V1ResourcePolicyRule
}
\details{
ResourcePolicyRule is a predicate that matches some resource requests, testing the request's verb and the target resource. A ResourcePolicyRule matches a resource request if and only if: (a) at least one member of verbs matches the request, (b) at least one member of apiGroups matches the request, (c) at least one member of resources matches the request, and (d) either (d1) the request does not specify a namespace (i.e., `Namespace==""`) and clusterScope is true or (d2) the request specifies a namespace and least one member of namespaces matches the request's namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourcePolicyRule-new}{\code{V1ResourcePolicyRule$new()}}
\item \href{#method-V1ResourcePolicyRule-to_list}{\code{V1ResourcePolicyRule$to_list()}}
\item \href{#method-V1ResourcePolicyRule-clone}{\code{V1ResourcePolicyRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourcePolicyRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourcePolicyRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourcePolicyRule$new(
  api_groups = NULL,
  cluster_scope = NULL,
  namespaces = NULL,
  resources = NULL,
  verbs = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourcePolicyRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourcePolicyRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourcePolicyRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourcePolicyRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourcePolicyRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourcePolicyRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
