% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceQuota.R
\name{V1ResourceQuota}
\alias{V1ResourceQuota}
\title{V1ResourceQuota}
\description{
V1ResourceQuota

V1ResourceQuota
}
\details{
ResourceQuota sets aggregate quota restrictions enforced per namespace
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceQuota-new}{\code{V1ResourceQuota$new()}}
\item \href{#method-V1ResourceQuota-to_list}{\code{V1ResourceQuota$to_list()}}
\item \href{#method-V1ResourceQuota-clone}{\code{V1ResourceQuota$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuota-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuota-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuota$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuota-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuota-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuota$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuota-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuota-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuota$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
