% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceQuotaSpec.R
\name{V1ResourceQuotaSpec}
\alias{V1ResourceQuotaSpec}
\title{V1ResourceQuotaSpec}
\description{
V1ResourceQuotaSpec

V1ResourceQuotaSpec
}
\details{
ResourceQuotaSpec defines the desired hard limits to enforce for Quota.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceQuotaSpec-new}{\code{V1ResourceQuotaSpec$new()}}
\item \href{#method-V1ResourceQuotaSpec-to_list}{\code{V1ResourceQuotaSpec$to_list()}}
\item \href{#method-V1ResourceQuotaSpec-clone}{\code{V1ResourceQuotaSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuotaSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuotaSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuotaSpec$new(hard = NULL, scope_selector = NULL, scopes = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuotaSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuotaSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuotaSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuotaSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuotaSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuotaSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
