% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceQuotaStatus.R
\name{V1ResourceQuotaStatus}
\alias{V1ResourceQuotaStatus}
\title{V1ResourceQuotaStatus}
\description{
V1ResourceQuotaStatus

V1ResourceQuotaStatus
}
\details{
ResourceQuotaStatus defines the enforced hard limits and observed use.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceQuotaStatus-new}{\code{V1ResourceQuotaStatus$new()}}
\item \href{#method-V1ResourceQuotaStatus-to_list}{\code{V1ResourceQuotaStatus$to_list()}}
\item \href{#method-V1ResourceQuotaStatus-clone}{\code{V1ResourceQuotaStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuotaStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuotaStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuotaStatus$new(hard = NULL, used = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuotaStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuotaStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuotaStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceQuotaStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceQuotaStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceQuotaStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
