% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceRequirements.R
\name{V1ResourceRequirements}
\alias{V1ResourceRequirements}
\title{V1ResourceRequirements}
\description{
V1ResourceRequirements

V1ResourceRequirements
}
\details{
ResourceRequirements describes the compute resource requirements.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceRequirements-new}{\code{V1ResourceRequirements$new()}}
\item \href{#method-V1ResourceRequirements-to_list}{\code{V1ResourceRequirements$to_list()}}
\item \href{#method-V1ResourceRequirements-clone}{\code{V1ResourceRequirements$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceRequirements-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceRequirements-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceRequirements$new(claims = NULL, limits = NULL, requests = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceRequirements-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceRequirements-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceRequirements$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceRequirements-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceRequirements-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceRequirements$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
