% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceRule.R
\name{V1ResourceRule}
\alias{V1ResourceRule}
\title{V1ResourceRule}
\description{
V1ResourceRule

V1ResourceRule
}
\details{
ResourceRule is the list of actions the subject is allowed to perform on resources. The list ordering isn't significant, may contain duplicates, and possibly be incomplete.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceRule-new}{\code{V1ResourceRule$new()}}
\item \href{#method-V1ResourceRule-to_list}{\code{V1ResourceRule$to_list()}}
\item \href{#method-V1ResourceRule-clone}{\code{V1ResourceRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceRule$new(
  api_groups = NULL,
  resource_names = NULL,
  resources = NULL,
  verbs = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
