% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ResourceStatus.R
\name{V1ResourceStatus}
\alias{V1ResourceStatus}
\title{V1ResourceStatus}
\description{
V1ResourceStatus

V1ResourceStatus
}
\details{
(no description)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ResourceStatus-new}{\code{V1ResourceStatus$new()}}
\item \href{#method-V1ResourceStatus-to_list}{\code{V1ResourceStatus$to_list()}}
\item \href{#method-V1ResourceStatus-clone}{\code{V1ResourceStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceStatus$new(name = NULL, resources = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ResourceStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ResourceStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ResourceStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
