% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Role.R
\name{V1Role}
\alias{V1Role}
\title{V1Role}
\description{
V1Role

V1Role
}
\details{
Role is a namespaced, logical grouping of PolicyRules that can be referenced as a unit by a RoleBinding.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Role-new}{\code{V1Role$new()}}
\item \href{#method-V1Role-to_list}{\code{V1Role$to_list()}}
\item \href{#method-V1Role-clone}{\code{V1Role$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Role-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Role-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Role$new(api_version = NULL, kind = NULL, metadata = NULL, rules = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Role-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Role-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Role$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Role-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Role-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Role$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
