% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RoleBinding.R
\name{V1RoleBinding}
\alias{V1RoleBinding}
\title{V1RoleBinding}
\description{
V1RoleBinding

V1RoleBinding
}
\details{
RoleBinding references a role, but does not contain it.  It can reference a Role in the same namespace or a ClusterRole in the global namespace. It adds who information via Subjects and namespace information by which namespace it exists in.  RoleBindings in a given namespace only have effect in that namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RoleBinding-new}{\code{V1RoleBinding$new()}}
\item \href{#method-V1RoleBinding-to_list}{\code{V1RoleBinding$to_list()}}
\item \href{#method-V1RoleBinding-clone}{\code{V1RoleBinding$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleBinding-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleBinding-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleBinding$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  role_ref = NULL,
  subjects = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleBinding-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleBinding-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleBinding$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleBinding-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleBinding-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleBinding$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
