% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RoleBindingList.R
\name{V1RoleBindingList}
\alias{V1RoleBindingList}
\title{V1RoleBindingList}
\description{
V1RoleBindingList

V1RoleBindingList
}
\details{
RoleBindingList is a collection of RoleBindings
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RoleBindingList-new}{\code{V1RoleBindingList$new()}}
\item \href{#method-V1RoleBindingList-to_list}{\code{V1RoleBindingList$to_list()}}
\item \href{#method-V1RoleBindingList-clone}{\code{V1RoleBindingList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleBindingList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleBindingList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleBindingList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleBindingList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleBindingList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleBindingList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleBindingList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleBindingList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleBindingList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
