% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RoleList.R
\name{V1RoleList}
\alias{V1RoleList}
\title{V1RoleList}
\description{
V1RoleList

V1RoleList
}
\details{
RoleList is a collection of Roles
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RoleList-new}{\code{V1RoleList$new()}}
\item \href{#method-V1RoleList-to_list}{\code{V1RoleList$to_list()}}
\item \href{#method-V1RoleList-clone}{\code{V1RoleList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleList$new(api_version = NULL, items = NULL, kind = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RoleList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RoleList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RoleList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
