% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RollingUpdateDaemonSet.R
\name{V1RollingUpdateDaemonSet}
\alias{V1RollingUpdateDaemonSet}
\title{V1RollingUpdateDaemonSet}
\description{
V1RollingUpdateDaemonSet

V1RollingUpdateDaemonSet
}
\details{
Spec to control the desired behavior of daemon set rolling update.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RollingUpdateDaemonSet-new}{\code{V1RollingUpdateDaemonSet$new()}}
\item \href{#method-V1RollingUpdateDaemonSet-to_list}{\code{V1RollingUpdateDaemonSet$to_list()}}
\item \href{#method-V1RollingUpdateDaemonSet-clone}{\code{V1RollingUpdateDaemonSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RollingUpdateDaemonSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RollingUpdateDaemonSet-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RollingUpdateDaemonSet$new(max_surge = NULL, max_unavailable = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RollingUpdateDaemonSet-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RollingUpdateDaemonSet-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RollingUpdateDaemonSet$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RollingUpdateDaemonSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RollingUpdateDaemonSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RollingUpdateDaemonSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
