% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RollingUpdateStatefulSetStrategy.R
\name{V1RollingUpdateStatefulSetStrategy}
\alias{V1RollingUpdateStatefulSetStrategy}
\title{V1RollingUpdateStatefulSetStrategy}
\description{
V1RollingUpdateStatefulSetStrategy

V1RollingUpdateStatefulSetStrategy
}
\details{
RollingUpdateStatefulSetStrategy is used to communicate parameter for RollingUpdateStatefulSetStrategyType.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RollingUpdateStatefulSetStrategy-new}{\code{V1RollingUpdateStatefulSetStrategy$new()}}
\item \href{#method-V1RollingUpdateStatefulSetStrategy-to_list}{\code{V1RollingUpdateStatefulSetStrategy$to_list()}}
\item \href{#method-V1RollingUpdateStatefulSetStrategy-clone}{\code{V1RollingUpdateStatefulSetStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RollingUpdateStatefulSetStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RollingUpdateStatefulSetStrategy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RollingUpdateStatefulSetStrategy$new(
  max_unavailable = NULL,
  partition = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RollingUpdateStatefulSetStrategy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RollingUpdateStatefulSetStrategy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RollingUpdateStatefulSetStrategy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RollingUpdateStatefulSetStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RollingUpdateStatefulSetStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RollingUpdateStatefulSetStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
