% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RuleWithOperations.R
\name{V1RuleWithOperations}
\alias{V1RuleWithOperations}
\title{V1RuleWithOperations}
\description{
V1RuleWithOperations

V1RuleWithOperations
}
\details{
RuleWithOperations is a tuple of Operations and Resources. It is recommended to make sure that all the tuple expansions are valid.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RuleWithOperations-new}{\code{V1RuleWithOperations$new()}}
\item \href{#method-V1RuleWithOperations-to_list}{\code{V1RuleWithOperations$to_list()}}
\item \href{#method-V1RuleWithOperations-clone}{\code{V1RuleWithOperations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuleWithOperations-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuleWithOperations-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuleWithOperations$new(
  api_groups = NULL,
  api_versions = NULL,
  operations = NULL,
  resources = NULL,
  scope = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuleWithOperations-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuleWithOperations-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuleWithOperations$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuleWithOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuleWithOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuleWithOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
