% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RuntimeClass.R
\name{V1RuntimeClass}
\alias{V1RuntimeClass}
\title{V1RuntimeClass}
\description{
V1RuntimeClass

V1RuntimeClass
}
\details{
RuntimeClass defines a class of container runtime supported in the cluster. The RuntimeClass is used to determine which container runtime is used to run all containers in a pod. RuntimeClasses are manually defined by a user or cluster provisioner, and referenced in the PodSpec. The Kubelet is responsible for resolving the RuntimeClassName reference before running the pod.  For more details, see https://kubernetes.io/docs/concepts/containers/runtime-class/
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RuntimeClass-new}{\code{V1RuntimeClass$new()}}
\item \href{#method-V1RuntimeClass-to_list}{\code{V1RuntimeClass$to_list()}}
\item \href{#method-V1RuntimeClass-clone}{\code{V1RuntimeClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuntimeClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuntimeClass-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuntimeClass$new(
  api_version = NULL,
  handler = NULL,
  kind = NULL,
  metadata = NULL,
  overhead = NULL,
  scheduling = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuntimeClass-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuntimeClass-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuntimeClass$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuntimeClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuntimeClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuntimeClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
