% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1RuntimeClassList.R
\name{V1RuntimeClassList}
\alias{V1RuntimeClassList}
\title{V1RuntimeClassList}
\description{
V1RuntimeClassList

V1RuntimeClassList
}
\details{
RuntimeClassList is a list of RuntimeClass objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1RuntimeClassList-new}{\code{V1RuntimeClassList$new()}}
\item \href{#method-V1RuntimeClassList-to_list}{\code{V1RuntimeClassList$to_list()}}
\item \href{#method-V1RuntimeClassList-clone}{\code{V1RuntimeClassList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuntimeClassList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuntimeClassList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuntimeClassList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuntimeClassList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuntimeClassList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuntimeClassList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1RuntimeClassList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1RuntimeClassList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1RuntimeClassList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
