% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Scale.R
\name{V1Scale}
\alias{V1Scale}
\title{V1Scale}
\description{
V1Scale

V1Scale
}
\details{
Scale represents a scaling request for a resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Scale-new}{\code{V1Scale$new()}}
\item \href{#method-V1Scale-to_list}{\code{V1Scale$to_list()}}
\item \href{#method-V1Scale-clone}{\code{V1Scale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Scale-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Scale-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Scale$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Scale-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Scale-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Scale$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Scale-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Scale-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Scale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
