% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ScaleIOVolumeSource.R
\name{V1ScaleIOVolumeSource}
\alias{V1ScaleIOVolumeSource}
\title{V1ScaleIOVolumeSource}
\description{
V1ScaleIOVolumeSource

V1ScaleIOVolumeSource
}
\details{
ScaleIOVolumeSource represents a persistent ScaleIO volume
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ScaleIOVolumeSource-new}{\code{V1ScaleIOVolumeSource$new()}}
\item \href{#method-V1ScaleIOVolumeSource-to_list}{\code{V1ScaleIOVolumeSource$to_list()}}
\item \href{#method-V1ScaleIOVolumeSource-clone}{\code{V1ScaleIOVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleIOVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleIOVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleIOVolumeSource$new(
  fs_type = NULL,
  gateway = NULL,
  protection_domain = NULL,
  read_only = NULL,
  secret_ref = NULL,
  ssl_enabled = NULL,
  storage_mode = NULL,
  storage_pool = NULL,
  system = NULL,
  volume_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleIOVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleIOVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleIOVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleIOVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleIOVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleIOVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
