% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ScaleSpec.R
\name{V1ScaleSpec}
\alias{V1ScaleSpec}
\title{V1ScaleSpec}
\description{
V1ScaleSpec

V1ScaleSpec
}
\details{
ScaleSpec describes the attributes of a scale subresource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ScaleSpec-new}{\code{V1ScaleSpec$new()}}
\item \href{#method-V1ScaleSpec-to_list}{\code{V1ScaleSpec$to_list()}}
\item \href{#method-V1ScaleSpec-clone}{\code{V1ScaleSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleSpec$new(replicas = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
