% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ScaleStatus.R
\name{V1ScaleStatus}
\alias{V1ScaleStatus}
\title{V1ScaleStatus}
\description{
V1ScaleStatus

V1ScaleStatus
}
\details{
ScaleStatus represents the current status of a scale subresource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ScaleStatus-new}{\code{V1ScaleStatus$new()}}
\item \href{#method-V1ScaleStatus-to_list}{\code{V1ScaleStatus$to_list()}}
\item \href{#method-V1ScaleStatus-clone}{\code{V1ScaleStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleStatus$new(replicas = NULL, selector = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScaleStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScaleStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScaleStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
