% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Scheduling.R
\name{V1Scheduling}
\alias{V1Scheduling}
\title{V1Scheduling}
\description{
V1Scheduling

V1Scheduling
}
\details{
Scheduling specifies the scheduling constraints for nodes supporting a RuntimeClass.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Scheduling-new}{\code{V1Scheduling$new()}}
\item \href{#method-V1Scheduling-to_list}{\code{V1Scheduling$to_list()}}
\item \href{#method-V1Scheduling-clone}{\code{V1Scheduling$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Scheduling-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Scheduling-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Scheduling$new(node_selector = NULL, tolerations = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Scheduling-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Scheduling-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Scheduling$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Scheduling-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Scheduling-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Scheduling$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
