% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ScopedResourceSelectorRequirement.R
\name{V1ScopedResourceSelectorRequirement}
\alias{V1ScopedResourceSelectorRequirement}
\title{V1ScopedResourceSelectorRequirement}
\description{
V1ScopedResourceSelectorRequirement

V1ScopedResourceSelectorRequirement
}
\details{
A scoped-resource selector requirement is a selector that contains values, a scope name, and an operator that relates the scope name and values.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ScopedResourceSelectorRequirement-new}{\code{V1ScopedResourceSelectorRequirement$new()}}
\item \href{#method-V1ScopedResourceSelectorRequirement-to_list}{\code{V1ScopedResourceSelectorRequirement$to_list()}}
\item \href{#method-V1ScopedResourceSelectorRequirement-clone}{\code{V1ScopedResourceSelectorRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScopedResourceSelectorRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScopedResourceSelectorRequirement-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScopedResourceSelectorRequirement$new(
  operator = NULL,
  scope_name = NULL,
  values = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScopedResourceSelectorRequirement-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScopedResourceSelectorRequirement-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScopedResourceSelectorRequirement$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScopedResourceSelectorRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScopedResourceSelectorRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScopedResourceSelectorRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
