% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ScopeSelector.R
\name{V1ScopeSelector}
\alias{V1ScopeSelector}
\title{V1ScopeSelector}
\description{
V1ScopeSelector

V1ScopeSelector
}
\details{
A scope selector represents the AND of the selectors represented by the scoped-resource selector requirements.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ScopeSelector-new}{\code{V1ScopeSelector$new()}}
\item \href{#method-V1ScopeSelector-to_list}{\code{V1ScopeSelector$to_list()}}
\item \href{#method-V1ScopeSelector-clone}{\code{V1ScopeSelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScopeSelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScopeSelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScopeSelector$new(match_expressions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScopeSelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScopeSelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScopeSelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ScopeSelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ScopeSelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ScopeSelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
