% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SeccompProfile.R
\name{V1SeccompProfile}
\alias{V1SeccompProfile}
\title{V1SeccompProfile}
\description{
V1SeccompProfile

V1SeccompProfile
}
\details{
SeccompProfile defines a pod/container's seccomp profile settings. Only one profile source may be set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SeccompProfile-new}{\code{V1SeccompProfile$new()}}
\item \href{#method-V1SeccompProfile-to_list}{\code{V1SeccompProfile$to_list()}}
\item \href{#method-V1SeccompProfile-clone}{\code{V1SeccompProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SeccompProfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SeccompProfile-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SeccompProfile$new(localhost_profile = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SeccompProfile-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SeccompProfile-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SeccompProfile$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SeccompProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SeccompProfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SeccompProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
