% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Secret.R
\name{V1Secret}
\alias{V1Secret}
\title{V1Secret}
\description{
V1Secret

V1Secret
}
\details{
Secret holds secret data of a certain type. The total bytes of the values in the Data field must be less than MaxSecretSize bytes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Secret-new}{\code{V1Secret$new()}}
\item \href{#method-V1Secret-to_list}{\code{V1Secret$to_list()}}
\item \href{#method-V1Secret-clone}{\code{V1Secret$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Secret-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Secret-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Secret$new(
  api_version = NULL,
  data = NULL,
  immutable = NULL,
  kind = NULL,
  metadata = NULL,
  string_data = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Secret-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Secret-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Secret$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Secret-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Secret-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Secret$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
