% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SecretKeySelector.R
\name{V1SecretKeySelector}
\alias{V1SecretKeySelector}
\title{V1SecretKeySelector}
\description{
V1SecretKeySelector

V1SecretKeySelector
}
\details{
SecretKeySelector selects a key of a Secret.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SecretKeySelector-new}{\code{V1SecretKeySelector$new()}}
\item \href{#method-V1SecretKeySelector-to_list}{\code{V1SecretKeySelector$to_list()}}
\item \href{#method-V1SecretKeySelector-clone}{\code{V1SecretKeySelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretKeySelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretKeySelector-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretKeySelector$new(key = NULL, name = NULL, optional = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretKeySelector-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretKeySelector-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretKeySelector$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretKeySelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretKeySelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretKeySelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
