% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SecretList.R
\name{V1SecretList}
\alias{V1SecretList}
\title{V1SecretList}
\description{
V1SecretList

V1SecretList
}
\details{
SecretList is a list of Secret.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SecretList-new}{\code{V1SecretList$new()}}
\item \href{#method-V1SecretList-to_list}{\code{V1SecretList$to_list()}}
\item \href{#method-V1SecretList-clone}{\code{V1SecretList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
