% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SecretProjection.R
\name{V1SecretProjection}
\alias{V1SecretProjection}
\title{V1SecretProjection}
\description{
V1SecretProjection

V1SecretProjection
}
\details{
Adapts a secret into a projected volume.

The contents of the target Secret's Data field will be presented in a projected volume as files using the keys in the Data field as the file names. Note that this is identical to a secret volume source without the default mode.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SecretProjection-new}{\code{V1SecretProjection$new()}}
\item \href{#method-V1SecretProjection-to_list}{\code{V1SecretProjection$to_list()}}
\item \href{#method-V1SecretProjection-clone}{\code{V1SecretProjection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretProjection-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretProjection-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretProjection$new(items = NULL, name = NULL, optional = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretProjection-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretProjection-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretProjection$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretProjection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretProjection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretProjection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
