% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SecretReference.R
\name{V1SecretReference}
\alias{V1SecretReference}
\title{V1SecretReference}
\description{
V1SecretReference

V1SecretReference
}
\details{
SecretReference represents a Secret Reference. It has enough information to retrieve secret in any namespace
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SecretReference-new}{\code{V1SecretReference$new()}}
\item \href{#method-V1SecretReference-to_list}{\code{V1SecretReference$to_list()}}
\item \href{#method-V1SecretReference-clone}{\code{V1SecretReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretReference$new(name = NULL, namespace = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
