% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SecretVolumeSource.R
\name{V1SecretVolumeSource}
\alias{V1SecretVolumeSource}
\title{V1SecretVolumeSource}
\description{
V1SecretVolumeSource

V1SecretVolumeSource
}
\details{
Adapts a Secret into a volume.

The contents of the target Secret's Data field will be presented in a volume as files using the keys in the Data field as the file names. Secret volumes support ownership management and SELinux relabeling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SecretVolumeSource-new}{\code{V1SecretVolumeSource$new()}}
\item \href{#method-V1SecretVolumeSource-to_list}{\code{V1SecretVolumeSource$to_list()}}
\item \href{#method-V1SecretVolumeSource-clone}{\code{V1SecretVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretVolumeSource$new(
  default_mode = NULL,
  items = NULL,
  optional = NULL,
  secret_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecretVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecretVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecretVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
