% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SecurityContext.R
\name{V1SecurityContext}
\alias{V1SecurityContext}
\title{V1SecurityContext}
\description{
V1SecurityContext

V1SecurityContext
}
\details{
SecurityContext holds security configuration that will be applied to a container. Some fields are present in both SecurityContext and PodSecurityContext.  When both are set, the values in SecurityContext take precedence.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SecurityContext-new}{\code{V1SecurityContext$new()}}
\item \href{#method-V1SecurityContext-to_list}{\code{V1SecurityContext$to_list()}}
\item \href{#method-V1SecurityContext-clone}{\code{V1SecurityContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecurityContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecurityContext-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecurityContext$new(
  allow_privilege_escalation = NULL,
  app_armor_profile = NULL,
  capabilities = NULL,
  privileged = NULL,
  proc_mount = NULL,
  read_only_root_filesystem = NULL,
  run_as_group = NULL,
  run_as_non_root = NULL,
  run_as_user = NULL,
  se_linux_options = NULL,
  seccomp_profile = NULL,
  windows_options = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecurityContext-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecurityContext-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecurityContext$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SecurityContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SecurityContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SecurityContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
