% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SelectableField.R
\name{V1SelectableField}
\alias{V1SelectableField}
\title{V1SelectableField}
\description{
V1SelectableField

V1SelectableField
}
\details{
SelectableField specifies the JSON path of a field that may be used with field selectors.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SelectableField-new}{\code{V1SelectableField$new()}}
\item \href{#method-V1SelectableField-to_list}{\code{V1SelectableField$to_list()}}
\item \href{#method-V1SelectableField-clone}{\code{V1SelectableField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelectableField-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelectableField-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelectableField$new(json_path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelectableField-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelectableField-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelectableField$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelectableField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelectableField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelectableField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
