% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SelfSubjectAccessReview.R
\name{V1SelfSubjectAccessReview}
\alias{V1SelfSubjectAccessReview}
\title{V1SelfSubjectAccessReview}
\description{
V1SelfSubjectAccessReview

V1SelfSubjectAccessReview
}
\details{
SelfSubjectAccessReview checks whether or the current user can perform an action.  Not filling in a spec.namespace means "in all namespaces".  Self is a special case, because users should always be able to check whether they can perform an action
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SelfSubjectAccessReview-new}{\code{V1SelfSubjectAccessReview$new()}}
\item \href{#method-V1SelfSubjectAccessReview-to_list}{\code{V1SelfSubjectAccessReview$to_list()}}
\item \href{#method-V1SelfSubjectAccessReview-clone}{\code{V1SelfSubjectAccessReview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectAccessReview-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectAccessReview-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectAccessReview$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectAccessReview-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectAccessReview-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectAccessReview$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectAccessReview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectAccessReview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectAccessReview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
