% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SelfSubjectAccessReviewSpec.R
\name{V1SelfSubjectAccessReviewSpec}
\alias{V1SelfSubjectAccessReviewSpec}
\title{V1SelfSubjectAccessReviewSpec}
\description{
V1SelfSubjectAccessReviewSpec

V1SelfSubjectAccessReviewSpec
}
\details{
SelfSubjectAccessReviewSpec is a description of the access request.  Exactly one of ResourceAuthorizationAttributes and NonResourceAuthorizationAttributes must be set
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SelfSubjectAccessReviewSpec-new}{\code{V1SelfSubjectAccessReviewSpec$new()}}
\item \href{#method-V1SelfSubjectAccessReviewSpec-to_list}{\code{V1SelfSubjectAccessReviewSpec$to_list()}}
\item \href{#method-V1SelfSubjectAccessReviewSpec-clone}{\code{V1SelfSubjectAccessReviewSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectAccessReviewSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectAccessReviewSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectAccessReviewSpec$new(
  non_resource_attributes = NULL,
  resource_attributes = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectAccessReviewSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectAccessReviewSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectAccessReviewSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectAccessReviewSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectAccessReviewSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectAccessReviewSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
