% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SelfSubjectReview.R
\name{V1SelfSubjectReview}
\alias{V1SelfSubjectReview}
\title{V1SelfSubjectReview}
\description{
V1SelfSubjectReview

V1SelfSubjectReview
}
\details{
SelfSubjectReview contains the user information that the kube-apiserver has about the user making this request. When using impersonation, users will receive the user info of the user being impersonated.  If impersonation or request header authentication is used, any extra keys will have their case ignored and returned as lowercase.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SelfSubjectReview-new}{\code{V1SelfSubjectReview$new()}}
\item \href{#method-V1SelfSubjectReview-to_list}{\code{V1SelfSubjectReview$to_list()}}
\item \href{#method-V1SelfSubjectReview-clone}{\code{V1SelfSubjectReview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectReview-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectReview-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectReview$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectReview-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectReview-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectReview$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectReview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectReview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectReview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
