% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SelfSubjectReviewStatus.R
\name{V1SelfSubjectReviewStatus}
\alias{V1SelfSubjectReviewStatus}
\title{V1SelfSubjectReviewStatus}
\description{
V1SelfSubjectReviewStatus

V1SelfSubjectReviewStatus
}
\details{
SelfSubjectReviewStatus is filled by the kube-apiserver and sent back to a user.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SelfSubjectReviewStatus-new}{\code{V1SelfSubjectReviewStatus$new()}}
\item \href{#method-V1SelfSubjectReviewStatus-to_list}{\code{V1SelfSubjectReviewStatus$to_list()}}
\item \href{#method-V1SelfSubjectReviewStatus-clone}{\code{V1SelfSubjectReviewStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectReviewStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectReviewStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectReviewStatus$new(user_info = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectReviewStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectReviewStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectReviewStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectReviewStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectReviewStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectReviewStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
