% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SelfSubjectRulesReview.R
\name{V1SelfSubjectRulesReview}
\alias{V1SelfSubjectRulesReview}
\title{V1SelfSubjectRulesReview}
\description{
V1SelfSubjectRulesReview

V1SelfSubjectRulesReview
}
\details{
SelfSubjectRulesReview enumerates the set of actions the current user can perform within a namespace. The returned list of actions may be incomplete depending on the server's authorization mode, and any errors experienced during the evaluation. SelfSubjectRulesReview should be used by UIs to show/hide actions, or to quickly let an end user reason about their permissions. It should NOT Be used by external systems to drive authorization decisions as this raises confused deputy, cache lifetime/revocation, and correctness concerns. SubjectAccessReview, and LocalAccessReview are the correct way to defer authorization decisions to the API server.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SelfSubjectRulesReview-new}{\code{V1SelfSubjectRulesReview$new()}}
\item \href{#method-V1SelfSubjectRulesReview-to_list}{\code{V1SelfSubjectRulesReview$to_list()}}
\item \href{#method-V1SelfSubjectRulesReview-clone}{\code{V1SelfSubjectRulesReview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectRulesReview-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectRulesReview-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectRulesReview$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectRulesReview-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectRulesReview-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectRulesReview$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SelfSubjectRulesReview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SelfSubjectRulesReview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SelfSubjectRulesReview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
