% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SELinuxOptions.R
\name{V1SELinuxOptions}
\alias{V1SELinuxOptions}
\title{V1SELinuxOptions}
\description{
V1SELinuxOptions

V1SELinuxOptions
}
\details{
SELinuxOptions are the labels to be applied to the container
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SELinuxOptions-new}{\code{V1SELinuxOptions$new()}}
\item \href{#method-V1SELinuxOptions-to_list}{\code{V1SELinuxOptions$to_list()}}
\item \href{#method-V1SELinuxOptions-clone}{\code{V1SELinuxOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SELinuxOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SELinuxOptions-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SELinuxOptions$new(level = NULL, role = NULL, type = NULL, user = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SELinuxOptions-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SELinuxOptions-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SELinuxOptions$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SELinuxOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SELinuxOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SELinuxOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
