% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServerAddressByClientCIDR.R
\name{V1ServerAddressByClientCIDR}
\alias{V1ServerAddressByClientCIDR}
\title{V1ServerAddressByClientCIDR}
\description{
V1ServerAddressByClientCIDR

V1ServerAddressByClientCIDR
}
\details{
ServerAddressByClientCIDR helps the client to determine the server address that they should use, depending on the clientCIDR that they match.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServerAddressByClientCIDR-new}{\code{V1ServerAddressByClientCIDR$new()}}
\item \href{#method-V1ServerAddressByClientCIDR-to_list}{\code{V1ServerAddressByClientCIDR$to_list()}}
\item \href{#method-V1ServerAddressByClientCIDR-clone}{\code{V1ServerAddressByClientCIDR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServerAddressByClientCIDR-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServerAddressByClientCIDR-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServerAddressByClientCIDR$new(client_cidr = NULL, server_address = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServerAddressByClientCIDR-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServerAddressByClientCIDR-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServerAddressByClientCIDR$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServerAddressByClientCIDR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServerAddressByClientCIDR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServerAddressByClientCIDR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
