% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Service.R
\name{V1Service}
\alias{V1Service}
\title{V1Service}
\description{
V1Service

V1Service
}
\details{
Service is a named abstraction of software service (for example, mysql) consisting of local port (for example 3306) that the proxy listens on, and the selector that determines which pods will answer requests sent through the proxy.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Service-new}{\code{V1Service$new()}}
\item \href{#method-V1Service-to_list}{\code{V1Service$to_list()}}
\item \href{#method-V1Service-clone}{\code{V1Service$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Service-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Service-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Service$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Service-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Service-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Service$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Service-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Service-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Service$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
