% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServiceAccount.R
\name{V1ServiceAccount}
\alias{V1ServiceAccount}
\title{V1ServiceAccount}
\description{
V1ServiceAccount

V1ServiceAccount
}
\details{
ServiceAccount binds together: * a name, understood by users, and perhaps by peripheral systems, for an identity * a principal that can be authenticated and authorized * a set of secrets
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServiceAccount-new}{\code{V1ServiceAccount$new()}}
\item \href{#method-V1ServiceAccount-to_list}{\code{V1ServiceAccount$to_list()}}
\item \href{#method-V1ServiceAccount-clone}{\code{V1ServiceAccount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccount-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccount-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccount$new(
  api_version = NULL,
  automount_service_account_token = NULL,
  image_pull_secrets = NULL,
  kind = NULL,
  metadata = NULL,
  secrets = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccount-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccount-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccount$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
