% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServiceAccountSubject.R
\name{V1ServiceAccountSubject}
\alias{V1ServiceAccountSubject}
\title{V1ServiceAccountSubject}
\description{
V1ServiceAccountSubject

V1ServiceAccountSubject
}
\details{
ServiceAccountSubject holds detailed information for service-account-kind subject.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServiceAccountSubject-new}{\code{V1ServiceAccountSubject$new()}}
\item \href{#method-V1ServiceAccountSubject-to_list}{\code{V1ServiceAccountSubject$to_list()}}
\item \href{#method-V1ServiceAccountSubject-clone}{\code{V1ServiceAccountSubject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccountSubject-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccountSubject-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccountSubject$new(name = NULL, namespace = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccountSubject-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccountSubject-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccountSubject$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccountSubject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccountSubject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccountSubject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
