% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServiceAccountTokenProjection.R
\name{V1ServiceAccountTokenProjection}
\alias{V1ServiceAccountTokenProjection}
\title{V1ServiceAccountTokenProjection}
\description{
V1ServiceAccountTokenProjection

V1ServiceAccountTokenProjection
}
\details{
ServiceAccountTokenProjection represents a projected service account token volume. This projection can be used to insert a service account token into the pods runtime filesystem for use against APIs (Kubernetes API Server or otherwise).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServiceAccountTokenProjection-new}{\code{V1ServiceAccountTokenProjection$new()}}
\item \href{#method-V1ServiceAccountTokenProjection-to_list}{\code{V1ServiceAccountTokenProjection$to_list()}}
\item \href{#method-V1ServiceAccountTokenProjection-clone}{\code{V1ServiceAccountTokenProjection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccountTokenProjection-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccountTokenProjection-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccountTokenProjection$new(
  audience = NULL,
  expiration_seconds = NULL,
  path = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccountTokenProjection-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccountTokenProjection-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccountTokenProjection$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceAccountTokenProjection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceAccountTokenProjection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceAccountTokenProjection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
