% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServiceBackendPort.R
\name{V1ServiceBackendPort}
\alias{V1ServiceBackendPort}
\title{V1ServiceBackendPort}
\description{
V1ServiceBackendPort

V1ServiceBackendPort
}
\details{
ServiceBackendPort is the service port being referenced.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServiceBackendPort-new}{\code{V1ServiceBackendPort$new()}}
\item \href{#method-V1ServiceBackendPort-to_list}{\code{V1ServiceBackendPort$to_list()}}
\item \href{#method-V1ServiceBackendPort-clone}{\code{V1ServiceBackendPort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceBackendPort-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceBackendPort-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceBackendPort$new(name = NULL, number = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceBackendPort-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceBackendPort-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceBackendPort$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceBackendPort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceBackendPort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceBackendPort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
