% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServiceList.R
\name{V1ServiceList}
\alias{V1ServiceList}
\title{V1ServiceList}
\description{
V1ServiceList

V1ServiceList
}
\details{
ServiceList holds a list of services.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServiceList-new}{\code{V1ServiceList$new()}}
\item \href{#method-V1ServiceList-to_list}{\code{V1ServiceList$to_list()}}
\item \href{#method-V1ServiceList-clone}{\code{V1ServiceList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
