% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServicePort.R
\name{V1ServicePort}
\alias{V1ServicePort}
\title{V1ServicePort}
\description{
V1ServicePort

V1ServicePort
}
\details{
ServicePort contains information on service's port.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServicePort-new}{\code{V1ServicePort$new()}}
\item \href{#method-V1ServicePort-to_list}{\code{V1ServicePort$to_list()}}
\item \href{#method-V1ServicePort-clone}{\code{V1ServicePort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServicePort-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServicePort-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServicePort$new(
  app_protocol = NULL,
  name = NULL,
  node_port = NULL,
  port = NULL,
  protocol = NULL,
  target_port = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServicePort-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServicePort-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServicePort$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServicePort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServicePort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServicePort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
