% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServiceSpec.R
\name{V1ServiceSpec}
\alias{V1ServiceSpec}
\title{V1ServiceSpec}
\description{
V1ServiceSpec

V1ServiceSpec
}
\details{
ServiceSpec describes the attributes that a user creates on a service.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServiceSpec-new}{\code{V1ServiceSpec$new()}}
\item \href{#method-V1ServiceSpec-to_list}{\code{V1ServiceSpec$to_list()}}
\item \href{#method-V1ServiceSpec-clone}{\code{V1ServiceSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceSpec$new(
  allocate_load_balancer_node_ports = NULL,
  cluster_ip = NULL,
  cluster_i_ps = NULL,
  external_i_ps = NULL,
  external_name = NULL,
  external_traffic_policy = NULL,
  health_check_node_port = NULL,
  internal_traffic_policy = NULL,
  ip_families = NULL,
  ip_family_policy = NULL,
  load_balancer_class = NULL,
  load_balancer_ip = NULL,
  load_balancer_source_ranges = NULL,
  ports = NULL,
  publish_not_ready_addresses = NULL,
  selector = NULL,
  session_affinity = NULL,
  session_affinity_config = NULL,
  traffic_distribution = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
