% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ServiceStatus.R
\name{V1ServiceStatus}
\alias{V1ServiceStatus}
\title{V1ServiceStatus}
\description{
V1ServiceStatus

V1ServiceStatus
}
\details{
ServiceStatus represents the current status of a service.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ServiceStatus-new}{\code{V1ServiceStatus$new()}}
\item \href{#method-V1ServiceStatus-to_list}{\code{V1ServiceStatus$to_list()}}
\item \href{#method-V1ServiceStatus-clone}{\code{V1ServiceStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceStatus$new(conditions = NULL, load_balancer = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ServiceStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ServiceStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ServiceStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
