% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SessionAffinityConfig.R
\name{V1SessionAffinityConfig}
\alias{V1SessionAffinityConfig}
\title{V1SessionAffinityConfig}
\description{
V1SessionAffinityConfig

V1SessionAffinityConfig
}
\details{
SessionAffinityConfig represents the configurations of session affinity.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SessionAffinityConfig-new}{\code{V1SessionAffinityConfig$new()}}
\item \href{#method-V1SessionAffinityConfig-to_list}{\code{V1SessionAffinityConfig$to_list()}}
\item \href{#method-V1SessionAffinityConfig-clone}{\code{V1SessionAffinityConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SessionAffinityConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SessionAffinityConfig-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SessionAffinityConfig$new(client_ip = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SessionAffinityConfig-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SessionAffinityConfig-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SessionAffinityConfig$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SessionAffinityConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SessionAffinityConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SessionAffinityConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
