% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SleepAction.R
\name{V1SleepAction}
\alias{V1SleepAction}
\title{V1SleepAction}
\description{
V1SleepAction

V1SleepAction
}
\details{
SleepAction describes a "sleep" action.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SleepAction-new}{\code{V1SleepAction$new()}}
\item \href{#method-V1SleepAction-to_list}{\code{V1SleepAction$to_list()}}
\item \href{#method-V1SleepAction-clone}{\code{V1SleepAction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SleepAction-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SleepAction-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SleepAction$new(seconds = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SleepAction-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SleepAction-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SleepAction$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SleepAction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SleepAction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SleepAction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
