% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatefulSet.R
\name{V1StatefulSet}
\alias{V1StatefulSet}
\title{V1StatefulSet}
\description{
V1StatefulSet

V1StatefulSet
}
\details{
StatefulSet represents a set of pods with consistent identities. Identities are defined as:
  - Network: A single stable DNS and hostname.
  - Storage: As many VolumeClaims as requested.

The StatefulSet guarantees that a given network identity will always map to the same storage identity.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSet-new}{\code{V1StatefulSet$new()}}
\item \href{#method-V1StatefulSet-to_list}{\code{V1StatefulSet$to_list()}}
\item \href{#method-V1StatefulSet-clone}{\code{V1StatefulSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSet-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSet$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSet-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSet-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSet$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
