% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatefulSetCondition.R
\name{V1StatefulSetCondition}
\alias{V1StatefulSetCondition}
\title{V1StatefulSetCondition}
\description{
V1StatefulSetCondition

V1StatefulSetCondition
}
\details{
StatefulSetCondition describes the state of a statefulset at a certain point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSetCondition-new}{\code{V1StatefulSetCondition$new()}}
\item \href{#method-V1StatefulSetCondition-to_list}{\code{V1StatefulSetCondition$to_list()}}
\item \href{#method-V1StatefulSetCondition-clone}{\code{V1StatefulSetCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
