% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatefulSetList.R
\name{V1StatefulSetList}
\alias{V1StatefulSetList}
\title{V1StatefulSetList}
\description{
V1StatefulSetList

V1StatefulSetList
}
\details{
StatefulSetList is a collection of StatefulSets.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSetList-new}{\code{V1StatefulSetList$new()}}
\item \href{#method-V1StatefulSetList-to_list}{\code{V1StatefulSetList$to_list()}}
\item \href{#method-V1StatefulSetList-clone}{\code{V1StatefulSetList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
