% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatefulSetOrdinals.R
\name{V1StatefulSetOrdinals}
\alias{V1StatefulSetOrdinals}
\title{V1StatefulSetOrdinals}
\description{
V1StatefulSetOrdinals

V1StatefulSetOrdinals
}
\details{
StatefulSetOrdinals describes the policy used for replica ordinal assignment in this StatefulSet.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSetOrdinals-new}{\code{V1StatefulSetOrdinals$new()}}
\item \href{#method-V1StatefulSetOrdinals-to_list}{\code{V1StatefulSetOrdinals$to_list()}}
\item \href{#method-V1StatefulSetOrdinals-clone}{\code{V1StatefulSetOrdinals$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetOrdinals-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetOrdinals-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetOrdinals$new(start = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetOrdinals-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetOrdinals-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetOrdinals$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetOrdinals-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetOrdinals-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetOrdinals$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
