% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1StatefulSetPersistentVolumeClaimRetentionPolicy.R
\name{V1StatefulSetPersistentVolumeClaimRetentionPolicy}
\alias{V1StatefulSetPersistentVolumeClaimRetentionPolicy}
\title{V1StatefulSetPersistentVolumeClaimRetentionPolicy}
\description{
V1StatefulSetPersistentVolumeClaimRetentionPolicy

V1StatefulSetPersistentVolumeClaimRetentionPolicy
}
\details{
StatefulSetPersistentVolumeClaimRetentionPolicy describes the policy used for PVCs created from the StatefulSet VolumeClaimTemplates.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-new}{\code{V1StatefulSetPersistentVolumeClaimRetentionPolicy$new()}}
\item \href{#method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-to_list}{\code{V1StatefulSetPersistentVolumeClaimRetentionPolicy$to_list()}}
\item \href{#method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-clone}{\code{V1StatefulSetPersistentVolumeClaimRetentionPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetPersistentVolumeClaimRetentionPolicy$new(
  when_deleted = NULL,
  when_scaled = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetPersistentVolumeClaimRetentionPolicy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetPersistentVolumeClaimRetentionPolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetPersistentVolumeClaimRetentionPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
