% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatefulSetSpec.R
\name{V1StatefulSetSpec}
\alias{V1StatefulSetSpec}
\title{V1StatefulSetSpec}
\description{
V1StatefulSetSpec

V1StatefulSetSpec
}
\details{
A StatefulSetSpec is the specification of a StatefulSet.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSetSpec-new}{\code{V1StatefulSetSpec$new()}}
\item \href{#method-V1StatefulSetSpec-to_list}{\code{V1StatefulSetSpec$to_list()}}
\item \href{#method-V1StatefulSetSpec-clone}{\code{V1StatefulSetSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetSpec$new(
  min_ready_seconds = NULL,
  ordinals = NULL,
  persistent_volume_claim_retention_policy = NULL,
  pod_management_policy = NULL,
  replicas = NULL,
  revision_history_limit = NULL,
  selector = NULL,
  service_name = NULL,
  template = NULL,
  update_strategy = NULL,
  volume_claim_templates = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
