% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatefulSetStatus.R
\name{V1StatefulSetStatus}
\alias{V1StatefulSetStatus}
\title{V1StatefulSetStatus}
\description{
V1StatefulSetStatus

V1StatefulSetStatus
}
\details{
StatefulSetStatus represents the current state of a StatefulSet.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSetStatus-new}{\code{V1StatefulSetStatus$new()}}
\item \href{#method-V1StatefulSetStatus-to_list}{\code{V1StatefulSetStatus$to_list()}}
\item \href{#method-V1StatefulSetStatus-clone}{\code{V1StatefulSetStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetStatus$new(
  available_replicas = NULL,
  collision_count = NULL,
  conditions = NULL,
  current_replicas = NULL,
  current_revision = NULL,
  observed_generation = NULL,
  ready_replicas = NULL,
  replicas = NULL,
  update_revision = NULL,
  updated_replicas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
