% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatefulSetUpdateStrategy.R
\name{V1StatefulSetUpdateStrategy}
\alias{V1StatefulSetUpdateStrategy}
\title{V1StatefulSetUpdateStrategy}
\description{
V1StatefulSetUpdateStrategy

V1StatefulSetUpdateStrategy
}
\details{
StatefulSetUpdateStrategy indicates the strategy that the StatefulSet controller will use to perform updates. It includes any additional parameters necessary to perform the update for the indicated strategy.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatefulSetUpdateStrategy-new}{\code{V1StatefulSetUpdateStrategy$new()}}
\item \href{#method-V1StatefulSetUpdateStrategy-to_list}{\code{V1StatefulSetUpdateStrategy$to_list()}}
\item \href{#method-V1StatefulSetUpdateStrategy-clone}{\code{V1StatefulSetUpdateStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetUpdateStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetUpdateStrategy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetUpdateStrategy$new(rolling_update = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetUpdateStrategy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetUpdateStrategy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetUpdateStrategy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatefulSetUpdateStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatefulSetUpdateStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatefulSetUpdateStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
