% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Status.R
\name{V1Status}
\alias{V1Status}
\title{V1Status}
\description{
V1Status

V1Status
}
\details{
Status is a return value for calls that don't return other objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Status-new}{\code{V1Status$new()}}
\item \href{#method-V1Status-to_list}{\code{V1Status$to_list()}}
\item \href{#method-V1Status-clone}{\code{V1Status$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Status-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Status-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Status$new(
  api_version = NULL,
  code = NULL,
  details = NULL,
  kind = NULL,
  message = NULL,
  metadata = NULL,
  reason = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Status-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Status-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Status$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Status-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Status-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Status$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
