% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatusCause.R
\name{V1StatusCause}
\alias{V1StatusCause}
\title{V1StatusCause}
\description{
V1StatusCause

V1StatusCause
}
\details{
StatusCause provides more information about an api.Status failure, including cases when multiple errors are encountered.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatusCause-new}{\code{V1StatusCause$new()}}
\item \href{#method-V1StatusCause-to_list}{\code{V1StatusCause$to_list()}}
\item \href{#method-V1StatusCause-clone}{\code{V1StatusCause$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatusCause-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatusCause-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatusCause$new(field = NULL, message = NULL, reason = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatusCause-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatusCause-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatusCause$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatusCause-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatusCause-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatusCause$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
