% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StatusDetails.R
\name{V1StatusDetails}
\alias{V1StatusDetails}
\title{V1StatusDetails}
\description{
V1StatusDetails

V1StatusDetails
}
\details{
StatusDetails is a set of additional properties that MAY be set by the server to provide additional information about a response. The Reason field of a Status object defines what attributes will be set. Clients must ignore fields that do not match the defined type of each attribute, and should assume that any attribute may be empty, invalid, or under defined.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StatusDetails-new}{\code{V1StatusDetails$new()}}
\item \href{#method-V1StatusDetails-to_list}{\code{V1StatusDetails$to_list()}}
\item \href{#method-V1StatusDetails-clone}{\code{V1StatusDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatusDetails-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatusDetails-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatusDetails$new(
  causes = NULL,
  group = NULL,
  kind = NULL,
  name = NULL,
  retry_after_seconds = NULL,
  uid = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatusDetails-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatusDetails-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatusDetails$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StatusDetails-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StatusDetails-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StatusDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
