% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StorageClass.R
\name{V1StorageClass}
\alias{V1StorageClass}
\title{V1StorageClass}
\description{
V1StorageClass

V1StorageClass
}
\details{
StorageClass describes the parameters for a class of storage for which PersistentVolumes can be dynamically provisioned.

StorageClasses are non-namespaced; the name of the storage class according to etcd is in ObjectMeta.Name.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StorageClass-new}{\code{V1StorageClass$new()}}
\item \href{#method-V1StorageClass-to_list}{\code{V1StorageClass$to_list()}}
\item \href{#method-V1StorageClass-clone}{\code{V1StorageClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StorageClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StorageClass-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StorageClass$new(
  allow_volume_expansion = NULL,
  allowed_topologies = NULL,
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  mount_options = NULL,
  parameters = NULL,
  provisioner = NULL,
  reclaim_policy = NULL,
  volume_binding_mode = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StorageClass-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StorageClass-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StorageClass$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StorageClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StorageClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StorageClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
