% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1StorageOSVolumeSource.R
\name{V1StorageOSVolumeSource}
\alias{V1StorageOSVolumeSource}
\title{V1StorageOSVolumeSource}
\description{
V1StorageOSVolumeSource

V1StorageOSVolumeSource
}
\details{
Represents a StorageOS persistent volume resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1StorageOSVolumeSource-new}{\code{V1StorageOSVolumeSource$new()}}
\item \href{#method-V1StorageOSVolumeSource-to_list}{\code{V1StorageOSVolumeSource$to_list()}}
\item \href{#method-V1StorageOSVolumeSource-clone}{\code{V1StorageOSVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StorageOSVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1StorageOSVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StorageOSVolumeSource$new(
  fs_type = NULL,
  read_only = NULL,
  secret_ref = NULL,
  volume_name = NULL,
  volume_namespace = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StorageOSVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1StorageOSVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StorageOSVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1StorageOSVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1StorageOSVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1StorageOSVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
