% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SubjectAccessReview.R
\name{V1SubjectAccessReview}
\alias{V1SubjectAccessReview}
\title{V1SubjectAccessReview}
\description{
V1SubjectAccessReview

V1SubjectAccessReview
}
\details{
SubjectAccessReview checks whether or not a user or group can perform an action.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SubjectAccessReview-new}{\code{V1SubjectAccessReview$new()}}
\item \href{#method-V1SubjectAccessReview-to_list}{\code{V1SubjectAccessReview$to_list()}}
\item \href{#method-V1SubjectAccessReview-clone}{\code{V1SubjectAccessReview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectAccessReview-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectAccessReview-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectAccessReview$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectAccessReview-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectAccessReview-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectAccessReview$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectAccessReview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectAccessReview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectAccessReview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
