% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SubjectAccessReviewSpec.R
\name{V1SubjectAccessReviewSpec}
\alias{V1SubjectAccessReviewSpec}
\title{V1SubjectAccessReviewSpec}
\description{
V1SubjectAccessReviewSpec

V1SubjectAccessReviewSpec
}
\details{
SubjectAccessReviewSpec is a description of the access request.  Exactly one of ResourceAuthorizationAttributes and NonResourceAuthorizationAttributes must be set
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SubjectAccessReviewSpec-new}{\code{V1SubjectAccessReviewSpec$new()}}
\item \href{#method-V1SubjectAccessReviewSpec-to_list}{\code{V1SubjectAccessReviewSpec$to_list()}}
\item \href{#method-V1SubjectAccessReviewSpec-clone}{\code{V1SubjectAccessReviewSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectAccessReviewSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectAccessReviewSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectAccessReviewSpec$new(
  extra = NULL,
  groups = NULL,
  non_resource_attributes = NULL,
  resource_attributes = NULL,
  uid = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectAccessReviewSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectAccessReviewSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectAccessReviewSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectAccessReviewSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectAccessReviewSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectAccessReviewSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
