% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SubjectRulesReviewStatus.R
\name{V1SubjectRulesReviewStatus}
\alias{V1SubjectRulesReviewStatus}
\title{V1SubjectRulesReviewStatus}
\description{
V1SubjectRulesReviewStatus

V1SubjectRulesReviewStatus
}
\details{
SubjectRulesReviewStatus contains the result of a rules check. This check can be incomplete depending on the set of authorizers the server is configured with and any errors experienced during evaluation. Because authorization rules are additive, if a rule appears in a list it's safe to assume the subject has that permission, even if that list is incomplete.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SubjectRulesReviewStatus-new}{\code{V1SubjectRulesReviewStatus$new()}}
\item \href{#method-V1SubjectRulesReviewStatus-to_list}{\code{V1SubjectRulesReviewStatus$to_list()}}
\item \href{#method-V1SubjectRulesReviewStatus-clone}{\code{V1SubjectRulesReviewStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectRulesReviewStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectRulesReviewStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectRulesReviewStatus$new(
  evaluation_error = NULL,
  incomplete = NULL,
  non_resource_rules = NULL,
  resource_rules = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectRulesReviewStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectRulesReviewStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectRulesReviewStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SubjectRulesReviewStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SubjectRulesReviewStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SubjectRulesReviewStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
