% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1SuccessPolicy.R
\name{V1SuccessPolicy}
\alias{V1SuccessPolicy}
\title{V1SuccessPolicy}
\description{
V1SuccessPolicy

V1SuccessPolicy
}
\details{
SuccessPolicy describes when a Job can be declared as succeeded based on the success of some indexes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1SuccessPolicy-new}{\code{V1SuccessPolicy$new()}}
\item \href{#method-V1SuccessPolicy-to_list}{\code{V1SuccessPolicy$to_list()}}
\item \href{#method-V1SuccessPolicy-clone}{\code{V1SuccessPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SuccessPolicy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1SuccessPolicy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SuccessPolicy$new(rules = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SuccessPolicy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1SuccessPolicy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SuccessPolicy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1SuccessPolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1SuccessPolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1SuccessPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
