% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Taint.R
\name{V1Taint}
\alias{V1Taint}
\title{V1Taint}
\description{
V1Taint

V1Taint
}
\details{
The node this Taint is attached to has the "effect" on any pod that does not tolerate the Taint.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Taint-new}{\code{V1Taint$new()}}
\item \href{#method-V1Taint-to_list}{\code{V1Taint$to_list()}}
\item \href{#method-V1Taint-clone}{\code{V1Taint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Taint-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Taint-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Taint$new(effect = NULL, key = NULL, time_added = NULL, value = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Taint-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Taint-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Taint$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Taint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Taint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Taint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
