% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TCPSocketAction.R
\name{V1TCPSocketAction}
\alias{V1TCPSocketAction}
\title{V1TCPSocketAction}
\description{
V1TCPSocketAction

V1TCPSocketAction
}
\details{
TCPSocketAction describes an action based on opening a socket
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TCPSocketAction-new}{\code{V1TCPSocketAction$new()}}
\item \href{#method-V1TCPSocketAction-to_list}{\code{V1TCPSocketAction$to_list()}}
\item \href{#method-V1TCPSocketAction-clone}{\code{V1TCPSocketAction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TCPSocketAction-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TCPSocketAction-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TCPSocketAction$new(host = NULL, port = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TCPSocketAction-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TCPSocketAction-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TCPSocketAction$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TCPSocketAction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TCPSocketAction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TCPSocketAction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
