% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Time.R
\name{V1Time}
\alias{V1Time}
\title{V1Time}
\description{
V1Time

V1Time
}
\details{
Time is a wrapper around time.Time which supports correct marshaling to YAML and JSON.  Wrappers are provided for many of the factory methods that the time package offers.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Underlying scalar / unstructured value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Time-new}{\code{V1Time$new()}}
\item \href{#method-V1Time-to_list}{\code{V1Time$to_list()}}
\item \href{#method-V1Time-clone}{\code{V1Time$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Time-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Time-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Time$new(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The underlying value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Time-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Time-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return the underlying value (no key wrapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Time$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Time-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Time-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Time$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
