% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TokenRequestSpec.R
\name{V1TokenRequestSpec}
\alias{V1TokenRequestSpec}
\title{V1TokenRequestSpec}
\description{
V1TokenRequestSpec

V1TokenRequestSpec
}
\details{
TokenRequestSpec contains client provided parameters of a token request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TokenRequestSpec-new}{\code{V1TokenRequestSpec$new()}}
\item \href{#method-V1TokenRequestSpec-to_list}{\code{V1TokenRequestSpec$to_list()}}
\item \href{#method-V1TokenRequestSpec-clone}{\code{V1TokenRequestSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenRequestSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenRequestSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenRequestSpec$new(
  audiences = NULL,
  bound_object_ref = NULL,
  expiration_seconds = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenRequestSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenRequestSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenRequestSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenRequestSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenRequestSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenRequestSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
