% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TokenRequestStatus.R
\name{V1TokenRequestStatus}
\alias{V1TokenRequestStatus}
\title{V1TokenRequestStatus}
\description{
V1TokenRequestStatus

V1TokenRequestStatus
}
\details{
TokenRequestStatus is the result of a token request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TokenRequestStatus-new}{\code{V1TokenRequestStatus$new()}}
\item \href{#method-V1TokenRequestStatus-to_list}{\code{V1TokenRequestStatus$to_list()}}
\item \href{#method-V1TokenRequestStatus-clone}{\code{V1TokenRequestStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenRequestStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenRequestStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenRequestStatus$new(expiration_timestamp = NULL, token = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenRequestStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenRequestStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenRequestStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenRequestStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenRequestStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenRequestStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
