% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TokenReview.R
\name{V1TokenReview}
\alias{V1TokenReview}
\title{V1TokenReview}
\description{
V1TokenReview

V1TokenReview
}
\details{
TokenReview attempts to authenticate a token to a known user. Note: TokenReview requests may be cached by the webhook token authenticator plugin in the kube-apiserver.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TokenReview-new}{\code{V1TokenReview$new()}}
\item \href{#method-V1TokenReview-to_list}{\code{V1TokenReview$to_list()}}
\item \href{#method-V1TokenReview-clone}{\code{V1TokenReview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReview-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReview-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReview$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReview-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReview-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReview$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
