% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TokenReviewSpec.R
\name{V1TokenReviewSpec}
\alias{V1TokenReviewSpec}
\title{V1TokenReviewSpec}
\description{
V1TokenReviewSpec

V1TokenReviewSpec
}
\details{
TokenReviewSpec is a description of the token authentication request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TokenReviewSpec-new}{\code{V1TokenReviewSpec$new()}}
\item \href{#method-V1TokenReviewSpec-to_list}{\code{V1TokenReviewSpec$to_list()}}
\item \href{#method-V1TokenReviewSpec-clone}{\code{V1TokenReviewSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReviewSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReviewSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReviewSpec$new(audiences = NULL, token = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReviewSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReviewSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReviewSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReviewSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReviewSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReviewSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
