% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TokenReviewStatus.R
\name{V1TokenReviewStatus}
\alias{V1TokenReviewStatus}
\title{V1TokenReviewStatus}
\description{
V1TokenReviewStatus

V1TokenReviewStatus
}
\details{
TokenReviewStatus is the result of the token authentication request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TokenReviewStatus-new}{\code{V1TokenReviewStatus$new()}}
\item \href{#method-V1TokenReviewStatus-to_list}{\code{V1TokenReviewStatus$to_list()}}
\item \href{#method-V1TokenReviewStatus-clone}{\code{V1TokenReviewStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReviewStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReviewStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReviewStatus$new(
  audiences = NULL,
  authenticated = NULL,
  error = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReviewStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReviewStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReviewStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TokenReviewStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TokenReviewStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TokenReviewStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
