% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Toleration.R
\name{V1Toleration}
\alias{V1Toleration}
\title{V1Toleration}
\description{
V1Toleration

V1Toleration
}
\details{
The pod this Toleration is attached to tolerates any taint that matches the triple <key,value,effect> using the matching operator <operator>.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Toleration-new}{\code{V1Toleration$new()}}
\item \href{#method-V1Toleration-to_list}{\code{V1Toleration$to_list()}}
\item \href{#method-V1Toleration-clone}{\code{V1Toleration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Toleration-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Toleration-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Toleration$new(
  effect = NULL,
  key = NULL,
  operator = NULL,
  toleration_seconds = NULL,
  value = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Toleration-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Toleration-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Toleration$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Toleration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Toleration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Toleration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
