% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TopologySelectorLabelRequirement.R
\name{V1TopologySelectorLabelRequirement}
\alias{V1TopologySelectorLabelRequirement}
\title{V1TopologySelectorLabelRequirement}
\description{
V1TopologySelectorLabelRequirement

V1TopologySelectorLabelRequirement
}
\details{
A topology selector requirement is a selector that matches given label. This is an alpha feature and may change in the future.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TopologySelectorLabelRequirement-new}{\code{V1TopologySelectorLabelRequirement$new()}}
\item \href{#method-V1TopologySelectorLabelRequirement-to_list}{\code{V1TopologySelectorLabelRequirement$to_list()}}
\item \href{#method-V1TopologySelectorLabelRequirement-clone}{\code{V1TopologySelectorLabelRequirement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySelectorLabelRequirement-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySelectorLabelRequirement-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySelectorLabelRequirement$new(key = NULL, values = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySelectorLabelRequirement-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySelectorLabelRequirement-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySelectorLabelRequirement$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySelectorLabelRequirement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySelectorLabelRequirement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySelectorLabelRequirement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
