% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TopologySelectorTerm.R
\name{V1TopologySelectorTerm}
\alias{V1TopologySelectorTerm}
\title{V1TopologySelectorTerm}
\description{
V1TopologySelectorTerm

V1TopologySelectorTerm
}
\details{
A topology selector term represents the result of label queries. A null or empty topology selector term matches no objects. The requirements of them are ANDed. It provides a subset of functionality as NodeSelectorTerm. This is an alpha feature and may change in the future.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TopologySelectorTerm-new}{\code{V1TopologySelectorTerm$new()}}
\item \href{#method-V1TopologySelectorTerm-to_list}{\code{V1TopologySelectorTerm$to_list()}}
\item \href{#method-V1TopologySelectorTerm-clone}{\code{V1TopologySelectorTerm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySelectorTerm-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySelectorTerm-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySelectorTerm$new(match_label_expressions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySelectorTerm-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySelectorTerm-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySelectorTerm$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySelectorTerm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySelectorTerm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySelectorTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
