% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TopologySpreadConstraint.R
\name{V1TopologySpreadConstraint}
\alias{V1TopologySpreadConstraint}
\title{V1TopologySpreadConstraint}
\description{
V1TopologySpreadConstraint

V1TopologySpreadConstraint
}
\details{
TopologySpreadConstraint specifies how to spread matching pods among the given topology.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TopologySpreadConstraint-new}{\code{V1TopologySpreadConstraint$new()}}
\item \href{#method-V1TopologySpreadConstraint-to_list}{\code{V1TopologySpreadConstraint$to_list()}}
\item \href{#method-V1TopologySpreadConstraint-clone}{\code{V1TopologySpreadConstraint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySpreadConstraint-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySpreadConstraint-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySpreadConstraint$new(
  label_selector = NULL,
  match_label_keys = NULL,
  max_skew = NULL,
  min_domains = NULL,
  node_affinity_policy = NULL,
  node_taints_policy = NULL,
  topology_key = NULL,
  when_unsatisfiable = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySpreadConstraint-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySpreadConstraint-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySpreadConstraint$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TopologySpreadConstraint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TopologySpreadConstraint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TopologySpreadConstraint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
