% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1TypedLocalObjectReference.R
\name{V1TypedLocalObjectReference}
\alias{V1TypedLocalObjectReference}
\title{V1TypedLocalObjectReference}
\description{
V1TypedLocalObjectReference

V1TypedLocalObjectReference
}
\details{
TypedLocalObjectReference contains enough information to let you locate the typed referenced object inside the same namespace.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1TypedLocalObjectReference-new}{\code{V1TypedLocalObjectReference$new()}}
\item \href{#method-V1TypedLocalObjectReference-to_list}{\code{V1TypedLocalObjectReference$to_list()}}
\item \href{#method-V1TypedLocalObjectReference-clone}{\code{V1TypedLocalObjectReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TypedLocalObjectReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1TypedLocalObjectReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TypedLocalObjectReference$new(api_group = NULL, kind = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TypedLocalObjectReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1TypedLocalObjectReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TypedLocalObjectReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1TypedLocalObjectReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1TypedLocalObjectReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1TypedLocalObjectReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
