% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1UncountedTerminatedPods.R
\name{V1UncountedTerminatedPods}
\alias{V1UncountedTerminatedPods}
\title{V1UncountedTerminatedPods}
\description{
V1UncountedTerminatedPods

V1UncountedTerminatedPods
}
\details{
UncountedTerminatedPods holds UIDs of Pods that have terminated but haven't been accounted in Job status counters.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1UncountedTerminatedPods-new}{\code{V1UncountedTerminatedPods$new()}}
\item \href{#method-V1UncountedTerminatedPods-to_list}{\code{V1UncountedTerminatedPods$to_list()}}
\item \href{#method-V1UncountedTerminatedPods-clone}{\code{V1UncountedTerminatedPods$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UncountedTerminatedPods-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1UncountedTerminatedPods-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UncountedTerminatedPods$new(failed = NULL, succeeded = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UncountedTerminatedPods-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1UncountedTerminatedPods-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UncountedTerminatedPods$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UncountedTerminatedPods-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1UncountedTerminatedPods-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UncountedTerminatedPods$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
