% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1UserInfo.R
\name{V1UserInfo}
\alias{V1UserInfo}
\title{V1UserInfo}
\description{
V1UserInfo

V1UserInfo
}
\details{
UserInfo holds the information about the user needed to implement the user.Info interface.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1UserInfo-new}{\code{V1UserInfo$new()}}
\item \href{#method-V1UserInfo-to_list}{\code{V1UserInfo$to_list()}}
\item \href{#method-V1UserInfo-clone}{\code{V1UserInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UserInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1UserInfo-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UserInfo$new(extra = NULL, groups = NULL, uid = NULL, username = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UserInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1UserInfo-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UserInfo$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UserInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1UserInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UserInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
