% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1UserSubject.R
\name{V1UserSubject}
\alias{V1UserSubject}
\title{V1UserSubject}
\description{
V1UserSubject

V1UserSubject
}
\details{
UserSubject holds detailed information for user-kind subject.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1UserSubject-new}{\code{V1UserSubject$new()}}
\item \href{#method-V1UserSubject-to_list}{\code{V1UserSubject$to_list()}}
\item \href{#method-V1UserSubject-clone}{\code{V1UserSubject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UserSubject-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1UserSubject-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UserSubject$new(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UserSubject-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1UserSubject-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UserSubject$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1UserSubject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1UserSubject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1UserSubject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
