% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1ValidatingAdmissionPolicyBindingList.R
\name{V1ValidatingAdmissionPolicyBindingList}
\alias{V1ValidatingAdmissionPolicyBindingList}
\title{V1ValidatingAdmissionPolicyBindingList}
\description{
V1ValidatingAdmissionPolicyBindingList

V1ValidatingAdmissionPolicyBindingList
}
\details{
ValidatingAdmissionPolicyBindingList is a list of ValidatingAdmissionPolicyBinding.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ValidatingAdmissionPolicyBindingList-new}{\code{V1ValidatingAdmissionPolicyBindingList$new()}}
\item \href{#method-V1ValidatingAdmissionPolicyBindingList-to_list}{\code{V1ValidatingAdmissionPolicyBindingList$to_list()}}
\item \href{#method-V1ValidatingAdmissionPolicyBindingList-clone}{\code{V1ValidatingAdmissionPolicyBindingList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingAdmissionPolicyBindingList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingAdmissionPolicyBindingList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingAdmissionPolicyBindingList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingAdmissionPolicyBindingList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingAdmissionPolicyBindingList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingAdmissionPolicyBindingList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingAdmissionPolicyBindingList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingAdmissionPolicyBindingList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingAdmissionPolicyBindingList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
