% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ValidatingAdmissionPolicyList.R
\name{V1ValidatingAdmissionPolicyList}
\alias{V1ValidatingAdmissionPolicyList}
\title{V1ValidatingAdmissionPolicyList}
\description{
V1ValidatingAdmissionPolicyList

V1ValidatingAdmissionPolicyList
}
\details{
ValidatingAdmissionPolicyList is a list of ValidatingAdmissionPolicy.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ValidatingAdmissionPolicyList-new}{\code{V1ValidatingAdmissionPolicyList$new()}}
\item \href{#method-V1ValidatingAdmissionPolicyList-to_list}{\code{V1ValidatingAdmissionPolicyList$to_list()}}
\item \href{#method-V1ValidatingAdmissionPolicyList-clone}{\code{V1ValidatingAdmissionPolicyList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingAdmissionPolicyList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingAdmissionPolicyList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingAdmissionPolicyList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingAdmissionPolicyList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingAdmissionPolicyList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingAdmissionPolicyList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingAdmissionPolicyList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingAdmissionPolicyList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingAdmissionPolicyList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
