% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ValidatingWebhook.R
\name{V1ValidatingWebhook}
\alias{V1ValidatingWebhook}
\title{V1ValidatingWebhook}
\description{
V1ValidatingWebhook

V1ValidatingWebhook
}
\details{
ValidatingWebhook describes an admission webhook and the resources and operations it applies to.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ValidatingWebhook-new}{\code{V1ValidatingWebhook$new()}}
\item \href{#method-V1ValidatingWebhook-to_list}{\code{V1ValidatingWebhook$to_list()}}
\item \href{#method-V1ValidatingWebhook-clone}{\code{V1ValidatingWebhook$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingWebhook-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingWebhook-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingWebhook$new(
  admission_review_versions = NULL,
  client_config = NULL,
  failure_policy = NULL,
  match_conditions = NULL,
  match_policy = NULL,
  name = NULL,
  namespace_selector = NULL,
  object_selector = NULL,
  rules = NULL,
  side_effects = NULL,
  timeout_seconds = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingWebhook-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingWebhook-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingWebhook$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingWebhook-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingWebhook-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingWebhook$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
