% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gen_model_V1ValidatingWebhookConfigurationList.R
\name{V1ValidatingWebhookConfigurationList}
\alias{V1ValidatingWebhookConfigurationList}
\title{V1ValidatingWebhookConfigurationList}
\description{
V1ValidatingWebhookConfigurationList

V1ValidatingWebhookConfigurationList
}
\details{
ValidatingWebhookConfigurationList is a list of ValidatingWebhookConfiguration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ValidatingWebhookConfigurationList-new}{\code{V1ValidatingWebhookConfigurationList$new()}}
\item \href{#method-V1ValidatingWebhookConfigurationList-to_list}{\code{V1ValidatingWebhookConfigurationList$to_list()}}
\item \href{#method-V1ValidatingWebhookConfigurationList-clone}{\code{V1ValidatingWebhookConfigurationList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingWebhookConfigurationList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingWebhookConfigurationList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingWebhookConfigurationList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingWebhookConfigurationList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingWebhookConfigurationList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingWebhookConfigurationList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidatingWebhookConfigurationList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidatingWebhookConfigurationList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidatingWebhookConfigurationList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
