% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Validation.R
\name{V1Validation}
\alias{V1Validation}
\title{V1Validation}
\description{
V1Validation

V1Validation
}
\details{
Validation specifies the CEL expression which is used to apply the validation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Validation-new}{\code{V1Validation$new()}}
\item \href{#method-V1Validation-to_list}{\code{V1Validation$to_list()}}
\item \href{#method-V1Validation-clone}{\code{V1Validation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Validation-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Validation-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Validation$new(
  expression = NULL,
  message = NULL,
  message_expression = NULL,
  reason = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Validation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Validation-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Validation$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Validation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Validation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Validation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
