% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1ValidationRule.R
\name{V1ValidationRule}
\alias{V1ValidationRule}
\title{V1ValidationRule}
\description{
V1ValidationRule

V1ValidationRule
}
\details{
ValidationRule describes a validation rule written in the CEL expression language.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1ValidationRule-new}{\code{V1ValidationRule$new()}}
\item \href{#method-V1ValidationRule-to_list}{\code{V1ValidationRule$to_list()}}
\item \href{#method-V1ValidationRule-clone}{\code{V1ValidationRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidationRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidationRule-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidationRule$new(
  field_path = NULL,
  message = NULL,
  message_expression = NULL,
  optional_old_self = NULL,
  reason = NULL,
  rule = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidationRule-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidationRule-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidationRule$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1ValidationRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1ValidationRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1ValidationRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
