% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1Volume.R
\name{V1Volume}
\alias{V1Volume}
\title{V1Volume}
\description{
V1Volume

V1Volume
}
\details{
Volume represents a named volume in a pod that may be accessed by any container in the pod.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1Volume-new}{\code{V1Volume$new()}}
\item \href{#method-V1Volume-to_list}{\code{V1Volume$to_list()}}
\item \href{#method-V1Volume-clone}{\code{V1Volume$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Volume-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1Volume-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Volume$new(
  aws_elastic_block_store = NULL,
  azure_disk = NULL,
  azure_file = NULL,
  cephfs = NULL,
  cinder = NULL,
  config_map = NULL,
  csi = NULL,
  downward_api = NULL,
  empty_dir = NULL,
  ephemeral = NULL,
  fc = NULL,
  flex_volume = NULL,
  flocker = NULL,
  gce_persistent_disk = NULL,
  git_repo = NULL,
  glusterfs = NULL,
  host_path = NULL,
  image = NULL,
  iscsi = NULL,
  name = NULL,
  nfs = NULL,
  persistent_volume_claim = NULL,
  photon_persistent_disk = NULL,
  portworx_volume = NULL,
  projected = NULL,
  quobyte = NULL,
  rbd = NULL,
  scale_io = NULL,
  secret = NULL,
  storageos = NULL,
  vsphere_volume = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Volume-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1Volume-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Volume$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1Volume-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1Volume-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1Volume$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
