% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeAttachment.R
\name{V1VolumeAttachment}
\alias{V1VolumeAttachment}
\title{V1VolumeAttachment}
\description{
V1VolumeAttachment

V1VolumeAttachment
}
\details{
VolumeAttachment captures the intent to attach or detach the specified volume to/from the specified node.

VolumeAttachment objects are non-namespaced.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeAttachment-new}{\code{V1VolumeAttachment$new()}}
\item \href{#method-V1VolumeAttachment-to_list}{\code{V1VolumeAttachment$to_list()}}
\item \href{#method-V1VolumeAttachment-clone}{\code{V1VolumeAttachment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachment-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachment-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachment$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachment-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachment-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachment$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
