% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeAttachmentSource.R
\name{V1VolumeAttachmentSource}
\alias{V1VolumeAttachmentSource}
\title{V1VolumeAttachmentSource}
\description{
V1VolumeAttachmentSource

V1VolumeAttachmentSource
}
\details{
VolumeAttachmentSource represents a volume that should be attached. Right now only PersistenVolumes can be attached via external attacher, in future we may allow also inline volumes in pods. Exactly one member can be set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeAttachmentSource-new}{\code{V1VolumeAttachmentSource$new()}}
\item \href{#method-V1VolumeAttachmentSource-to_list}{\code{V1VolumeAttachmentSource$to_list()}}
\item \href{#method-V1VolumeAttachmentSource-clone}{\code{V1VolumeAttachmentSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachmentSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachmentSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachmentSource$new(
  inline_volume_spec = NULL,
  persistent_volume_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachmentSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachmentSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachmentSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachmentSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachmentSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachmentSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
