% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeAttachmentStatus.R
\name{V1VolumeAttachmentStatus}
\alias{V1VolumeAttachmentStatus}
\title{V1VolumeAttachmentStatus}
\description{
V1VolumeAttachmentStatus

V1VolumeAttachmentStatus
}
\details{
VolumeAttachmentStatus is the status of a VolumeAttachment request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeAttachmentStatus-new}{\code{V1VolumeAttachmentStatus$new()}}
\item \href{#method-V1VolumeAttachmentStatus-to_list}{\code{V1VolumeAttachmentStatus$to_list()}}
\item \href{#method-V1VolumeAttachmentStatus-clone}{\code{V1VolumeAttachmentStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachmentStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachmentStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachmentStatus$new(
  attach_error = NULL,
  attached = NULL,
  attachment_metadata = NULL,
  detach_error = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachmentStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachmentStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachmentStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeAttachmentStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeAttachmentStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeAttachmentStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
