% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeDevice.R
\name{V1VolumeDevice}
\alias{V1VolumeDevice}
\title{V1VolumeDevice}
\description{
V1VolumeDevice

V1VolumeDevice
}
\details{
volumeDevice describes a mapping of a raw block device within a container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeDevice-new}{\code{V1VolumeDevice$new()}}
\item \href{#method-V1VolumeDevice-to_list}{\code{V1VolumeDevice$to_list()}}
\item \href{#method-V1VolumeDevice-clone}{\code{V1VolumeDevice$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeDevice-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeDevice-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeDevice$new(device_path = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeDevice-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeDevice-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeDevice$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeDevice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeDevice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeDevice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
