% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeError.R
\name{V1VolumeError}
\alias{V1VolumeError}
\title{V1VolumeError}
\description{
V1VolumeError

V1VolumeError
}
\details{
VolumeError captures an error encountered during a volume operation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeError-new}{\code{V1VolumeError$new()}}
\item \href{#method-V1VolumeError-to_list}{\code{V1VolumeError$to_list()}}
\item \href{#method-V1VolumeError-clone}{\code{V1VolumeError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeError-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeError-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeError$new(message = NULL, time = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeError-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeError-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeError$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
