% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeMount.R
\name{V1VolumeMount}
\alias{V1VolumeMount}
\title{V1VolumeMount}
\description{
V1VolumeMount

V1VolumeMount
}
\details{
VolumeMount describes a mounting of a Volume within a container.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeMount-new}{\code{V1VolumeMount$new()}}
\item \href{#method-V1VolumeMount-to_list}{\code{V1VolumeMount$to_list()}}
\item \href{#method-V1VolumeMount-clone}{\code{V1VolumeMount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeMount-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeMount-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeMount$new(
  mount_path = NULL,
  mount_propagation = NULL,
  name = NULL,
  read_only = NULL,
  recursive_read_only = NULL,
  sub_path = NULL,
  sub_path_expr = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeMount-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeMount-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeMount$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeMount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeMount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeMount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
