% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeMountStatus.R
\name{V1VolumeMountStatus}
\alias{V1VolumeMountStatus}
\title{V1VolumeMountStatus}
\description{
V1VolumeMountStatus

V1VolumeMountStatus
}
\details{
VolumeMountStatus shows status of volume mounts.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeMountStatus-new}{\code{V1VolumeMountStatus$new()}}
\item \href{#method-V1VolumeMountStatus-to_list}{\code{V1VolumeMountStatus$to_list()}}
\item \href{#method-V1VolumeMountStatus-clone}{\code{V1VolumeMountStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeMountStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeMountStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeMountStatus$new(
  mount_path = NULL,
  name = NULL,
  read_only = NULL,
  recursive_read_only = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeMountStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeMountStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeMountStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeMountStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeMountStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeMountStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
