% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeNodeAffinity.R
\name{V1VolumeNodeAffinity}
\alias{V1VolumeNodeAffinity}
\title{V1VolumeNodeAffinity}
\description{
V1VolumeNodeAffinity

V1VolumeNodeAffinity
}
\details{
VolumeNodeAffinity defines constraints that limit what nodes this volume can be accessed from.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeNodeAffinity-new}{\code{V1VolumeNodeAffinity$new()}}
\item \href{#method-V1VolumeNodeAffinity-to_list}{\code{V1VolumeNodeAffinity$to_list()}}
\item \href{#method-V1VolumeNodeAffinity-clone}{\code{V1VolumeNodeAffinity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeNodeAffinity-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeNodeAffinity-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeNodeAffinity$new(required = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeNodeAffinity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeNodeAffinity-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeNodeAffinity$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeNodeAffinity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeNodeAffinity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeNodeAffinity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
