% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeProjection.R
\name{V1VolumeProjection}
\alias{V1VolumeProjection}
\title{V1VolumeProjection}
\description{
V1VolumeProjection

V1VolumeProjection
}
\details{
Projection that may be projected along with other supported volume types. Exactly one of these fields must be set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeProjection-new}{\code{V1VolumeProjection$new()}}
\item \href{#method-V1VolumeProjection-to_list}{\code{V1VolumeProjection$to_list()}}
\item \href{#method-V1VolumeProjection-clone}{\code{V1VolumeProjection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeProjection-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeProjection-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeProjection$new(
  cluster_trust_bundle = NULL,
  config_map = NULL,
  downward_api = NULL,
  secret = NULL,
  service_account_token = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeProjection-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeProjection-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeProjection$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeProjection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeProjection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeProjection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
