% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VolumeResourceRequirements.R
\name{V1VolumeResourceRequirements}
\alias{V1VolumeResourceRequirements}
\title{V1VolumeResourceRequirements}
\description{
V1VolumeResourceRequirements

V1VolumeResourceRequirements
}
\details{
VolumeResourceRequirements describes the storage resource requirements for a volume.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VolumeResourceRequirements-new}{\code{V1VolumeResourceRequirements$new()}}
\item \href{#method-V1VolumeResourceRequirements-to_list}{\code{V1VolumeResourceRequirements$to_list()}}
\item \href{#method-V1VolumeResourceRequirements-clone}{\code{V1VolumeResourceRequirements$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeResourceRequirements-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeResourceRequirements-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeResourceRequirements$new(limits = NULL, requests = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeResourceRequirements-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeResourceRequirements-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeResourceRequirements$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VolumeResourceRequirements-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VolumeResourceRequirements-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VolumeResourceRequirements$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
