% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1VsphereVirtualDiskVolumeSource.R
\name{V1VsphereVirtualDiskVolumeSource}
\alias{V1VsphereVirtualDiskVolumeSource}
\title{V1VsphereVirtualDiskVolumeSource}
\description{
V1VsphereVirtualDiskVolumeSource

V1VsphereVirtualDiskVolumeSource
}
\details{
Represents a vSphere volume resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1VsphereVirtualDiskVolumeSource-new}{\code{V1VsphereVirtualDiskVolumeSource$new()}}
\item \href{#method-V1VsphereVirtualDiskVolumeSource-to_list}{\code{V1VsphereVirtualDiskVolumeSource$to_list()}}
\item \href{#method-V1VsphereVirtualDiskVolumeSource-clone}{\code{V1VsphereVirtualDiskVolumeSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VsphereVirtualDiskVolumeSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1VsphereVirtualDiskVolumeSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VsphereVirtualDiskVolumeSource$new(
  fs_type = NULL,
  storage_policy_id = NULL,
  storage_policy_name = NULL,
  volume_path = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VsphereVirtualDiskVolumeSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1VsphereVirtualDiskVolumeSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VsphereVirtualDiskVolumeSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1VsphereVirtualDiskVolumeSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1VsphereVirtualDiskVolumeSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1VsphereVirtualDiskVolumeSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
