% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1WatchEvent.R
\name{V1WatchEvent}
\alias{V1WatchEvent}
\title{V1WatchEvent}
\description{
V1WatchEvent

V1WatchEvent
}
\details{
Event represents a single event to a watched resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1WatchEvent-new}{\code{V1WatchEvent$new()}}
\item \href{#method-V1WatchEvent-to_list}{\code{V1WatchEvent$to_list()}}
\item \href{#method-V1WatchEvent-clone}{\code{V1WatchEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WatchEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1WatchEvent-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WatchEvent$new(object = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WatchEvent-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1WatchEvent-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WatchEvent$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WatchEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1WatchEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WatchEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
