% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1WebhookConversion.R
\name{V1WebhookConversion}
\alias{V1WebhookConversion}
\title{V1WebhookConversion}
\description{
V1WebhookConversion

V1WebhookConversion
}
\details{
WebhookConversion describes how to call a conversion webhook
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1WebhookConversion-new}{\code{V1WebhookConversion$new()}}
\item \href{#method-V1WebhookConversion-to_list}{\code{V1WebhookConversion$to_list()}}
\item \href{#method-V1WebhookConversion-clone}{\code{V1WebhookConversion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WebhookConversion-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1WebhookConversion-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WebhookConversion$new(
  client_config = NULL,
  conversion_review_versions = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WebhookConversion-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1WebhookConversion-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WebhookConversion$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WebhookConversion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1WebhookConversion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WebhookConversion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
