% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1WeightedPodAffinityTerm.R
\name{V1WeightedPodAffinityTerm}
\alias{V1WeightedPodAffinityTerm}
\title{V1WeightedPodAffinityTerm}
\description{
V1WeightedPodAffinityTerm

V1WeightedPodAffinityTerm
}
\details{
The weights of all of the matched WeightedPodAffinityTerm fields are added per-node to find the most preferred node(s)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1WeightedPodAffinityTerm-new}{\code{V1WeightedPodAffinityTerm$new()}}
\item \href{#method-V1WeightedPodAffinityTerm-to_list}{\code{V1WeightedPodAffinityTerm$to_list()}}
\item \href{#method-V1WeightedPodAffinityTerm-clone}{\code{V1WeightedPodAffinityTerm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WeightedPodAffinityTerm-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1WeightedPodAffinityTerm-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WeightedPodAffinityTerm$new(pod_affinity_term = NULL, weight = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WeightedPodAffinityTerm-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1WeightedPodAffinityTerm-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WeightedPodAffinityTerm$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WeightedPodAffinityTerm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1WeightedPodAffinityTerm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WeightedPodAffinityTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
