% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V1WindowsSecurityContextOptions.R
\name{V1WindowsSecurityContextOptions}
\alias{V1WindowsSecurityContextOptions}
\title{V1WindowsSecurityContextOptions}
\description{
V1WindowsSecurityContextOptions

V1WindowsSecurityContextOptions
}
\details{
WindowsSecurityContextOptions contain Windows-specific options and credentials.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V1WindowsSecurityContextOptions-new}{\code{V1WindowsSecurityContextOptions$new()}}
\item \href{#method-V1WindowsSecurityContextOptions-to_list}{\code{V1WindowsSecurityContextOptions$to_list()}}
\item \href{#method-V1WindowsSecurityContextOptions-clone}{\code{V1WindowsSecurityContextOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WindowsSecurityContextOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-V1WindowsSecurityContextOptions-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WindowsSecurityContextOptions$new(
  gmsa_credential_spec = NULL,
  gmsa_credential_spec_name = NULL,
  host_process = NULL,
  run_as_user_name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WindowsSecurityContextOptions-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V1WindowsSecurityContextOptions-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WindowsSecurityContextOptions$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V1WindowsSecurityContextOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V1WindowsSecurityContextOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V1WindowsSecurityContextOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
