% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2ContainerResourceMetricSource.R
\name{V2ContainerResourceMetricSource}
\alias{V2ContainerResourceMetricSource}
\title{V2ContainerResourceMetricSource}
\description{
V2ContainerResourceMetricSource

V2ContainerResourceMetricSource
}
\details{
ContainerResourceMetricSource indicates how to scale on a resource metric known to Kubernetes, as specified in requests and limits, describing each pod in the current scale target (e.g. CPU or memory).  The values will be averaged together before being compared to the target.  Such metrics are built in to Kubernetes, and have special scaling options on top of those available to normal per-pod metrics using the "pods" source.  Only one "target" type should be set.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ContainerResourceMetricSource-new}{\code{V2ContainerResourceMetricSource$new()}}
\item \href{#method-V2ContainerResourceMetricSource-to_list}{\code{V2ContainerResourceMetricSource$to_list()}}
\item \href{#method-V2ContainerResourceMetricSource-clone}{\code{V2ContainerResourceMetricSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ContainerResourceMetricSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2ContainerResourceMetricSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ContainerResourceMetricSource$new(
  container = NULL,
  name = NULL,
  target = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ContainerResourceMetricSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2ContainerResourceMetricSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ContainerResourceMetricSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ContainerResourceMetricSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ContainerResourceMetricSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ContainerResourceMetricSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
