% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2ContainerResourceMetricStatus.R
\name{V2ContainerResourceMetricStatus}
\alias{V2ContainerResourceMetricStatus}
\title{V2ContainerResourceMetricStatus}
\description{
V2ContainerResourceMetricStatus

V2ContainerResourceMetricStatus
}
\details{
ContainerResourceMetricStatus indicates the current value of a resource metric known to Kubernetes, as specified in requests and limits, describing a single container in each pod in the current scale target (e.g. CPU or memory).  Such metrics are built in to Kubernetes, and have special scaling options on top of those available to normal per-pod metrics using the "pods" source.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ContainerResourceMetricStatus-new}{\code{V2ContainerResourceMetricStatus$new()}}
\item \href{#method-V2ContainerResourceMetricStatus-to_list}{\code{V2ContainerResourceMetricStatus$to_list()}}
\item \href{#method-V2ContainerResourceMetricStatus-clone}{\code{V2ContainerResourceMetricStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ContainerResourceMetricStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2ContainerResourceMetricStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ContainerResourceMetricStatus$new(
  container = NULL,
  current = NULL,
  name = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ContainerResourceMetricStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2ContainerResourceMetricStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ContainerResourceMetricStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ContainerResourceMetricStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ContainerResourceMetricStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ContainerResourceMetricStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
