% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2CrossVersionObjectReference.R
\name{V2CrossVersionObjectReference}
\alias{V2CrossVersionObjectReference}
\title{V2CrossVersionObjectReference}
\description{
V2CrossVersionObjectReference

V2CrossVersionObjectReference
}
\details{
CrossVersionObjectReference contains enough information to let you identify the referred resource.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2CrossVersionObjectReference-new}{\code{V2CrossVersionObjectReference$new()}}
\item \href{#method-V2CrossVersionObjectReference-to_list}{\code{V2CrossVersionObjectReference$to_list()}}
\item \href{#method-V2CrossVersionObjectReference-clone}{\code{V2CrossVersionObjectReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2CrossVersionObjectReference-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2CrossVersionObjectReference-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2CrossVersionObjectReference$new(api_version = NULL, kind = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2CrossVersionObjectReference-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2CrossVersionObjectReference-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2CrossVersionObjectReference$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2CrossVersionObjectReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2CrossVersionObjectReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2CrossVersionObjectReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
