% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2ExternalMetricSource.R
\name{V2ExternalMetricSource}
\alias{V2ExternalMetricSource}
\title{V2ExternalMetricSource}
\description{
V2ExternalMetricSource

V2ExternalMetricSource
}
\details{
ExternalMetricSource indicates how to scale on a metric not associated with any Kubernetes object (for example length of queue in cloud messaging service, or QPS from loadbalancer running outside of cluster).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ExternalMetricSource-new}{\code{V2ExternalMetricSource$new()}}
\item \href{#method-V2ExternalMetricSource-to_list}{\code{V2ExternalMetricSource$to_list()}}
\item \href{#method-V2ExternalMetricSource-clone}{\code{V2ExternalMetricSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ExternalMetricSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2ExternalMetricSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ExternalMetricSource$new(metric = NULL, target = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ExternalMetricSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2ExternalMetricSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ExternalMetricSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ExternalMetricSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ExternalMetricSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ExternalMetricSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
