% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2ExternalMetricStatus.R
\name{V2ExternalMetricStatus}
\alias{V2ExternalMetricStatus}
\title{V2ExternalMetricStatus}
\description{
V2ExternalMetricStatus

V2ExternalMetricStatus
}
\details{
ExternalMetricStatus indicates the current value of a global metric not associated with any Kubernetes object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ExternalMetricStatus-new}{\code{V2ExternalMetricStatus$new()}}
\item \href{#method-V2ExternalMetricStatus-to_list}{\code{V2ExternalMetricStatus$to_list()}}
\item \href{#method-V2ExternalMetricStatus-clone}{\code{V2ExternalMetricStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ExternalMetricStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2ExternalMetricStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ExternalMetricStatus$new(current = NULL, metric = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ExternalMetricStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2ExternalMetricStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ExternalMetricStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ExternalMetricStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ExternalMetricStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ExternalMetricStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
