% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HorizontalPodAutoscaler.R
\name{V2HorizontalPodAutoscaler}
\alias{V2HorizontalPodAutoscaler}
\title{V2HorizontalPodAutoscaler}
\description{
V2HorizontalPodAutoscaler

V2HorizontalPodAutoscaler
}
\details{
HorizontalPodAutoscaler is the configuration for a horizontal pod autoscaler, which automatically manages the replica count of any resource implementing the scale subresource based on the metrics specified.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HorizontalPodAutoscaler-new}{\code{V2HorizontalPodAutoscaler$new()}}
\item \href{#method-V2HorizontalPodAutoscaler-to_list}{\code{V2HorizontalPodAutoscaler$to_list()}}
\item \href{#method-V2HorizontalPodAutoscaler-clone}{\code{V2HorizontalPodAutoscaler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscaler-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscaler-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscaler$new(
  api_version = NULL,
  kind = NULL,
  metadata = NULL,
  spec = NULL,
  status = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscaler-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscaler-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscaler$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscaler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscaler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
