% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HorizontalPodAutoscalerBehavior.R
\name{V2HorizontalPodAutoscalerBehavior}
\alias{V2HorizontalPodAutoscalerBehavior}
\title{V2HorizontalPodAutoscalerBehavior}
\description{
V2HorizontalPodAutoscalerBehavior

V2HorizontalPodAutoscalerBehavior
}
\details{
HorizontalPodAutoscalerBehavior configures the scaling behavior of the target in both Up and Down directions (scaleUp and scaleDown fields respectively).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HorizontalPodAutoscalerBehavior-new}{\code{V2HorizontalPodAutoscalerBehavior$new()}}
\item \href{#method-V2HorizontalPodAutoscalerBehavior-to_list}{\code{V2HorizontalPodAutoscalerBehavior$to_list()}}
\item \href{#method-V2HorizontalPodAutoscalerBehavior-clone}{\code{V2HorizontalPodAutoscalerBehavior$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerBehavior-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerBehavior-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerBehavior$new(scale_down = NULL, scale_up = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerBehavior-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerBehavior-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerBehavior$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerBehavior-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerBehavior-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerBehavior$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
