% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HorizontalPodAutoscalerCondition.R
\name{V2HorizontalPodAutoscalerCondition}
\alias{V2HorizontalPodAutoscalerCondition}
\title{V2HorizontalPodAutoscalerCondition}
\description{
V2HorizontalPodAutoscalerCondition

V2HorizontalPodAutoscalerCondition
}
\details{
HorizontalPodAutoscalerCondition describes the state of a HorizontalPodAutoscaler at a certain point.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HorizontalPodAutoscalerCondition-new}{\code{V2HorizontalPodAutoscalerCondition$new()}}
\item \href{#method-V2HorizontalPodAutoscalerCondition-to_list}{\code{V2HorizontalPodAutoscalerCondition$to_list()}}
\item \href{#method-V2HorizontalPodAutoscalerCondition-clone}{\code{V2HorizontalPodAutoscalerCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerCondition-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerCondition$new(
  last_transition_time = NULL,
  message = NULL,
  reason = NULL,
  status = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerCondition-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerCondition-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerCondition$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
