% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HorizontalPodAutoscalerList.R
\name{V2HorizontalPodAutoscalerList}
\alias{V2HorizontalPodAutoscalerList}
\title{V2HorizontalPodAutoscalerList}
\description{
V2HorizontalPodAutoscalerList

V2HorizontalPodAutoscalerList
}
\details{
HorizontalPodAutoscalerList is a list of horizontal pod autoscaler objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HorizontalPodAutoscalerList-new}{\code{V2HorizontalPodAutoscalerList$new()}}
\item \href{#method-V2HorizontalPodAutoscalerList-to_list}{\code{V2HorizontalPodAutoscalerList$to_list()}}
\item \href{#method-V2HorizontalPodAutoscalerList-clone}{\code{V2HorizontalPodAutoscalerList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerList-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerList-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerList$new(
  api_version = NULL,
  items = NULL,
  kind = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerList-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerList-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerList$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
