% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HorizontalPodAutoscalerSpec.R
\name{V2HorizontalPodAutoscalerSpec}
\alias{V2HorizontalPodAutoscalerSpec}
\title{V2HorizontalPodAutoscalerSpec}
\description{
V2HorizontalPodAutoscalerSpec

V2HorizontalPodAutoscalerSpec
}
\details{
HorizontalPodAutoscalerSpec describes the desired functionality of the HorizontalPodAutoscaler.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HorizontalPodAutoscalerSpec-new}{\code{V2HorizontalPodAutoscalerSpec$new()}}
\item \href{#method-V2HorizontalPodAutoscalerSpec-to_list}{\code{V2HorizontalPodAutoscalerSpec$to_list()}}
\item \href{#method-V2HorizontalPodAutoscalerSpec-clone}{\code{V2HorizontalPodAutoscalerSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerSpec$new(
  behavior = NULL,
  max_replicas = NULL,
  metrics = NULL,
  min_replicas = NULL,
  scale_target_ref = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
