% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HorizontalPodAutoscalerStatus.R
\name{V2HorizontalPodAutoscalerStatus}
\alias{V2HorizontalPodAutoscalerStatus}
\title{V2HorizontalPodAutoscalerStatus}
\description{
V2HorizontalPodAutoscalerStatus

V2HorizontalPodAutoscalerStatus
}
\details{
HorizontalPodAutoscalerStatus describes the current status of a horizontal pod autoscaler.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HorizontalPodAutoscalerStatus-new}{\code{V2HorizontalPodAutoscalerStatus$new()}}
\item \href{#method-V2HorizontalPodAutoscalerStatus-to_list}{\code{V2HorizontalPodAutoscalerStatus$to_list()}}
\item \href{#method-V2HorizontalPodAutoscalerStatus-clone}{\code{V2HorizontalPodAutoscalerStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerStatus$new(
  conditions = NULL,
  current_metrics = NULL,
  current_replicas = NULL,
  desired_replicas = NULL,
  last_scale_time = NULL,
  observed_generation = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HorizontalPodAutoscalerStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HorizontalPodAutoscalerStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HorizontalPodAutoscalerStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
