% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HPAScalingPolicy.R
\name{V2HPAScalingPolicy}
\alias{V2HPAScalingPolicy}
\title{V2HPAScalingPolicy}
\description{
V2HPAScalingPolicy

V2HPAScalingPolicy
}
\details{
HPAScalingPolicy is a single policy which must hold true for a specified past interval.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HPAScalingPolicy-new}{\code{V2HPAScalingPolicy$new()}}
\item \href{#method-V2HPAScalingPolicy-to_list}{\code{V2HPAScalingPolicy$to_list()}}
\item \href{#method-V2HPAScalingPolicy-clone}{\code{V2HPAScalingPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HPAScalingPolicy-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HPAScalingPolicy-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HPAScalingPolicy$new(period_seconds = NULL, type = NULL, value = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HPAScalingPolicy-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HPAScalingPolicy-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HPAScalingPolicy$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HPAScalingPolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HPAScalingPolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HPAScalingPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
