% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2HPAScalingRules.R
\name{V2HPAScalingRules}
\alias{V2HPAScalingRules}
\title{V2HPAScalingRules}
\description{
V2HPAScalingRules

V2HPAScalingRules
}
\details{
HPAScalingRules configures the scaling behavior for one direction. These Rules are applied after calculating DesiredReplicas from metrics for the HPA. They can limit the scaling velocity by specifying scaling policies. They can prevent flapping by specifying the stabilization window, so that the number of replicas is not set instantly, instead, the safest value from the stabilization window is chosen.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2HPAScalingRules-new}{\code{V2HPAScalingRules$new()}}
\item \href{#method-V2HPAScalingRules-to_list}{\code{V2HPAScalingRules$to_list()}}
\item \href{#method-V2HPAScalingRules-clone}{\code{V2HPAScalingRules$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HPAScalingRules-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2HPAScalingRules-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HPAScalingRules$new(
  policies = NULL,
  select_policy = NULL,
  stabilization_window_seconds = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HPAScalingRules-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2HPAScalingRules-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HPAScalingRules$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2HPAScalingRules-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2HPAScalingRules-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2HPAScalingRules$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
