% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2MetricIdentifier.R
\name{V2MetricIdentifier}
\alias{V2MetricIdentifier}
\title{V2MetricIdentifier}
\description{
V2MetricIdentifier

V2MetricIdentifier
}
\details{
MetricIdentifier defines the name and optionally selector for a metric
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2MetricIdentifier-new}{\code{V2MetricIdentifier$new()}}
\item \href{#method-V2MetricIdentifier-to_list}{\code{V2MetricIdentifier$to_list()}}
\item \href{#method-V2MetricIdentifier-clone}{\code{V2MetricIdentifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricIdentifier-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricIdentifier-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricIdentifier$new(name = NULL, selector = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricIdentifier-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricIdentifier-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricIdentifier$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricIdentifier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricIdentifier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricIdentifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
