% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2MetricSpec.R
\name{V2MetricSpec}
\alias{V2MetricSpec}
\title{V2MetricSpec}
\description{
V2MetricSpec

V2MetricSpec
}
\details{
MetricSpec specifies how to scale based on a single metric (only `type` and one other matching field should be set at once).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2MetricSpec-new}{\code{V2MetricSpec$new()}}
\item \href{#method-V2MetricSpec-to_list}{\code{V2MetricSpec$to_list()}}
\item \href{#method-V2MetricSpec-clone}{\code{V2MetricSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricSpec-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricSpec$new(
  container_resource = NULL,
  external = NULL,
  object = NULL,
  pods = NULL,
  resource = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
