% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2MetricStatus.R
\name{V2MetricStatus}
\alias{V2MetricStatus}
\title{V2MetricStatus}
\description{
V2MetricStatus

V2MetricStatus
}
\details{
MetricStatus describes the last-read state of a single metric.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2MetricStatus-new}{\code{V2MetricStatus$new()}}
\item \href{#method-V2MetricStatus-to_list}{\code{V2MetricStatus$to_list()}}
\item \href{#method-V2MetricStatus-clone}{\code{V2MetricStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricStatus$new(
  container_resource = NULL,
  external = NULL,
  object = NULL,
  pods = NULL,
  resource = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
