% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2MetricTarget.R
\name{V2MetricTarget}
\alias{V2MetricTarget}
\title{V2MetricTarget}
\description{
V2MetricTarget

V2MetricTarget
}
\details{
MetricTarget defines the target value, average value, or average utilization of a specific metric
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2MetricTarget-new}{\code{V2MetricTarget$new()}}
\item \href{#method-V2MetricTarget-to_list}{\code{V2MetricTarget$to_list()}}
\item \href{#method-V2MetricTarget-clone}{\code{V2MetricTarget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricTarget-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricTarget-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricTarget$new(
  average_utilization = NULL,
  average_value = NULL,
  type = NULL,
  value = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricTarget-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricTarget-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricTarget$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2MetricTarget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2MetricTarget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2MetricTarget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
