% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2ObjectMetricSource.R
\name{V2ObjectMetricSource}
\alias{V2ObjectMetricSource}
\title{V2ObjectMetricSource}
\description{
V2ObjectMetricSource

V2ObjectMetricSource
}
\details{
ObjectMetricSource indicates how to scale on a metric describing a kubernetes object (for example, hits-per-second on an Ingress object).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ObjectMetricSource-new}{\code{V2ObjectMetricSource$new()}}
\item \href{#method-V2ObjectMetricSource-to_list}{\code{V2ObjectMetricSource$to_list()}}
\item \href{#method-V2ObjectMetricSource-clone}{\code{V2ObjectMetricSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ObjectMetricSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2ObjectMetricSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ObjectMetricSource$new(described_object = NULL, metric = NULL, target = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ObjectMetricSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2ObjectMetricSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ObjectMetricSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ObjectMetricSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ObjectMetricSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ObjectMetricSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
