% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2ObjectMetricStatus.R
\name{V2ObjectMetricStatus}
\alias{V2ObjectMetricStatus}
\title{V2ObjectMetricStatus}
\description{
V2ObjectMetricStatus

V2ObjectMetricStatus
}
\details{
ObjectMetricStatus indicates the current value of a metric describing a kubernetes object (for example, hits-per-second on an Ingress object).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ObjectMetricStatus-new}{\code{V2ObjectMetricStatus$new()}}
\item \href{#method-V2ObjectMetricStatus-to_list}{\code{V2ObjectMetricStatus$to_list()}}
\item \href{#method-V2ObjectMetricStatus-clone}{\code{V2ObjectMetricStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ObjectMetricStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2ObjectMetricStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ObjectMetricStatus$new(
  current = NULL,
  described_object = NULL,
  metric = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ObjectMetricStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2ObjectMetricStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ObjectMetricStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ObjectMetricStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ObjectMetricStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ObjectMetricStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
