% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2PodsMetricSource.R
\name{V2PodsMetricSource}
\alias{V2PodsMetricSource}
\title{V2PodsMetricSource}
\description{
V2PodsMetricSource

V2PodsMetricSource
}
\details{
PodsMetricSource indicates how to scale on a metric describing each pod in the current scale target (for example, transactions-processed-per-second). The values will be averaged together before being compared to the target value.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2PodsMetricSource-new}{\code{V2PodsMetricSource$new()}}
\item \href{#method-V2PodsMetricSource-to_list}{\code{V2PodsMetricSource$to_list()}}
\item \href{#method-V2PodsMetricSource-clone}{\code{V2PodsMetricSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2PodsMetricSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2PodsMetricSource-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2PodsMetricSource$new(metric = NULL, target = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2PodsMetricSource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2PodsMetricSource-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2PodsMetricSource$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2PodsMetricSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2PodsMetricSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2PodsMetricSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
