% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2PodsMetricStatus.R
\name{V2PodsMetricStatus}
\alias{V2PodsMetricStatus}
\title{V2PodsMetricStatus}
\description{
V2PodsMetricStatus

V2PodsMetricStatus
}
\details{
PodsMetricStatus indicates the current value of a metric describing each pod in the current scale target (for example, transactions-processed-per-second).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2PodsMetricStatus-new}{\code{V2PodsMetricStatus$new()}}
\item \href{#method-V2PodsMetricStatus-to_list}{\code{V2PodsMetricStatus$to_list()}}
\item \href{#method-V2PodsMetricStatus-clone}{\code{V2PodsMetricStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2PodsMetricStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2PodsMetricStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2PodsMetricStatus$new(current = NULL, metric = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2PodsMetricStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2PodsMetricStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2PodsMetricStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2PodsMetricStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2PodsMetricStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2PodsMetricStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
