% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model_V2ResourceMetricStatus.R
\name{V2ResourceMetricStatus}
\alias{V2ResourceMetricStatus}
\title{V2ResourceMetricStatus}
\description{
V2ResourceMetricStatus

V2ResourceMetricStatus
}
\details{
ResourceMetricStatus indicates the current value of a resource metric known to Kubernetes, as specified in requests and limits, describing each pod in the current scale target (e.g. CPU or memory).  Such metrics are built in to Kubernetes, and have special scaling options on top of those available to normal per-pod metrics using the "pods" source.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ResourceMetricStatus-new}{\code{V2ResourceMetricStatus$new()}}
\item \href{#method-V2ResourceMetricStatus-to_list}{\code{V2ResourceMetricStatus$to_list()}}
\item \href{#method-V2ResourceMetricStatus-clone}{\code{V2ResourceMetricStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ResourceMetricStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-V2ResourceMetricStatus-new}{}}}
\subsection{Method \code{new()}}{
Construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ResourceMetricStatus$new(current = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ResourceMetricStatus-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-V2ResourceMetricStatus-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert to a plain named list (JSON keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ResourceMetricStatus$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ResourceMetricStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ResourceMetricStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ResourceMetricStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
