% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_VersionApi.R
\name{VersionApi}
\alias{VersionApi}
\title{VersionApi}
\description{
VersionApi

VersionApi
}
\details{
Typed Kubernetes API client for the `Version` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VersionApi-new}{\code{VersionApi$new()}}
\item \href{#method-VersionApi-get_code_version}{\code{VersionApi$get_code_version()}}
\item \href{#method-VersionApi-clone}{\code{VersionApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-VersionApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCodeVersion
get the code version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionApi-get_code_version"></a>}}
\if{latex}{\out{\hypertarget{method-VersionApi-get_code_version}{}}}
\subsection{Method \code{get_code_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionApi$get_code_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VersionApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
