% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_WellKnownApi.R
\name{WellKnownApi}
\alias{WellKnownApi}
\title{WellKnownApi}
\description{
WellKnownApi

WellKnownApi
}
\details{
Typed Kubernetes API client for the `WellKnown` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WellKnownApi-new}{\code{WellKnownApi$new()}}
\item \href{#method-WellKnownApi-get_service_account_issuer_open_id_configuration}{\code{WellKnownApi$get_service_account_issuer_open_id_configuration()}}
\item \href{#method-WellKnownApi-clone}{\code{WellKnownApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WellKnownApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-WellKnownApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WellKnownApi$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getServiceAccountIssuerOpenIDConfiguration
get service account issuer OpenID configuration, also known as the 'OIDC discovery doc'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WellKnownApi-get_service_account_issuer_open_id_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-WellKnownApi-get_service_account_issuer_open_id_configuration}{}}}
\subsection{Method \code{get_service_account_issuer_open_id_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WellKnownApi$get_service_account_issuer_open_id_configuration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WellKnownApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WellKnownApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WellKnownApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
