% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_AppsV1Api.R
\name{AppsV1Api}
\alias{AppsV1Api}
\title{AppsV1Api}
\description{
AppsV1Api

AppsV1Api
}
\details{
Typed Kubernetes API client for the `AppsV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AppsV1Api-new}{\code{AppsV1Api$new()}}
\item \href{#method-AppsV1Api-get_api_resources}{\code{AppsV1Api$get_api_resources()}}
\item \href{#method-AppsV1Api-list_controller_revision_for_all_namespaces}{\code{AppsV1Api$list_controller_revision_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_daemon_set_for_all_namespaces}{\code{AppsV1Api$list_daemon_set_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_deployment_for_all_namespaces}{\code{AppsV1Api$list_deployment_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_namespaced_controller_revision}{\code{AppsV1Api$list_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-create_namespaced_controller_revision}{\code{AppsV1Api$create_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_controller_revision}{\code{AppsV1Api$delete_collection_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-read_namespaced_controller_revision}{\code{AppsV1Api$read_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-replace_namespaced_controller_revision}{\code{AppsV1Api$replace_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-patch_namespaced_controller_revision}{\code{AppsV1Api$patch_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-delete_namespaced_controller_revision}{\code{AppsV1Api$delete_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-list_namespaced_daemon_set}{\code{AppsV1Api$list_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-create_namespaced_daemon_set}{\code{AppsV1Api$create_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_daemon_set}{\code{AppsV1Api$delete_collection_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-read_namespaced_daemon_set}{\code{AppsV1Api$read_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-replace_namespaced_daemon_set}{\code{AppsV1Api$replace_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-patch_namespaced_daemon_set}{\code{AppsV1Api$patch_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-delete_namespaced_daemon_set}{\code{AppsV1Api$delete_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-read_namespaced_daemon_set_status}{\code{AppsV1Api$read_namespaced_daemon_set_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_daemon_set_status}{\code{AppsV1Api$replace_namespaced_daemon_set_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_daemon_set_status}{\code{AppsV1Api$patch_namespaced_daemon_set_status()}}
\item \href{#method-AppsV1Api-list_namespaced_deployment}{\code{AppsV1Api$list_namespaced_deployment()}}
\item \href{#method-AppsV1Api-create_namespaced_deployment}{\code{AppsV1Api$create_namespaced_deployment()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_deployment}{\code{AppsV1Api$delete_collection_namespaced_deployment()}}
\item \href{#method-AppsV1Api-read_namespaced_deployment}{\code{AppsV1Api$read_namespaced_deployment()}}
\item \href{#method-AppsV1Api-replace_namespaced_deployment}{\code{AppsV1Api$replace_namespaced_deployment()}}
\item \href{#method-AppsV1Api-patch_namespaced_deployment}{\code{AppsV1Api$patch_namespaced_deployment()}}
\item \href{#method-AppsV1Api-delete_namespaced_deployment}{\code{AppsV1Api$delete_namespaced_deployment()}}
\item \href{#method-AppsV1Api-read_namespaced_deployment_scale}{\code{AppsV1Api$read_namespaced_deployment_scale()}}
\item \href{#method-AppsV1Api-replace_namespaced_deployment_scale}{\code{AppsV1Api$replace_namespaced_deployment_scale()}}
\item \href{#method-AppsV1Api-patch_namespaced_deployment_scale}{\code{AppsV1Api$patch_namespaced_deployment_scale()}}
\item \href{#method-AppsV1Api-read_namespaced_deployment_status}{\code{AppsV1Api$read_namespaced_deployment_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_deployment_status}{\code{AppsV1Api$replace_namespaced_deployment_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_deployment_status}{\code{AppsV1Api$patch_namespaced_deployment_status()}}
\item \href{#method-AppsV1Api-list_namespaced_replica_set}{\code{AppsV1Api$list_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-create_namespaced_replica_set}{\code{AppsV1Api$create_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_replica_set}{\code{AppsV1Api$delete_collection_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-read_namespaced_replica_set}{\code{AppsV1Api$read_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-replace_namespaced_replica_set}{\code{AppsV1Api$replace_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-patch_namespaced_replica_set}{\code{AppsV1Api$patch_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-delete_namespaced_replica_set}{\code{AppsV1Api$delete_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-read_namespaced_replica_set_scale}{\code{AppsV1Api$read_namespaced_replica_set_scale()}}
\item \href{#method-AppsV1Api-replace_namespaced_replica_set_scale}{\code{AppsV1Api$replace_namespaced_replica_set_scale()}}
\item \href{#method-AppsV1Api-patch_namespaced_replica_set_scale}{\code{AppsV1Api$patch_namespaced_replica_set_scale()}}
\item \href{#method-AppsV1Api-read_namespaced_replica_set_status}{\code{AppsV1Api$read_namespaced_replica_set_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_replica_set_status}{\code{AppsV1Api$replace_namespaced_replica_set_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_replica_set_status}{\code{AppsV1Api$patch_namespaced_replica_set_status()}}
\item \href{#method-AppsV1Api-list_namespaced_stateful_set}{\code{AppsV1Api$list_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-create_namespaced_stateful_set}{\code{AppsV1Api$create_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-delete_collection_namespaced_stateful_set}{\code{AppsV1Api$delete_collection_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-read_namespaced_stateful_set}{\code{AppsV1Api$read_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-replace_namespaced_stateful_set}{\code{AppsV1Api$replace_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-patch_namespaced_stateful_set}{\code{AppsV1Api$patch_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-delete_namespaced_stateful_set}{\code{AppsV1Api$delete_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-read_namespaced_stateful_set_scale}{\code{AppsV1Api$read_namespaced_stateful_set_scale()}}
\item \href{#method-AppsV1Api-replace_namespaced_stateful_set_scale}{\code{AppsV1Api$replace_namespaced_stateful_set_scale()}}
\item \href{#method-AppsV1Api-patch_namespaced_stateful_set_scale}{\code{AppsV1Api$patch_namespaced_stateful_set_scale()}}
\item \href{#method-AppsV1Api-read_namespaced_stateful_set_status}{\code{AppsV1Api$read_namespaced_stateful_set_status()}}
\item \href{#method-AppsV1Api-replace_namespaced_stateful_set_status}{\code{AppsV1Api$replace_namespaced_stateful_set_status()}}
\item \href{#method-AppsV1Api-patch_namespaced_stateful_set_status}{\code{AppsV1Api$patch_namespaced_stateful_set_status()}}
\item \href{#method-AppsV1Api-list_replica_set_for_all_namespaces}{\code{AppsV1Api$list_replica_set_for_all_namespaces()}}
\item \href{#method-AppsV1Api-list_stateful_set_for_all_namespaces}{\code{AppsV1Api$list_stateful_set_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_controller_revision_list_for_all_namespaces}{\code{AppsV1Api$watch_controller_revision_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_daemon_set_list_for_all_namespaces}{\code{AppsV1Api$watch_daemon_set_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_deployment_list_for_all_namespaces}{\code{AppsV1Api$watch_deployment_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_namespaced_controller_revision_list}{\code{AppsV1Api$watch_namespaced_controller_revision_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_controller_revision}{\code{AppsV1Api$watch_namespaced_controller_revision()}}
\item \href{#method-AppsV1Api-watch_namespaced_daemon_set_list}{\code{AppsV1Api$watch_namespaced_daemon_set_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_daemon_set}{\code{AppsV1Api$watch_namespaced_daemon_set()}}
\item \href{#method-AppsV1Api-watch_namespaced_deployment_list}{\code{AppsV1Api$watch_namespaced_deployment_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_deployment}{\code{AppsV1Api$watch_namespaced_deployment()}}
\item \href{#method-AppsV1Api-watch_namespaced_replica_set_list}{\code{AppsV1Api$watch_namespaced_replica_set_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_replica_set}{\code{AppsV1Api$watch_namespaced_replica_set()}}
\item \href{#method-AppsV1Api-watch_namespaced_stateful_set_list}{\code{AppsV1Api$watch_namespaced_stateful_set_list()}}
\item \href{#method-AppsV1Api-watch_namespaced_stateful_set}{\code{AppsV1Api$watch_namespaced_stateful_set()}}
\item \href{#method-AppsV1Api-watch_replica_set_list_for_all_namespaces}{\code{AppsV1Api$watch_replica_set_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-watch_stateful_set_list_for_all_namespaces}{\code{AppsV1Api$watch_stateful_set_list_for_all_namespaces()}}
\item \href{#method-AppsV1Api-clone}{\code{AppsV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getAppsV1APIResources
get available resources
listAppsV1ControllerRevisionForAllNamespaces
list or watch objects of kind ControllerRevision
listAppsV1DaemonSetForAllNamespaces
list or watch objects of kind DaemonSet
listAppsV1DeploymentForAllNamespaces
list or watch objects of kind Deployment
listAppsV1NamespacedControllerRevision
list or watch objects of kind ControllerRevision
createAppsV1NamespacedControllerRevision
create a ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_controller_revision_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_controller_revision_for_all_namespaces}{}}}
\subsection{Method \code{list_controller_revision_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_controller_revision_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_daemon_set_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_daemon_set_for_all_namespaces}{}}}
\subsection{Method \code{list_daemon_set_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_daemon_set_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_deployment_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_deployment_for_all_namespaces}{}}}
\subsection{Method \code{list_deployment_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_deployment_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_controller_revision}{}}}
\subsection{Method \code{list_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_controller_revision()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_controller_revision}{}}}
\subsection{Method \code{create_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_controller_revision(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedControllerRevision
delete collection of ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_controller_revision}{}}}
\subsection{Method \code{delete_collection_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_controller_revision(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedControllerRevision
read the specified ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_controller_revision}{}}}
\subsection{Method \code{read_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_controller_revision(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedControllerRevision
replace the specified ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_controller_revision}{}}}
\subsection{Method \code{replace_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_controller_revision(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedControllerRevision
partially update the specified ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_controller_revision}{}}}
\subsection{Method \code{patch_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_controller_revision(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1NamespacedControllerRevision
delete a ControllerRevision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_controller_revision}{}}}
\subsection{Method \code{delete_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_controller_revision(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
listAppsV1NamespacedDaemonSet
list or watch objects of kind DaemonSet
createAppsV1NamespacedDaemonSet
create a DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_daemon_set}{}}}
\subsection{Method \code{list_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_daemon_set()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_daemon_set}{}}}
\subsection{Method \code{create_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_daemon_set(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedDaemonSet
delete collection of DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_daemon_set}{}}}
\subsection{Method \code{delete_collection_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_daemon_set(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedDaemonSet
read the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_daemon_set}{}}}
\subsection{Method \code{read_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_daemon_set(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedDaemonSet
replace the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_daemon_set}{}}}
\subsection{Method \code{replace_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_daemon_set(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDaemonSet
partially update the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_daemon_set}{}}}
\subsection{Method \code{patch_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_daemon_set(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1NamespacedDaemonSet
delete a DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_daemon_set}{}}}
\subsection{Method \code{delete_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_daemon_set(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedDaemonSetStatus
read status of the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_daemon_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_daemon_set_status}{}}}
\subsection{Method \code{read_namespaced_daemon_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_daemon_set_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedDaemonSetStatus
replace status of the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_daemon_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_daemon_set_status}{}}}
\subsection{Method \code{replace_namespaced_daemon_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_daemon_set_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDaemonSetStatus
partially update status of the specified DaemonSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_daemon_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_daemon_set_status}{}}}
\subsection{Method \code{patch_namespaced_daemon_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_daemon_set_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
listAppsV1NamespacedDeployment
list or watch objects of kind Deployment
createAppsV1NamespacedDeployment
create a Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_deployment}{}}}
\subsection{Method \code{list_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_deployment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_deployment}{}}}
\subsection{Method \code{create_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_deployment(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedDeployment
delete collection of Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_deployment}{}}}
\subsection{Method \code{delete_collection_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_deployment(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedDeployment
read the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_deployment}{}}}
\subsection{Method \code{read_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_deployment(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedDeployment
replace the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_deployment}{}}}
\subsection{Method \code{replace_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_deployment(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDeployment
partially update the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_deployment}{}}}
\subsection{Method \code{patch_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_deployment(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1NamespacedDeployment
delete a Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_deployment}{}}}
\subsection{Method \code{delete_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_deployment(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedDeploymentScale
read scale of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_deployment_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_deployment_scale}{}}}
\subsection{Method \code{read_namespaced_deployment_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_deployment_scale(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedDeploymentScale
replace scale of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_deployment_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_deployment_scale}{}}}
\subsection{Method \code{replace_namespaced_deployment_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_deployment_scale(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDeploymentScale
partially update scale of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_deployment_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_deployment_scale}{}}}
\subsection{Method \code{patch_namespaced_deployment_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_deployment_scale(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
readAppsV1NamespacedDeploymentStatus
read status of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_deployment_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_deployment_status}{}}}
\subsection{Method \code{read_namespaced_deployment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_deployment_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedDeploymentStatus
replace status of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_deployment_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_deployment_status}{}}}
\subsection{Method \code{replace_namespaced_deployment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_deployment_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedDeploymentStatus
partially update status of the specified Deployment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_deployment_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_deployment_status}{}}}
\subsection{Method \code{patch_namespaced_deployment_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_deployment_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
listAppsV1NamespacedReplicaSet
list or watch objects of kind ReplicaSet
createAppsV1NamespacedReplicaSet
create a ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_replica_set}{}}}
\subsection{Method \code{list_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_replica_set()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_replica_set}{}}}
\subsection{Method \code{create_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_replica_set(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedReplicaSet
delete collection of ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_replica_set}{}}}
\subsection{Method \code{delete_collection_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_replica_set(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedReplicaSet
read the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_replica_set}{}}}
\subsection{Method \code{read_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_replica_set(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedReplicaSet
replace the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_replica_set}{}}}
\subsection{Method \code{replace_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_replica_set(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedReplicaSet
partially update the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_replica_set}{}}}
\subsection{Method \code{patch_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_replica_set(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1NamespacedReplicaSet
delete a ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_replica_set}{}}}
\subsection{Method \code{delete_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_replica_set(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedReplicaSetScale
read scale of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_replica_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_replica_set_scale}{}}}
\subsection{Method \code{read_namespaced_replica_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_replica_set_scale(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedReplicaSetScale
replace scale of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_replica_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_replica_set_scale}{}}}
\subsection{Method \code{replace_namespaced_replica_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_replica_set_scale(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedReplicaSetScale
partially update scale of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_replica_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_replica_set_scale}{}}}
\subsection{Method \code{patch_namespaced_replica_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_replica_set_scale(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
readAppsV1NamespacedReplicaSetStatus
read status of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_replica_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_replica_set_status}{}}}
\subsection{Method \code{read_namespaced_replica_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_replica_set_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedReplicaSetStatus
replace status of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_replica_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_replica_set_status}{}}}
\subsection{Method \code{replace_namespaced_replica_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_replica_set_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedReplicaSetStatus
partially update status of the specified ReplicaSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_replica_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_replica_set_status}{}}}
\subsection{Method \code{patch_namespaced_replica_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_replica_set_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
listAppsV1NamespacedStatefulSet
list or watch objects of kind StatefulSet
createAppsV1NamespacedStatefulSet
create a StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_namespaced_stateful_set}{}}}
\subsection{Method \code{list_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_namespaced_stateful_set()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-create_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-create_namespaced_stateful_set}{}}}
\subsection{Method \code{create_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$create_namespaced_stateful_set(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1CollectionNamespacedStatefulSet
delete collection of StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_collection_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_collection_namespaced_stateful_set}{}}}
\subsection{Method \code{delete_collection_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_collection_namespaced_stateful_set(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedStatefulSet
read the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_stateful_set}{}}}
\subsection{Method \code{read_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_stateful_set(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedStatefulSet
replace the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_stateful_set}{}}}
\subsection{Method \code{replace_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_stateful_set(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedStatefulSet
partially update the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_stateful_set}{}}}
\subsection{Method \code{patch_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_stateful_set(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteAppsV1NamespacedStatefulSet
delete a StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-delete_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-delete_namespaced_stateful_set}{}}}
\subsection{Method \code{delete_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$delete_namespaced_stateful_set(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readAppsV1NamespacedStatefulSetScale
read scale of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_stateful_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_stateful_set_scale}{}}}
\subsection{Method \code{read_namespaced_stateful_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_stateful_set_scale(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedStatefulSetScale
replace scale of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_stateful_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_stateful_set_scale}{}}}
\subsection{Method \code{replace_namespaced_stateful_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_stateful_set_scale(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedStatefulSetScale
partially update scale of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_stateful_set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_stateful_set_scale}{}}}
\subsection{Method \code{patch_namespaced_stateful_set_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_stateful_set_scale(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
readAppsV1NamespacedStatefulSetStatus
read status of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-read_namespaced_stateful_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-read_namespaced_stateful_set_status}{}}}
\subsection{Method \code{read_namespaced_stateful_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$read_namespaced_stateful_set_status(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceAppsV1NamespacedStatefulSetStatus
replace status of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-replace_namespaced_stateful_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-replace_namespaced_stateful_set_status}{}}}
\subsection{Method \code{replace_namespaced_stateful_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$replace_namespaced_stateful_set_status(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchAppsV1NamespacedStatefulSetStatus
partially update status of the specified StatefulSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-patch_namespaced_stateful_set_status"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-patch_namespaced_stateful_set_status}{}}}
\subsection{Method \code{patch_namespaced_stateful_set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$patch_namespaced_stateful_set_status(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
listAppsV1ReplicaSetForAllNamespaces
list or watch objects of kind ReplicaSet
listAppsV1StatefulSetForAllNamespaces
list or watch objects of kind StatefulSet
watchAppsV1ControllerRevisionListForAllNamespaces
watch individual changes to a list of ControllerRevision. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1DaemonSetListForAllNamespaces
watch individual changes to a list of DaemonSet. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1DeploymentListForAllNamespaces
watch individual changes to a list of Deployment. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1NamespacedControllerRevisionList
watch individual changes to a list of ControllerRevision. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1NamespacedControllerRevision
watch changes to an object of kind ControllerRevision. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_replica_set_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_replica_set_for_all_namespaces}{}}}
\subsection{Method \code{list_replica_set_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_replica_set_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-list_stateful_set_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-list_stateful_set_for_all_namespaces}{}}}
\subsection{Method \code{list_stateful_set_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$list_stateful_set_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_controller_revision_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_controller_revision_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_controller_revision_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_controller_revision_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_daemon_set_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_daemon_set_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_daemon_set_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_daemon_set_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_deployment_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_deployment_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_deployment_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_deployment_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_controller_revision_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_controller_revision_list}{}}}
\subsection{Method \code{watch_namespaced_controller_revision_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_controller_revision_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_controller_revision"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_controller_revision}{}}}
\subsection{Method \code{watch_namespaced_controller_revision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_controller_revision(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchAppsV1NamespacedDaemonSetList
watch individual changes to a list of DaemonSet. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1NamespacedDaemonSet
watch changes to an object of kind DaemonSet. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_daemon_set_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_daemon_set_list}{}}}
\subsection{Method \code{watch_namespaced_daemon_set_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_daemon_set_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_daemon_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_daemon_set}{}}}
\subsection{Method \code{watch_namespaced_daemon_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_daemon_set(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchAppsV1NamespacedDeploymentList
watch individual changes to a list of Deployment. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1NamespacedDeployment
watch changes to an object of kind Deployment. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_deployment_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_deployment_list}{}}}
\subsection{Method \code{watch_namespaced_deployment_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_deployment_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_deployment"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_deployment}{}}}
\subsection{Method \code{watch_namespaced_deployment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_deployment(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchAppsV1NamespacedReplicaSetList
watch individual changes to a list of ReplicaSet. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1NamespacedReplicaSet
watch changes to an object of kind ReplicaSet. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_replica_set_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_replica_set_list}{}}}
\subsection{Method \code{watch_namespaced_replica_set_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_replica_set_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_replica_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_replica_set}{}}}
\subsection{Method \code{watch_namespaced_replica_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_replica_set(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchAppsV1NamespacedStatefulSetList
watch individual changes to a list of StatefulSet. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1NamespacedStatefulSet
watch changes to an object of kind StatefulSet. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_stateful_set_list"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_stateful_set_list}{}}}
\subsection{Method \code{watch_namespaced_stateful_set_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_stateful_set_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_namespaced_stateful_set"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_namespaced_stateful_set}{}}}
\subsection{Method \code{watch_namespaced_stateful_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_namespaced_stateful_set(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{watchAppsV1ReplicaSetListForAllNamespaces
watch individual changes to a list of ReplicaSet. deprecated: use the 'watch' parameter with a list operation instead.
watchAppsV1StatefulSetListForAllNamespaces
watch individual changes to a list of StatefulSet. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_replica_set_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_replica_set_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_replica_set_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_replica_set_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-watch_stateful_set_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-watch_stateful_set_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_stateful_set_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$watch_stateful_set_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppsV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AppsV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppsV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
