% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CertificatesV1alpha1Api.R
\name{CertificatesV1alpha1Api}
\alias{CertificatesV1alpha1Api}
\title{CertificatesV1alpha1Api}
\description{
CertificatesV1alpha1Api

CertificatesV1alpha1Api
}
\details{
Typed Kubernetes API client for the `CertificatesV1alpha1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CertificatesV1alpha1Api-new}{\code{CertificatesV1alpha1Api$new()}}
\item \href{#method-CertificatesV1alpha1Api-get_api_resources}{\code{CertificatesV1alpha1Api$get_api_resources()}}
\item \href{#method-CertificatesV1alpha1Api-list_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$list_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-create_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$create_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-delete_collection_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$delete_collection_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-read_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$read_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-replace_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$replace_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-patch_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$patch_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-delete_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$delete_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-watch_cluster_trust_bundle_list}{\code{CertificatesV1alpha1Api$watch_cluster_trust_bundle_list()}}
\item \href{#method-CertificatesV1alpha1Api-watch_cluster_trust_bundle}{\code{CertificatesV1alpha1Api$watch_cluster_trust_bundle()}}
\item \href{#method-CertificatesV1alpha1Api-clone}{\code{CertificatesV1alpha1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCertificatesV1alpha1APIResources
get available resources
listCertificatesV1alpha1ClusterTrustBundle
list or watch objects of kind ClusterTrustBundle
createCertificatesV1alpha1ClusterTrustBundle
create a ClusterTrustBundle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-list_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-list_cluster_trust_bundle}{}}}
\subsection{Method \code{list_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$list_cluster_trust_bundle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-create_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-create_cluster_trust_bundle}{}}}
\subsection{Method \code{create_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$create_cluster_trust_bundle(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCertificatesV1alpha1CollectionClusterTrustBundle
delete collection of ClusterTrustBundle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-delete_collection_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-delete_collection_cluster_trust_bundle}{}}}
\subsection{Method \code{delete_collection_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$delete_collection_cluster_trust_bundle(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readCertificatesV1alpha1ClusterTrustBundle
read the specified ClusterTrustBundle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-read_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-read_cluster_trust_bundle}{}}}
\subsection{Method \code{read_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$read_cluster_trust_bundle(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceCertificatesV1alpha1ClusterTrustBundle
replace the specified ClusterTrustBundle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-replace_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-replace_cluster_trust_bundle}{}}}
\subsection{Method \code{replace_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$replace_cluster_trust_bundle(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchCertificatesV1alpha1ClusterTrustBundle
partially update the specified ClusterTrustBundle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-patch_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-patch_cluster_trust_bundle}{}}}
\subsection{Method \code{patch_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$patch_cluster_trust_bundle(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCertificatesV1alpha1ClusterTrustBundle
delete a ClusterTrustBundle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-delete_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-delete_cluster_trust_bundle}{}}}
\subsection{Method \code{delete_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$delete_cluster_trust_bundle(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchCertificatesV1alpha1ClusterTrustBundleList
watch individual changes to a list of ClusterTrustBundle. deprecated: use the 'watch' parameter with a list operation instead.
watchCertificatesV1alpha1ClusterTrustBundle
watch changes to an object of kind ClusterTrustBundle. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-watch_cluster_trust_bundle_list"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-watch_cluster_trust_bundle_list}{}}}
\subsection{Method \code{watch_cluster_trust_bundle_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$watch_cluster_trust_bundle_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-watch_cluster_trust_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-watch_cluster_trust_bundle}{}}}
\subsection{Method \code{watch_cluster_trust_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$watch_cluster_trust_bundle(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1alpha1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1alpha1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1alpha1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
