% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CertificatesV1Api.R
\name{CertificatesV1Api}
\alias{CertificatesV1Api}
\title{CertificatesV1Api}
\description{
CertificatesV1Api

CertificatesV1Api
}
\details{
Typed Kubernetes API client for the `CertificatesV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CertificatesV1Api-new}{\code{CertificatesV1Api$new()}}
\item \href{#method-CertificatesV1Api-get_api_resources}{\code{CertificatesV1Api$get_api_resources()}}
\item \href{#method-CertificatesV1Api-list_certificate_signing_request}{\code{CertificatesV1Api$list_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-create_certificate_signing_request}{\code{CertificatesV1Api$create_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-delete_collection_certificate_signing_request}{\code{CertificatesV1Api$delete_collection_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-read_certificate_signing_request}{\code{CertificatesV1Api$read_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-replace_certificate_signing_request}{\code{CertificatesV1Api$replace_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-patch_certificate_signing_request}{\code{CertificatesV1Api$patch_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-delete_certificate_signing_request}{\code{CertificatesV1Api$delete_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-read_certificate_signing_request_approval}{\code{CertificatesV1Api$read_certificate_signing_request_approval()}}
\item \href{#method-CertificatesV1Api-replace_certificate_signing_request_approval}{\code{CertificatesV1Api$replace_certificate_signing_request_approval()}}
\item \href{#method-CertificatesV1Api-patch_certificate_signing_request_approval}{\code{CertificatesV1Api$patch_certificate_signing_request_approval()}}
\item \href{#method-CertificatesV1Api-read_certificate_signing_request_status}{\code{CertificatesV1Api$read_certificate_signing_request_status()}}
\item \href{#method-CertificatesV1Api-replace_certificate_signing_request_status}{\code{CertificatesV1Api$replace_certificate_signing_request_status()}}
\item \href{#method-CertificatesV1Api-patch_certificate_signing_request_status}{\code{CertificatesV1Api$patch_certificate_signing_request_status()}}
\item \href{#method-CertificatesV1Api-watch_certificate_signing_request_list}{\code{CertificatesV1Api$watch_certificate_signing_request_list()}}
\item \href{#method-CertificatesV1Api-watch_certificate_signing_request}{\code{CertificatesV1Api$watch_certificate_signing_request()}}
\item \href{#method-CertificatesV1Api-clone}{\code{CertificatesV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCertificatesV1APIResources
get available resources
listCertificatesV1CertificateSigningRequest
list or watch objects of kind CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-list_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-list_certificate_signing_request}{}}}
\subsection{Method \code{list_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$list_certificate_signing_request(
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createCertificatesV1CertificateSigningRequest
create a CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-create_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-create_certificate_signing_request}{}}}
\subsection{Method \code{create_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$create_certificate_signing_request(
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCertificatesV1CollectionCertificateSigningRequest
delete collection of CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-delete_collection_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-delete_collection_certificate_signing_request}{}}}
\subsection{Method \code{delete_collection_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$delete_collection_certificate_signing_request(
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readCertificatesV1CertificateSigningRequest
read the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-read_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-read_certificate_signing_request}{}}}
\subsection{Method \code{read_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$read_certificate_signing_request(name, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceCertificatesV1CertificateSigningRequest
replace the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-replace_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-replace_certificate_signing_request}{}}}
\subsection{Method \code{replace_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$replace_certificate_signing_request(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchCertificatesV1CertificateSigningRequest
partially update the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-patch_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-patch_certificate_signing_request}{}}}
\subsection{Method \code{patch_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$patch_certificate_signing_request(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteCertificatesV1CertificateSigningRequest
delete a CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-delete_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-delete_certificate_signing_request}{}}}
\subsection{Method \code{delete_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$delete_certificate_signing_request(
  name,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readCertificatesV1CertificateSigningRequestApproval
read approval of the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-read_certificate_signing_request_approval"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-read_certificate_signing_request_approval}{}}}
\subsection{Method \code{read_certificate_signing_request_approval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$read_certificate_signing_request_approval(
  name,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceCertificatesV1CertificateSigningRequestApproval
replace approval of the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-replace_certificate_signing_request_approval"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-replace_certificate_signing_request_approval}{}}}
\subsection{Method \code{replace_certificate_signing_request_approval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$replace_certificate_signing_request_approval(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchCertificatesV1CertificateSigningRequestApproval
partially update approval of the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-patch_certificate_signing_request_approval"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-patch_certificate_signing_request_approval}{}}}
\subsection{Method \code{patch_certificate_signing_request_approval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$patch_certificate_signing_request_approval(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
readCertificatesV1CertificateSigningRequestStatus
read status of the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-read_certificate_signing_request_status"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-read_certificate_signing_request_status}{}}}
\subsection{Method \code{read_certificate_signing_request_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$read_certificate_signing_request_status(name, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceCertificatesV1CertificateSigningRequestStatus
replace status of the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-replace_certificate_signing_request_status"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-replace_certificate_signing_request_status}{}}}
\subsection{Method \code{replace_certificate_signing_request_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$replace_certificate_signing_request_status(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchCertificatesV1CertificateSigningRequestStatus
partially update status of the specified CertificateSigningRequest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-patch_certificate_signing_request_status"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-patch_certificate_signing_request_status}{}}}
\subsection{Method \code{patch_certificate_signing_request_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$patch_certificate_signing_request_status(
  name,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
watchCertificatesV1CertificateSigningRequestList
watch individual changes to a list of CertificateSigningRequest. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-watch_certificate_signing_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-watch_certificate_signing_request_list}{}}}
\subsection{Method \code{watch_certificate_signing_request_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$watch_certificate_signing_request_list(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchCertificatesV1CertificateSigningRequest
watch changes to an object of kind CertificateSigningRequest. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-watch_certificate_signing_request"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-watch_certificate_signing_request}{}}}
\subsection{Method \code{watch_certificate_signing_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$watch_certificate_signing_request(
  name,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CertificatesV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CertificatesV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CertificatesV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
