% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CoordinationV1alpha1Api.R
\name{CoordinationV1alpha1Api}
\alias{CoordinationV1alpha1Api}
\title{CoordinationV1alpha1Api}
\description{
CoordinationV1alpha1Api

CoordinationV1alpha1Api
}
\details{
Typed Kubernetes API client for the `CoordinationV1alpha1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordinationV1alpha1Api-new}{\code{CoordinationV1alpha1Api$new()}}
\item \href{#method-CoordinationV1alpha1Api-get_api_resources}{\code{CoordinationV1alpha1Api$get_api_resources()}}
\item \href{#method-CoordinationV1alpha1Api-list_lease_candidate_for_all_namespaces}{\code{CoordinationV1alpha1Api$list_lease_candidate_for_all_namespaces()}}
\item \href{#method-CoordinationV1alpha1Api-list_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$list_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-create_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$create_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-delete_collection_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$delete_collection_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-read_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$read_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-replace_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$replace_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-patch_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$patch_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-delete_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$delete_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-watch_lease_candidate_list_for_all_namespaces}{\code{CoordinationV1alpha1Api$watch_lease_candidate_list_for_all_namespaces()}}
\item \href{#method-CoordinationV1alpha1Api-watch_namespaced_lease_candidate_list}{\code{CoordinationV1alpha1Api$watch_namespaced_lease_candidate_list()}}
\item \href{#method-CoordinationV1alpha1Api-watch_namespaced_lease_candidate}{\code{CoordinationV1alpha1Api$watch_namespaced_lease_candidate()}}
\item \href{#method-CoordinationV1alpha1Api-clone}{\code{CoordinationV1alpha1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCoordinationV1alpha1APIResources
get available resources
listCoordinationV1alpha1LeaseCandidateForAllNamespaces
list or watch objects of kind LeaseCandidate
listCoordinationV1alpha1NamespacedLeaseCandidate
list or watch objects of kind LeaseCandidate
createCoordinationV1alpha1NamespacedLeaseCandidate
create a LeaseCandidate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-list_lease_candidate_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-list_lease_candidate_for_all_namespaces}{}}}
\subsection{Method \code{list_lease_candidate_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$list_lease_candidate_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-list_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-list_namespaced_lease_candidate}{}}}
\subsection{Method \code{list_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$list_namespaced_lease_candidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-create_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-create_namespaced_lease_candidate}{}}}
\subsection{Method \code{create_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$create_namespaced_lease_candidate(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCoordinationV1alpha1CollectionNamespacedLeaseCandidate
delete collection of LeaseCandidate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-delete_collection_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-delete_collection_namespaced_lease_candidate}{}}}
\subsection{Method \code{delete_collection_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$delete_collection_namespaced_lease_candidate(
  dry_run = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readCoordinationV1alpha1NamespacedLeaseCandidate
read the specified LeaseCandidate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-read_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-read_namespaced_lease_candidate}{}}}
\subsection{Method \code{read_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$read_namespaced_lease_candidate(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceCoordinationV1alpha1NamespacedLeaseCandidate
replace the specified LeaseCandidate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-replace_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-replace_namespaced_lease_candidate}{}}}
\subsection{Method \code{replace_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$replace_namespaced_lease_candidate(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchCoordinationV1alpha1NamespacedLeaseCandidate
partially update the specified LeaseCandidate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-patch_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-patch_namespaced_lease_candidate}{}}}
\subsection{Method \code{patch_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$patch_namespaced_lease_candidate(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCoordinationV1alpha1NamespacedLeaseCandidate
delete a LeaseCandidate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-delete_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-delete_namespaced_lease_candidate}{}}}
\subsection{Method \code{delete_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$delete_namespaced_lease_candidate(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchCoordinationV1alpha1LeaseCandidateListForAllNamespaces
watch individual changes to a list of LeaseCandidate. deprecated: use the 'watch' parameter with a list operation instead.
watchCoordinationV1alpha1NamespacedLeaseCandidateList
watch individual changes to a list of LeaseCandidate. deprecated: use the 'watch' parameter with a list operation instead.
watchCoordinationV1alpha1NamespacedLeaseCandidate
watch changes to an object of kind LeaseCandidate. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-watch_lease_candidate_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-watch_lease_candidate_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_lease_candidate_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$watch_lease_candidate_list_for_all_namespaces()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-watch_namespaced_lease_candidate_list"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-watch_namespaced_lease_candidate_list}{}}}
\subsection{Method \code{watch_namespaced_lease_candidate_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$watch_namespaced_lease_candidate_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-watch_namespaced_lease_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-watch_namespaced_lease_candidate}{}}}
\subsection{Method \code{watch_namespaced_lease_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$watch_namespaced_lease_candidate(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1alpha1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1alpha1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1alpha1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
