% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_CoordinationV1Api.R
\name{CoordinationV1Api}
\alias{CoordinationV1Api}
\title{CoordinationV1Api}
\description{
CoordinationV1Api

CoordinationV1Api
}
\details{
Typed Kubernetes API client for the `CoordinationV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordinationV1Api-new}{\code{CoordinationV1Api$new()}}
\item \href{#method-CoordinationV1Api-get_api_resources}{\code{CoordinationV1Api$get_api_resources()}}
\item \href{#method-CoordinationV1Api-list_lease_for_all_namespaces}{\code{CoordinationV1Api$list_lease_for_all_namespaces()}}
\item \href{#method-CoordinationV1Api-list_namespaced_lease}{\code{CoordinationV1Api$list_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-create_namespaced_lease}{\code{CoordinationV1Api$create_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-delete_collection_namespaced_lease}{\code{CoordinationV1Api$delete_collection_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-read_namespaced_lease}{\code{CoordinationV1Api$read_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-replace_namespaced_lease}{\code{CoordinationV1Api$replace_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-patch_namespaced_lease}{\code{CoordinationV1Api$patch_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-delete_namespaced_lease}{\code{CoordinationV1Api$delete_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-watch_lease_list_for_all_namespaces}{\code{CoordinationV1Api$watch_lease_list_for_all_namespaces()}}
\item \href{#method-CoordinationV1Api-watch_namespaced_lease_list}{\code{CoordinationV1Api$watch_namespaced_lease_list()}}
\item \href{#method-CoordinationV1Api-watch_namespaced_lease}{\code{CoordinationV1Api$watch_namespaced_lease()}}
\item \href{#method-CoordinationV1Api-clone}{\code{CoordinationV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getCoordinationV1APIResources
get available resources
listCoordinationV1LeaseForAllNamespaces
list or watch objects of kind Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-list_lease_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-list_lease_for_all_namespaces}{}}}
\subsection{Method \code{list_lease_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$list_lease_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listCoordinationV1NamespacedLease
list or watch objects of kind Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-list_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-list_namespaced_lease}{}}}
\subsection{Method \code{list_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$list_namespaced_lease(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createCoordinationV1NamespacedLease
create a Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-create_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-create_namespaced_lease}{}}}
\subsection{Method \code{create_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$create_namespaced_lease(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteCoordinationV1CollectionNamespacedLease
delete collection of Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-delete_collection_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-delete_collection_namespaced_lease}{}}}
\subsection{Method \code{delete_collection_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$delete_collection_namespaced_lease(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readCoordinationV1NamespacedLease
read the specified Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-read_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-read_namespaced_lease}{}}}
\subsection{Method \code{read_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$read_namespaced_lease(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceCoordinationV1NamespacedLease
replace the specified Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-replace_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-replace_namespaced_lease}{}}}
\subsection{Method \code{replace_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$replace_namespaced_lease(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchCoordinationV1NamespacedLease
partially update the specified Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-patch_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-patch_namespaced_lease}{}}}
\subsection{Method \code{patch_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$patch_namespaced_lease(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteCoordinationV1NamespacedLease
delete a Lease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-delete_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-delete_namespaced_lease}{}}}
\subsection{Method \code{delete_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$delete_namespaced_lease(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
watchCoordinationV1LeaseListForAllNamespaces
watch individual changes to a list of Lease. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-watch_lease_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-watch_lease_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_lease_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$watch_lease_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchCoordinationV1NamespacedLeaseList
watch individual changes to a list of Lease. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-watch_namespaced_lease_list"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-watch_namespaced_lease_list}{}}}
\subsection{Method \code{watch_namespaced_lease_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$watch_namespaced_lease_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchCoordinationV1NamespacedLease
watch changes to an object of kind Lease. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-watch_namespaced_lease"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-watch_namespaced_lease}{}}}
\subsection{Method \code{watch_namespaced_lease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$watch_namespaced_lease(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinationV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinationV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinationV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
