% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_DiscoveryV1Api.R
\name{DiscoveryV1Api}
\alias{DiscoveryV1Api}
\title{DiscoveryV1Api}
\description{
DiscoveryV1Api

DiscoveryV1Api
}
\details{
Typed Kubernetes API client for the `DiscoveryV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiscoveryV1Api-new}{\code{DiscoveryV1Api$new()}}
\item \href{#method-DiscoveryV1Api-get_api_resources}{\code{DiscoveryV1Api$get_api_resources()}}
\item \href{#method-DiscoveryV1Api-list_endpoint_slice_for_all_namespaces}{\code{DiscoveryV1Api$list_endpoint_slice_for_all_namespaces()}}
\item \href{#method-DiscoveryV1Api-list_namespaced_endpoint_slice}{\code{DiscoveryV1Api$list_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-create_namespaced_endpoint_slice}{\code{DiscoveryV1Api$create_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-delete_collection_namespaced_endpoint_slice}{\code{DiscoveryV1Api$delete_collection_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-read_namespaced_endpoint_slice}{\code{DiscoveryV1Api$read_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-replace_namespaced_endpoint_slice}{\code{DiscoveryV1Api$replace_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-patch_namespaced_endpoint_slice}{\code{DiscoveryV1Api$patch_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-delete_namespaced_endpoint_slice}{\code{DiscoveryV1Api$delete_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-watch_endpoint_slice_list_for_all_namespaces}{\code{DiscoveryV1Api$watch_endpoint_slice_list_for_all_namespaces()}}
\item \href{#method-DiscoveryV1Api-watch_namespaced_endpoint_slice_list}{\code{DiscoveryV1Api$watch_namespaced_endpoint_slice_list()}}
\item \href{#method-DiscoveryV1Api-watch_namespaced_endpoint_slice}{\code{DiscoveryV1Api$watch_namespaced_endpoint_slice()}}
\item \href{#method-DiscoveryV1Api-clone}{\code{DiscoveryV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getDiscoveryV1APIResources
get available resources
listDiscoveryV1EndpointSliceForAllNamespaces
list or watch objects of kind EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-list_endpoint_slice_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-list_endpoint_slice_for_all_namespaces}{}}}
\subsection{Method \code{list_endpoint_slice_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$list_endpoint_slice_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
listDiscoveryV1NamespacedEndpointSlice
list or watch objects of kind EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-list_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-list_namespaced_endpoint_slice}{}}}
\subsection{Method \code{list_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$list_namespaced_endpoint_slice(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createDiscoveryV1NamespacedEndpointSlice
create an EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-create_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-create_namespaced_endpoint_slice}{}}}
\subsection{Method \code{create_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$create_namespaced_endpoint_slice(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteDiscoveryV1CollectionNamespacedEndpointSlice
delete collection of EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-delete_collection_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-delete_collection_namespaced_endpoint_slice}{}}}
\subsection{Method \code{delete_collection_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$delete_collection_namespaced_endpoint_slice(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readDiscoveryV1NamespacedEndpointSlice
read the specified EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-read_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-read_namespaced_endpoint_slice}{}}}
\subsection{Method \code{read_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$read_namespaced_endpoint_slice(name, namespace, pretty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replaceDiscoveryV1NamespacedEndpointSlice
replace the specified EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-replace_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-replace_namespaced_endpoint_slice}{}}}
\subsection{Method \code{replace_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$replace_namespaced_endpoint_slice(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchDiscoveryV1NamespacedEndpointSlice
partially update the specified EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-patch_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-patch_namespaced_endpoint_slice}{}}}
\subsection{Method \code{patch_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$patch_namespaced_endpoint_slice(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deleteDiscoveryV1NamespacedEndpointSlice
delete an EndpointSlice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-delete_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-delete_namespaced_endpoint_slice}{}}}
\subsection{Method \code{delete_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$delete_namespaced_endpoint_slice(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
watchDiscoveryV1EndpointSliceListForAllNamespaces
watch individual changes to a list of EndpointSlice. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-watch_endpoint_slice_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-watch_endpoint_slice_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_endpoint_slice_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$watch_endpoint_slice_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchDiscoveryV1NamespacedEndpointSliceList
watch individual changes to a list of EndpointSlice. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-watch_namespaced_endpoint_slice_list"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-watch_namespaced_endpoint_slice_list}{}}}
\subsection{Method \code{watch_namespaced_endpoint_slice_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$watch_namespaced_endpoint_slice_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchDiscoveryV1NamespacedEndpointSlice
watch changes to an object of kind EndpointSlice. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-watch_namespaced_endpoint_slice"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-watch_namespaced_endpoint_slice}{}}}
\subsection{Method \code{watch_namespaced_endpoint_slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$watch_namespaced_endpoint_slice(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscoveryV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiscoveryV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscoveryV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
