% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_PolicyV1Api.R
\name{PolicyV1Api}
\alias{PolicyV1Api}
\title{PolicyV1Api}
\description{
PolicyV1Api

PolicyV1Api
}
\details{
Typed Kubernetes API client for the `PolicyV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PolicyV1Api-new}{\code{PolicyV1Api$new()}}
\item \href{#method-PolicyV1Api-get_api_resources}{\code{PolicyV1Api$get_api_resources()}}
\item \href{#method-PolicyV1Api-list_namespaced_pod_disruption_budget}{\code{PolicyV1Api$list_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-create_namespaced_pod_disruption_budget}{\code{PolicyV1Api$create_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-delete_collection_namespaced_pod_disruption_budget}{\code{PolicyV1Api$delete_collection_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-read_namespaced_pod_disruption_budget}{\code{PolicyV1Api$read_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-replace_namespaced_pod_disruption_budget}{\code{PolicyV1Api$replace_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-patch_namespaced_pod_disruption_budget}{\code{PolicyV1Api$patch_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-delete_namespaced_pod_disruption_budget}{\code{PolicyV1Api$delete_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-read_namespaced_pod_disruption_budget_status}{\code{PolicyV1Api$read_namespaced_pod_disruption_budget_status()}}
\item \href{#method-PolicyV1Api-replace_namespaced_pod_disruption_budget_status}{\code{PolicyV1Api$replace_namespaced_pod_disruption_budget_status()}}
\item \href{#method-PolicyV1Api-patch_namespaced_pod_disruption_budget_status}{\code{PolicyV1Api$patch_namespaced_pod_disruption_budget_status()}}
\item \href{#method-PolicyV1Api-list_pod_disruption_budget_for_all_namespaces}{\code{PolicyV1Api$list_pod_disruption_budget_for_all_namespaces()}}
\item \href{#method-PolicyV1Api-watch_namespaced_pod_disruption_budget_list}{\code{PolicyV1Api$watch_namespaced_pod_disruption_budget_list()}}
\item \href{#method-PolicyV1Api-watch_namespaced_pod_disruption_budget}{\code{PolicyV1Api$watch_namespaced_pod_disruption_budget()}}
\item \href{#method-PolicyV1Api-watch_pod_disruption_budget_list_for_all_namespaces}{\code{PolicyV1Api$watch_pod_disruption_budget_list_for_all_namespaces()}}
\item \href{#method-PolicyV1Api-clone}{\code{PolicyV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getPolicyV1APIResources
get available resources
listPolicyV1NamespacedPodDisruptionBudget
list or watch objects of kind PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-list_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-list_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{list_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$list_namespaced_pod_disruption_budget(
  namespace,
  pretty = NULL,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
createPolicyV1NamespacedPodDisruptionBudget
create a PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-create_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-create_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{create_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$create_namespaced_pod_disruption_budget(
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
deletePolicyV1CollectionNamespacedPodDisruptionBudget
delete collection of PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-delete_collection_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-delete_collection_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{delete_collection_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$delete_collection_namespaced_pod_disruption_budget(
  namespace,
  pretty = NULL,
  continue = NULL,
  dry_run = NULL,
  field_selector = NULL,
  grace_period_seconds = NULL,
  label_selector = NULL,
  limit = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{body}}{= NULL
readPolicyV1NamespacedPodDisruptionBudget
read the specified PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-read_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-read_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{read_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$read_namespaced_pod_disruption_budget(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replacePolicyV1NamespacedPodDisruptionBudget
replace the specified PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-replace_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-replace_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{replace_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$replace_namespaced_pod_disruption_budget(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchPolicyV1NamespacedPodDisruptionBudget
partially update the specified PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-patch_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-patch_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{patch_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$patch_namespaced_pod_disruption_budget(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
deletePolicyV1NamespacedPodDisruptionBudget
delete a PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-delete_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-delete_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{delete_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$delete_namespaced_pod_disruption_budget(
  name,
  namespace,
  pretty = NULL,
  dry_run = NULL,
  grace_period_seconds = NULL,
  orphan_dependents = NULL,
  propagation_policy = NULL,
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{grace_period_seconds}}{= NULL}

\item{\code{orphan_dependents}}{= NULL}

\item{\code{propagation_policy}}{= NULL}

\item{\code{body}}{= NULL
readPolicyV1NamespacedPodDisruptionBudgetStatus
read status of the specified PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-read_namespaced_pod_disruption_budget_status"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-read_namespaced_pod_disruption_budget_status}{}}}
\subsection{Method \code{read_namespaced_pod_disruption_budget_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$read_namespaced_pod_disruption_budget_status(
  name,
  namespace,
  pretty = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL
replacePolicyV1NamespacedPodDisruptionBudgetStatus
replace status of the specified PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-replace_namespaced_pod_disruption_budget_status"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-replace_namespaced_pod_disruption_budget_status}{}}}
\subsection{Method \code{replace_namespaced_pod_disruption_budget_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$replace_namespaced_pod_disruption_budget_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL
patchPolicyV1NamespacedPodDisruptionBudgetStatus
partially update status of the specified PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-patch_namespaced_pod_disruption_budget_status"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-patch_namespaced_pod_disruption_budget_status}{}}}
\subsection{Method \code{patch_namespaced_pod_disruption_budget_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$patch_namespaced_pod_disruption_budget_status(
  name,
  namespace,
  body,
  pretty = NULL,
  dry_run = NULL,
  field_manager = NULL,
  field_validation = NULL,
  force = NULL,
  content_type = "application/json-patch+json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{= NULL}

\item{\code{dry_run}}{= NULL}

\item{\code{field_manager}}{= NULL}

\item{\code{field_validation}}{= NULL}

\item{\code{force}}{= NULL}

\item{\code{content_type}}{= "application/json-patch+json"
listPolicyV1PodDisruptionBudgetForAllNamespaces
list or watch objects of kind PodDisruptionBudget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-list_pod_disruption_budget_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-list_pod_disruption_budget_for_all_namespaces}{}}}
\subsection{Method \code{list_pod_disruption_budget_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$list_pod_disruption_budget_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchPolicyV1NamespacedPodDisruptionBudgetList
watch individual changes to a list of PodDisruptionBudget. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-watch_namespaced_pod_disruption_budget_list"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-watch_namespaced_pod_disruption_budget_list}{}}}
\subsection{Method \code{watch_namespaced_pod_disruption_budget_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$watch_namespaced_pod_disruption_budget_list(
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchPolicyV1NamespacedPodDisruptionBudget
watch changes to an object of kind PodDisruptionBudget. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-watch_namespaced_pod_disruption_budget"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-watch_namespaced_pod_disruption_budget}{}}}
\subsection{Method \code{watch_namespaced_pod_disruption_budget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$watch_namespaced_pod_disruption_budget(
  name,
  namespace,
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL
watchPolicyV1PodDisruptionBudgetListForAllNamespaces
watch individual changes to a list of PodDisruptionBudget. deprecated: use the 'watch' parameter with a list operation instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-watch_pod_disruption_budget_list_for_all_namespaces"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-watch_pod_disruption_budget_list_for_all_namespaces}{}}}
\subsection{Method \code{watch_pod_disruption_budget_list_for_all_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$watch_pod_disruption_budget_list_for_all_namespaces(
  allow_watch_bookmarks = NULL,
  continue = NULL,
  field_selector = NULL,
  label_selector = NULL,
  limit = NULL,
  pretty = NULL,
  resource_version = NULL,
  resource_version_match = NULL,
  send_initial_events = NULL,
  timeout_seconds = NULL,
  watch = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_watch_bookmarks}}{= NULL}

\item{\code{continue}}{= NULL}

\item{\code{field_selector}}{= NULL}

\item{\code{label_selector}}{= NULL}

\item{\code{limit}}{= NULL}

\item{\code{pretty}}{= NULL}

\item{\code{resource_version}}{= NULL}

\item{\code{resource_version_match}}{= NULL}

\item{\code{send_initial_events}}{= NULL}

\item{\code{timeout_seconds}}{= NULL}

\item{\code{watch}}{= NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolicyV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PolicyV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolicyV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
