% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_SchedulingV1Api.R
\name{SchedulingV1Api}
\alias{SchedulingV1Api}
\title{SchedulingV1Api}
\description{
SchedulingV1Api

SchedulingV1Api
}
\details{
Typed Kubernetes API client for the `SchedulingV1` surface. See generated methods for available operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{The backing [ApiClient].}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SchedulingV1Api-new}{\code{SchedulingV1Api$new()}}
\item \href{#method-SchedulingV1Api-get_api_resources}{\code{SchedulingV1Api$get_api_resources()}}
\item \href{#method-SchedulingV1Api-list_priority_class}{\code{SchedulingV1Api$list_priority_class()}}
\item \href{#method-SchedulingV1Api-create_priority_class}{\code{SchedulingV1Api$create_priority_class()}}
\item \href{#method-SchedulingV1Api-delete_collection_priority_class}{\code{SchedulingV1Api$delete_collection_priority_class()}}
\item \href{#method-SchedulingV1Api-read_priority_class}{\code{SchedulingV1Api$read_priority_class()}}
\item \href{#method-SchedulingV1Api-replace_priority_class}{\code{SchedulingV1Api$replace_priority_class()}}
\item \href{#method-SchedulingV1Api-patch_priority_class}{\code{SchedulingV1Api$patch_priority_class()}}
\item \href{#method-SchedulingV1Api-delete_priority_class}{\code{SchedulingV1Api$delete_priority_class()}}
\item \href{#method-SchedulingV1Api-watch_priority_class_list}{\code{SchedulingV1Api$watch_priority_class_list()}}
\item \href{#method-SchedulingV1Api-watch_priority_class}{\code{SchedulingV1Api$watch_priority_class()}}
\item \href{#method-SchedulingV1Api-clone}{\code{SchedulingV1Api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-new"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$new(api_client = ApiClient$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An [ApiClient].
getSchedulingV1APIResources
get available resources
listSchedulingV1PriorityClass
list or watch objects of kind PriorityClass
createSchedulingV1PriorityClass
create a PriorityClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-get_api_resources"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-get_api_resources}{}}}
\subsection{Method \code{get_api_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$get_api_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-list_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-list_priority_class}{}}}
\subsection{Method \code{list_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$list_priority_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-create_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-create_priority_class}{}}}
\subsection{Method \code{create_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$create_priority_class(
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteSchedulingV1CollectionPriorityClass
delete collection of PriorityClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-delete_collection_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-delete_collection_priority_class}{}}}
\subsection{Method \code{delete_collection_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$delete_collection_priority_class(dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
readSchedulingV1PriorityClass
read the specified PriorityClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-read_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-read_priority_class}{}}}
\subsection{Method \code{read_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$read_priority_class(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{replaceSchedulingV1PriorityClass
replace the specified PriorityClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-replace_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-replace_priority_class}{}}}
\subsection{Method \code{replace_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$replace_priority_class(
  name,
  body,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
patchSchedulingV1PriorityClass
partially update the specified PriorityClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-patch_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-patch_priority_class}{}}}
\subsection{Method \code{patch_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$patch_priority_class(
  name,
  dry_run = NULL,
  field_validation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL}

\item{\code{field_validation}}{= NULL
deleteSchedulingV1PriorityClass
delete a PriorityClass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-delete_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-delete_priority_class}{}}}
\subsection{Method \code{delete_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$delete_priority_class(name, dry_run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{= NULL
watchSchedulingV1PriorityClassList
watch individual changes to a list of PriorityClass. deprecated: use the 'watch' parameter with a list operation instead.
watchSchedulingV1PriorityClass
watch changes to an object of kind PriorityClass. deprecated: use the 'watch' parameter with a list operation instead, filtered to a single item with the 'fieldSelector' parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-watch_priority_class_list"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-watch_priority_class_list}{}}}
\subsection{Method \code{watch_priority_class_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$watch_priority_class_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-watch_priority_class"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-watch_priority_class}{}}}
\subsection{Method \code{watch_priority_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$watch_priority_class(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SchedulingV1Api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SchedulingV1Api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchedulingV1Api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
